% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campBrainSubsetSCE}
\alias{campBrainSubsetSCE}
\title{Example Single Cell RNA-Seq data in SCtkExperiment Object, Campbell, et al.
subset}
\format{
SCtkExperiment
}
\source{
DOI: 10.1038/nn.4495
}
\usage{
campBrainSubsetSCE
}
\description{
A subset of 500 cells from a single cell RNA-Seq experiment from Campbell, et
al. Nat. Neuroscience 2017. This contains information such as the diet of the mice, 
sex and proposed cell type for each cell.
}
\examples{
data("campBrainSubsetSCE")
}
\keyword{datasets}
