% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger_dir_to_sctke.R
\name{cellranger_dir_to_sctke}
\alias{cellranger_dir_to_sctke}
\title{Convert Cell Ranger outs Directory to a SingleCelltkExperiment}
\usage{
cellranger_dir_to_sctke(path = ".")
}
\arguments{
\item{path}{The path to the Cell Ranger outs directory. Defaults to the cwd.}
}
\value{
a SingleCelltkExperiment object to use in the Single Cell Toolkit.
raw data, PCA, tSNE, and clustering results are extracted and stored in the
object.
path <- "~/Dropbox/grad_school/johnson_lab/20170919_CellrangertoSCTK/JC-PBMC-HIGH/outs/"
}
\description{
Convert Cell Ranger outs Directory to a SingleCelltkExperiment
}
