% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASTPlotFunctions.R
\name{checkMASTResult}
\alias{checkMASTResult}
\title{Check if the specified MAST result in SCtkExperiment object is complete.
But does not garantee the biological correctness.}
\usage{
checkMASTResult(inSCE, useResult)
}
\arguments{
\item{inSCE}{SCtkExperiment object. `runMAST()` has to be run in advance.}

\item{useResult}{character. A string specifying the `comparisonName` used.}
}
\description{
Check if the specified MAST result in SCtkExperiment object is complete.
But does not garantee the biological correctness.
}
