% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComBat_SCE.R
\name{ComBat_SCE}
\alias{ComBat_SCE}
\title{ComBat_SCE}
\usage{
ComBat_SCE(SCEdata, batch, use_assay = "logcounts",
  par.prior = "Parametric", covariates = NULL, mean.only = FALSE,
  ref.batch = NULL)
}
\arguments{
\item{SCEdata}{SCtkExperiment object. Required}

\item{batch}{The name of a column in colData to use as the batch variable.
Required}

\item{use_assay}{The assay to use for ComBat. The default is "logcounts"}

\item{par.prior}{TRUE indicates parametric adjustments will be used, FALSE
indicates non-parametric adjustments will be used. Accepted parameters:
"Parametric" or "Non-parametric"}

\item{covariates}{List of other column names in colData to be added to the
ComBat model as covariates}

\item{mean.only}{If TRUE ComBat only corrects the mean of the batch effect}

\item{ref.batch}{If given, will use the selected batch as a reference for
batch adjustment.}
}
\value{
ComBat matrix based on inputs. You can save this matrix into the
SCtkExperiment with assay()
}
\description{
Run ComBat on a SCtkExperiment object
}
