% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGeneIDs.R
\name{convertGeneIDs}
\alias{convertGeneIDs}
\title{Convert Gene IDs}
\usage{
convertGeneIDs(inSCE, inSymbol, outSymbol, database = "org.Hs.eg.db")
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{inSymbol}{The input symbol type}

\item{outSymbol}{The output symbol type}

\item{database}{The org.*.eg.db database to use. The default is org.Hs.eg.db}
}
\value{
A \linkS4class{SingleCellExperiment} object with converted gene IDs.
}
\description{
Convert the gene IDs in a SingleCellExperiment object using Bioconductor
org.*.eg.db data packages. Because annotation databases do not have a 1:1
relationship, this tool removes rows with no corresponding annotation in
your desired annotation, and remove any duplicate annotations after
conversion.
}
\examples{
if(requireNamespace("org.Mm.eg.db", quietly = TRUE)) {
  #convert mouse gene symbols to ensembl IDs
  library("org.Mm.eg.db")
  sample(rownames(mouseBrainSubsetSCE), 50)
  mouseBrainSubsetSymbol <- convertGeneIDs(inSCE = mouseBrainSubsetSCE,
                                           inSymbol = "SYMBOL",
                                           outSymbol = "ENSEMBL",
                                           database = "org.Mm.eg.db")
  sample(rownames(mouseBrainSubsetSymbol), 50)
}

}
