% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{createSCESet}
\alias{createSCESet}
\title{Create a SCESet object}
\usage{
createSCESet(countfile, annotfile)
}
\arguments{
\item{countfile}{The path to a text file that contains a header row of sample
names, and rows of raw counts per gene for those samples}

\item{annotfile}{The path to a text file that contains columns of annotation
information for each sample in the countfile. This file should have the same
number of rows as there are columns in the countfile.}
}
\value{
a SCESet object
}
\description{
From a file of counts and a file of annotation information, create a SCESet
object.
}

