% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSCEHeatmap.R
\name{dataAnnotationColor}
\alias{dataAnnotationColor}
\title{Generate distinct colors for all categorical col/rowData entries.}
\usage{
dataAnnotationColor(inSCE, axis = NULL, colorGen = grDevices::rainbow)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object.}

\item{axis}{Choose from \code{"col"} or \code{"row"}.}

\item{colorGen}{A function that generates color code vector by giving an 
integer for the number of colors. Alternatively, 
\code{\link{celda::distinctColors}}. Default \code{\link{rainbow}}.}
}
\value{
A \code{\link{list}} object containing distinct colors mapped to all 
possible categorical entries in \code{rowData} or \code{colData}.
}
\description{
Generate distinct colors for all categorical col/rowData entries.
}
\author{
Yichen Wang
}
