% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{.addSeuratToMetaDataSCE}
\alias{.addSeuratToMetaDataSCE}
\title{.addSeuratToMetaDataSCE
Adds the input seurat object to the metadata slot of the input sce object (after removing the data matrices)}
\usage{
.addSeuratToMetaDataSCE(inSCE, seuratObject)
}
\arguments{
\item{inSCE}{(sce) object to which seurat object should be added in the metadata slot (copy to)}

\item{seuratObject}{seurat object which should be added to the metadata slot of sce object (copy from)}
}
\value{
Updated \code{SingleCellExperiment} object which now contains the seurat object in its metadata slot (excluding data matrices)
}
\description{
.addSeuratToMetaDataSCE
Adds the input seurat object to the metadata slot of the input sce object (after removing the data matrices)
}
