% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQCUtils.R
\name{.checkCell}
\alias{.checkCell}
\title{Perform comprehensive single cell QC}
\usage{
.checkCell(FilterFile, FilterDir, basepath, Reference, process)
}
\arguments{
\item{FilterFile}{The cell QC object as rendered in R}

\item{FilterDir}{The cell QC file location}

\item{basepath}{The base/root directory of the QC files.}

\item{Reference}{Reference library for MitoGenes}

\item{process}{The method that we are looking at, be it CellRanger, Seurat, SCE, AnnData, or otherwise.}
}
\value{
0 if the check is completed successfully, else a premature stop.
}
\description{
Utility functions for runQC - checking cell input for validity
}
