% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEAnalysis.R
\name{.checkDiffExpResultExists}
\alias{.checkDiffExpResultExists}
\title{Check if the specified MAST result in SingleCellExperiment object is
complete. But does not garantee the biological correctness.}
\usage{
.checkDiffExpResultExists(inSCE, useResult, labelBy = NULL)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object. a
differential expression analysis function has to be run in advance.}

\item{useResult}{character. A string specifying the \code{analysisName}
used when running a differential expression analysis function.}

\item{labelBy}{A single character for a column of \code{rowData(inSCE)} as
where to search for the labeling text. Default \code{NULL}.}
}
\value{
Stop point if found
}
\description{
Check if the specified MAST result in SingleCellExperiment object is
complete. But does not garantee the biological correctness.
}
