% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQCUtils.R
\name{.checkDroplet}
\alias{.checkDroplet}
\title{Perform comprehensive single cell QC}
\usage{
.checkDroplet(RawFile, RawDir, basepath, Reference, process)
}
\arguments{
\item{RawFile}{The droplet QC object as rendered in R}

\item{RawDir}{The droplet QC file location}

\item{basepath}{The base/root directory of the QC files.}

\item{Reference}{Reference library for MitoGenes}

\item{process}{The method that we are looking at, be it CellRanger, Seurat, SCE, AnnData, or otherwise.}
}
\value{
0 if the check is completed successfully, else a premature stop.
}
\description{
Utility functions for runQC - checking droplet input for validity
}
