% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASTPlotFunctions.R
\name{.checkMASTResultExists}
\alias{.checkMASTResultExists}
\title{Check if the specified MAST result in SCtkExperiment object is complete.
But does not garantee the biological correctness.}
\usage{
.checkMASTResultExists(inSCE, useResult)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object.
\code{runMAST()} has to be run in advance.}

\item{useResult}{character. A string specifying the \code{comparisonName}
used when running \code{runMAST()}.}
}
\description{
Check if the specified MAST result in SCtkExperiment object is complete.
But does not garantee the biological correctness.
}
