% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctkQCUtils.R
\name{.convert_hex_to_int}
\alias{.convert_hex_to_int}
\title{Generate HTAN manifest file for droplet and cell count data}
\usage{
.convert_hex_to_int(hex)
}
\arguments{
\item{dropletSCE}{A \link[SingleCellExperiment]{SingleCellExperiment} object containing
droplet count matrix data}

\item{cellSCE}{A \link[SingleCellExperiment]{SingleCellExperiment} object containing
cell count matrix data}

\item{samplename}{The sample name of the \link[SingleCellExperiment]{SingleCellExperiment} objects}

\item{htan_patient_id}{The HTAN patient id of the sample in \link[SingleCellExperiment]{SingleCellExperiment} object}

\item{dir}{The output directory of the SCTK QC pipeline.}

\item{dataType}{Type of the input data. It can be one of "Droplet", "Cell" or "Both".}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object which combines all
objects in sceList. The colData is merged.
}
\description{
Generate HTAN manifest file for droplet and cell count data
}
\examples{
data(scExample, package = "singleCellTK")
generateHTANMeta(sce, dir = ".", samplename = "Sample")
}
