% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopHVG.R
\name{.dfFromHVGMetric}
\alias{.dfFromHVGMetric}
\title{Check if specified method has already been performed, and extract the metric.
Will be used by `getTopHVG` to rank the top, and `plotTopHVG` for the axis}
\usage{
.dfFromHVGMetric(
  inSCE,
  method = c("vst", "mean.var.plot", "dispersion", "modelGeneVar")
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object}

\item{method}{Specify which method to use for variable gene extraction
from Seurat \code{"vst"}, \code{"mean.var.plot"}, \code{"dispersion"} or 
Scran \code{"modelGeneVar"}. Default \code{"vst"}}
}
\value{
data.frame object of HVG metrics calculated by \code{method}, 
containing columns of \code{"mean"}, \code{"v_rank"}, \code{"v_plot"}
}
\description{
Check if specified method has already been performed, and extract the metric.
Will be used by `getTopHVG` to rank the top, and `plotTopHVG` for the axis
}
