% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSCEHeatmap.R
\name{.extractSCEAnnotation}
\alias{.extractSCEAnnotation}
\title{Extract columns from row/colData and transfer to factors}
\usage{
.extractSCEAnnotation(inSCE, axis = NULL, columns = NULL, index = NULL)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object.}

\item{axis}{Choose from \code{"col"} or \code{"row"}.}

\item{columns}{character vector. The columns needed to be extracted. If
\code{NULL}, will return an empty \code{data.frame} with matched row
names. Default \code{NULL}.}

\item{index}{Valid index to subset the col/row.}
}
\value{
A \code{data.frame} object.
}
\description{
Extract columns from row/colData and transfer to factors
}
