% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEAnalysis.R
\name{.formatDEAList}
\alias{.formatDEAList}
\title{Helper function for differential expression analysis methods that accepts
multiple ways of conditional subsetting and returns stable index format.
Meanwhile it does all the input checkings.}
\usage{
.formatDEAList(
  inSCE,
  useAssay,
  index1 = NULL,
  index2 = NULL,
  class = NULL,
  classGroup1 = NULL,
  classGroup2 = NULL,
  groupName1,
  groupName2,
  analysisName,
  covariates = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object. Required.}

\item{useAssay}{character. A string specifying which assay to use. Required.}

\item{index1}{Any type of indices that can subset a
\linkS4class{SingleCellExperiment} inherited object by cells. Specifies
which cells are of interests. Default \code{NULL}.}

\item{index2}{Any type of indices that can subset a
\linkS4class{SingleCellExperiment} inherited object by cells. specifies
the control group against those specified by \code{index1}. If
\code{NULL} when using index specification, \code{index1} cells will be
compared with all other cells. Default \code{NULL}.}

\item{class}{A vector/factor with \code{ncol(inSCE)} elements, or a character
scalar that specifies a column name of \code{colData(inSCE)}. Default
\code{NULL}.}

\item{classGroup1}{a vector specifying which "levels" given in \code{class}
are of interests. Default \code{NULL}.}

\item{classGroup2}{a vector specifying which "levels" given in \code{class}
is the control group against those specified by \code{classGroup1}. If
\code{NULL} when using annotation specification, \code{classGroup1} cells
will be compared with all other cells.}

\item{groupName1}{A character scalar naming the group of interests. Required.}

\item{groupName2}{A character scalar naming the control group. Required.}

\item{analysisName}{A character scalar naming the DEG analysis. Required}

\item{covariates}{A character vector of additional covariates used in linear
regression methods such as Limma and DESeq2. Default \code{NULL}}

\item{overwrite}{A logical scalar. Whether to overwrite result if exists.
Default \code{FALSE}.}
}
\value{
A list object with part of formatted DE analysis information
}
\description{
Helper function for differential expression analysis methods that accepts
multiple ways of conditional subsetting and returns stable index format.
Meanwhile it does all the input checkings.
}
\author{
Yichen Wang
}
