% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSCAN.R
\name{.getClustersLineData}
\alias{.getClustersLineData}
\title{Extract edges that connects to the clusters of interest}
\usage{
.getClustersLineData(line.data, useCluster)
}
\arguments{
\item{line.data}{data.frame of three columns. First column should be named
"edge", the other two are coordinates of one of the vertices that this
edge connects.}

\item{useCluster}{Vector of clusters, that are going to be the vertices for
edge finding.}
}
\value{
data.frame, subset rows of line.data
}
\description{
Extract edges that connects to the clusters of interest
}
