% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{.getComponentNames}
\alias{.getComponentNames}
\title{.getComponentNames
Creates a list of PC/IC components to populate the picker for PC/IC heatmap generation}
\usage{
.getComponentNames(maxComponents, component = c("PC", "IC"))
}
\arguments{
\item{maxComponents}{Number of components to return for the picker}

\item{component}{Which component to use. Choices are \code{PC} or \code{IC}.}
}
\value{
List of component names (appended with \code{PC} or \code{IC})
}
\description{
.getComponentNames
Creates a list of PC/IC components to populate the picker for PC/IC heatmap generation
}
