% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{.getPCAComponentNames}
\alias{.getPCAComponentNames}
\title{.getPCAComponentNames
Creates a list of PC components to populate the picker for PC heatmap generation}
\usage{
.getPCAComponentNames(maxComponents)
}
\arguments{
\item{maxComponents;}{number of components to return for the picker}
}
\value{
componentNames; list of component names (appended with "PC")
}
\description{
.getPCAComponentNames
Creates a list of PC components to populate the picker for PC heatmap generation
}
