% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotting.R
\name{.ggDensity}
\alias{.ggDensity}
\title{Density plot plotting tool.}
\usage{
.ggDensity(
  value,
  groupBy = NULL,
  xlab = NULL,
  ylab = NULL,
  axisSize = 10,
  axisLabelSize = 10,
  defaultTheme = TRUE,
  title = NULL,
  titleSize = 18,
  cutoff = NULL
)
}
\arguments{
\item{value}{Numeric value that will be plotted via density plot.}

\item{groupBy}{Groupings for each numeric value. A user may input a vector
equal length to the number of the samples in the SingleCellExperiment
object, or can be retrieved from the colData slot. Default NULL.}

\item{xlab}{Character vector. Label for x-axis. Default NULL.}

\item{ylab}{Character vector. Label for y-axis. Default NULL.}

\item{axisSize}{Size of x/y-axis ticks. Default 10.}

\item{axisLabelSize}{Size of x/y-axis labels. Default 10.}

\item{defaultTheme}{Removes grid in plot and sets axis title size to 10
when TRUE. Default TRUE.}

\item{title}{Title of plot. Default NULL.}

\item{titleSize}{Size of title of plot. Default 15.}

\item{cutoff}{Numeric value. The plot will be annotated with a vertical line
if set. Default NULL.}
}
\value{
density plot, in .ggplot.
}
\description{
Visualizes values stored in the specified slot of a
 SingleCellExperiment object via a density plot.
}
