% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotting.R
\name{.ggScatter}
\alias{.ggScatter}
\title{Plot results of reduced dimensions data.}
\usage{
.ggScatter(
  inSCE,
  colorBy,
  shape,
  reducedDimName,
  conditionClass = NULL,
  useAssay,
  labelClusters = FALSE,
  xlab,
  ylab,
  dim1 = NULL,
  dim2 = NULL,
  dotsize = 2,
  transparency = 1,
  defaultTheme = TRUE,
  title = NULL,
  titleSize = 15,
  legendTitle = NULL
)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object with saved dimension reduction
components or a variable with saved results. Required}

\item{colorBy}{color by a condition(any column of the annotation data).}

\item{shape}{add shapes to each condition.}

\item{reducedDimName}{saved dimension reduction name in the SCtkExperiment
object. Required.}

\item{conditionClass}{class of the annotation data used in colorBy. Options
are NULL, "factor" or "numeric". If NULL, class will default to the original
class. Default NULL.}

\item{useAssay}{Indicate which assay to use. The default is "logcounts"}

\item{labelClusters}{Logical. Whether the cluster labels are plotted.
Default FALSE.}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{dim1}{1st dimension to be used for plotting. Default is NULL.}

\item{dim2}{2nd dimension to be used for plotting. Default is NULL.}

\item{dotsize}{size of dots. Default 2.}

\item{transparency}{transparency of the dots, values will be 0-1. Default 1.}

\item{defaultTheme}{removes grid in plot and sets axis title size to 10
when TRUE. Default TRUE.}

\item{title}{title of plot. Default NULL.}

\item{titleSize}{size of title of plot. Default 15.}

\item{legendTitle}{title of legend. Default NULL.}
}
\value{
a ggplot of the reduced dimensions.
}
\description{
Plot results of reduced dimensions data and colors the plots by
 the input vector.
}
