% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopHVG.R
\name{.parseUseFeatureSubset}
\alias{.parseUseFeatureSubset}
\title{parse `useFeatureSubset` in other functions such as `scaterPCA`, `getUMAP`..
Do checks, and return logical vector. Or character vector as needed by Seurat
methods}
\usage{
.parseUseFeatureSubset(
  inSCE,
  useFeatureSubset,
  altExpObj = NULL,
  returnType = c("logical", "character")
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object}

\item{useFeatureSubset}{Subset of feature to use. A character string 
indicating a \code{rowData} variable that stores the logical vector of HVG 
selection, or a vector that can subset the rows of \code{inSCE}.}

\item{altExpObj}{A \linkS4class{SingleCellExperiment} object, extracted from
\code{altExps} of \code{inSCE}, can be \code{identical()} to \code{inSCE}.
Used when functions like \code{\link{scaterPCA}} allows \code{useAltExp}, so
that the output vector match to the rownames of \code{altExpObj}}

\item{returnType}{What type of vector should be returned}
}
\value{
\code{NULL} if \code{is.null(useFeatureSubset)}. Otherwise, a vector
subsetting the features.
}
\description{
parse `useFeatureSubset` in other functions such as `scaterPCA`, `getUMAP`..
Do checks, and return logical vector. Or character vector as needed by Seurat
methods
}
