% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{.rdsToSce}
\alias{.rdsToSce}
\title{.rdsToSce
Reads rds file (from a local path) and loads into sce object 
*Only to be used for first time initialization of the rds file into R environment*}
\usage{
.rdsToSce(filePath)
}
\arguments{
\item{filePath;}{path of the rds file to load}
}
\value{
sce object
}
\description{
.rdsToSce
Reads rds file (from a local path) and loads into sce object 
*Only to be used for first time initialization of the rds file into R environment*
}
