% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCluster.R
\name{.relabelClusterBySize}
\alias{.relabelClusterBySize}
\title{Reassign cluster label by group size}
\usage{
.relabelClusterBySize(x)
}
\arguments{
\item{x}{categorical vector, factor}
}
\value{
factor
}
\description{
Reassign cluster label by group size
}
