% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{.rowNamesSCE}
\alias{.rowNamesSCE}
\title{.rowNamesSCE
Retrieves a list of genenames/rownames/featurenames from sce object}
\usage{
.rowNamesSCE(inSCE)
}
\arguments{
\item{inSCE}{sce object from which the genenames/rownames/featurenames should be extracted}
}
\value{
list() of genenames/rownames/featurenames
}
\description{
.rowNamesSCE
Retrieves a list of genenames/rownames/featurenames from sce object
}
