% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{.rowNamesSeurat}
\alias{.rowNamesSeurat}
\title{.rowNamesSeurat
Retrieves a list of genenames/rownames/featurenames from seurat object}
\usage{
.rowNamesSeurat(seuratObject)
}
\arguments{
\item{seuratObject;}{seurat object from which the genenames/rownames/featurenames should be extracted}
}
\value{
list() of genenames/rownames/featurenames
}
\description{
.rowNamesSeurat
Retrieves a list of genenames/rownames/featurenames from seurat object
}
