% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQCUtils.R
\name{.runCell}
\alias{.runCell}
\title{Perform comprehensive single cell QC}
\usage{
.runCell(cellSCE, preproc, geneSetCollection, Params, cellQCAlgos, mitoInfo)
}
\arguments{
\item{cellSCE}{The cell SCE object as rendered in R}

\item{preproc}{The method that we are looking at, be it CellRanger, Seurat, SCE, AnnData, or otherwise.}

\item{geneSetCollection}{Reference library of genes}

\item{Params}{The list of parameters we are using.}

\item{cellQCAlgos}{The list of various QC algorithms to use.}

\item{mitoInfo}{Reference library for MitoGenes}
}
\value{
No objects are returned, just runs the cell QC and throws an error if problems occur in subfunctions
}
\description{
QC execution function wrapper for cell SCE objects
}
