% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQCUtils.R
\name{.runDroplet}
\alias{.runDroplet}
\title{Perform comprehensive single cell QC}
\usage{
.runDroplet(
  dropletSCE,
  geneSetCollection,
  Params,
  mitoInfo,
  cellQCAlgos,
  detectCell,
  cellCalling
)
}
\arguments{
\item{dropletSCE}{The cell SCE object as rendered in R}

\item{geneSetCollection}{Reference library of genes}

\item{Params}{The list of parameters we are using.}

\item{mitoInfo}{Reference library for MitoGenes}

\item{cellQCAlgos}{The list of various QC algorithms to use.}

\item{detectCell}{Whether or not to detect the cell SCE object, or not}

\item{cellCalling}{The cell algorithm to be used.}
}
\value{
No objects are returned, just runs the cell QC and throws an error if problems occur in subfunctions
}
\description{
QC execution function wrapper for droplet SCE objects
}
