% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{.sceToSeurat}
\alias{.sceToSeurat}
\title{.sceToSeurat
Converts a sce object to seurat object (using rds filepath)
*Only to be used for first time initialization of seurat object*}
\usage{
.sceToSeurat(filePath)
}
\arguments{
\item{filePath}{path of the rds file to convert to seurat object}
}
\value{
seurat object
}
\description{
.sceToSeurat
Converts a sce object to seurat object (using rds filepath)
*Only to be used for first time initialization of seurat object*
}
