% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{.seuratGetVariableFeatures}
\alias{.seuratGetVariableFeatures}
\title{.seuratGetVariableFeatures
Retrieves the requested number of variable feature names}
\usage{
.seuratGetVariableFeatures(sceObject, geneNamesSeurat, numberOfFeatures)
}
\arguments{
\item{sceObject;}{sce object from which to extract the variable feature names}

\item{geneNamesSeurat;}{a list of rowames/genenames/featurenames for seurat object for consistent formatting}

\item{numberOfFeatures;}{numerical value indicating how many feature names should be retrieved (default is 100)}
}
\value{
list() of variable feature names
}
\description{
.seuratGetVariableFeatures
Retrieves the requested number of variable feature names
}
