% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{.seuratInvalidate}
\alias{.seuratInvalidate}
\title{.seuratInvalidate
Removes seurat data from the input SingleCellExperiment object specified by the task in the Seurat workflow.}
\usage{
.seuratInvalidate(
  inSCE,
  scaleData = TRUE,
  varFeatures = TRUE,
  PCA = TRUE,
  ICA = TRUE,
  tSNE = TRUE,
  UMAP = TRUE,
  clusters = TRUE
)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object to remove Seurat data from.}

\item{scaleData}{Remove scaled data from seurat. Default \code{TRUE}.}

\item{varFeatures}{Remove variable features from seurat. Default \code{TRUE}.}

\item{PCA}{Remove PCA from seurat. Default \code{TRUE}.}

\item{ICA}{Remove ICA from seurat. Default \code{TRUE}.}

\item{tSNE}{Remove tSNE from seurat. Default \code{TRUE}.}

\item{UMAP}{Remove UMAP from seurat. Default \code{TRUE}.}

\item{clusters}{Remove clusters from seurat. Default \code{TRUE}.}
}
\value{
Updated SingleCellExperiment object containing the Seurat object in the metadata slot with the data removed
}
\description{
.seuratInvalidate
Removes seurat data from the input SingleCellExperiment object specified by the task in the Seurat workflow.
}
