% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSCAN.R
\name{.viridisPseudoTimeColor}
\alias{.viridisPseudoTimeColor}
\title{Basing on pseudotime is presented within range from 0 to 100
Use the viridis color scale to match with `plotTSCANResult` embedding color}
\usage{
.viridisPseudoTimeColor(x)
}
\arguments{
\item{x}{numeric vector of pseudotime}
}
\description{
Basing on pseudotime is presented within range from 0 to 100
Use the viridis color scale to match with `plotTSCANResult` embedding color
}
