% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichRSCE.R
\name{enrichRSCE}
\alias{enrichRSCE}
\title{enrichR
Given a list of genes this function runs the enrichR() to perform Gene
enrichment}
\usage{
enrichRSCE(inSCE, useAssay, glist, db = NULL)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object. Required}

\item{useAssay}{The assay to use in the enrichment analysis. Required}

\item{glist}{selected genes for enrichment analysis using enrichR(). Required}

\item{db}{selected database name from the enrichR database list.}
}
\value{
enrichRSCE(): returns a data.frame of enrichment terms overlapping in
the respective databases along with p-values, z-scores etc.,
}
\description{
enrichR
Given a list of genes this function runs the enrichR() to perform Gene
enrichment
}
\examples{
enrichRSCE(mouseBrainSubsetSCE, "counts", "Cmtm5",
           "GO_Cellular_Component_2017")
}
