% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctkTagging.R
\name{expData}
\alias{expData}
\title{expData
Get data item from an input \code{SingleCellExperiment} object. The data item can be an \code{assay}, \code{altExp} (subset) or a \code{reducedDim}, which is retrieved based on the name of the data item.}
\usage{
expData(inSCE, assayName)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{assayName}{Specify the name of the data item to retrieve.}
}
\value{
Specified data item.
}
\description{
expData
Get data item from an input \code{SingleCellExperiment} object. The data item can be an \code{assay}, \code{altExp} (subset) or a \code{reducedDim}, which is retrieved based on the name of the data item.
}
\examples{
data(scExample, package = "singleCellTK")
mat <- expData(sce, "counts")
}
