% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctkTagging.R
\name{expDataNames,ANY-method}
\alias{expDataNames,ANY-method}
\title{expDataNames
Get names of all the data items in the input \code{SingleCellExperiment} object including assays, altExps and reducedDims.}
\usage{
\S4method{expDataNames}{ANY}(inSCE)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}
}
\value{
A combined \code{vector} of \code{assayNames}, \code{altExpNames} and \code{reducedDimNames}.
}
\description{
expDataNames
Get names of all the data items in the input \code{SingleCellExperiment} object including assays, altExps and reducedDims.
}
\examples{
data(scExample, package = "singleCellTK")
expDataNames(sce)
}
