% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSCEtoTXT.R
\name{exportSCEtoTXT}
\alias{exportSCEtoTXT}
\title{Export a \link[SingleCellExperiment]{SingleCellExperiment} object to flat text files}
\usage{
exportSCEtoTXT(
  sce,
  outputDir = "./",
  outputType = "txt",
  overwrite = FALSE,
  gzipped = TRUE
)
}
\arguments{
\item{sce}{\link[SingleCellExperiment]{SingleCellExperiment} object to be
exported.}

\item{outputDir}{Name of the directory to store the exported file(s).}

\item{outputType}{The desired type of the output files. Default \code{"txt"}
which
writes the \link[SingleCellExperiment]{SingleCellExperiment} object as
tab delimited text files.}

\item{overwrite}{Boolean. Whether to overwrite the output files. Default
\code{FALSE}.}

\item{gzipped}{Boolean. \code{TRUE} if the output files are to be
gzip compressed. \code{FALSE} otherwise. Default
\code{TRUE} to save disk space.}
}
\description{
Writes all assays, colData, rowData, reducedDims, and altExps objects in a
\link[SingleCellExperiment]{SingleCellExperiment} to text files.
The items in the 'metadata' slot remain stored in list and are saved in an RDS file.
}
\examples{
data(sce_chcl, package = "scds")
exportSCEtoTXT(sce_chcl, "sce_chcl")

}
