% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSCEHeatmap.R
\name{extractData}
\alias{extractData}
\title{Extract columns from row/colData and transfer to factors}
\usage{
extractData(inSCE, axis = NULL, columns = NULL, index = NULL)
}
\arguments{
\item{inSCE}{SingleCellExperiment object.}

\item{axis}{Choose from "col" or "row"}

\item{columns}{character vector, default NULL (return NULL). The columns
needed to be extracted.}

\item{index}{Valid index to subset the col/row.}
}
\value{
A data.frame object
}
\description{
Extract columns from row/colData and transfer to factors
}
