% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{featureNameDedup}
\alias{featureNameDedup}
\title{Resolve duplicated feature names in a matrix}
\usage{
featureNameDedup(countmat)
}
\arguments{
\item{countmat}{matrix, with row names as feature names that need to be
resolved.}
}
\value{
The same matrix as input with rowname duplication resolved.
}
\description{
Adds '-1', '-2', ... '-i' to multiple duplicated feature names
}
