% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{filterSCData}
\alias{filterSCData}
\title{Filter Genes and Samples from a Single Cell Object}
\usage{
filterSCData(insceset, deletesamples = NULL, remove_noexpress = TRUE,
  remove_bottom = 0.5, minimum_detect_genes = 1700)
}
\arguments{
\item{insceset}{Input single cell object, required}

\item{deletesamples}{List of samples to delete from the object.}

\item{remove_noexpress}{Remove genes that have no expression across all
samples. The default is true}

\item{remove_bottom}{Fraction of low expression genes to remove from the
single cell object. This occurs after remove_noexpress. The default is 0.50.}

\item{minimum_detect_genes}{Minimum number of genes with at least 1
count to include a sample in the single cell object. The default is 1700.}
}
\value{
The filtered single cell object.
}
\description{
Filter Genes and Samples from a Single Cell Object
}
\examples{
data("GSE60361_subset_sce")
GSE60361_subset_sce <- filterSCData(GSE60361_subset_sce,
                                    deletesamples="X1772063061_G11")
}
