% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBiomarker.R
\name{findBiomarker}
\alias{findBiomarker}
\title{Identify biomarkers}
\usage{
findBiomarker(inSCE, condition, useAssay, gene_subset = seq(100),
  nfolds = 3, nrepeats = 3, seed = 99, percent_top_biomarker = 0.2,
  model_name = c("logistic regression", "random forest"),
  level_selected_1va = NULL, level_selected_1v1 = NULL)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object. Required}

\item{condition}{Which condition is the target condition}

\item{useAssay}{useAssay Indicate which assay to use}

\item{gene_subset}{gene names that will be used as biomarker pool, default is first 100 genes}

\item{nfolds}{number of splits in CV}

\item{nrepeats}{number of CVs with different random splits}

\item{seed}{for repeatable research}

\item{percent_top_biomarker}{Top importance percentage to pick biomarker}

\item{model_name}{one of "logistic regression", "random forest"}
}
\value{
A R list containing biomarker info, importantce score plot, and ROC plot
}
\description{
Identify biomarkers
}
\examples{
data("mouseBrainSubsetSCE")
p <- findBiomarker(mouseBrainSubsetSCE,
                    condition = "level1class",
                    useAssay = "logcounts",
                    gene_subset = seq(100),
                    nfolds = 3,
                    nrepeats = 3,
                    seed = 99,
                    percent_top_biomarker = 0.2,
                    model_name = "logistic regression")



}
