% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMAST.R
\name{findMarkerDiffExp}
\alias{findMarkerDiffExp}
\title{Find the marker gene set for each cluster
With an input SingleCellExperiment object and specifying the clustering
labels, this function iteratively call the differential expression analysis
on each cluster against all the others.}
\usage{
findMarkerDiffExp(
  inSCE,
  useAssay = "logcounts",
  cluster = "cluster",
  log2fcThreshold = NULL,
  fdrThreshold = 1,
  useThresh = FALSE,
  freqExpressed = 0.1
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object.}

\item{useAssay}{character. A string specifying which assay to use for the
MAST calculations. Default \code{"logcounts"}.}

\item{cluster}{One single character to specify a column in
\code{colData(inSCE)} for the clustering label. Alternatively, a vector or
a factor is also acceptable. Default \code{"cluster"}.}

\item{log2fcThreshold}{Only out put DEGs with the absolute values of log2FC
larger than this value. Default \code{NULL}}

\item{fdrThreshold}{Only out put DEGs with FDR value smaller than this
value. Default \code{1}}

\item{useThresh}{Whether to use adaptive thresholding to filter genes.
Default \code{FALSE}.}

\item{freqExpressed}{A numeric threshold that the genes expressed in less
than this fraction of cells will be removed. Default \code{0.1}.}
}
\value{
The input \linkS4class{SingleCellExperiment} object with
\code{metadata(inSCE)$findMarker} updated with a data.table of the up-
regulated DEGs for each cluster.
}
\description{
Find the marker gene set for each cluster
With an input SingleCellExperiment object and specifying the clustering
labels, this function iteratively call the differential expression analysis
on each cluster against all the others.
}
\author{
Yichen Wang
}
