% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCluster.R
\name{getCluster}
\alias{getCluster}
\title{Get clustering with a chosen approach}
\usage{
getCluster(
  inSCE,
  useAssay = NULL,
  useReducedDim = NULL,
  useAltExp = NULL,
  method = c("scranSNN", "KMeans"),
  ...
)
}
\arguments{
\item{inSCE}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}
object.}

\item{useAssay}{A single \code{character}, specifying which
\code{\link[SummarizedExperiment]{assay}} to perform the clustering algorithm
on. Default \code{NULL}.}

\item{useReducedDim}{A single \code{character}, specifying which
low-dimension representation (\code{\link[SingleCellExperiment]{reducedDim}})
to perform the clustering algorithm on. Default \code{NULL}.}

\item{useAltExp}{A single \code{character}, specifying the assay in which
\code{\link[SingleCellExperiment]{altExp}} to perform the clustering algorithm
on. Default \code{NULL}.}

\item{method}{A single character, for which method to use. Choose from
\code{"scranSNN"}, \code{"KMeans"}. Default \code{"scranSNN"}.}

\item{...}{Other parameters passed to specified method.}
}
\value{
The input \code{\link[SingleCellExperiment]{SingleCellExperiment}}
object with \code{factor} cluster labeling updated in
\code{colData(inSCE)[[clusterName]]}.
}
\description{
Get clustering with a chosen approach
}
