% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HtmlExport.R
\name{getHtmlReportPerQC}
\alias{getHtmlReportPerQC}
\title{Get .html report of the output of the selected QC algorithm}
\usage{
getHtmlReportPerQC(
  inSCE,
  QCtype = "DecontX",
  output_file = NULL,
  output_dir = NULL
)
}
\arguments{
\item{inSCE}{A \link[SingleCellExperiment]{SingleCellExperiment} object containing
the count matrix (full droplets or filtered matrix, depends on the selected QC algorithm) with the output from at least one of these functions:
runQCMetrics, runScrublet, runDoubletCells, runCxds, runBcds, runCxdsBcdsHybrid, runDecontX, runBarcodeRankDrops, runEmptyDrops}

\item{output_file}{name of the generated file. If NULL/default then the output file name will be based on the name of the selected QC algorithm name .}

\item{output_dir}{name of the output directory to save the rendered file. If NULL/default the file is stored to the current working directory}

\item{algorithm}{Character. Specifies which QC algorithm report to generate.
Available options are "BarcodeRankDrops", "EmptyDrops", "QCMetrics", "Scrublet", "DoubletCells", "Cxds", "Bcds", "CxdsBcdsHybrid", "DoubletFinder"  and "DecontX".}
}
\value{
.html file
}
\description{
A  function to generate .html Rmarkdown report for the specified QC algorithm output
}
\examples{
inSCE <- runDecontX(inSCE)
getHtmlReportPerQC(inSCE = outSCE, QCtype = "DecontX")
}
