% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCA.R
\name{getPCA}
\alias{getPCA}
\title{Get PCA components for a SCtkE object}
\usage{
getPCA(count_data, use_assay = "logcounts", reducedDimName = "PCA")
}
\arguments{
\item{count_data}{SCtkE object}

\item{use_assay}{Indicate which assay to use for PCA. Default is "counts"}

\item{reducedDimName}{Store the PCA data with this name. The default is PCA.
The toolkit will store data with the pattern <ASSAY>_<ALGORITHM>.}
}
\value{
A SCtkE object with the specified reduecedDim and pca_variances
updated
}
\description{
Selects the 500 most variable genes in the SCE, performs
PCA based on them and stores the PCA values in the reducedDims slot of the
SCE object.
}
\examples{
data("GSE60361_subset_sce")
#add a CPM assay
assay(GSE60361_subset_sce, "cpm") <- apply(assay(GSE60361_subset_sce,
                                                 "counts"),
                                           2, function(x) {
                                             x / (sum(x) / 1000000)
                                           })
GSE60361_subset_sce <- getPCA(GSE60361_subset_sce,
                              use_assay = "cpm",
                              reducedDimName = "PCA_cpm")
reducedDims(GSE60361_subset_sce)

}
