% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSummaryStats.R
\name{getSampleSummaryStats}
\alias{getSampleSummaryStats}
\title{Store table of SCTK QC outputs to metadata.}
\usage{
getSampleSummaryStats(inSCE, statsName)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object with saved
\link{assay} data and/or \link{colData} data. Required.}

\item{statsName}{A \code{character} value indicating the slot
that stores the stats table within the metadata of the
SingleCellExperiment object. Required.}
}
\value{
A matrix/array object. Contains a summary table for QC statistics
generated from SingleCellTK.
}
\description{
Stores a table of QC metrics generated from
 QC algorithms within the metadata slot of the SingleCellExperiment object.
}
\examples{
data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sce <- runCellQC(sce)
sce <- sampleSummaryStats(sce, simple = FALSE)
getSampleSummaryStats(sce, statsName = "sctk_qc")
}
