% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUMAP.R
\name{getUMAP}
\alias{getUMAP}
\title{Uniform Manifold Approximation and Projection(UMAP) algorithm for
dimension reduction.}
\usage{
getUMAP(
  inSCE,
  useAssay = "logcounts",
  reducedDimName = "UMAP",
  n_neighbors = 5,
  n_iterations = 200,
  alpha = 1,
  metric = "euclidean",
  run_pca = TRUE
)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object. Required}

\item{useAssay}{Indicate which assay to use. The default is "logcounts".}

\item{reducedDimName}{a name to store the results of the dimension reduction
coordinates obtained from this method. This is stored in the SingleCellExperiment
object in the reducedDims slot. Required.}

\item{n_neighbors}{specify the number of nearest neighbors. Default is 5.}

\item{n_iterations}{number of iterations performed during layout optimization.
Default is 200.}

\item{alpha}{initial value of "learning rate" of layout optimization. Default is 1.}

\item{metric}{distance metric. Default is euclidean, other options are 'manhattan', 'cosine', 'pearson'.}

\item{run_pca}{run UMAP on PCA components? Default is TRUE.}
}
\value{
a SCtkExperiment object with the reduced dimensions updated under
reducedDimName specified.
}
\description{
Uniform Manifold Approximation and Projection(UMAP) algorithm for
dimension reduction.
}
\examples{
umap_res <- getUMAP(inSCE = mouseBrainSubsetSCE, useAssay = "counts",
                    reducedDimName = "UMAP", n_neighbors = 3, n_iterations = 200,
                    alpha = 1, metric = "euclidean",  run_pca = TRUE)
reducedDims(umap_res)

}
