% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE36552}
\alias{GSE36552}
\title{Example Single Cell RNA-Seq data in SCESet Object, GSE36552}
\format{List of two data frames, with counts and annotations. Use them as
input to createSCESet()}
\source{
DOI: 10.1038/nsmb.2660
}
\usage{
GSE36552
}
\description{
124 Single-cell transcriptome profiling from embryonic stem cells derived
from donated human pre-implatation embryos.
}
\examples{
library(scater)
data("GSE36552")
GSE36552_SCESet <- createSCESet(countfile = GSE36552$counts,
                                annotfile = GSE36552$annot,
                                inputdataframes = TRUE)
}
\keyword{datasets}
