% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE36552_sce}
\alias{GSE36552_sce}
\title{Example Single Cell RNA-Seq data in SCtkExperiment Object, GSE36552}
\format{SCtkExperiment}
\source{
DOI: 10.1038/nsmb.2660
}
\usage{
GSE36552_sce
}
\description{
124 Single-cell transcriptome profiling from embryonic stem cells derived
from donated human pre-implatation embryos.
}
\examples{
data("GSE36552_sce")
}
\keyword{datasets}
