% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE60361_subset}
\alias{GSE60361_subset}
\title{Example Single Cell RNA-Seq data in SCESet Object, GSE60361 subest}
\format{List of two data frames, with counts and annotations. Use them as
input to createSCESet()}
\source{
DOI: 10.1126/science.aaa1934
}
\usage{
GSE60361_subset
}
\description{
A subset of 30 samples from a single cell RNA-Seq experiment from Zeisel, et
al. Science 2015. The data was produced from cells from the mouse
somatosensory cortex (S1) and hippocampus (CA1). 15 of the cells were
identified as oligodendrocytes and 15 of the cell were identified as
microglia.
}
\examples{
library(scater)
data("GSE60361_subset")
GSE60361_SCESet <- createSCESet(countfile = GSE60361_subset$counts,
                                annotfile = GSE60361_subset$annot,
                                inputdataframes = TRUE)
}
\keyword{datasets}
