% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE60361_subset_sce}
\alias{GSE60361_subset_sce}
\title{Example Single Cell RNA-Seq data in SingleCelltkExperiment Object, GSE60361
subest}
\format{SingleCelltkExperiment}
\source{
DOI: 10.1126/science.aaa1934
}
\usage{
GSE60361_subset_sce
}
\description{
A subset of 30 samples from a single cell RNA-Seq experiment from Zeisel, et
al. Science 2015. The data was produced from cells from the mouse
somatosensory cortex (S1) and hippocampus (CA1). 15 of the cells were
identified as oligodendrocytes and 15 of the cell were identified as
microglia.
}
\examples{
data("GSE60361_subset_sce")
}
\keyword{datasets}
