% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE66507}
\alias{GSE66507}
\title{Example Single Cell RNA-Seq data in SCESet Object, GSE66507}
\format{List of two data frames, with counts and annotations. Use them as
input to createSCESet()}
\source{
DOI: 10.1242/dev.123547
}
\usage{
GSE66507
}
\description{
30 Single-cells from embryonic stem cells separated into three different
tissue types.
}
\examples{
library(scater)
data("GSE66507")
GSE66507_SCESet <- createSCESet(countfile = GSE66507$counts,
                                annotfile = GSE66507$annot,
                                inputdataframes = TRUE)
}
\keyword{datasets}
