% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE66507_sce}
\alias{GSE66507_sce}
\title{Example Single Cell RNA-Seq data in SingleCelltkExperiment Object, GSE66507}
\format{SingleCelltkExperiment}
\source{
DOI: 10.1242/dev.123547
}
\usage{
GSE66507_sce
}
\description{
30 Single-cells from embryonic stem cells separated into three different
tissue types.
}
\examples{
data("GSE66507_sce")
}
\keyword{datasets}
