% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE73121}
\alias{GSE73121}
\title{Example Single Cell RNA-Seq data in SCESet Object, GSE73121}
\format{List of two data frames, with counts and annotations. Use them as
input to createSCESet()}
\source{
DOI: 10.1186/s13059-016-0945-9
}
\usage{
GSE73121
}
\description{
117 Single-cell transcriptome profiling for metastatic renal cell carcinoma
patient-derived cells
}
\examples{
library(scater)
data("GSE73121")
GSE73121_SCESet <- createSCESet(countfile = GSE73121$counts,
                                annotfile = GSE73121$annot,
                                inputdataframes = TRUE)
}
\keyword{datasets}
