% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE73121_sce}
\alias{GSE73121_sce}
\title{Example Single Cell RNA-Seq data in SingleCellExperiment Object, GSE73121}
\format{SingleCellExperiment}
\source{
DOI: 10.1186/s13059-016-0945-9
}
\usage{
GSE73121_sce
}
\description{
117 Single-cell transcriptome profiling for metastatic renal cell carcinoma
patient-derived cells
}
\examples{
data("GSE73121_sce")
}
\keyword{datasets}
