% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSVA.R
\name{GSVA_plot}
\alias{GSVA_plot}
\title{GSVA_plot}
\usage{
GSVA_plot(SCEdata, gsva_data, plot_type, condition = NULL)
}
\arguments{
\item{SCEdata}{SCESet object}

\item{gsva_data}{GSVA data to plot. Required.}

\item{plot_type}{The type of plot to use, "Violin" or "Heatmap". Required.}

\item{condition}{The condition(s) to use for the Violin plot, or the condition(s)
to add as color bars above the Heatmap. Required for Violin, optional for Heatmap.}
}
\value{
The requested plot of the GSVA results.
}
\description{
Plot GSVA results.
}
