% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSVA.R
\name{GSVA_sce}
\alias{GSVA_sce}
\title{GSVA_sce}
\usage{
GSVA_sce(SCEdata, use_assay = "logcounts", pathway_source, pathway_names)
}
\arguments{
\item{SCEdata}{SCESet object}

\item{use_assay}{The assay to use for the MAST calculations. The default is
"logcounts"}

\item{pathway_source}{The pathway source if "Manual Input", the pathway_names
should be rowData annotations that are logical vectors. If, "MSigDB c2 (Human, Entrez ID only)",
the pathway_names should be pathways from MSigDB c2 or "ALL" to run on all
available pathways.}

\item{pathway_names}{List of pathway names to run, depending on pathway_source
parameter.}
}
\value{
A data.frame of pathway activity scores from GSVA.
}
\description{
Run GSVA analysis on a SCESet object.
}
