% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSVA.R
\name{gsvaPlot}
\alias{gsvaPlot}
\title{gsvaPlot}
\usage{
gsvaPlot(SCEdata, gsvaData, plotType, condition = NULL)
}
\arguments{
\item{SCEdata}{SCESet object}

\item{gsvaData}{GSVA data to plot. Required.}

\item{plotType}{The type of plot to use, "Violin" or "Heatmap". Required.}

\item{condition}{The condition(s) to use for the Violin plot, or the
condition(s) to add as color bars above the Heatmap. Required for Violin,
optional for Heatmap.}
}
\value{
The requested plot of the GSVA results.
}
\description{
Plot GSVA results.
}
