% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSVA.R
\name{gsvaSCE}
\alias{gsvaSCE}
\title{gsvaSCE}
\usage{
gsvaSCE(SCEdata, useAssay = "logcounts", pathwaySource, pathwayNames, ...)
}
\arguments{
\item{SCEdata}{SCESet object}

\item{useAssay}{The assay to use for the MAST calculations. The default is
"logcounts"}

\item{pathwaySource}{The pathway source if "Manual Input", the pathwayNames
should be rowData annotations that are (0,1) vectors. If, "MSigDB c2 (Human,
Entrez ID only)", the pathwayNames should be pathways from MSigDB c2 or "ALL"
to run on all available pathways.}

\item{pathwayNames}{List of pathway names to run, depending on pathwaySource
parameter.}

\item{...}{Parameters to pass to gsva}
}
\value{
A data.frame of pathway activity scores from GSVA.
}
\description{
Run GSVA analysis on a SCESet object.
}
