% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maits_sce}
\alias{maits_sce}
\title{Example Single Cell RNA-Seq MAITs data from MAST package}
\format{SCtkExperiment}
\source{
DOI: 10.1186/s13059-015-0844-5
}
\usage{
maits_sce
}
\description{
96 Single-cell transcriptome profiling from Mucosal Associated Invariant T
cells (MAITs), measured on the Fluidigm C1.
}
\examples{
data("maits_sce")
}
\keyword{datasets}
