% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maits_SCESet}
\alias{maits_SCESet}
\title{Example Single Cell RNA-Seq MAITS data from MAST package}
\format{List of three data frames, with counts, features and annotations.
Use them as input to createSCESet()}
\source{
DOI: 10.1186/s13059-015-0844-5
}
\usage{
maits_SCESet
}
\description{
96 Single-cell transcriptome profiling from Mucosal Associated Invariant T
cells (MAITs), measured on the Fluidigm C1.
}
\examples{
library(scater)
data("maits_SCESet")
maits_SCESet <- createSCESet(countfile = maits_SCESet$counts,
                             annotfile = maits_SCESet$annot,
                             featurefile = maits_SCESet$features,
                             inputdataframes = TRUE)
}
\keyword{datasets}
