% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST_plot_functions.R
\name{MASTregression}
\alias{MASTregression}
\title{MAST linear model}
\usage{
MASTregression(SCEdata, use_assay = "logcounts", fcHurdleSig,
  samplesize = 49, threshP = FALSE, variable)
}
\arguments{
\item{SCEdata}{SCtkExperiment object}

\item{use_assay}{The assay to use for the MAST calculations. The default is
"logcounts"}

\item{fcHurdleSig}{The filtered result from hurdle model}

\item{samplesize}{The number of most significant genes}

\item{threshP}{Plot threshold values from adaptive thresholding. Default is
FALSE}

\item{variable}{Select the condition of interest}
}
\description{
Run MAST analysis on a SCtkExperiment object.
}
