% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASTPlotFunctions.R
\name{MASTviolin}
\alias{MASTviolin}
\title{MAST Violin}
\usage{
MASTviolin(SCEdata, useAssay = "logcounts", fcHurdleSig, samplesize = 49,
  threshP = FALSE, variable)
}
\arguments{
\item{SCEdata}{SCtkExperiment object}

\item{useAssay}{The assay to use for the MAST calculations. The default is
"logcounts"}

\item{fcHurdleSig}{The filtered result from hurdle model}

\item{samplesize}{The number of most significant genes}

\item{threshP}{Plot threshold values from adaptive thresholding. Default is
FALSE}

\item{variable}{Select the condition of interest}
}
\value{
A ggplot object of MAST violin plots.
}
\description{
Run MAST analysis on a SCtkExperiment object.
}
