% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MitoGenes}
\alias{MitoGenes}
\title{List of mitochondrial genes of multiple reference}
\format{
A list
}
\usage{
data("MitoGenes")
}
\description{
A list of gene set that contains mitochondrial genes of multiple reference
(hg38, hg19, mm10 and mm9). It contains multiple types of gene identifier:
gene symbol, entrez ID, ensemble ID and ensemble transcript ID. It's used 
for the function 'importMitoGeneSet'.
}
\examples{
data("MitoGenes")
}
\keyword{datasets}
