% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdb_table}
\alias{msigdb_table}
\title{MSigDB gene get Cctegory table}
\format{
A data.frame.
}
\usage{
msigdb_table
}
\description{
A table of gene set categories that can be download from MSigDB. The
categories and descriptions can be found here:
https://www.gsea-msigdb.org/gsea/msigdb/collections.jsp. The IDs in the
first column can be used to retrieve the gene sets for these categories
using the \link{importGeneSetsFromMSigDB} function.
}
\examples{
data("msigdb_table")
}
\keyword{datasets}
