% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignSingleCellData.R
\name{parse_rsubread_logs}
\alias{parse_rsubread_logs}
\title{Parse Rsubread Logs for Mapping and Feature Count Statistics}
\usage{
parse_rsubread_logs(align_log = NULL, featurecount_log = NULL,
  sample_name = NULL)
}
\arguments{
\item{align_log}{Path to a log file created by the Rsubread align function}

\item{featurecount_log}{Path to a log file created by the Rsubread feature
count function}

\item{sample_name}{Sample name corresponding to the two log files}
}
\value{
A single line of a data frame with alignment and feature count
information
}
\description{
Parse Rsubread Logs for Mapping and Feature Count Statistics
}
