% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{pca_variances<-}
\alias{pca_variances<-}
\title{Set PCA variances}
\usage{
pca_variances(x, ...) <- value
}
\arguments{
\item{x}{SCtkE object}

\item{...}{other parameters}

\item{value}{PCA variances DataFrame()}
}
\description{
Set PCA variances
}
