% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{pcaVariances<-}
\alias{pcaVariances<-}
\title{Set PCA variances}
\usage{
pcaVariances(x, ...) <- value
}
\arguments{
\item{x}{SCtkE object}

\item{...}{other parameters}

\item{value}{PCA variances DataFrame()}
}
\value{
A SCtkExperiment object with the pcaVariances slot set.
}
\description{
Set PCA variances
}
\examples{
data("mouseBrainSubsetSCE")
pcaVariances(mouseBrainSubsetSCE)
#getPCA() sets the pcaVariances
newSCE <- getPCA(mouseBrainSubsetSCE, useAssay = "counts")

#alternatively, set the pcaVariances directly
pca <- prcomp(assay(mouseBrainSubsetSCE, "logcounts"))
percentVar <- pca$sdev ^ 2 / sum(pca$sdev ^ 2)
pcaVariances(mouseBrainSubsetSCE) <- DataFrame(percentVar)
}
