% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getset.R
\name{pcaVariances}
\alias{pcaVariances}
\alias{pcaVariances,SCtkExperiment-method}
\alias{pcaVariances<-,SCtkExperiment-method}
\title{Get PCA variances}
\usage{
pcaVariances(x, ...)

\S4method{pcaVariances}{SCtkExperiment}(x)

\S4method{pcaVariances}{SCtkExperiment}(x) <- value
}
\arguments{
\item{x}{SCtkE object}

\item{...}{other parameters}

\item{value}{The DataFrame of pcaVariances}
}
\value{
A data frame of percent variation explained by each PC.

A SCtkExperiment object with the pcaVariances object set.
}
\description{
Get PCA variances

Get PCA variances

Set PCA variances
}
\examples{
data("mouseBrainSubsetSCE")
pcaVariances(mouseBrainSubsetSCE)
}
