% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{plot_d3DiffEx}
\alias{plot_d3DiffEx}
\title{Plot Interactive Differential Expression}
\usage{
plot_d3DiffEx(inSCESet, condition, geneList, clusterRow = TRUE,
  clusterCol = TRUE)
}
\arguments{
<<<<<<< HEAD
\item{inSCESet}{Input data object that contains the data to be plotted.
Required}

\item{condition}{The condition used for plotting the heatmap. Required}

\item{geneList}{The list of genes to put in the heatmap. Required}

\item{clusterRow}{Cluster the rows. The default is TRUE}

\item{clusterCol}{Cluster the columns. The default is TRUE}
}
\value{
A d3heatmap object is plotted
=======
\item{clusterCol}{}
>>>>>>> Added some customization options for the heatmap in the differential expression tab.
}
\description{
Plot Interactive Differential Expression
}
