% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBatchEffect.R
\name{plotBatchEffect}
\alias{plotBatchEffect}
\title{Visualize batch effect with different kinds of diagnostic plots}
\usage{
plotBatchEffect(inSCE, useAssay = "logcounts", batch, condition = NULL,
  covariates = NULL, log_dat = FALSE, method)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object. Required}

\item{useAssay}{Indicate which assay to use for PCA. Default is "logcounts"}

\item{batch}{The column in the annotation data that corresponds to batch.
Required}

\item{condition}{The column in the annotation data that corresponds to
condition. Optional}

\item{log_dat}{Default FALSE, if TRUE log transform the assay}

\item{method}{Type of plot to generate}
}
\value{
A plot of degree of batch effect.
}
\description{
Visualize batch effect with different kinds of diagnostic plots
}
\examples{
if(requireNamespace("bladderbatch", quietly = TRUE)) {
  library(bladderbatch)
  data(bladderdata)
  dat <- as(as(bladderEset, "SummarizedExperiment"), "SCtkExperiment")
  plotBatchEffect(dat, useAssay="exprs", batch="batch", condition = "cancer",
                  method="Explained Variation")
}
}
