% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBiomarker.R
\name{plotBiomarker}
\alias{plotBiomarker}
\title{Given a set of genes, return a ggplot of expression
values.}
\usage{
plotBiomarker(inSCE, gene, binary = "Binary", visual = "PCA",
  shape = "No Shape", x = "PC1", y = "PC2", useAssay = "counts",
  reducedDimName = "PCA")
}
\arguments{
\item{inSCE}{Input SCtkExperiment object. Required}

\item{gene}{genelist to run the method on.}

\item{binary}{binary/continuous color for the expression.}

\item{visual}{Type of visualization (PCA, tSNE or UMAP). Default: "PCA"}

\item{shape}{shape parameter for the ggplot.}

\item{x}{x coordinate for PCA}

\item{y}{y coordinate for PCA}

\item{useAssay}{Indicate which assay to use. The default is "logcounts".}

\item{reducedDimName}{a name to store the results of the dimension reduction
coordinates obtained from this method. This is stored in the SingleCellExperiment
object in the reducedDims slot. Required.}
}
\value{
A Biomarker plot
}
\description{
Given a set of genes, return a ggplot of expression
values.
}
\examples{
data("mouseBrainSubsetSCE")
plotBiomarker(mouseBrainSubsetSCE, gene="C1qa", shape="level1class")

}
