% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBubble.R
\name{plotBubble}
\alias{plotBubble}
\title{Plot Bubble plot}
\usage{
plotBubble(
  inSCE,
  useAssay = "logcounts",
  featureNames,
  displayName = NULL,
  groupNames = "cluster",
  title = "",
  xlab = NULL,
  ylab = NULL,
  colorLow = "white",
  colorHigh = "blue",
  scale = FALSE
)
}
\arguments{
\item{inSCE}{The single cell experiment to use.}

\item{useAssay}{The assay to use.}

\item{featureNames}{A string or vector of strings with each gene to aggregate.}

\item{displayName}{A string that is the name of the column used for genes.}

\item{groupNames}{The name of a colData entry that can be used as groupNames.}

\item{title}{The title of the bubble plot}

\item{xlab}{The x-axis label}

\item{ylab}{The y-axis label}

\item{colorLow}{The color to be used for lowest value of mean expression}

\item{colorHigh}{The color to be used for highest value of mean expression}

\item{scale}{Option to scale the data. Default: \code{FALSE}. Selected assay will not be scaled.}
}
\value{
A ggplot of the bubble plot.
}
\description{
Plot a bubble plot with the color of the plot being the mean expression
and the size of the dot being the percent of cells in the cluster expressing the gene.
}
\examples{
data("scExample")
plotBubble(inSCE=sce, useAssay="counts", featureNames=c("B2M", "MALAT1"),
displayName="feature_name", groupNames="type", title="cell type test",
xlab="gene", ylab="cluster", colorLow="white", colorHigh="blue")
}
\seealso{
Other plot functions: 
\code{\link{computeHeatmap}()},
\code{\link{discreteColorPalette}()},
\code{\link{distinctColors}()},
\code{\link{plotBarcodeRankDropsResults}()},
\code{\link{plotBarcodeRankScatter}()},
\code{\link{plotBatchCorrCompare}()},
\code{\link{plotBatchVariance}()},
\code{\link{plotBcdsResults}()},
\code{\link{plotClusterAbundance}()},
\code{\link{plotCxdsResults}()},
\code{\link{plotDEGHeatmap}()},
\code{\link{plotDEGRegression}()},
\code{\link{plotDEGViolin}()},
\code{\link{plotDEGVolcano}()},
\code{\link{plotDecontXResults}()},
\code{\link{plotDimRed}()},
\code{\link{plotDoubletFinderResults}()},
\code{\link{plotEmptyDropsResults}()},
\code{\link{plotEmptyDropsScatter}()},
\code{\link{plotFindMarkerHeatmap}()},
\code{\link{plotMASTThresholdGenes}()},
\code{\link{plotPCA}()},
\code{\link{plotPathway}()},
\code{\link{plotRunPerCellQCResults}()},
\code{\link{plotSCEBarAssayData}()},
\code{\link{plotSCEBarColData}()},
\code{\link{plotSCEBatchFeatureMean}()},
\code{\link{plotSCEDensity}()},
\code{\link{plotSCEDensityAssayData}()},
\code{\link{plotSCEDensityColData}()},
\code{\link{plotSCEDimReduceColData}()},
\code{\link{plotSCEDimReduceFeatures}()},
\code{\link{plotSCEHeatmap}()},
\code{\link{plotSCEScatter}()},
\code{\link{plotSCEViolin}()},
\code{\link{plotSCEViolinAssayData}()},
\code{\link{plotSCEViolinColData}()},
\code{\link{plotScDblFinderResults}()},
\code{\link{plotScanpyDotPlot}()},
\code{\link{plotScanpyEmbedding}()},
\code{\link{plotScanpyHVG}()},
\code{\link{plotScanpyHeatmap}()},
\code{\link{plotScanpyMarkerGenes}()},
\code{\link{plotScanpyMarkerGenesDotPlot}()},
\code{\link{plotScanpyMarkerGenesHeatmap}()},
\code{\link{plotScanpyMarkerGenesMatrixPlot}()},
\code{\link{plotScanpyMarkerGenesViolin}()},
\code{\link{plotScanpyMatrixPlot}()},
\code{\link{plotScanpyPCA}()},
\code{\link{plotScanpyPCAGeneRanking}()},
\code{\link{plotScanpyPCAVariance}()},
\code{\link{plotScanpyViolin}()},
\code{\link{plotScdsHybridResults}()},
\code{\link{plotScrubletResults}()},
\code{\link{plotSeuratElbow}()},
\code{\link{plotSeuratGenes}()},
\code{\link{plotSeuratHVG}()},
\code{\link{plotSeuratHeatmap}()},
\code{\link{plotSeuratJackStraw}()},
\code{\link{plotSeuratReduction}()},
\code{\link{plotSoupXResults}()},
\code{\link{plotTSCANClusterDEG}()},
\code{\link{plotTSCANClusterPseudo}()},
\code{\link{plotTSCANDimReduceFeatures}()},
\code{\link{plotTSCANPseudotimeGenes}()},
\code{\link{plotTSCANPseudotimeHeatmap}()},
\code{\link{plotTSCANResults}()},
\code{\link{plotTSNE}()},
\code{\link{plotTopHVG}()},
\code{\link{plotUMAP}()}
}
\concept{plot functions}
