% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{plotDiffEx}
\alias{plotDiffEx}
\title{Plot Differential Expression}
\usage{
plotDiffEx(inSCESet, useAssay = "logcounts", condition, geneList,
  clusterRow = TRUE, clusterCol = TRUE, displayRowLabels = TRUE,
  displayColumnLabels = TRUE, displayRowDendrograms = TRUE,
  displayColumnDendrograms = TRUE, annotationColors = NULL,
  columnTitle = "Differential Expression")
}
\arguments{
\item{inSCESet}{Input data object that contains the data to be plotted.
Required}

\item{useAssay}{Indicate which assay to use. Default is "logcounts"}

\item{condition}{The condition used for plotting the heatmap. Required}

\item{geneList}{The list of genes to put in the heatmap. Required}

\item{clusterRow}{Cluster the rows. The default is TRUE}

\item{clusterCol}{Cluster the columns. The default is TRUE}

\item{displayRowLabels}{Display the row labels on the heatmap. The default
is TRUE.}

\item{displayColumnLabels}{Display the column labels on the heatmap. The
default is TRUE}

\item{displayRowDendrograms}{Display the row dendrograms on the heatmap. The
default is TRUE}

\item{displayColumnDendrograms}{Display the column dendrograms on the
heatmap. The default is TRUE.}

\item{annotationColors}{Set of annotation colors for color bar. If null,
no color bar is shown. default is NULL.}

\item{columnTitle}{Title to be displayed at top of heatmap.}
}
\value{
ComplexHeatmap object for the provided geneList annotated with the
condition.
}
\description{
Plot Differential Expression
}
\examples{
data("mouseBrainSubsetSCE")
res <- scDiffEx(mouseBrainSubsetSCE,
                useAssay = "logcounts",
                "level1class",
                diffexmethod = "limma")
plotDiffEx(mouseBrainSubsetSCE, condition = "level1class",
            geneList = rownames(res)[1:50], annotationColors = "auto")

}
