% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDimRedBaseline.R
\name{plotDimRedAssay}
\alias{plotDimRedAssay}
\title{Plot results of reduced dimensions data of counts stored in assay.}
\usage{
plotDimRedAssay(
  inSCE,
  feature,
  shape = "No Shape",
  reducedDimName = NULL,
  useAssay = "logcounts",
  comp1 = NULL,
  comp2 = NULL,
  pcX = NULL,
  pcY = NULL,
  background = TRUE,
  size = 2,
  title = NULL,
  titleSize = 15
)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object with saved dimension reduction components
or a variable with saved results. Required}

\item{feature}{name of feature stored in assay of singleCellExperiment object.
Plot will be colored based on feature value.}

\item{shape}{add shapes to each condition.}

\item{reducedDimName}{saved dimension reduction name in the SCtkExperiment object. Required.}

\item{useAssay}{Indicate which assay to use. The default is "logcounts"}

\item{comp1}{label for x-axis}

\item{comp2}{label for y-axis}

\item{pcX}{PCA component to be used for plotting(if applicable).
Default is first PCA component for PCA data and NULL otherwise.}

\item{pcY}{PCA component to be used for plotting(if applicable).
Default is second PCA component for PCA data and NULL otherwise.}

\item{background}{adds grid to plot when TRUE. Default TRUE.}

\item{size}{size of dots. Default 2.}

\item{title}{title of plot. Default NULL.}

\item{titleSize}{size of title of plot. Default 15.}
}
\value{
a ggplot of the reduced dimensions.
}
\description{
Plot results of reduced dimensions data of counts stored in assay.
}
\examples{
plotDimRedAssay(inSCE = mouseBrainSubsetSCE, feature = "Sox2",
           shape = "No Shape", reducedDimName = "TSNE_counts",
           useAssay = "counts", comp1 = "tSNE1", comp2 = "tSNE2")
}
