% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDimRedBaseline.R
\name{plotDimRedBaseline}
\alias{plotDimRedBaseline}
\title{Plot results of reduced dimensions data.}
\usage{
plotDimRedBaseline(
  inSCE,
  colorBy,
  shape,
  reducedDimName,
  conditionClass = NULL,
  useAssay,
  labelClusters = FALSE,
  comp1,
  comp2,
  dim1 = NULL,
  dim2 = NULL,
  background = FALSE,
  size = 2,
  title = NULL,
  titleSize = 15
)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object with saved dimension reduction components
or a variable with saved results. Required}

\item{colorBy}{color by a condition(any column of the annotation data).}

\item{shape}{add shapes to each condition.}

\item{reducedDimName}{saved dimension reduction name in the SCtkExperiment object. Required.}

\item{conditionClass}{class of the annotation data used in colorBy. Options are
NULL, "factor" or "numeric". If NULL, class will default to the original class.
Default NULL.}

\item{useAssay}{Indicate which assay to use. The default is "logcounts"}

\item{labelClusters}{Logical. Whether the cluster labels are plotted.
Default FALSE.}

\item{comp1}{label for x-axis}

\item{comp2}{label for y-axis}

\item{dim1}{1st dimension to be used for plotting. Default is NULL.}

\item{dim2}{2nd dimension to be used for plotting. Default is NULL.
Default is second PCA component for PCA data and NULL otherwise.}

\item{background}{adds grid to plot when TRUE. Default TRUE.}

\item{size}{size of dots. Default 2.}

\item{title}{title of plot. Default NULL.}

\item{titleSize}{size of title of plot. Default 15.}
}
\value{
a ggplot of the reduced dimensions.
}
\description{
Plot results of reduced dimensions data.
}
\examples{
plotDimRedBaseline(inSCE = mouseBrainSubsetSCE, colorBy = "No Color",
           shape = "No Shape", reducedDimName = "TSNE_counts",
           useAssay = "counts", comp1 = "tSNE1", comp2 = "tSNE2")

}
