% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDimRedBaseline.R
\name{plotDimRedColData}
\alias{plotDimRedColData}
\title{Plot results of reduced dimensions data of annotations in ColData.}
\usage{
plotDimRedColData(
  inSCE,
  colorBy = "No Color",
  shape = "No Shape",
  reducedDimName = NULL,
  conditionClass = NULL,
  comp1 = NULL,
  comp2 = NULL,
  dim1 = NULL,
  dim2 = NULL,
  background = TRUE,
  size = 2,
  title = NULL,
  titleSize = 15,
  labelClusters = TRUE
)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object with saved dimension reduction components
or a variable with saved results. Required}

\item{colorBy}{color by a condition(any column of the annotation data).}

\item{shape}{add shapes to each condition.}

\item{reducedDimName}{saved dimension reduction name in the SCtkExperiment object. Required.}

\item{conditionClass}{class of the annotation data used in colorBy. Options are
NULL, "factor" or "numeric". If NULL, class will default to the original class.
Default NULL.}

\item{comp1}{label for x-axis}

\item{comp2}{label for y-axis}

\item{dim1}{1st dimension to be used for plotting. Default is NULL.}

\item{dim2}{2nd dimension to be used for plotting. Default is NULL.}

\item{background}{adds grid to plot when TRUE. Default TRUE.}

\item{size}{size of dots. Default 2.}

\item{title}{title of plot. Default NULL.}

\item{titleSize}{size of title of plot. Default 15.}

\item{labelClusters}{Logical. Whether the cluster labels are plotted.
Default FALSE.}
}
\value{
a ggplot of the reduced dimensions.
}
\description{
Plot results of reduced dimensions data of annotations in ColData.
}
\examples{
plotDimRedColData(inSCE = mouseBrainSubsetSCE, colorBy = "tissue",
           shape = "No Shape", conditionClass = "factor",
           reducedDimName = "TSNE_counts",
           comp1 = "tSNE1", comp2 = "tSNE2", labelClusters = TRUE)
}
