% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDimRedQC.R
\name{plotDimRedQC}
\alias{plotDimRedQC}
\title{Dimension reduction plot of QC outputs.}
\usage{
plotDimRedQC(
  inSCE,
  outputs,
  log.counts = TRUE,
  reducedDimName = "TSNE",
  runDimensionReduction = FALSE,
  useAssay = "counts",
  size = 0.5,
  figRows = NULL,
  figCols = NULL
)
}
\arguments{
\item{inSCE}{A \link[SingleCellExperiment]{SingleCellExperiment} object.
Needs \code{counts} in assays slot.}

\item{outputs}{Names of QC values in the colData that will be plotted.}

\item{log.counts}{Determines whether counts should be log transformed 
prior to dimension reduction. Default TRUE.}

\item{reducedDimName}{Determines which dimension reduction coordinates 
to plot.}

\item{runDimensionReduction}{Determines whether dimension reduction needs
to be run prior to plotting. Default FALSE. If TRUE, will run dimension
reduction algorithm specified in "reducedDimName".}

\item{useAssay}{Indicate which assay to use. Default "counts".}

\item{size}{Numeric. Size of dots in plot. Default 0.5.}

\item{figRows}{Number of rows in figure.}

\item{figCols}{Number of columns in figure.}
}
\description{
Visualizes QC output values via a scatterplot.
 Users may either define specific outputs to plot, or can specify
 all outputs from a specific QC algorithm. Currently supports TSNE and UMAP.
 \code{fileName} is defined.
}
\examples{
data("mouseBrainSubsetSCE")
plotDimRedQC(inSCE = mouseBrainSubsetSCE, outputs = c("tissue","group"), 
runDimensionReduction=TRUE)
}
