% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plotHeatmap}
\usage{
plotHeatmap(
  inSCE,
  useAssay,
  dims = 1:dims,
  nfeatures = 30,
  cells = NULL,
  reduction = "pca",
  disp.min = -2.5,
  disp.max = NULL,
  balanced = TRUE,
  projected = FALSE,
  ncol = NULL,
  fast = FALSE,
  raster = TRUE,
  slot = "scale.data",
  assays = NULL,
  combine = TRUE,
  externalReduction = NULL
)
}
\arguments{
\item{inSCE}{}

\item{useAssay}{}

\item{dims}{}

\item{nfeatures}{}

\item{cells}{}

\item{reduction}{}

\item{disp.min}{}

\item{disp.max}{}

\item{balanced}{}

\item{projected}{}

\item{ncol}{}

\item{fast}{}

\item{raster}{}

\item{slot}{}

\item{assays}{}

\item{externalReduction}{}
}
\value{

}
\description{
plotHeatmap
}
