% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmapMulti.R
\name{plotHeatmapMulti}
\alias{plotHeatmapMulti}
\title{plotHeatmapMulti}
\usage{
plotHeatmapMulti(plots, components = NULL, nCol = NULL)
}
\arguments{
\item{plots}{Input list of plot objects.}

\item{components}{Specify the components against which the heatmaps should
be plotted. Default is \code{NULL} which will include all available
in input list of plots.}

\item{nCol}{Specify the number of columns in the output plot. Default
is \code{NULL} which will auto-compute the number of columns.}
}
\value{
Heatmap plot object.
}
\description{
plotHeatmapMulti
}
