% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASTPlotFunctions.R
\name{plotMASTHeatmap}
\alias{plotMASTHeatmap}
\title{Heatmap visualization of DEG result called by MAST}
\usage{
plotMASTHeatmap(
  inSCE,
  useResult = NULL,
  onlyPos = FALSE,
  log2fcThreshold = 1,
  fdrThreshold = 0.05,
  ...
)
}
\arguments{
\item{inSCE}{SingleCellExperiment object. Should be output of runMAST(),
with DEG information saved in metadata.}

\item{useResult}{character or numeric, default `NULL`. Specify which
comparison result to plot. Character refers to metadata(inSCE)$MAST; numeric
stands for index.}

\item{onlyPos}{logical, default `FALSE`. Whether to only plot DEG with
positive log2_FC value.}

\item{log2fcThreshold}{numeric, default `1`. Only plot DEGs with the
absolute values of log2FC larger than this value.}

\item{fdrThreshold}{numeric, default `0.05`. Only plot DEGs with FDR value
smaller than this value.}

\item{...}{Other arguments passed to `plotSCEHeatmap()`}
}
\value{
A ComplexHeatmap::Heatmap object
}
\description{
runMAST() has to be run in advance so that information is stored in the
metadata of the input SCE object. This function wraps plotSCEHeatmap.
}
