% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASTPlotFunctions.R
\name{plotMASTRegression}
\alias{plotMASTRegression}
\title{plot the linear regression to show visualize the expression the of DEGs
identified by MAST}
\usage{
plotMASTRegression(inSCE, useResult, threshP = FALSE, nrow = 6, ncol = 6)
}
\arguments{
\item{inSCE}{SCtkExperiment object. `runMAST()` has to be run in advance.}

\item{useResult}{character. A string specifying the `comparisonName` used.}

\item{threshP}{logical, default FALSE. Whether to plot threshold values from
adaptive thresholding, instead of using the assay used by `runMAST()`}

\item{nrow}{integer, default 6. Number of rows in the plot grid.}

\item{ncol}{integer, default 6. Number of columns in the plot grid.}
}
\value{
A ggplot object of MAST linear regression
}
\description{
plot the linear regression to show visualize the expression the of DEGs
identified by MAST
}
