% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASTPlotFunctions.R
\name{plotMASTViolin}
\alias{plotMASTViolin}
\title{plot the violin plot to show visualize the expression distribution of DEGs
identified by MAST}
\usage{
plotMASTViolin(inSCE, useResult, threshP = FALSE, nrow = 6, ncol = 6)
}
\arguments{
\item{inSCE}{SCtkExperiment object. `runMAST()` has to be run in advance.}

\item{useResult}{character. A string specifying the `comparisonName` used.}

\item{threshP}{logical, default FALSE. Whether to plot threshold values from
adaptive thresholding, instead of using the assay used by `runMAST()`}

\item{nrow}{integer, default 6. Number of rows in the plot grid.}

\item{ncol}{integer, default 6. Number of columns in the plot grid.}
}
\value{
A ggplot object of MAST violin plot
}
\description{
plot the violin plot to show visualize the expression distribution of DEGs
identified by MAST
}
