% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMusicResults.R
\name{plotMusicResults}
\alias{plotMusicResults}
\title{Plotting function of runMusic.R}
\usage{
plotMusicResults(
  inSCE,
  analysisType = c("EstCellProp", "PreGroupedClustProp", "SingleCellClust"),
  heatmapTitle = NULL,
  analysisName = NULL
)
}
\arguments{
\item{inSCE}{A \link[SingleCellExperiment]{SingleCellExperiment} object.}

\item{analysisType}{Character. Specify which function to run 
Available options are  "EstCellProp","PreGroupedClustProp","SingleCellClust"}

\item{heatmapTitle}{Character. Title for heatmap; Default is NULL}

\item{analysisName}{Character. User-defined analysis name.
This will be used as the slot name and results can be stored and retrived from SCE object using this name}
}
\value{
SingleCellExperiment object containing the outputs of the
 specified algorithms in the \link{colData}
of \code{inSCE}.
}
\description{
A wrapper that plots heatmap and cluster plots for results from runMusic.R
}
