% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpyFunctions.R
\name{plotScanpyDotPlot}
\alias{plotScanpyDotPlot}
\title{plotScanpyDotPlot}
\usage{
plotScanpyDotPlot(
  inSCE,
  useAssay = NULL,
  features,
  groupBy,
  standardScale = NULL,
  title = "",
  vmin = NULL,
  vmax = NULL,
  colorBarTitle = "Mean expression in group"
)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{useAssay}{Assay to use for plotting. By default it will use counts assay.}

\item{features}{Genes to plot. Sometimes is useful to pass a specific list of
var names (e.g. genes). The var_names could be a dictionary or a list.}

\item{groupBy}{The key of the observation grouping to consider.}

\item{standardScale}{Whether or not to standardize the given dimension 
between 0 and 1, meaning for each variable or group, subtract the minimum and 
divide each by its maximum. Default \code{NULL} means that it doesn't perform
any scaling.}

\item{title}{Provide title for the figure.}

\item{vmin}{The value representing the lower limit of the color scale. 
Values smaller than vmin are plotted with the same color as vmin. 
Default \code{NULL}}

\item{vmax}{The value representing the upper limit of the color scale. 
Values larger than vmax are plotted with the same color as vmax. 
Default \code{NULL}}

\item{colorBarTitle}{Title for the color bar.}
}
\value{
plot object
}
\description{
plotScanpyDotPlot
}
\examples{
data(scExample, package = "singleCellTK")
\dontrun{
sce <- runScanpyNormalizeData(sce, useAssay = "counts")
sce <- runScanpyFindHVG(sce, useAssay = "scanpyNormData", method = "seurat")
sce <- runScanpyScaleData(sce, useAssay = "scanpyNormData")
sce <- runScanpyPCA(sce, useAssay = "scanpyScaledData")
sce <- runScanpyFindClusters(sce, useReducedDim = "scanpyPCA")
sce <- runScanpyUMAP(sce, useReducedDim = "scanpyPCA")
markers <- c("MALAT1" ,"RPS27" ,"CST3")
plotScanpyDotPlot(sce, features = markers, groupBy = 'Scanpy_louvain_1')
}
}
\seealso{
Other plot functions: 
\code{\link{computeHeatmap}()},
\code{\link{discreteColorPalette}()},
\code{\link{distinctColors}()},
\code{\link{plotBarcodeRankDropsResults}()},
\code{\link{plotBarcodeRankScatter}()},
\code{\link{plotBatchCorrCompare}()},
\code{\link{plotBatchVariance}()},
\code{\link{plotBcdsResults}()},
\code{\link{plotBubble}()},
\code{\link{plotClusterAbundance}()},
\code{\link{plotCxdsResults}()},
\code{\link{plotDEGHeatmap}()},
\code{\link{plotDEGRegression}()},
\code{\link{plotDEGViolin}()},
\code{\link{plotDEGVolcano}()},
\code{\link{plotDecontXResults}()},
\code{\link{plotDimRed}()},
\code{\link{plotDoubletFinderResults}()},
\code{\link{plotEmptyDropsResults}()},
\code{\link{plotEmptyDropsScatter}()},
\code{\link{plotFindMarkerHeatmap}()},
\code{\link{plotMASTThresholdGenes}()},
\code{\link{plotPCA}()},
\code{\link{plotPathway}()},
\code{\link{plotRunPerCellQCResults}()},
\code{\link{plotSCEBarAssayData}()},
\code{\link{plotSCEBarColData}()},
\code{\link{plotSCEBatchFeatureMean}()},
\code{\link{plotSCEDensity}()},
\code{\link{plotSCEDensityAssayData}()},
\code{\link{plotSCEDensityColData}()},
\code{\link{plotSCEDimReduceColData}()},
\code{\link{plotSCEDimReduceFeatures}()},
\code{\link{plotSCEHeatmap}()},
\code{\link{plotSCEScatter}()},
\code{\link{plotSCEViolin}()},
\code{\link{plotSCEViolinAssayData}()},
\code{\link{plotSCEViolinColData}()},
\code{\link{plotScDblFinderResults}()},
\code{\link{plotScanpyEmbedding}()},
\code{\link{plotScanpyHVG}()},
\code{\link{plotScanpyHeatmap}()},
\code{\link{plotScanpyMarkerGenes}()},
\code{\link{plotScanpyMarkerGenesDotPlot}()},
\code{\link{plotScanpyMarkerGenesHeatmap}()},
\code{\link{plotScanpyMarkerGenesMatrixPlot}()},
\code{\link{plotScanpyMarkerGenesViolin}()},
\code{\link{plotScanpyMatrixPlot}()},
\code{\link{plotScanpyPCA}()},
\code{\link{plotScanpyPCAGeneRanking}()},
\code{\link{plotScanpyPCAVariance}()},
\code{\link{plotScanpyViolin}()},
\code{\link{plotScdsHybridResults}()},
\code{\link{plotScrubletResults}()},
\code{\link{plotSeuratElbow}()},
\code{\link{plotSeuratGenes}()},
\code{\link{plotSeuratHVG}()},
\code{\link{plotSeuratHeatmap}()},
\code{\link{plotSeuratJackStraw}()},
\code{\link{plotSeuratReduction}()},
\code{\link{plotSoupXResults}()},
\code{\link{plotTSCANClusterDEG}()},
\code{\link{plotTSCANClusterPseudo}()},
\code{\link{plotTSCANDimReduceFeatures}()},
\code{\link{plotTSCANPseudotimeGenes}()},
\code{\link{plotTSCANPseudotimeHeatmap}()},
\code{\link{plotTSCANResults}()},
\code{\link{plotTSNE}()},
\code{\link{plotTopHVG}()},
\code{\link{plotUMAP}()}
}
\concept{plot functions}
