% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpyFunctions.R
\name{plotScanpyEmbedding}
\alias{plotScanpyEmbedding}
\title{plotScanpyEmbedding}
\usage{
plotScanpyEmbedding(
  inSCE,
  reducedDimName,
  useAssay = NULL,
  color = NULL,
  legend = "right margin",
  title = ""
)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{reducedDimName}{Name of reducedDims object containing embeddings.
Eg. scanpyUMAP.}

\item{useAssay}{Specify name of assay to use. Default is \code{NULL},
which will use scaled assay by default.}

\item{color}{Keys for annotations of observations/cells or variables/genes.}

\item{legend}{Location of legend, either 'on data', 'right margin' or a 
valid keyword for the loc parameter of Legend.}

\item{title}{Provide title for panels either as string or list of strings}
}
\value{
plot object
}
\description{
plotScanpyEmbedding
}
\examples{
data(scExample, package = "singleCellTK")
\dontrun{
sce <- runScanpyNormalizeData(sce, useAssay = "counts")
sce <- runScanpyFindHVG(sce, useAssay = "scanpyNormData", method = "seurat")
sce <- runScanpyScaleData(sce, useAssay = "scanpyNormData")
sce <- runScanpyPCA(sce, useAssay = "scanpyScaledData")
sce <- runScanpyFindClusters(sce, useReducedDim = "scanpyPCA")
sce <- runScanpyUMAP(sce, useReducedDim = "scanpyPCA")
plotScanpyEmbedding(sce, reducedDimName = "scanpyUMAP", color = 'Scanpy_louvain_1')
}
}
\seealso{
Other plot functions: 
\code{\link{computeHeatmap}()},
\code{\link{discreteColorPalette}()},
\code{\link{distinctColors}()},
\code{\link{plotBarcodeRankDropsResults}()},
\code{\link{plotBarcodeRankScatter}()},
\code{\link{plotBatchCorrCompare}()},
\code{\link{plotBatchVariance}()},
\code{\link{plotBcdsResults}()},
\code{\link{plotBubble}()},
\code{\link{plotClusterAbundance}()},
\code{\link{plotCxdsResults}()},
\code{\link{plotDEGHeatmap}()},
\code{\link{plotDEGRegression}()},
\code{\link{plotDEGViolin}()},
\code{\link{plotDEGVolcano}()},
\code{\link{plotDecontXResults}()},
\code{\link{plotDimRed}()},
\code{\link{plotDoubletFinderResults}()},
\code{\link{plotEmptyDropsResults}()},
\code{\link{plotEmptyDropsScatter}()},
\code{\link{plotFindMarkerHeatmap}()},
\code{\link{plotMASTThresholdGenes}()},
\code{\link{plotPCA}()},
\code{\link{plotPathway}()},
\code{\link{plotRunPerCellQCResults}()},
\code{\link{plotSCEBarAssayData}()},
\code{\link{plotSCEBarColData}()},
\code{\link{plotSCEBatchFeatureMean}()},
\code{\link{plotSCEDensity}()},
\code{\link{plotSCEDensityAssayData}()},
\code{\link{plotSCEDensityColData}()},
\code{\link{plotSCEDimReduceColData}()},
\code{\link{plotSCEDimReduceFeatures}()},
\code{\link{plotSCEHeatmap}()},
\code{\link{plotSCEScatter}()},
\code{\link{plotSCEViolin}()},
\code{\link{plotSCEViolinAssayData}()},
\code{\link{plotSCEViolinColData}()},
\code{\link{plotScDblFinderResults}()},
\code{\link{plotScanpyDotPlot}()},
\code{\link{plotScanpyHVG}()},
\code{\link{plotScanpyHeatmap}()},
\code{\link{plotScanpyMarkerGenes}()},
\code{\link{plotScanpyMarkerGenesDotPlot}()},
\code{\link{plotScanpyMarkerGenesHeatmap}()},
\code{\link{plotScanpyMarkerGenesMatrixPlot}()},
\code{\link{plotScanpyMarkerGenesViolin}()},
\code{\link{plotScanpyMatrixPlot}()},
\code{\link{plotScanpyPCA}()},
\code{\link{plotScanpyPCAGeneRanking}()},
\code{\link{plotScanpyPCAVariance}()},
\code{\link{plotScanpyViolin}()},
\code{\link{plotScdsHybridResults}()},
\code{\link{plotScrubletResults}()},
\code{\link{plotSeuratElbow}()},
\code{\link{plotSeuratGenes}()},
\code{\link{plotSeuratHVG}()},
\code{\link{plotSeuratHeatmap}()},
\code{\link{plotSeuratJackStraw}()},
\code{\link{plotSeuratReduction}()},
\code{\link{plotSoupXResults}()},
\code{\link{plotTSCANClusterDEG}()},
\code{\link{plotTSCANClusterPseudo}()},
\code{\link{plotTSCANDimReduceFeatures}()},
\code{\link{plotTSCANPseudotimeGenes}()},
\code{\link{plotTSCANPseudotimeHeatmap}()},
\code{\link{plotTSCANResults}()},
\code{\link{plotTSNE}()},
\code{\link{plotTopHVG}()},
\code{\link{plotUMAP}()}
}
\concept{plot functions}
