% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{renameClusters}
\alias{renameClusters}
\title{Add together two numbers}
\usage{
renameClusters(inSCE, clusterName, from, to, newClusterName = NULL)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{clusterName}{Input}

\item{From}{value to describe what to map}

\item{To}{value to describe what to map to}

\item{newCluster}{value to change name of cluster}
}
\value{
inSCE object with changed values
}
\description{
Add together two numbers
}
