% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlReports.R
\name{reportDiffExp}
\alias{reportDiffExp}
\title{Get runDEAnalysis .html report}
\usage{
reportDiffExp(
  inSCE,
  study,
  useReducedDim,
  featureDisplay = NULL,
  output_file = NULL,
  output_dir = NULL
)
}
\arguments{
\item{inSCE}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}
object containing the output from \code{\link{runDEAnalysis}} function}

\item{study}{The specific analysis to visualize, used as \code{analysisName}
argument when running differential expression.}

\item{useReducedDim}{Specify an embedding for visualizing the relation ship
between the conditions.}

\item{featureDisplay}{The feature ID type to use for displaying. Should
exists as a variable name of \code{rowData}. Default \code{NULL} use rownames
of \code{inSCE}.}

\item{output_file}{name of the generated file. If \code{NULL} then the output
file name will be based on the name of the Rmarkdown template. Default
\code{NULL}.}

\item{output_dir}{name of the output directory to save the rendered file. If
\code{NULL} the file is stored to the current working directory.
Default \code{NULL}.}
}
\value{
Saves the HTML report in the specified output directory.
}
\description{
A  function to generate .html Rmarkdown report containing the
visualizations of the \code{\link{runDEAnalysis}} function output
}
\seealso{
Other report functions: 
\code{\link{reportCellQC}()},
\code{\link{reportClusterAbundance}()},
\code{\link{reportDiffAbundanceFET}()},
\code{\link{reportDropletQC}()},
\code{\link{reportFindMarker}()},
\code{\link{reportQCTool}()},
\code{\link{reportSeurat}()},
\code{\link{reportSeuratClustering}()},
\code{\link{reportSeuratDimRed}()},
\code{\link{reportSeuratFeatureSelection}()},
\code{\link{reportSeuratMarkerSelection}()},
\code{\link{reportSeuratNormalization}()},
\code{\link{reportSeuratResults}()},
\code{\link{reportSeuratRun}()},
\code{\link{reportSeuratScaling}()}
}
\concept{report functions}
