% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlReports.R
\name{reportSeuratNormalization}
\alias{reportSeuratNormalization}
\title{Generates an HTML report for Seurat Normalization
 and returns the SCE object with the results computed and stored
 inside the object.}
\usage{
reportSeuratNormalization(
  inSCE,
  outputFile = NULL,
  outputPath = NULL,
  subtitle = NULL,
  authors = NULL,
  showSession = FALSE,
  pdf = FALSE,
  forceRun = TRUE
)
}
\arguments{
\item{inSCE}{Input \code{\link[SingleCellExperiment]{SingleCellExperiment}}
object previously passed through \code{reportSeuratRun()}.}

\item{outputFile}{Specify the name of the generated output HTML file.
If \code{NULL} then the output file name will be based on the name of the
Rmarkdown template. Default \code{NULL}.}

\item{outputPath}{Specify the name of the output directory to save the
rendered HTML file. If \code{NULL} the file is stored to the current
working directory. Default \code{NULL}.}

\item{subtitle}{A character value specifying the subtitle to use in the
report. Default \code{NULL}.}

\item{authors}{A character value specifying the names of the authors to use
in the report. Default \code{NULL}.}

\item{showSession}{A logical value indicating if session information
should be displayed or not. Default is \code{FALSE}.}

\item{pdf}{A logical value indicating if a pdf should also be generated for
each figure in the report. Default is \code{FALSE}.}

\item{forceRun}{A logical value indicating if all computations previously
computed should be re-calculated regardless if these computations are
available in the input object. Default is \code{TRUE}.}
}
\value{
A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object
 with computations stored.
}
\description{
Generates an HTML report for Seurat Normalization
 and returns the SCE object with the results computed and stored
 inside the object.
}
\seealso{
Other report functions: 
\code{\link{reportCellQC}()},
\code{\link{reportClusterAbundance}()},
\code{\link{reportDiffAbundanceFET}()},
\code{\link{reportDiffExp}()},
\code{\link{reportDropletQC}()},
\code{\link{reportFindMarker}()},
\code{\link{reportQCTool}()},
\code{\link{reportSeurat}()},
\code{\link{reportSeuratClustering}()},
\code{\link{reportSeuratDimRed}()},
\code{\link{reportSeuratFeatureSelection}()},
\code{\link{reportSeuratMarkerSelection}()},
\code{\link{reportSeuratResults}()},
\code{\link{reportSeuratRun}()},
\code{\link{reportSeuratScaling}()}
}
\concept{report functions}
