% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEAnalysis.R
\name{runANOVA}
\alias{runANOVA}
\title{Perform differential expression analysis on SCE with ANOVA}
\usage{
runANOVA(
  inSCE,
  useAssay = "logcounts",
  index1 = NULL,
  index2 = NULL,
  class = NULL,
  classGroup1 = NULL,
  classGroup2 = NULL,
  analysisName,
  groupName1,
  groupName2,
  covariates = NULL,
  onlyPos = FALSE,
  log2fcThreshold = 0.25,
  fdrThreshold = 0.05,
  overwrite = FALSE
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object.}

\item{useAssay}{character. A string specifying which assay to use for ANOVA.
Default \code{"logcounts"}.}

\item{index1}{Any type of indices that can subset a
\linkS4class{SingleCellExperiment} inherited object by cells. Specifies
which cells are of interests. Default \code{NULL}.}

\item{index2}{Any type of indices that can subset a
\linkS4class{SingleCellExperiment} inherited object by cells. specifies
the control group against those specified by \code{index1}. If
\code{NULL} when using index specification, \code{index1} cells will be
compared with all other cells. Default \code{NULL}.}

\item{class}{A vector/factor with \code{ncol(inSCE)} elements, or a character
scalar that specifies a column name of \code{colData(inSCE)}. Default
\code{NULL}.}

\item{classGroup1}{a vector specifying which "levels" given in \code{class}
are of interests. Default \code{NULL}.}

\item{classGroup2}{a vector specifying which "levels" given in \code{class}
is the control group against those specified by \code{classGroup1}. If
\code{NULL} when using annotation specification, \code{classGroup1} cells
will be compared with all other cells.}

\item{analysisName}{A character scalar naming the DEG analysis. Required}

\item{groupName1}{A character scalar naming the group of interests. Required.}

\item{groupName2}{A character scalar naming the control group. Required.}

\item{covariates}{A character vector of additional covariates to use when
building the model. All covariates must exist in
\code{names(colData(inSCE))}. Default \code{NULL}.}

\item{onlyPos}{Whether to only output DEG with positive log2_FC value.
Default \code{FALSE}.}

\item{log2fcThreshold}{Only out put DEGs with the absolute values of log2FC
greater than this value. Default \code{0.25}}

\item{fdrThreshold}{Only out put DEGs with FDR value less than this
value. Default \code{0.05}}

\item{overwrite}{A logical scalar. Whether to overwrite result if exists.
Default \code{FALSE}.}
}
\value{
The input \linkS4class{SingleCellExperiment} object with
\code{metadata(inSCE)$diffExp} updated with the results: a list named by
\code{analysisName}, with \code{$groupNames} containing the naming of the
two conditions, \code{$useAssay} storing the assay name that was used for
calculation, \code{$select} storing the cell selection indices (logical) for
each condition, \code{$result} storing a \code{\link{data.frame}} of
the DEGs summary, and \code{$method} storing \code{"ANOVA"}.
}
\description{
Condition specification allows two methods:
1. Index level selection. Arguments \code{index1} and \code{index2} will be
used.
2. Annotation level selection. Arguments \code{class}, \code{classGroup1} and
\code{classGroup2} will be used.
}
\details{
NOTE that ANOVA method does not produce Log2FC value, but P-value and FDR
only.
}
\examples{
data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
library(scater)
sce <- logNormCounts(sce)
sce <- runANOVA(inSCE = sce, groupName1 = "Sample1",
 groupName2 = "Sample2", index1 = 1:20, index2 = 21:40,
 analysisName = "ANOVA", fdrThreshold = NULL)

}
