% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runComBat.R
\name{runComBat}
\alias{runComBat}
\title{runComBat}
\usage{
runComBat(
  inSCE,
  batch = "batch",
  useAssay = "logcounts",
  par.prior = TRUE,
  covariates = NULL,
  mean.only = FALSE,
  ref.batch = NULL,
  assayName = "ComBat"
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object. Required}

\item{batch}{A character scalar, for the name of a column in
\code{\link[SummarizedExperiment]{colData}} to use as the batch variable.
Default \code{"batch"}.}

\item{useAssay}{The assay to correct. Default \code{"logcounts"}.}

\item{par.prior}{A logical scalar. TRUE indicates parametric adjustments
will be used, FALSE indicates non-parametric adjustments will be used.
Default \code{TRUE}.}

\item{covariates}{List of other column names in colData to be added to the
ComBat model as covariates. Default \code{NULL}.}

\item{mean.only}{If TRUE ComBat only corrects the mean of the batch effect.
Default \code{FALSE}.}

\item{ref.batch}{If given, will use the selected batch as a reference for
batch adjustment. Default \code{NULL}.}

\item{assayName}{A character scalar for the new assay name to save the
corrected expression. Default \code{"ComBat"}}
}
\value{
Input SCTK object with corrected assay updated at \code{assay(inSCE, assayName)}.
}
\description{
Run ComBat batch correction method on a SCtkExperiment object
}
\examples{
if(requireNamespace("bladderbatch", quietly = TRUE)) {
  library(bladderbatch)
  data(bladderdata)

  #subset for testing
  dat <- bladderEset[1:50,]
  dat <- as(as(dat, "SummarizedExperiment"), "SCtkExperiment")

  # parametric adjustment
  dat <- runComBat(inSCE = dat, useAssay = "exprs",
                   batch = "batch", covariates = NULL,
                   assayName = "parametric_combat")

  # non-parametric adjustment, mean-only version
  dat <- runComBat(inSCE = dat, useAssay = "exprs",
                   batch = "batch", par.prior = FALSE,
                   mean.only = TRUE, covariates = NULL,
                   assayName = "nonparametric_combat_meanonly")

  # reference-batch version, with covariates
  dat <- runComBat(inSCE = dat, useAssay = "exprs",
                   batch = "batch", covariates = "cancer",
                   ref.batch = 3, assayName = "refbatch_combat_wcov")
}
}
