% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDimensionalityReduction.R
\name{runDimensionalityReduction}
\alias{runDimensionalityReduction}
\title{Wrapper function to run one of the three available dimensionality
reduction algorithms integrated within the singleCellTK from 'scaterPCA',
'seuratPCA' and 'seuratICA'.}
\usage{
runDimensionalityReduction(
  inSCE,
  useAssay,
  reducedDimName,
  method = c("scaterPCA", "seuratPCA", "seuratICA"),
  nPCs = 10
)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{useAssay}{Specify the name of the assay that should be used.}

\item{reducedDimName}{Specify the name of the output reducedDim.}

\item{method}{Specify a method from 'scaterPCA', 'seuratPCA' or 'seuratICA'.}

\item{nPCs}{Specify the number of dimensions to use with 'seuratPCA' or 
'seuratICA' methods. Not applicable for 'scaterPCA' method. Default
is \code{10}.}
}
\value{
A \linkS4class{SingleCellExperiment} object with PCA computation
updated in \code{reducedDim(inSCE, reducedDimName)}.
}
\description{
Wrapper function to run one of the three available dimensionality
reduction algorithms integrated within the singleCellTK from 'scaterPCA',
'seuratPCA' and 'seuratICA'.
}
