% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDimRed.R
\name{runDimRed}
\alias{runDimRed}
\title{Run PCA or tSNE
  
Use this function to run limma differential expression and load an interactive D3 heatmap}
\usage{
runDimRed(method, count_data, colorClusters, pc1, pc2)
}
\arguments{
\item{method}{Either PCA or tSNE}

\item{count_data}{A SCE object}

\item{colorClusters}{The variable to color clusters by}

\item{pc1}{User choice for the first principal component}

\item{pc2}{User choice for the second prinicipal component}
}
\value{
A reduced dimension object
}
\description{
Run PCA or tSNE
  
Use this function to run limma differential expression and load an interactive D3 heatmap
}
