% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLIGER.R
\name{runLIGER}
\alias{runLIGER}
\title{Apply LIGER batch effect correction method to SingleCellExperiment object}
\usage{
runLIGER(inSCE, exprs = "logcounts", batchKey = "batch",
  reducedDimName = "LIGER", nComponents = 20L, lambda = 5,
  resolution = 1)
}
\arguments{
\item{inSCE}{SingleCellExperiment object. An object that stores your dataset
and analysis procedures.}

\item{exprs}{character, default `"logcounts"`. A string indicating the name 
of the assay requiring batch correction in "inSCE", should exist in 
`assayNames(inSCE)`.}

\item{batchKey}{character, default `"batch"`. A string indicating the 
field of `colData(inSCE)` that defines different batches.}

\item{reducedDimName}{character, default `"LIGER"`. The name for the 
corrected low-dimensional representation.}

\item{nComponents}{integer, default `20L`. Number of principle components or 
dimensionality (factors, for this algorithm) to generate in the resulting 
reducedDim.}

\item{lambda}{numeric, default `5.0`. Algorithmic parameter, the penalty 
parameter which limits the dataset-specific component of the factorization.}

\item{resolution}{numeric, default `1.0`. Algorithmic paramter, the 
clustering resolution, increasing this increases the number of communities 
detected.}
}
\value{
SingleCellExperiment object with `reducedDim(inSCE, reducedDimName)` 
updated with corrected low-dimentional representation.
}
\description{
LIGER relies on integrative non-negative matrix factorization to identify 
shared and dataset-specific factors.
}
\examples{
 
data('sceBatches', package = 'singleCellTK')
sceBatches
## class: SingleCellExperiment 
## dim: 27610 1820 
## metadata(0):
## assays(3): normcounts logcounts
## rownames(27610): GCG MALAT1 ... LOC102724004 LOC102724238
## rowData names(0):
## colnames(1820): reads.12732 reads.12733 ... Sample_1598 Sample_1600
## colData names(2): cell_type1 batch
## reducedDimNames(5): PCA
## spikeNames(0):
sceCorr <- runLIGER(sceBatches)
}
\references{
Joshua Welch, et al., 2018
}
