% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMAST.R
\name{runMAST}
\alias{runMAST}
\title{runMAST}
\usage{
runMAST(
  inSCE,
  useAssay = "logcounts",
  index1 = NULL,
  index2 = NULL,
  class = NULL,
  classGroup1 = NULL,
  classGroup2 = NULL,
  comparisonName,
  groupName1,
  groupName2,
  useThresh = FALSE,
  freqExpressed = 0.1,
  onlyPos = FALSE,
  log2fcThreshold = NULL,
  fdrThreshold = 1
)
}
\arguments{
\item{inSCE}{SCtkExperiment object}

\item{useAssay}{character, default "logcounts". A string specifying which
assay to use for the MAST calculations.}

\item{index1}{character/numeric/logical, default NULL. A vector that
specifies which cells are of interests.}

\item{index2}{character/numeric/logical, default NULL. A vector that
specifies which cells are being compared with those specified by `index1`.
If NULL, `index1` cells will be comapred with all other cells.}

\item{class}{character/numeric or factor, default NULL. A vector/factor of
`ncol(inSCE)` elements, or a single string that specifies a colname of
`colData(inSCE)`.}

\item{classGroup1}{character/numeric, default NULL. Specifying which
"levels" given in `class` are of interests.}

\item{classGroup2}{character/numeric, default NULL. Specifying which
"levels" given in `class` will be compared with those specified by
`classGroup1`. If NULL, `classGroup1` cells will be compared with all other
cells.}

\item{comparisonName}{character. A string naming the DEG experiment.}

\item{groupName1}{character. A string naming the first comparison group.}

\item{groupName2}{character. A string naming the second comparison group.}

\item{useThresh}{logical, default FALSE. Whether to use adaptive
thresholding to filter genes.}

\item{freqExpressed}{numeric, default 0.1. A threshold that the genes
expressed in less than this fraction of cells will be removed.}

\item{onlyPos}{logical, default FALSE. Whether to only output DEG with
positive log2_FC value.}

\item{log2fcThreshold}{numeric, default NULL. Only out put DEGs with the
absolute values of log2FC larger than this value.}

\item{fdrThreshold}{numeric, default 1. Only out put DEGs with FDR value
smaller than this value.}
}
\value{
The input SCtkExperiment object with `metadata(inSCE)$MAST`` updated
with the results: a list named by `comparisonName`, with `groupNames`
containing the naming of the two conditions, `useAssay` storing the assay
name that was used for calculation, `select` storing the cell selection
indices (logical) for each condition, and `result` storing a `data.frame` of
the DEGs summary.
}
\description{
Run MAST for differential expression analysis on a SCtkExperiment object.
}
\examples{
data(sceBatch)
sce.deg <- runMAST(sceBatch, class = "cell_type", comparisonName = 'aVSb',
                   classGroup1 = 'alpha', classGroup2 = 'beta')
}
