% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPCA.R
\name{runPCA}
\alias{runPCA}
\title{Run multiple PCA approach}
\usage{
runPCA(plot.type, method, countm, annotm, featurem, involving.variables,
  additional.variables, colorClusters)
}
\arguments{
\item{plot.type}{Either 'Single Plot' or 'Paired Plot'}

\item{method}{regular PCA/randomized PCA/...}

\item{countm}{A count matrix}

\item{annotm}{An annotation matrix}

\item{featurem}{A feature annotation matrix}

\item{involving.variables}{User choice for the prinicipal components}

\item{additional.variables}{User choice for feature}

\item{colorClusters}{The variable to color clusters by}
}
\value{
A reduced dimension object
}
\description{
Use this function to run Principle component Analysis using different
approach and load plot
}
