% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCGEN.R
\name{runSCGEN}
\alias{runSCGEN}
\title{Apply scGen batch effect correction method to SingleCellExperiment object}
\usage{
runSCGEN(
  inSCE,
  exprs = "logcounts",
  batchKey = "batch",
  cellTypeKey = "cell_type",
  assayName = "SCGEN",
  nEpochs = 50L
)
}
\arguments{
\item{inSCE}{SingleCellExperiment object. An object that stores your dataset
and analysis procedures.}

\item{exprs}{character, default `"logcounts"`. A string indicating the name 
of the assay requiring batch correction in "inSCE", should exist in 
`assayNames(inSCE)`.}

\item{batchKey}{character, default `"batch"`. A string indicating the field 
of `colData(inSCE)` that defines different batches.}

\item{cellTypeKey}{character, default `"cell_type"`. A string indicating the 
field of `colData(inSCE)` that defines different cell types.}

\item{assayName}{character, default `"SCGEN"`. The name for the corrected 
full-sized expression matrix.}

\item{nEpochs}{integer, default `100L`. Algorithmic parameter, number of 
epochs to iterate and optimize network weights.}
}
\description{
scGen is a generative model to predict single-cell perturbation response 
across cell types, studies and species. It works by combining variational 
autoencoders and latent space vector arithmetics for high-dimensional single-
cell gene expression data.
}
\details{
Result does not look fine for now. Time consuming also even it allocates 32 
cores.
}
\examples{
data('sceBatches', package = 'singleCellTK')
sceBatches
## class: SingleCellExperiment 
## dim: 27610 1820 
## metadata(0):
## assays(3): normcounts logcounts
## rownames(27610): GCG MALAT1 ... LOC102724004 LOC102724238
## rowData names(0):
## colnames(1820): reads.12732 reads.12733 ... Sample_1598 Sample_1600
## colData names(2): cell_type1 batch
## reducedDimNames(5): PCA
## spikeNames(0):
sceCorr <- runSCGEN(sceBatches)
}
\references{
Lotfollahi, Mohammad et al., 2019
}
