% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSeurat.R
\name{runSeurat}
\alias{runSeurat}
\title{A wrapper function run complete Seurat workflow.}
\usage{
runSeurat(inSCE, useAssay)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{useAssay}{Specify the assay to use with Seurat workflow.}
}
\value{
\code{SingleCellExperiment} with results from Seurat workflow stored.
}
\description{
A wrapper function run complete Seurat workflow.
}
\examples{
data(sce_chcl, package = "scds")
sce_chcl <- runSeurat(inSCE = sce_chcl, useAssay = "counts")
}
