% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveBiomarkerRes.R
\name{saveBiomarkerRes}
\alias{saveBiomarkerRes}
\title{saveBiomarkerRes
Save biomarker gene information with a custom name when provided with diffex results.}
\usage{
saveBiomarkerRes(inSCE, diffex, biomarkerName, method, ntop = 25,
  logFC = NULL, pVal = NULL)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object. Required}

\item{diffex}{results table saved from the differential expression analysis.
Required.}

\item{biomarkerName}{name of the biomarker result to be saved under in rowData().
Required.}

\item{method}{name of the diffex method used to generate the results. Options
are DESeq2, Limma and ANOVA. Required}

\item{ntop}{number of top N genes. Default is 25. Required}

\item{logFC}{logfold-change cutoff applied to save biomarker results. Optional}

\item{pVal}{adjusted p-value cutoff. Optional}
}
\value{
a new SCE object with the diffex result saved in the rowData using
the "biomarkerName"
}
\description{
saveBiomarkerRes
Save biomarker gene information with a custom name when provided with diffex results.
}
\examples{
data("mouseBrainSubsetSCE")
res <- scDiffEx(mouseBrainSubsetSCE,
                useAssay = "logcounts",
                condition = "level1class",
                ntop = length(rownames(mouseBrainSubsetSCE)),
                usesig = FALSE,
                diffexmethod = "limma")
sceObj <- saveDiffExResults(mouseBrainSubsetSCE, res, "Limma_Level1class", "limma")
bioMarkRes <- saveBiomarkerRes(sceObj, res, "bioMarker", "limma", logFC = 0, pVal = 0.05)
}
