% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveDiffExResults.R
\name{saveDiffExResults}
\alias{saveDiffExResults}
\title{saveDiffExResults
Save Differential Expression Results with a custom name.}
\usage{
saveDiffExResults(inSCE, diffex, name, method)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object. Required}

\item{diffex}{results table saved from the differential expression analysis.
Required}

\item{name}{name of the result to be saved under in rowData(). Required}

\item{method}{name of the diffex method used to generate the results. Options
are DESeq2, limma and ANOVA. Required}
}
\value{
a new SCE object with the diffex result saved in the rowData using the "name"
}
\description{
saveDiffExResults
Save Differential Expression Results with a custom name.
}
\examples{
data("mouseBrainSubsetSCE")
res <- scDiffEx(mouseBrainSubsetSCE,
                useAssay = "logcounts",
                condition = "level1class",
                ntop = length(rownames(mouseBrainSubsetSCE)),
                usesig = FALSE,
                diffexmethod = "limma")
sceObj <- saveDiffExResults(mouseBrainSubsetSCE, res, "Limma_Level1class", "limma")
}
