% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scater_logNormCounts.R
\name{scater_logNormCounts}
\alias{scater_logNormCounts}
\title{scater_logNormCounts
Uses \link[scater]{logNormCounts} to log normalize input data}
\usage{
scater_logNormCounts(
  inSCE,
  logAssayName = "ScaterLogNormCounts",
  useAssay = "counts"
)
}
\arguments{
\item{inSCE}{Input SingleCellExperiment object}

\item{logAssayName}{New assay name for log normalized data}

\item{useAssay}{Input assay}
}
\value{
inSCE Updated SingleCellExperiment object that contains the new log normalized data
}
\description{
scater_logNormCounts
Uses \link[scater]{logNormCounts} to log normalize input data
}
\examples{
data(sce_chcl, package = "scds")
sce_chcl <- scater_logNormCounts(sce_chcl,"logcounts", "counts")
}
\author{
Irzam Sarfraz
}
