% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scater_logNormCounts.R
\name{scaterlogNormCounts}
\alias{scaterlogNormCounts}
\title{scaterlogNormCounts
Log-normalizes counts by library size, as \code{scater::logNormCounts}
did, using \code{\link[scrapper]{normalizeRnaCounts.se}}. Existing
\code{sizeFactors(inSCE)} are used (after centering) when present.}
\usage{
scaterlogNormCounts(
  inSCE,
  assayName = "ScaterLogNormCounts",
  useAssay = "counts"
)
}
\arguments{
\item{inSCE}{Input SingleCellExperiment object}

\item{assayName}{New assay name for log normalized data}

\item{useAssay}{Input assay}
}
\value{
inSCE Updated SingleCellExperiment object that contains the new log normalized data
}
\description{
scaterlogNormCounts
Log-normalizes counts by library size, as \code{scater::logNormCounts}
did, using \code{\link[scrapper]{normalizeRnaCounts.se}}. Existing
\code{sizeFactors(inSCE)} are used (after centering) when present.
}
\examples{
data(sce_chcl, package = "scds")
sce_chcl <- scaterlogNormCounts(sce_chcl,"logcounts", "counts")
}
\author{
Irzam Sarfraz
}
