% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{scDiffEx}
\alias{scDiffEx}
\title{Perform differential expression on a SCESet object using DESeq}
\usage{
scDiffEx(inSCESet, condition, significance, ntop, usesig)
}
\arguments{
\item{inSCESet}{Input SCESet object. Required}

\item{condition}{The name of the condition to use for differential
expression. Must be a name of a column from pData that contains two labels.}

\item{significance}{FDR corrected significance cutoff for differentially
expressed genes}
}
\value{
A list of differentially expressed genes.
}
\description{
Perform differential expression on a SCESet object using DESeq
}
\examples{

scDiffEx(newsceset_david, "age", "0.1")
}

