% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{scDiffEx_deseq}
\alias{scDiffEx_deseq}
\title{Perform differential expression analysis with DESeq}
\usage{
scDiffEx_deseq(inSCESet, condition)
}
\arguments{
\item{inSCESet}{Input SCtkExperiment object. Required}

\item{condition}{A factor for the condition to use for differential
expression. Must be a two level factor. Required}
}
\value{
A data frame of gene names and adjusted p-values
}
\description{
Returns a data frame of gene names and adjusted p-values
}
