% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{scDiffEx_deseq2}
\alias{scDiffEx_deseq2}
\title{Perform differential expression analysis with DESeq2}
\usage{
scDiffEx_deseq2(inSCESet, use_assay = "counts", condition,
  analysis_type = "biomarker", levelofinterest = NULL,
  controlLevel = NULL, covariates = NULL)
}
\arguments{
\item{inSCESet}{Input SCtkExperiment object. Required}

\item{use_assay}{Indicate which assay to use. Default is "counts"}

\item{condition}{The name of the condition to use for differential
expression. Must be a name of a column from colData that contains at least
two labels. Required}

\item{analysis_type}{Choose "biomarker" to compare the levelofinterest to all
other samples. Choose "contrast" to compare the levelofinterest to a
controlLevel (see below). Choose "fullreduced" to perform DESeq2 in LRT mode
comparing the model with condition to a model without condition.}

\item{levelofinterest}{If the condition has more than two labels,
levelofinterest should contain one factor of interest from condition.}

\item{controlLevel}{If the condition has more than two labels, controlLevel
should contain one factor from condition to use as the control.}

\item{covariates}{Additional covariates to add to the model.}
}
\value{
A data frame of gene names and adjusted p-values
}
\description{
Returns a data frame of gene names and adjusted p-values
}
\examples{
data("mouse_brain_subset_sce")
#subset to 100 genes
subset <- mouse_brain_subset_sce[rownames(mouse_brain_subset_sce)[order(rowSums(assay(mouse_brain_subset_sce, "counts")), decreasing = TRUE)][1:100], ]
res <- scDiffEx_deseq2(subset, condition = "level1class")

}
