% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{scDiffExANOVA}
\alias{scDiffExANOVA}
\title{Perform ANOVA analysis}
\usage{
scDiffExANOVA(inSCESet, useAssay = "logcounts", condition,
  covariates = NULL, adjust = "fdr")
}
\arguments{
\item{inSCESet}{Input SCtkExperiment object. Required}

\item{useAssay}{Indicate which assay to use. Default is "logcounts"}

\item{condition}{The name of the condition to use for differential
expression. Must be a name of a column from colData that contains at least
two labels. Required}

\item{covariates}{Additional covariates to add to the model.}

\item{adjust}{Method for p-value correction. See options in p.adjust().
The default is fdr.}
}
\value{
A data frame of gene names and adjusted p-values
}
\description{
Returns a data frame of gene names and adjusted p-values
}
\examples{
data("mouseBrainSubsetSCE")
res <- scDiffExANOVA(mouseBrainSubsetSCE, condition = "level1class")

}
