% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{scDiffExDESeq2}
\alias{scDiffExDESeq2}
\title{Perform differential expression analysis with DESeq2}
\usage{
scDiffExDESeq2(inSCESet, useAssay = "counts", condition,
  analysisType = "biomarker", levelofinterest = NULL, controlLevel = NULL,
  covariates = NULL, adjust = "fdr")
}
\arguments{
\item{inSCESet}{Input SCtkExperiment object. Required}

\item{useAssay}{Indicate which assay to use. Default is "counts"}

\item{condition}{The name of the condition to use for differential
expression. Must be a name of a column from colData that contains at least
two labels. Required}

\item{analysisType}{Choose "biomarker" to compare the levelofinterest to all
other samples. Choose "contrast" to compare the levelofinterest to a
controlLevel (see below). Choose "fullreduced" to perform DESeq2 in LRT mode
comparing the model with condition to a model without condition.}

\item{levelofinterest}{If the condition has more than two labels,
levelofinterest should contain one factor of interest from condition.}

\item{controlLevel}{If the condition has more than two labels, controlLevel
should contain one factor from condition to use as the control.}

\item{covariates}{Additional covariates to add to the model.}

\item{adjust}{Method for p-value correction. See options in p.adjust().
The default is fdr.}
}
\value{
A data frame of gene names and adjusted p-values
}
\description{
Returns a data frame of gene names and adjusted p-values
}
\examples{
data("mouseBrainSubsetSCE")
#sort first 100 expressed genes
ord <- rownames(mouseBrainSubsetSCE)[
  order(rowSums(assay(mouseBrainSubsetSCE, "counts")), 
        decreasing = TRUE)][1:100]
#subset to those first 100 genes
subset <- mouseBrainSubsetSCE[ord, ]
res <- scDiffExDESeq2(subset, condition = "level1class")

}
