% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{scDiffExlimma}
\alias{scDiffExlimma}
\title{Perform differential expression analysis with limma}
\usage{
scDiffExlimma(inSCESet, useAssay = "logcounts", condition,
  analysisType = "biomarker", levelofinterest = NULL, covariates = NULL,
  adjust = "fdr")
}
\arguments{
\item{inSCESet}{Input SCtkExperiment object. Required}

\item{useAssay}{Indicate which assay to use. Default is "logcounts"}

\item{condition}{The name of the condition to use for differential
expression. Must be a name of a column from colData that contains at least
two labels. Required}

\item{analysisType}{If there are more than two levels in your condition
variable, select the analysisType. Choose "biomarker" to compare the
levelofinterest to all other samples. Choose "coef" to select a coefficient
of interset with levelofinterest (see below). Choose "allcoef" to test if
any coefficient is different from zero.}

\item{levelofinterest}{If the condition has more than two labels,
levelofinterest should contain one factor of interest from condition.}

\item{covariates}{Additional covariates to add to the model.}

\item{adjust}{Method for p-value correction. See options in p.adjust().
The default is fdr.}
}
\value{
A data frame of gene names and adjusted p-values
}
\description{
Returns a data frame of gene names and adjusted p-values
}
\examples{
data("mouseBrainSubsetSCE")
res <- scDiffExlimma(mouseBrainSubsetSCE, condition = "level1class")

}
