% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scran_modelGeneVar.R
\name{scran_modelGeneVar}
\alias{scran_modelGeneVar}
\title{scran_modelGeneVar
Generates and stores variability data from scran::modelGeneVar in the input singleCellExperiment object}
\usage{
scran_modelGeneVar(inSCE, assayName)
}
\arguments{
\item{inSCE}{a singleCellExperiment object}

\item{assayName}{selected assay to compute variable features from}
}
\value{
inSCE updated singleCellExperiment object that contains variable feature metrics in rowData
}
\description{
scran_modelGeneVar
Generates and stores variability data from scran::modelGeneVar in the input singleCellExperiment object
}
\examples{
data(sce_chcl, package = "scds")
sce_chcl <- scran_modelGeneVar(sce_chcl, "counts")
}
\author{
Irzam Sarfraz
}
