% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scran_modelGeneVar.R
\name{scranModelGeneVar}
\alias{scranModelGeneVar}
\title{scranModelGeneVar
Generates and stores variability data from scran::modelGeneVar in the input singleCellExperiment object}
\usage{
scranModelGeneVar(inSCE, assayName)
}
\arguments{
\item{inSCE}{a singleCellExperiment object}

\item{assayName}{selected assay to compute variable features from}
}
\value{
inSCE updated singleCellExperiment object that contains variable feature metrics in rowData
}
\description{
scranModelGeneVar
Generates and stores variability data from scran::modelGeneVar in the input singleCellExperiment object
}
\examples{
data(sce_chcl, package = "scds")
sce_chcl <- scranModelGeneVar(sce_chcl, "counts")
}
\author{
Irzam Sarfraz
}
