% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate_setup.R
\name{sctk_py_install}
\alias{sctk_py_install}
\title{Installs Python packages}
\usage{
sctk_py_install(packages = c("scipy", "scrublet"), ...)
}
\arguments{
\item{packages}{List of packages to install}

\item{...}{Other parameters to pass to \code{\link[reticulate]{py_install}}. Can be useful for
selecting which environment to install these packages into. If no parameters are supplied,
then packages will be installed into default Python environment chosen by \code{\link{reticulate}}.}
}
\description{
Install all Python packages used in the \code{\link{singleCellTK}} package
using \code{\link[reticulate]{py_install}} from package \code{\link{reticulate}}.
}
