% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctkTagging.R
\name{sctkAltExp<-}
\alias{sctkAltExp<-}
\title{sctkAltExp
Store assays using tags to identify the type of assay stored. To be used
within the singleCellTK as a replacement for assay<- setter function.}
\usage{
sctkAltExp(inSCE, e, tag = NULL) <- value
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{tag}{Specify the tag to store against the input assay. Default
is \code{NULL}, which will set the tag to 'uncategorized'.}

\item{value}{Input matrix-type assay to store.}
}
\description{
sctkAltExp
Store assays using tags to identify the type of assay stored. To be used
within the singleCellTK as a replacement for assay<- setter function.
}
