% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{sctkAssay<-,ANY,ANY,CharacterOrNULL-method}
\alias{sctkAssay<-,ANY,ANY,CharacterOrNULL-method}
\title{sctkAssay
Store assays using tags to identify the type of assay stored. To be used
within the singleCellTK as a replacement for assay<- setter function.}
\usage{
\S4method{sctkAssay}{ANY,ANY,CharacterOrNULL}(inSCE, assayName, tag = NULL) <- value
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{assayName}{Specify the name of the input assay.}

\item{tag}{Specify the tag to store against the input assay. Default
is \code{NULL}, which will set the tag to 'uncategorized'.}

\item{value}{Input matrix-type assay to store.}
}
\description{
sctkAssay
Store assays using tags to identify the type of assay stored. To be used
within the singleCellTK as a replacement for assay<- setter function.
}
