% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctkTagging.R
\name{sctkAssay}
\alias{sctkAssay}
\title{sctkAssay
Store assays using tags to identify the type of assay stored. To be used
within the singleCellTK as a replacement for assay<- setter function.}
\usage{
sctkAssay(inSCE, assayName)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{assayName}{Specify the name of the input assay.}
}
\description{
sctkAssay
Store assays using tags to identify the type of assay stored. To be used
within the singleCellTK as a replacement for assay<- setter function.
}
