% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctkTagging.R
\name{sctkAssayNames}
\alias{sctkAssayNames}
\title{sctkAssayNames
Store assays using tags to identify the type of assay stored. To be used
within the singleCellTK as a replacement for assay<- setter function.}
\usage{
sctkAssayNames(inSCE)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}
}
\description{
sctkAssayNames
Store assays using tags to identify the type of assay stored. To be used
within the singleCellTK as a replacement for assay<- setter function.
}
