% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctkTagging.R
\name{sctkDeleteTag}
\alias{sctkDeleteTag}
\title{sctkDeleteTag}
\usage{
sctkDeleteTag(inSCE, assay)
}
\arguments{
\item{inSCE}{input SCE}

\item{assay}{assay to remove tag from}
}
\value{
Input SCE
}
\description{
sctkDeleteTag
}
