% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SCtkExperiment-class}
\alias{SCtkExperiment-class}
\title{A lightweight S4 extension to the SingleCellExperiment class to store
additional information.}
\arguments{
\item{value}{The DataFrame of pcaVariances}
}
\value{
A SingleCellExperiment like object with an addition pcaVariances
slot.
}
\description{
A lightweight S4 extension to the SingleCellExperiment class to store
additional information.
}
\section{Slots}{

\describe{
\item{\code{pcaVariances}}{The percent variation contained in each PCA dimension}
}}

\examples{
data("mouseBrainSubsetSCE")
counts_mat <- assay(mouseBrainSubsetSCE, "counts")
sample_annot <- colData(mouseBrainSubsetSCE)
row_annot <- rowData(mouseBrainSubsetSCE)
newSCE <- SCtkExperiment(assays=list(counts=counts_mat),
                         colData=sample_annot,
                         rowData=row_annot)
newSCE <- getPCA(newSCE, useAssay = "counts")
#View the percent variation of the PCA
pcaVariances(newSCE)

}
