% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{SCtkExperiment}
\alias{SCtkExperiment}
\title{Create a SCtkExperiment}
\usage{
SCtkExperiment(..., pca_variances = S4Vectors::DataFrame())
}
\arguments{
\item{...}{SingleCellExperiment and SummarizedExperiment components}

\item{pca_variances}{The percent variation contained in each PCA dimension}
}
\value{
A SingleCellExperiment like object with an addition pca_variances
slot.
}
\description{
Create a SCtkExperiment
}
\examples{
data("GSE60361_subset_sce")
counts_mat <- assay(GSE60361_subset_sce, "counts")
sample_annot <- colData(GSE60361_subset_sce)
row_annot <- rowData(GSE60361_subset_sce)
newSCE <- SCtkExperiment(assays=list(counts=counts_mat),
                         colData=sample_annot,
                         rowData=row_annot)
newSCE <- getPCA(newSCE, use_assay = "counts")
#View the percent variation of the PCA
pca_variances(newSCE)

}
