% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctkTagging.R
\name{sctkGetTag}
\alias{sctkGetTag}
\title{sctkGetTag}
\usage{
sctkGetTag(inSCE, assayType, redDims = FALSE)
}
\arguments{
\item{inSCE}{Input SCE}

\item{assayType}{type of assay}

\item{redDims}{redims iff required}
}
\value{
List
}
\description{
sctkGetTag
}
