% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctkTagging.R
\name{sctkSetTag}
\alias{sctkSetTag}
\title{sctkSetTag}
\usage{
sctkSetTag(inSCE, assayType, assays)
}
\arguments{
\item{inSCE}{input SCE}

\item{assayType}{type of tag}

\item{assays}{assay to set tag}
}
\value{
Input SCE
}
\description{
sctkSetTag
}
