% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctkTagging.R
\name{sctkSetTagExternal}
\alias{sctkSetTagExternal}
\title{sctkSetTagExternal}
\usage{
sctkSetTagExternal(inSCE, assayType, assays)
}
\arguments{
\item{inSCE}{Input SingleCellExperiment}

\item{assayType}{Type of assay}

\item{assays}{assays to set to}
}
\value{
Input SCE
}
\description{
sctkSetTagExternal
}
