% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SEG}
\alias{SEG}
\title{Stably Expressed Gene (SEG) list obect, with SEG sets for human and mouse.}
\format{
list, with two entries \code{"human"} and \code{"mouse"}, each is a
charactor vector.
}
\source{
\code{data('segList', package='scMerge')}
}
\usage{
SEG
}
\description{
The two gene sets came from dataset called `segList` of package `scMerge`.
}
\examples{
data('SEG')
humanSEG <- SEG$human
}
\keyword{datasets}
