% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSummaryStats.R
\name{setSampleSummaryStats}
\alias{setSampleSummaryStats}
\title{Store table of SCTK QC outputs to metadata. Executed within
`sampleSummaryStats` function.}
\usage{
setSampleSummaryStats(inSCE, slot, stats)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object with saved
\link{assay} data and/or \link{colData} data. Required.}

\item{slot}{A \code{character} value which specifies the
desired slot to store the stats table within the metadata of
the SingleCellExperiment object. Required.}

\item{stats}{Input stats table that will be stored within the
SingleCellExperiment object. Required.}
}
\value{
A SingleCellExperiment object with a summary table for QC statistics
in the `sampleSummary` slot of metadata.
}
\description{
Stores a table of QC metrics generated from
 QC algorithms within the metadata slot of the SingleCellExperiment object.
}
