% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{setSampleSummaryStatsTable<-}
\alias{setSampleSummaryStatsTable<-}
\title{Setter function which stores table of SCTK QC outputs to metadata.}
\usage{
setSampleSummaryStatsTable(inSCE, ...) <- value
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object with saved
\link{assay} data and/or \link{colData} data. Required.}

\item{...}{Other arguments passed to the function.}

\item{value}{The sample summary table of SCTK QC outputs}
}
\value{
A SingleCellExperiment object which contains a summary table for QC statistics
generated from SingleCellTK.
}
\description{
Stores table of QC metrics generated from
 QC algorithms within the metadata slot of the SingleCellExperiment object.
}
