% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{setSCERowNames}
\alias{setSCERowNames}
\title{Set rownames of SCE with a character vector or a rowData column}
\usage{
setSCERowNames(inSCE, rowNames, dedup = TRUE)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object}

\item{rowNames}{Character. Either of length equal to \code{nrow(inSCE)}, or
a single character specifying a column in \code{rowData(inSCE)}.}

\item{dedup}{Logical. Whether to deduplicate the specified rowNames. Default
\code{TRUE}}
}
\value{
The input SCE object with rownames updated.
}
\description{
Users can set rownames of an SCE object with either a character
vector where the length equals to \code{nrow(inSCE)}, or a single character 
specifying a column in \code{rowData(inSCE)}. Users can set 
\code{dedup = TRUE} to remove duplicated entries in the specification, by 
adding \code{-1, -2, ..., -i} suffix to the duplication of the same 
identifier.
}
\examples{
data("scExample", package = "singleCellTK")
head(rownames(sce))
sce <- setSCERowNames(sce, "feature_name")
head(rownames(sce))
}
