% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratComputeHeatmap}
\alias{seuratComputeHeatmap}
\title{seuratComputeHeatmap
Computes the heatmap plot object from the pca slot in the input sce object}
\usage{
seuratComputeHeatmap(sceObject, geneNamesSeurat, dims)
}
\arguments{
\item{sceObject;}{sce object from which to compute heatmap (pca should be computed)}

\item{geneNamesSeurat;}{a list of rowames/genenames/featurenames for seurat object for consistent formatting}

\item{dims;}{numerical value indicating how many components to use for the computation of heatmap plot object (default is number of pca components computed)}
}
\value{
plot object
}
\description{
seuratComputeHeatmap
Computes the heatmap plot object from the pca slot in the input sce object
}
