% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratComputeJackStraw}
\alias{seuratComputeJackStraw}
\title{seuratComputeJackStraw
Compute jackstraw plot and store the computations in the input sce object}
\usage{
seuratComputeJackStraw(inSCE, useAssay, geneNamesSeurat, dims)
}
\arguments{
\item{inSCE}{(sce) object on which to compute and store jackstraw plot}

\item{useAssay}{which assay to use from sce object}

\item{geneNamesSeurat}{a list of rowames/genenames/featurenames for seurat object for consistent formatting}

\item{dims}{numeric value of how many components to use for jackstraw plot (default = number of computed principal components)}
}
\value{
sceObject updated sce object with jackstraw computations stored in it
}
\description{
seuratComputeJackStraw
Compute jackstraw plot and store the computations in the input sce object
}
