% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratElbowPlot}
\alias{seuratElbowPlot}
\title{seuratElbowPlot
Computes the plot object for elbow plot from the pca slot in the input sce object}
\usage{
seuratElbowPlot(inSCE, useAssay, geneNamesSeurat, significantPC)
}
\arguments{
\item{inSCE}{(sce) object from which to compute the elbow plot (pca should be computed)}

\item{useAssay}{which assay to use from sce object}

\item{geneNamesSeurat}{a list of rowames/genenames/featurenames for seurat object for consistent formatting}

\item{significantPC}{a numerical value indicating the number of significant principal components (used to alter the color of the significant components)}
}
\value{
plot object
}
\description{
seuratElbowPlot
Computes the plot object for elbow plot from the pca slot in the input sce object
}
