% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratFindClusters}
\alias{seuratFindClusters}
\title{seuratFindClusters
Computes the clusters from the input sce object and stores them back in sce object}
\usage{
seuratFindClusters(
  inSCE,
  useAssay,
  useReduction = c("pca", "ica"),
  dims = 10,
  algorithm = c("louvain", "multilevel", "SLM"),
  groupSingletons = TRUE
)
}
\arguments{
\item{inSCE}{(sce) object from which clusters should be computed and stored in}

\item{useAssay}{Assay containing scaled counts to use for clustering.}

\item{useReduction}{Reduction method to use for computing clusters. One of "pca" or "ica". Default \code{"pca"}.}

\item{dims}{numeric value of how many components to use for computing clusters. Default \code{10}.}

\item{algorithm}{selected algorithm to compute clusters. One of "louvain", "multilevel", or "SLM". Use \code{louvain} for "original Louvain algorithm" and \code{multilevel} for "Louvain algorithm with multilevel refinement". Default \code{louvain}.}

\item{groupSingletons}{boolean if singletons should be grouped together or not. Default \code{TRUE}.}
}
\value{
sceObject; updated sce object which now contains the computed clusters
}
\description{
seuratFindClusters
Computes the clusters from the input sce object and stores them back in sce object
}
