% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratHeatmapPlot}
\alias{seuratHeatmapPlot}
\title{seuratHeatmapPlot
Modifies the heatmap plot object so it contains specified number of heatmaps in a single plot}
\usage{
seuratHeatmapPlot(plotObject, dims, ncol, labels)
}
\arguments{
\item{plotObject}{plot object computed from seuratComputeHeatmap() function}

\item{dims}{numerical value of how many heatmaps to draw (default is 0)}

\item{ncol}{numerical value indicating that in how many columns should the heatmaps be distrbuted (default is 2)}

\item{labels}{list() of labels to draw on heatmaps}
}
\value{
modified plot object
}
\description{
seuratHeatmapPlot
Modifies the heatmap plot object so it contains specified number of heatmaps in a single plot
}
