% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratICA}
\alias{seuratICA}
\title{seuratICA
Computes ICA on the input sce object and stores the calculated independent components within the sce object}
\usage{
seuratICA(sceObject, geneNamesSeurat, nics)
}
\arguments{
\item{sceObject}{sce object on which to compute ICA}

\item{geneNamesSeurat}{a list of rowames/genenames/featurenames for seurat object for consistent formatting}

\item{nics}{numeric value of how many components to compute (default is 20)}
}
\value{
sceObject updated sce object which now contains the computed independent components
}
\description{
seuratICA
Computes ICA on the input sce object and stores the calculated independent components within the sce object
}
