% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratJackStrawPlot}
\alias{seuratJackStrawPlot}
\title{seuratJackStrawPlot
Computes the plot object for jackstraw plot from the pca slot in the input sce object}
\usage{
seuratJackStrawPlot(sceObject, geneNamesSeurat, dims)
}
\arguments{
\item{sceObject;}{sce object from which to compute the jackstraw plot (pca should be computed)}

\item{geneNamesSeurat;}{a list of rowames/genenames/featurenames for seurat object for consistent formatting}

\item{dims;}{a numerical value indicating how many components to use in the computation of jackstraw plot from pca (default is number of pca components computed)}
}
\value{
plot object
}
\description{
seuratJackStrawPlot
Computes the plot object for jackstraw plot from the pca slot in the input sce object
}
