% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratNormalizeData}
\alias{seuratNormalizeData}
\title{seuratNormalizeData
Wrapper for NormalizeData() function from seurat library
Normalizes the sce object according to the input parameters}
\usage{
seuratNormalizeData(
  sceObject,
  geneNamesSeurat,
  normalizationMethod,
  scaleFactor
)
}
\arguments{
\item{sceObject;}{sce object to normalize}

\item{geneNamesSeurat;}{a list of rowames/genenames/featurenames for seurat object for consistent formatting}

\item{normalizationMethod;}{selected normalization method (default is "LogNormalize")}

\item{scaleFactor;}{numeric value that represents the scaling factor (default is 10000)}
}
\value{
sceObject; normalized sce object
}
\description{
seuratNormalizeData
Wrapper for NormalizeData() function from seurat library
Normalizes the sce object according to the input parameters
}
