% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratPlotHVG}
\alias{seuratPlotHVG}
\title{seuratPlotHVG
Plot highly variable genes from input sce object (must have highly variable genes computations stored)}
\usage{
seuratPlotHVG(sceObject, geneNamesSeurat)
}
\arguments{
\item{sceObject}{sce object that contains the highly variable genes computations}

\item{geneNamesSeurat}{a list of rowames/genenames/featurenames for seurat object for consistent formatting}
}
\value{
plot object
}
\description{
seuratPlotHVG
Plot highly variable genes from input sce object (must have highly variable genes computations stored)
}
