% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratReductionPlot}
\alias{seuratReductionPlot}
\title{seuratReductionPlot
Plots the selected dimensionality reduction method}
\usage{
seuratReductionPlot(inSCE, useReduction = c("pca", "ica", "tsne", "umap"))
}
\arguments{
\item{inSCE}{(sce) object which has the selected dimensionality reduction algorithm already computed and stored}

\item{useReduction}{Dimentionality reduction to plot. One of "pca", "ica", "tsne", or "umap". Default \code{"umap"}.}
}
\value{
plot object
}
\description{
seuratReductionPlot
Plots the selected dimensionality reduction method
}
