% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratRunTSNE}
\alias{seuratRunTSNE}
\title{seuratRunTSNE
Computes tSNE from the given sce object and stores the tSNE computations back into the sce object}
\usage{
seuratRunTSNE(inSCE, useAssay, geneNamesSeurat, reduction, dims)
}
\arguments{
\item{inSCE}{(sce) object on which to compute the tSNE}

\item{useAssay}{which assay to use from sce object}

\item{geneNamesSeurat}{a list of rowames/genenames/featurenames for seurat object for consistent formatting}

\item{reduction}{selected reduction algorithm to use for computing tSNE ("pca" or "ica", default is "pca")}

\item{dims}{numerical value of how many reduction components to use for tSNE computation (default is 10)}
}
\value{
Updated sce object with tSNE computations stored
}
\description{
seuratRunTSNE
Computes tSNE from the given sce object and stores the tSNE computations back into the sce object
}
