% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratRunUMAP}
\alias{seuratRunUMAP}
\title{seuratRunUMAP
Computes UMAP from the given sce object and stores the UMAP computations back into the sce object}
\usage{
seuratRunUMAP(
  inSCE,
  useReduction = c("pca", "ica"),
  reducedDimName = "seuratUMAP",
  dims = 10
)
}
\arguments{
\item{inSCE}{(sce) object on which to compute the UMAP}

\item{useReduction}{Reduction to use for computing UMAP. One of "pca" or "ica". Default is \code{"pca"}.}

\item{reducedDimName}{Name of new reducedDims object containing Seurat UMAP Default \code{seuratUMAP}.}

\item{dims}{numerical value of how many reduction components to use for UMAP computation. Default \code{10}.}
}
\value{
Updated sce object with UMAP computations stored
}
\description{
seuratRunUMAP
Computes UMAP from the given sce object and stores the UMAP computations back into the sce object
}
