% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratScaleData}
\alias{seuratScaleData}
\title{seuratScaleData
Scales the input sce object according to the input parameters}
\usage{
seuratScaleData(
  inSCE,
  useAssay,
  geneNamesSeurat,
  model.use,
  do.scale,
  do.center,
  scale.max
)
}
\arguments{
\item{inSCE}{(sce) object to scale}

\item{useAssay}{which assay to use from sce object}

\item{geneNamesSeurat}{a list of rowames/genenames/featurenames for seurat object for consistent formatting}

\item{model.use}{selected model to use for scaling data (default is "linear")}

\item{do.scale}{boolean if data should be scaled or not (TRUE or FALSE, default is TRUE)}

\item{do.center}{boolean if data should be centered or not (TRUE or FALSE, default is TRUE)}

\item{scale.max}{maximum numeric value to return for scaled data (default is 10)}
}
\value{
sceObject scaled sce object
}
\description{
seuratScaleData
Scales the input sce object according to the input parameters
}
