% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratSCTransform}
\alias{seuratSCTransform}
\title{seuratSCTransform
Runs the \code{SCTransform} function from \code{Seurat} library to transform/normalize the input data}
\usage{
seuratSCTransform(inSCE, normAssayName = "SCTCounts", useAssay = "counts")
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object}

\item{normAssayName}{Name for the output data assay. Default \code{"SCTCounts"}.}

\item{useAssay}{Name for the input data assay. Default \code{"counts"}.}
}
\description{
seuratSCTransform
Runs the \code{SCTransform} function from \code{Seurat} library to transform/normalize the input data
}
