% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratVariableFeatures}
\alias{seuratVariableFeatures}
\title{Get variable feature names after running seuratFindHVG function}
\usage{
seuratVariableFeatures(inSCE)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}
}
\value{
A list of variable feature names.
}
\description{
Get variable feature names after running seuratFindHVG function
}
