% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{simpleLog}
\alias{simpleLog}
\title{A decorator that prints the arguments to the decorated function}
\usage{
simpleLog(f)
}
\arguments{
\item{f}{A function to decorate}
}
\value{
Prints message
}
\description{
A decorator that prints the arguments to the decorated function
}
