% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_SVM.R
\name{singlecell_SVM}
\alias{singlecell_SVM}
\title{To implement SVM for quality control}
\usage{
singlecell_SVM(train_set, train_label, test_set, var)
}
\arguments{
\item{train_set}{The training dataset (cells x features)}

\item{train_label}{The training label annotation specifying the high/low quality of the training dataset.}

\item{test_set}{The test dataset (cells x features)}

\item{var}{feature names used for predicting cell quality ##}
}
\description{
To implement SVM for quality control
}
