% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{SingleCelltkExperiment}
\alias{SingleCelltkExperiment}
\title{Create a SingleCelltkExperiment}
\usage{
SingleCelltkExperiment(..., pca_variances = DataFrame())
}
\arguments{
\item{...}{SingleCellExperiment and SummarizedExperiment components}

\item{pca_variances}{The percent variation contained in each PCA dimension}
}
\description{
Create a SingleCelltkExperiment
}
