% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SRP063840}
\alias{SRP063840}
\title{Example Single Cell RNA-Seq data in SCESet Object}
\format{List of two data frames, with counts and annotations. Use them as
input to createSCESet()}
\source{
DOI: 10.1186/s13059-016-0945-9
}
\usage{
SRP063840
}
\description{
117 Single-cell transcriptome profiling for metastatic renal cell carcinoma
patient-derived cells
}
\examples{
library(scater)
data("SRP063840")
SRP063840_SCESet <- createSCESet(countfile = SRP063840$counts,
                                annotfile = SRP063840$annot,
                                inputdataframes = TRUE)
}
\keyword{datasets}
