% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downsample.Matrix.R
\name{subDiffEx_anova}
\alias{subDiffEx_anova}
\title{Runs ANOVA on all genes in a simulated dataset with more than 2 conditions,
and adjusts for FDR.}
\usage{
subDiffEx_anova(countMatrix, condition)
}
\arguments{
\item{countMatrix}{Matrix. A simulated counts matrix, sans labels.}

\item{condition}{Factor. The condition labels for the simulated cells.}
}
\value{
A vector of fdr-adjusted p-values for all genes. Nonviable results
(such as for genes with 0 counts in a simulated dataset) are coerced to 1.
}
\description{
Runs ANOVA on all genes in a simulated dataset with more than 2 conditions,
and adjusts for FDR.
}
\examples{
data("mouse_brain_subset_sce")
#sort first 100 expressed genes
ord <- rownames(mouse_brain_subset_sce)[
  order(rowSums(assay(mouse_brain_subset_sce, "counts")), 
        decreasing = TRUE)][1:100]
#subset to those first 100 genes
subset <- mouse_brain_subset_sce[ord, ]
res <- generateSimulatedData(totalReads = 1000, cells=10,
                             originalData = assay(subset, "counts"),
                             realLabels = colData(subset)[, "level2class"])
realLabels <- res[1, ]
output <- res[-1, ]
#fdr <- subDiffEx_anova(output, realLabels)

}
