% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownsampleMatrix.R
\name{subDiffExANOVA}
\alias{subDiffExANOVA}
\title{Runs ANOVA on all genes in a simulated dataset with more than 2 conditions,
and adjusts for FDR.}
\usage{
subDiffExANOVA(countMatrix, condition)
}
\arguments{
\item{countMatrix}{Matrix. A simulated counts matrix, sans labels.}

\item{condition}{Factor. The condition labels for the simulated cells.}
}
\value{
A vector of fdr-adjusted p-values for all genes. Nonviable results
(such as for genes with 0 counts in a simulated dataset) are coerced to 1.
}
\description{
Runs ANOVA on all genes in a simulated dataset with more than 2 conditions,
and adjusts for FDR.
}
\examples{
data("mouseBrainSubsetSCE")
#sort first 100 expressed genes
ord <- rownames(mouseBrainSubsetSCE)[
  order(rowSums(assay(mouseBrainSubsetSCE, "counts")), 
        decreasing = TRUE)][1:100]
# subset to those first 100 genes
subset <- mouseBrainSubsetSCE[ord, ]
res <- generateSimulatedData(totalReads = 1000, cells=10,
                             originalData = assay(subset, "counts"),
                             realLabels = colData(subset)[, "level2class"])
realLabels <- res[1, ]
output <- res[-1, ]
fdr <- subDiffExANOVA(output, realLabels)

}
