% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownsampleMatrix.R
\name{subDiffExttest}
\alias{subDiffExttest}
\title{Runs t-tests on all genes in a simulated dataset with 2 conditions, and
adjusts for FDR.}
\usage{
subDiffExttest(dataset, class.labels, test.type = "t.equalvar")
}
\arguments{
\item{dataset}{Matrix. A simulated counts matrix, sans labels.}

\item{class.labels}{Factor. The condition labels for the simulated cells.
Will be coerced into 1's and 0's.}

\item{test.type}{Type of test to perform. The default is t.equalvar.}
}
\value{
A vector of fdr-adjusted p-values for all genes. Nonviable results
(such as for genes with 0 counts in a simulated dataset) are coerced to 1.
}
\description{
Runs t-tests on all genes in a simulated dataset with 2 conditions, and
adjusts for FDR.
}
\examples{
data("mouseBrainSubsetSCE")
#sort first 100 expressed genes
ord <- rownames(mouseBrainSubsetSCE)[
  order(rowSums(assay(mouseBrainSubsetSCE, "counts")), 
        decreasing = TRUE)][1:100]
#subset to those first 100 genes
subset <- mouseBrainSubsetSCE[ord, ]
res <- generateSimulatedData(totalReads = 1000, cells=10,
                             originalData = assay(subset, "counts"),
                             realLabels = colData(subset)[, "level1class"])
realLabels <- res[1, ]
output <- res[-1, ]
fdr <- subDiffExttest(output, realLabels)

}
