% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{summarizeTable}
\alias{summarizeTable}
\title{Summarize SingleCelltkExperiment}
\usage{
summarizeTable(indata)
}
\arguments{
\item{indata}{Input SingleCelltkExperiment}
}
\value{
A data.frame object of summary metrics.
}
\description{
Creates a table of summary metrics from an input SingleCelltkExperiment.
}
