% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASTPlotFunctions.R
\name{thresholdGenes}
\alias{thresholdGenes}
\title{thresholdGenes}
\usage{
thresholdGenes(SCEdata, useAssay = "logcounts")
}
\arguments{
\item{SCEdata}{SCtkExperiment object}

\item{useAssay}{The assay to use for the MAST calculations. The default is
"logcounts"}
}
\value{
list of thresholded counts (on natural scale), thresholds, bins,
densities estimated on each bin, and the original data from
MAST::thresholdSCRNACountMatrix
}
\description{
thresholdGenes
}
\examples{
data("mouseBrainSubsetSCE")
res <- thresholdGenes(mouseBrainSubsetSCE)

}
