% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQCViolinPlot.R
\name{violinPlotQC}
\alias{violinPlotQC}
\title{Violin plot of QC outputs.}
\usage{
violinPlotQC(inSCE, outputnames, sample = NULL, useAssay = "counts",
  fileName = NULL, fileWidth = 20, fileHeight = 20,
  fileUnits = "cm", figRows = NULL, figCols = NULL)
}
\arguments{
\item{inSCE}{A \link[SingleCellExperiment]{SingleCellExperiment} object.
Needs \code{counts} in assays slot.}

\item{outputnames}{Names of QC values in the colData that will be plotted.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.}

\item{useAssay}{Indicate which assay to use. Default "counts".}

\item{fileName}{Desired file name of the output plot file. (e.g. .pdf, .png) 
If NULL, a plot object will be outputted, with no file output. Default NULL.}

\item{fileWidth}{Width of output plot file.}

\item{fileHeight}{Height of output plot file.}

\item{fileUnits}{Units of output plot file dimensions. Default "cm".}

\item{figRows}{Number of rows in figure.}

\item{figCols}{Number of columns in figure.}
}
\description{
Visualizes QC output values via a violin plot.
 Users may either define specific outputs to plot, or can specify
 all outputs from a specific QC algorithm. Will save plot if parameter 
 \code{fileName} is defined.
}
\examples{
data("mouseBrainSubsetSCE")
violinPlotQC(inSCE = mouseBrainSubsetSCE, 
 outputnames = "total.mRNA.mol")
}
