% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visPlot.R
\name{visPlot}
\alias{visPlot}
\title{visPlot}
\usage{
visPlot(inSCE, useAssay, method, condition, glist)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object. Required}

\item{useAssay}{The assay to use in the visualization plot. Required}

\item{method}{Visualization method. Available options are boxplot,
scatterplot, or heatmap. Required}

\item{condition}{colData annotation of the experiment. Required}

\item{glist}{selected genes for visualization. Required}
}
\value{
A visualization plot
}
\description{
Given a plotting method with condition and gene list, return
the respective visualization plot(s).
}
\examples{
visPlot(mouseBrainSubsetSCE, "logcounts", "boxplot", "level1class", "C1qa")
visPlot(mouseBrainSubsetSCE, "counts", "scatterplot", "age", "Cmtm5")
visPlot(mouseBrainSubsetSCE, "counts", "heatmap", "level1class",
        c("Cmtm5", "C1qa"))
}
