% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceTxtExport.R
\name{writeSCE}
\alias{writeSCE}
\title{Export \link[SingleCellExperiment]{SingleCellExperiment} object}
\usage{
writeSCE(
  sce,
  outputDir = "./",
  outputType = "txt",
  overwrite = FALSE,
  gzipped = TRUE
)
}
\arguments{
\item{sce}{\link[SingleCellExperiment]{SingleCellExperiment} object to be
exported.}

\item{outputDir}{Name of the directory to store the exported file(s).}

\item{outputType}{The desired type of the output files. Default \code{"txt"}
which
writes the \link[SingleCellExperiment]{SingleCellExperiment} object as
tab delimited text files.}

\item{overwrite}{Boolean. Whether to overwrite the output files. Default
\code{FALSE}.}

\item{gzipped}{Boolean. \code{TRUE} if the output files are to be
gzip compressed. \code{FALSE} otherwise. Default
\code{TRUE} to save disk space.}
}
\description{
Save \link[SingleCellExperiment]{SingleCellExperiment} to
 disk.
}
\examples{
data(sce_chcl, package = "scds")
writeSCE(sce_chcl, "sce_chcl")

if (!requireNamespace("scRNAseq", quietly = TRUE)) {
  BiocManager::install("scRNAseq")
}
sce <- scRNAseq::ReprocessedAllenData("tophat_counts")
writeSCE(sce, "ReprocessedAllenData")
}
