% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctkQCUtils.R
\name{combineSCE}
\alias{combineSCE}
\title{Combine a list of SingleCellExperiment objects as one SingleCellExperiment object}
\usage{
combineSCE(sceList)
}
\arguments{
\item{sceList}{A list contains \link[SingleCellExperiment]{SingleCellExperiment} objects}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object which combines all 
objects in sceList. The colData is merged.
}
\description{
Combine a list of SingleCellExperiment objects as one SingleCellExperiment object
}
