% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSCE.R
\name{combineSCE}
\alias{combineSCE}
\title{Combine a list of SingleCellExperiment objects as one SingleCellExperiment object}
\usage{
combineSCE(sceList, by.r, by.c, combined)
}
\arguments{
\item{sceList}{A list contains \link[SingleCellExperiment]{SingleCellExperiment} objects}

\item{by.r}{Specifications of the columns used for merging rowData. See 'Details'.}

\item{by.c}{Specifications of the columns used for merging colData. See 'Details'.}

\item{combined}{logical; if TRUE, it will combine the list of SingleCellExperiment objects. See 'Details'.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object which combines all
objects in sceList. The colData is merged.
}
\description{
Combine a list of SingleCellExperiment objects as one SingleCellExperiment object
}
\examples{
combinedsce <- combineSCE(list(sce,sce), by.r = NULL, by.c = NULL, combined = TRUE)
}
