% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctkQCUtils.R
\name{constructSCE}
\alias{constructSCE}
\title{Create SingleCellExperiment object from csv or txt input}
\usage{
constructSCE(data, samplename)
}
\arguments{
\item{data}{A \link[data.table]{data.table} object containing the count matrix.}

\item{samplename}{The sample name of the data.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object containing
the count matrix.
}
\description{
Create SingleCellExperiment object from csv or txt input
}
