% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene_ids.R
\name{convert_gene_ids}
\alias{convert_gene_ids}
\title{Convert gene IDs}
\usage{
convert_gene_ids(inSCESet, in_symbol, out_symbol, database = "org.Hs.eg.db")
}
\arguments{
\item{inSCESet}{Input SCtkExperiment object. Required}

\item{in_symbol}{The input symbol type}

\item{out_symbol}{The output symbol type}

\item{database}{The org.*.eg.db database to use. The default is org.Hs.eg.db}
}
\value{
A SCtkExperiment with converted gene IDs.
}
\description{
Convert the gene IDs in a SingleCellExperiment object using Bioconductor
org.*.eg.db data packages. Because annotation databases do not have a 1:1
relationship, this tool removes rows with no corresponding annotation in
your desired annotation, and remove any duplicate annotations after
conversion.
}
\examples{
if(requireNamespace("org.Mm.eg.db", quietly = TRUE)) {
  #convert mouse gene symbols to ensembl IDs
  library("org.Mm.eg.db")
  sample(rownames(GSE60361_subset_sce), 50)
  GSE60361_subset_symbol <- convert_gene_ids(inSCESet = GSE60361_subset_sce,
                                             in_symbol = "SYMBOL",
                                             out_symbol = "ENSEMBL",
                                             database = "org.Mm.eg.db")
  sample(rownames(GSE60361_subset_symbol), 50)
}

}
