% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{convertSCEToSeurat}
\alias{convertSCEToSeurat}
\title{convertSCEToSeurat
Converts sce object to seurat while retaining all assays and metadata}
\usage{
convertSCEToSeurat(
  inSCE,
  countsAssay = NULL,
  normAssay = NULL,
  scaledAssay = NULL
)
}
\arguments{
\item{inSCE}{A \code{SingleCellExperiment} object to convert to a Seurat object.}

\item{countsAssay}{Which assay to use from sce object for raw counts. Default \code{NULL}.}

\item{normAssay}{Which assay to use from sce object for normalized data. Default \code{NULL}.}

\item{scaledAssay}{Which assay to use from sce object for scaled data. Default \code{NULL}}
}
\value{
Updated seurat object that contains all data from the input sce object
}
\description{
convertSCEToSeurat
Converts sce object to seurat while retaining all assays and metadata
}
\examples{
data(scExample, package = "singleCellTK")
seurat <- convertSCEToSeurat(sce)
}
