% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{convertSeuratToSCE}
\alias{convertSeuratToSCE}
\title{convertSeuratToSCE
Converts the input seurat object to a sce object}
\usage{
convertSeuratToSCE(seuratObject)
}
\arguments{
\item{seuratObject}{input object}
}
\value{
inSCE output object
}
\description{
convertSeuratToSCE
Converts the input seurat object to a sce object
}
