% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{createSCE}
\alias{createSCE}
\title{Create a SingleCelltkExperiment object}
\usage{
createSCE(countfile = NULL, annotfile = NULL, featurefile = NULL,
  inputdataframes = FALSE)
}
\arguments{
\item{countfile}{The path to a text file that contains a header row of sample
names, and rows of raw counts per gene for those samples.}

\item{annotfile}{The path to a text file that contains columns of annotation
information for each sample in the countfile. This file should have the same
number of rows as there are columns in the countfile.}

\item{featurefile}{The path to a text file that contains columns of
annotation information for each gene in the count matrix. This file should
have the same genes in the same order as countfile. This is optional.}

\item{inputdataframes}{If TRUE, countfile and annotfile are read as data
frames instead of file paths. The default is FALSE.}
}
\value{
a SingleCelltkExperiment object
}
\description{
From a file of counts and a file of annotation information, create a
SingleCelltkExperiment object.
}
\examples{
\dontrun{
GSE60361_sce <- createSCE(countfile = "/path/to/input_counts.txt",
                          annotfile = "/path/to/input_annots.txt")
}
}
