% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSCEHeatmap.R
\name{dataAnnotationColor}
\alias{dataAnnotationColor}
\title{Generate distinct colors for all categorical col/rowData entries.}
\usage{
dataAnnotationColor(inSCE, axis = NULL, colorGen = grDevices::rainbow)
}
\arguments{
\item{inSCE}{SingleCellExperiment object.}

\item{axis}{Choose from "col" or "row"}

\item{colorGen}{function, default `rainbow`. A function that generates color
code vector by giving an integer for the number of colors. Alternatively,
`celda::distinctColors`.}
}
\value{
An list object containing distinct colors mapped to all possible
categorical entries in row/colData.
}
\description{
Generate distinct colors for all categorical col/rowData entries.
}
\author{
Yichen Wang
}
