% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{.computeSignificantPC}
\alias{.computeSignificantPC}
\title{.computeSignificantPC
Computes the significant principal components from an input sce object (must contain pca slot) using stdev}
\usage{
.computeSignificantPC(inSCE)
}
\arguments{
\item{inSCE}{(sce) object with pca computed}
}
\value{
A numerical value indicating how many number of components are considered significant
}
\description{
.computeSignificantPC
Computes the significant principal components from an input sce object (must contain pca slot) using stdev
}
