% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{.updateAssaySCE}
\alias{.updateAssaySCE}
\title{.updateAssaySCE
Update/Modify/Add an assay in the provided sce object}
\usage{
.updateAssaySCE(
  inSCE,
  seuratObject,
  assaySlotSCE,
  seuratDataSlot,
  seuratAssaySlot = "RNA"
)
}
\arguments{
\item{inSCE}{(sce) object to which we have to add assay to (copy to)}

\item{seuratObject}{from which we have to copy the assay (copy from)}

\item{assaySlotSCE}{the counts slot from assay in sce object}

\item{assaySlotSeurat}{the counts slot from assay in seurat object}

\item{slotSeurat}{assay to retrieve from seurat object (default is 'RNA')}
}
\description{
.updateAssaySCE
Update/Modify/Add an assay in the provided sce object
}
