% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{.updateAssaySCE}
\alias{.updateAssaySCE}
\title{.updateAssaySCE
Update/Modify/Add an assay in the provided sce object}
\usage{
.updateAssaySCE(inSCE, geneNames, seuratObject, assaySlotSCE, assaySlotSeurat)
}
\arguments{
\item{inSCE}{(sce) object to which we have to add assay to (copy to)}

\item{geneNames}{for consistent formatting}

\item{seuratObject}{from which we have to copy the assay (copy from)}

\item{assaySlotSCE}{the assay slot in sce object}

\item{assaySlotSeurat}{the assay slot in seurat object}
}
\description{
.updateAssaySCE
Update/Modify/Add an assay in the provided sce object
}
