% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downsample.Matrix.R
\name{Downsample}
\alias{Downsample}
\title{Downsample Data}
\usage{
Downsample(datamatrix, newcounts = c(4, 16, 64, 256, 1024, 4096, 16384, 65536,
  262144), byBatch = FALSE, batch = NULL, iterations = 10)
}
\arguments{
\item{datamatrix}{}

\item{newcounts}{}

\item{byBatch}{}

\item{batch}{}

\item{iterations}{}
}
\value{
Downsampled matrix
}
\description{
Downsample Data
}

