% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSCEtoAnndata.R
\name{exportSCEtoAnnData}
\alias{exportSCEtoAnnData}
\title{Export a \link[SingleCellExperiment]{SingleCellExperiment} R object as 
Python annData object}
\usage{
exportSCEtoAnnData(
  sce,
  useAssay = "counts",
  outputDir = "./",
  sample = "sample",
  overwrite = TRUE,
  compression = c("None", "lzf", "gzip"),
  compressionOpts = NULL,
  forceDense = c("False", "True")
)
}
\arguments{
\item{sce}{\link[SingleCellExperiment]{SingleCellExperiment} R object to be
 exported.
 @param useAssay Character, default `"counts"`. The name of assay of
interests that will be set as the primary matrix of the output AnnData.}

\item{outputDir}{Path to the directory where .h5ad outputs will be written}

\item{overwrite}{Boolean. Default \code{TRUE}.}

\item{compression}{Default \code{None}.If output file compression is required, this variable accepts
'gzip' or 'lzf' as inputs.}

\item{forceDense}{Default \code{False} Write sparse data as a dense matrix.
Refer anndata.write_h5ad documentation for details}

\item{compression_opts}{Integer. Default \code{NULL} Sets the compression level}
}
\description{
Writes all assays, colData, rowData, reducedDims, and altExps objects in a
\link[SingleCellExperiment]{SingleCellExperiment} to a Python annData object in the .h5ad format
All parameters of Anndata.write_h5ad function (https://icb-anndata.readthedocs-hosted.com/en/stable/anndata.AnnData.write_h5ad.html)
are available as parameters to this export function and set to defaults. Defaults can be
overridden at function call.
}
\examples{
data(sce_chcl, package = "scds")
exportSCEtoAnnData(sce=sce_chcl, compression="gzip")

}
