% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctkTagging.R
\name{expSetDataTag}
\alias{expSetDataTag}
\title{expSetDataTag
Set tag to an assay or a data item in the input SCE object.}
\usage{
expSetDataTag(inSCE, assayType, assays)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{assayType}{Specify a \code{character(1)} value as a tag that should be set against a data item.}

\item{assays}{Specify name(s) \code{character()} of data item(s) against which the tag should be set.}
}
\value{
The input \code{SingleCellExperiment} object with tag information stored in the metadata slot.
}
\description{
expSetDataTag
Set tag to an assay or a data item in the input SCE object.
}
\examples{
data(scExample, package = "singleCellTK")
sce <- expSetDataTag(sce, "raw", "counts")
}
