% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctkTagging.R
\name{expTaggedData}
\alias{expTaggedData}
\title{expTaggedData
Returns a list of names of data items from the input \code{SingleCellExperiment} object based upon the input parameters.}
\usage{
expTaggedData(inSCE, tags = NULL, redDims = FALSE)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{tags}{A \code{character()} value indicating if the data items should be returned separated by the specified tags. Default is \code{NULL} indicating that returned names of the data items are simply returned as a list with default tag as "uncategorized".}

\item{redDims}{A \code{logical} value indicating if \code{reducedDims} should be returned as well separated with 'redDims' tag.}
}
\value{
A \code{list} of names of data items specfied by the other parameters.
}
\description{
expTaggedData
Returns a list of names of data items from the input \code{SingleCellExperiment} object based upon the input parameters.
}
