% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{filterSCData}
\alias{filterSCData}
\title{Filter Genes and Samples from a Single Cell Object}
\usage{
filterSCData(inSCE, useAssay = "counts", deletesamples = NULL,
  removeNoExpress = TRUE, removeBottom = 0.5, minimumDetectGenes = 1700,
  filterSpike = TRUE)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object. Required}

\item{useAssay}{Indicate which assay to use for filtering. Default is
"counts"}

\item{deletesamples}{List of samples to delete from the object.}

\item{removeNoExpress}{Remove genes that have no expression across all
samples. The default is true}

\item{removeBottom}{Fraction of low expression genes to remove from the
single cell object. This occurs after removeNoExpress. The default is 0.50.}

\item{minimumDetectGenes}{Minimum number of genes with at least 1
count to include a sample in the single cell object. The default is 1700.}

\item{filterSpike}{Apply filtering to Spike in controls (indicated by
isSpike).
The default is TRUE.}
}
\value{
The filtered single cell object.
}
\description{
Filter Genes and Samples from a Single Cell Object
}
\examples{
data("mouseBrainSubsetSCE")
mouseBrainSubsetSCE <- filterSCData(mouseBrainSubsetSCE,
                                    deletesamples="X1772063061_G11")
}
