% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarker.R
\name{findMarkerDiffExp}
\alias{findMarkerDiffExp}
\title{Find the marker gene set for each cluster}
\usage{
findMarkerDiffExp(
  inSCE,
  useAssay = "logcounts",
  method = c("wilcox", "MAST", "DESeq2", "Limma", "ANOVA"),
  cluster = "cluster",
  covariates = NULL,
  log2fcThreshold = NULL,
  fdrThreshold = 0.05,
  minClustExprPerc = NULL,
  maxCtrlExprPerc = NULL,
  minMeanExpr = NULL
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object.}

\item{useAssay}{character. A string specifying which assay to use for the
MAST calculations. Default \code{"logcounts"}.}

\item{method}{A single character for specific differential expression
analysis method. Choose from \code{'wilcox'}, \code{'MAST'}, \code{'DESeq2'},
\code{'Limma'}, and \code{'ANOVA'}. Default \code{"wilcox"}.}

\item{cluster}{One single character to specify a column in
\code{colData(inSCE)} for the clustering label. Alternatively, a vector or
a factor is also acceptable. Default \code{"cluster"}.}

\item{covariates}{A character vector of additional covariates to use when
building the model. All covariates must exist in
\code{names(colData(inSCE))}. Not applicable when \code{method} is
\code{"MAST"} method. Default \code{NULL}.}

\item{log2fcThreshold}{Only out put DEGs with the absolute values of log2FC
larger than this value. Default \code{NULL}}

\item{fdrThreshold}{Only out put DEGs with FDR value smaller than this
value. Default \code{NULL}}

\item{minClustExprPerc}{A numeric scalar. The minimum cutoff of the
percentage of cells in the cluster of interests that expressed the marker
gene. From 0 to 1. Default \code{NULL}.}

\item{maxCtrlExprPerc}{A numeric scalar. The maximum cutoff of the
percentage of cells out of the cluster (control group) that expressed the
marker gene. From 0 to 1. Default \code{NULL}.}

\item{minMeanExpr}{A numeric scalar. The minimum cutoff of the mean
expression value of the marker in the cluster of interests. Default
\code{NULL}.}
}
\value{
The input \linkS4class{SingleCellExperiment} object with
\code{metadata(inSCE)$findMarker} updated with a data.table of the up-
regulated DEGs for each cluster.
}
\description{
With an input SingleCellExperiment object and specifying the
clustering labels, this function iteratively call the differential expression
analysis on each cluster against all the others.
}
\examples{
data("mouseBrainSubsetSCE", package = "singleCellTK")
mouseBrainSubsetSCE <- findMarkerDiffExp(mouseBrainSubsetSCE,
                                         useAssay = "logcounts",
                                         cluster = "level1class")
}
