% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiomarker.R
\name{getBiomarker}
\alias{getBiomarker}
\title{get Biomarker}
\usage{
getBiomarker(count_data, gene, binary = "Binary")
}
\arguments{
\item{count_data}{A SCE object}

\item{gene}{gene list}

\item{binary}{binary or gradient}
}
\value{
A PCA plot
}
\description{
Use this function to get expression or binary data of gene list
}
