% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiomarker.R
\name{getBiomarker}
\alias{getBiomarker}
\title{getBiomarker}
\usage{
getBiomarker(count_data, gene, binary = "Binary")
}
\arguments{
\item{count_data}{A SCtkExperiment object}

\item{gene}{gene list}

\item{binary}{"Binary" for binary expression or "Continuous" for a gradient.
Default: "Binary"}
}
\value{
A data.frame of expression values
}
\description{
Given a list of genes and a SCtkExperiment object, return the binary or
continuous expression of the genes.
}
