% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiomarker.R
\name{getBiomarker}
\alias{getBiomarker}
\title{Given a list of genes and a SingleCellExperiment object, return
the binary or continuous expression of the genes.}
\usage{
getBiomarker(inSCE, gene, binary = "Binary", useAssay = "counts")
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{gene}{gene list}

\item{binary}{"Binary" for binary expression or "Continuous" for a gradient.
Default: "Binary"}

\item{useAssay}{Indicate which assay to use. The default is "counts".}
}
\value{
getBiomarker(): A data.frame of expression values
}
\description{
Given a list of genes and a SingleCellExperiment object, return
the binary or continuous expression of the genes.
}
\examples{
getBiomarker(mouseBrainSubsetSCE, gene="C1qa")

}
