% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterInputData.R
\name{getClusterInputData}
\alias{getClusterInputData}
\title{Get data to use as input clustering algorithms}
\usage{
getClusterInputData(inSCE, inputData, useAssay = "logcounts",
  reducedDimName = NULL)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object. Required}

\item{inputData}{A string ("Raw Data", "PCA Components", "tSNE Components", "UMAP Components")}

\item{useAssay}{Indicate which assay to use for PCA. Default is "logcounts"}

\item{reducedDimName}{If clustering on PCA, t-SNE or UMAP data, dimension name.
The toolkit will store data with the pattern <ASSAY>_<ALGORITHM>.}
}
\value{
Cluster input data
}
\description{
Get data to use as input clustering algorithms
}
\examples{
data("mouseBrainSubsetSCE")
getClusterInputData(mouseBrainSubsetSCE, "PCA Components",
                    useAssay = "logcounts", reducedDimName = "PCA_logcounts")

}
