% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichRSCE.R
\name{getEnrichRResult<-}
\alias{getEnrichRResult<-}
\alias{getEnrichRResult}
\alias{getEnrichRResult,SingleCellExperiment-method}
\alias{getEnrichRResult<-,SingleCellExperiment-method}
\title{Get or Set EnrichR Result}
\usage{
getEnrichRResult(inSCE) <- value

getEnrichRResult(inSCE)

\S4method{getEnrichRResult}{SingleCellExperiment}(inSCE)

\S4method{getEnrichRResult}{SingleCellExperiment}(inSCE) <- value
}
\arguments{
\item{inSCE}{A \linkS4class{SingleCellExperiment} object.}

\item{value}{The EnrichR result table}
}
\value{
For getter method, a data.frame of the EnrichR result;
For setter method, \code{inSCE} with EnrichR results updated.
}
\description{
Get or Set EnrichR Result
}
\examples{
data("mouseBrainSubsetSCE")
mouseBrainSubsetSCE <- runEnrichR(mouseBrainSubsetSCE, glist = "Cmtm5", 
                                  db = "GO_Cellular_Component_2017")
result <- getEnrichRResult(mouseBrainSubsetSCE)
}
\seealso{
\code{\link{runEnrichR}}
}
