% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{getGenesetNamesFromCollection}
\alias{getGenesetNamesFromCollection}
\title{List geneset names from geneSetCollection}
\usage{
getGenesetNamesFromCollection(inSCE, geneSetCollectionName)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{geneSetCollectionName}{The name of an imported geneSetCollection.}
}
\value{
A character vector of available genesets from the collection.
}
\description{
List geneset names from geneSetCollection
}
\seealso{
Other accessor functions: 
\code{\link{getBiomarker}()},
\code{\link{getDEGTopTable}()},
\code{\link{getFindMarkerTopTable}()},
\code{\link{getMSigDBTable}()},
\code{\link{getPathwayResultNames}()},
\code{\link{getSceParams}()},
\code{\link{getSeuratVariableFeatures}()},
\code{\link{getTopHVG}()},
\code{\link{sctkListGeneSetCollections}()},
\code{\link{setRowNames}()},
\code{\link{setSCTKDisplayRow}()}
}
\concept{accessor functions}
