% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importGeneSets.R
\name{getMSigDBTable}
\alias{getMSigDBTable}
\title{Shows MSigDB categories}
\usage{
getMSigDBTable()
}
\value{
data.frame, containing MSigDB categories
}
\description{
Returns a data.frame that shows MSigDB categories and
subcategories as well as descriptions for each. The entries in the ID
column in this table can be used as input for \link{importGeneSetsFromMSigDB}.
}
\seealso{
\link{importGeneSetsFromMSigDB} for importing MSigDB gene sets.
}
\author{
Joshua D. Campbell
}
