% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCA.R
\name{getPCA}
\alias{getPCA}
\title{getPCA
A wrapper to \link[scater]{runPCA} function to compute principal component analysis (PCA) from a given \code{SingleCellExperiment} object.}
\usage{
getPCA(
  inSCE,
  useAssay = "logcounts",
  reducedDimName = "PCA",
  ndim = 50,
  ntop = 500
)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{useAssay}{Assay to use for PCA computation.}

\item{reducedDimName}{Name to use for the reduced output assay.}

\item{ndim}{Number of principal components to obtain from the PCA computation. Default \code{50}.}

\item{ntop}{Number of top features to use with PCA. Default \code{500}.}
}
\value{

}
\description{
getPCA
A wrapper to \link[scater]{runPCA} function to compute principal component analysis (PCA) from a given \code{SingleCellExperiment} object.
}
\examples{
data(sce_chcl, package = "scds")
sce_chcl <- getPCA(sce_chcl, useAssay = "logcounts")
}
