% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCA.R
\name{getPCA}
\alias{getPCA}
\title{Get PCA components for a feature counts table}
\usage{
getPCA(count_data)
}
\arguments{
\item{count_data}{Matrix of expresion values}
}
\value{
A reduced dimension object
}
\description{
Selects the 500 most variable genes in the feature count, performs
PCA based on them and outputs the principal components in a data frame
and their variances as percentVar attribute
}
