% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCA.R
\name{getPCA}
\alias{getPCA}
\title{Get and plot PCA components for a SingleCellExperiment object}
\usage{
getPCA(inSCE, useAssay = "logcounts", reducedDimName = "PCA")
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{useAssay}{Indicate which assay to use for PCA. Default is "counts"}

\item{reducedDimName}{Store the PCA data with this name. The default is PCA.
The toolkit will store data with the pattern <ASSAY>_<ALGORITHM>.}
}
\value{
A \linkS4class{SingleCellExperiment} object with the specified
reducedDim
}
\description{
Selects the 500 most variable genes in the SCE, performs
PCA based on them and stores the values in the reducedDims slot of
the SCE object.
}
\examples{
data("mouseBrainSubsetSCE")
#add a CPM assay
assay(mouseBrainSubsetSCE, "cpm") <- apply(assay(mouseBrainSubsetSCE,
                                                 "counts"),
                                           2, function(x) {
                                             x / (sum(x) / 1000000)
                                           })
mouseBrainSubsetSCE <- getPCA(mouseBrainSubsetSCE,
                              useAssay = "cpm",
                              reducedDimName = "PCA_cpm")
reducedDims(mouseBrainSubsetSCE)

}
