% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCA.R
\name{getPCA}
\alias{getPCA}
\title{Get PCA components for a SCtkE object}
\usage{
getPCA(countData, useAssay = "logcounts", reducedDimName = "PCA")
}
\arguments{
\item{countData}{SCtkE object}

\item{useAssay}{Indicate which assay to use for PCA. Default is "counts"}

\item{reducedDimName}{Store the PCA data with this name. The default is PCA.
The toolkit will store data with the pattern <ASSAY>_<ALGORITHM>.}
}
\value{
A SCtkE object with the specified reduecedDim and pcaVariances
updated
}
\description{
Selects the 500 most variable genes in the SCE, performs
PCA based on them and stores the PCA values in the reducedDims slot of the
SCE object.
}
\examples{
data("mouseBrainSubsetSCE")
#add a CPM assay
assay(mouseBrainSubsetSCE, "cpm") <- apply(assay(mouseBrainSubsetSCE,
                                                 "counts"),
                                           2, function(x) {
                                             x / (sum(x) / 1000000)
                                           })
mouseBrainSubsetSCE <- getPCA(mouseBrainSubsetSCE,
                              useAssay = "cpm",
                              reducedDimName = "PCA_cpm")
reducedDims(mouseBrainSubsetSCE)

}
