% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCA.R
\name{getPCA}
\alias{getPCA}
\title{Get PCA components for a SCtkE object}
\usage{
getPCA(count_data, use_assay = "counts")
}
\arguments{
\item{count_data}{SCtkE object}

\item{use_assay}{Indicate which assay to use for PCA. Default is "counts"}
}
\value{
A SCtkE object with reducedDim "PCA" and pca_variances updated
}
\description{
Selects the 500 most variable genes in the SCE, performs
PCA based on them and stores the PCA values in the reducedDims slot of the
SCE object.
}
