% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCA.R
\name{getPCA}
\alias{getPCA}
\title{Perform PCA on a SingleCellExperiment Object
A wrapper to \link[scater]{runPCA} function to compute principal component
analysis (PCA) from a given \linkS4class{SingleCellExperiment} object.}
\usage{
getPCA(
  inSCE,
  useAssay = "logcounts",
  useAltExp = NULL,
  reducedDimName = "PCA",
  ndim = 50,
  scale = TRUE,
  ntop = NULL
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{useAssay}{Assay to use for PCA computation. If \code{useAltExp} is
specified, \code{useAssay} has to exist in
\code{assays(altExp(inSCE, useAltExp))}. Default \code{"logcounts"}}

\item{useAltExp}{The subset to use for PCA computation, usually for the
selected.variable features. Default \code{NULL}.}

\item{reducedDimName}{Name to use for the reduced output assay. Default
\code{"PCA"}.}

\item{ndim}{Number of principal components to obtain from the PCA
computation. Default \code{50}.}

\item{scale}{Logical scalar, whether to standardize the expression values.
Default \code{TRUE}.}

\item{ntop}{Number of top features to use as a further variable feature
selection. Default \code{NULL}.}
}
\value{
A \linkS4class{SingleCellExperiment} object with PCA computation
updated in \code{reducedDim(inSCE, reducedDimName)}.
}
\description{
Perform PCA on a SingleCellExperiment Object
A wrapper to \link[scater]{runPCA} function to compute principal component
analysis (PCA) from a given \linkS4class{SingleCellExperiment} object.
}
\examples{
data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
logcounts(sce) <- log(counts(sce) + 1)
sce <- getPCA(sce, ntop = 500)
}
