% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{getSeuratVariableFeatures}
\alias{getSeuratVariableFeatures}
\title{Get variable feature names after running runSeuratFindHVG function}
\usage{
getSeuratVariableFeatures(inSCE)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}
}
\value{
A list of variable feature names.
}
\description{
Get variable feature names after running runSeuratFindHVG function
}
