% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopHVG.R
\name{getTopHVG}
\alias{getTopHVG}
\alias{setTopHVG}
\title{Get or set top HVG after calculation}
\usage{
getTopHVG(
  inSCE,
  method = c("vst", "dispersion", "mean.var.plot", "modelGeneVar", "seurat",
    "seurat_v3", "cell_ranger"),
  hvgNumber = 2000,
  useFeatureSubset = NULL,
  featureDisplay = metadata(inSCE)$featureDisplay
)

setTopHVG(
  inSCE,
  method = c("vst", "dispersion", "mean.var.plot", "modelGeneVar", "seurat",
    "seurat_v3", "cell_ranger"),
  hvgNumber = 2000,
  featureSubsetName = NULL,
  genes = NULL,
  genesBy = NULL,
  altExp = FALSE
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object}

\item{method}{Specify which method to use for variable gene extraction
from Seurat \code{"vst"}, \code{"mean.var.plot"}, \code{"dispersion"} or
Scran \code{"modelGeneVar"} or Scanpy \code{"seurat"}, \code{"cell_ranger"}, 
\code{"seurat_v3"}. Default \code{"vst"}}

\item{hvgNumber}{Specify the number of top variable genes to extract.}

\item{useFeatureSubset}{Get the feature names in the HVG list set by
\code{setTopHVG}. Will ignore \code{method} and \code{hvgNumber} if not
\code{NULL}. Default \code{NULL}.}

\item{featureDisplay}{A character string for the \code{rowData} variable name
to indicate what type of feature ID should be displayed. If set by
\code{\link{setSCTKDisplayRow}}, will by default use it. If \code{NULL}, will
use \code{rownames(inSCE)}.}

\item{featureSubsetName}{A character string for the \code{rowData} variable
name to store a logical index of selected features. Default \code{NULL}, will
be determined basing on other parameters.}

\item{genes}{A customized character vector of gene list to be set as a
\code{rowData} variable. Will ignore \code{method} and \code{hvgNumber} if
set. Default \code{NULL}.}

\item{genesBy}{If setting customized \code{genes}, where should it be found
in \code{rowData}? Leave \code{NULL} for matching \code{rownames}. Default
\code{NULL}.}

\item{altExp}{\code{TRUE} for also creating a subset \code{inSCE} object with
the selected HVGs and store this subset in the \code{altExps} slot, named by
\code{hvgListName}. Default \code{FALSE}.}
}
\value{
\item{getTopHVG}{A character vector of the top \code{hvgNumber} variable
feature names}
\item{setTopHVG}{The input \code{inSCE} object with the logical vector of
HVG selection updated in \code{rowData}, and related parameter updated in
\code{metadata}. If \code{altExp} is \code{TRUE}, an \code{altExp} is also
added}
}
\description{
Extracts or select the top variable genes from an input
\linkS4class{SingleCellExperiment} object. Note that the variability metrics
must be computed using the \code{runFeatureSelection} method before
extracting the feature names of the top variable features. \code{getTopHVG}
only returns a character vector of the HVG selection, while with
\code{setTopHVG}, a logical vector of the selection will be saved in the
\code{rowData}, and optionally, a subset object for the HVGs can be stored
in the \code{altExps} slot at the same time.
}
\examples{
data("scExample", package = "singleCellTK")
sce <- runSeuratFindHVG(sce)
hvgs <- getTopHVG(sce, hvgNumber = 10)
sce <- setTopHVG(sce, method = "vst", hvgNumber = 5)
}
\seealso{
\code{\link{runFeatureSelection}}, \code{\link{runSeuratFindHVG}},
\code{\link{runModelGeneVar}}, \code{\link{plotTopHVG}}
}
\author{
Irzam Sarfraz, Yichen Wang
}
