% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopHVG.R
\name{getTopHVG}
\alias{getTopHVG}
\title{getTopHVG
Extracts the top variable genes from an input singleCellExperiment object}
\usage{
getTopHVG(inSCE, method, n = 2000)
}
\arguments{
\item{inSCE}{an input singleCellExperiment object}

\item{method}{represents which method to use for variable gene extraction
from either Seurat "vst", "mean.var.plot", "dispersion" or Scran
"modelGeneVar"}

\item{n}{number of top variable genes to extract}
}
\value{
list of top variable gene names
}
\description{
getTopHVG
Extracts the top variable genes from an input singleCellExperiment object
}
\examples{
data(sce_chcl, package = "scds")
sce_chcl <- scran_modelGeneVar(sce_chcl, "counts")
# return top 10 variable genes
topGenes <- getTopHVG(sce_chcl, "modelGeneVar", 10)
}
\author{
Irzam Sarfraz
}
