% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSCAN.R
\name{getTSCANResults}
\alias{getTSCANResults}
\alias{getTSCANResults,SingleCellExperiment-method}
\alias{getTSCANResults<-}
\alias{getTSCANResults<-,SingleCellExperiment-method}
\alias{listTSCANResults}
\alias{listTSCANResults,SingleCellExperiment-method}
\title{getTSCANResults accessor function}
\usage{
getTSCANResults(x, analysisName = NULL, pathName = NULL)

\S4method{getTSCANResults}{SingleCellExperiment}(x, analysisName = NULL, pathName = NULL)

getTSCANResults(x, analysisName, pathName = NULL) <- value

\S4method{getTSCANResults}{SingleCellExperiment}(x, analysisName, pathName = NULL) <- value

listTSCANResults(x)

\S4method{listTSCANResults}{SingleCellExperiment}(x)
}
\arguments{
\item{x}{Input \linkS4class{SingleCellExperiment} object.}

\item{analysisName}{Algorithm name implemented}

\item{pathName}{Sub folder name within the \code{analysisName}}

\item{value}{Value to be stored within the \code{pathName} or 
\code{analysisName}}
}
\value{
Get or set TSCAN results
}
\description{
getTSCANResults accessor function
}
