% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTSNE.R
\name{getTSNE}
\alias{getTSNE}
\title{Get t-SNE components for a feature counts table
  
Selects the 500 most variable genes in the feature count, performs
t-SNE based on them and outputs the principal components in a data frame
and their variances as percentVar attribute}
\usage{
getTSNE(count_data)
}
\arguments{
\item{count_data}{Matrix of expresion values}
}
\value{
A reduced dimension object
}
\description{
Get t-SNE components for a feature counts table
  
Selects the 500 most variable genes in the feature count, performs
t-SNE based on them and outputs the principal components in a data frame
and their variances as percentVar attribute
}
