% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTSNE.R
\name{getTSNE}
\alias{getTSNE}
\title{Run t-SNE dimensionality reduction method on the assay data.}
\usage{
getTSNE(
  inSCE,
  useAssay = "logcounts",
  reducedDimName = "TSNE",
  n_iterations = 1000,
  perplexity = NULL
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{useAssay}{Indicate which assay to use. The default is "logcounts".}

\item{reducedDimName}{a name to store the results of the dimension reductions}

\item{n_iterations}{maximum iterations. Default is 1000}

\item{perplexity}{perplexity parameter. Default is 5}
}
\value{
A \linkS4class{SingleCellExperiment} object with the specified
reducedDim updated
}
\description{
Run t-SNE dimensionality reduction method on the assay data.
}
\examples{
data("mouseBrainSubsetSCE")
#add a CPM assay
assay(mouseBrainSubsetSCE, "cpm") <- apply(
  assay(mouseBrainSubsetSCE, "counts"), 2, function(x) {
    x / (sum(x) / 1000000)
  })
mouseBrainSubsetSCE <- getTSNE(mouseBrainSubsetSCE, useAssay = "cpm",
                               reducedDimName = "TSNE_cpm")
reducedDims(mouseBrainSubsetSCE)

}
