% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTSNE.R
\name{getTSNE}
\alias{getTSNE}
\title{Get t-SNE components for a SCE object}
\usage{
getTSNE(count_data, use_assay = "counts")
}
\arguments{
\item{count_data}{SCE object}

\item{use_assay}{Indicate which assay to use for PCA. Default is "counts"}
}
\value{
A SCE object with reducedDim "TSNE" updated
}
\description{
Selects the 500 most variable genes in the feature count, performs
t-SNE based on them and stores the TSNE values in the reducedDims slot of the
SCE object.
}
