% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTSNE.R
\name{getTSNE}
\alias{getTSNE}
\title{Get t-SNE components for a SCE object}
\usage{
getTSNE(countData, useAssay = "logcounts", reducedDimName = "TSNE")
}
\arguments{
\item{countData}{SCE object}

\item{useAssay}{Indicate which assay to use for PCA. Default is "counts"}

\item{reducedDimName}{Store the t-SNE data with this name. The default is
TSNE. The toolkit will store data with the pattern <ASSAY>_<ALGORITHM>.}
}
\value{
A SCE object with the specified reducedDim updated
}
\description{
Selects the 500 most variable genes in the feature count, performs
t-SNE based on them and stores the t-SNE values in the reducedDims slot of the
SCE object.
}
\examples{
data("mouseBrainSubsetSCE")
#add a CPM assay
assay(mouseBrainSubsetSCE, "cpm") <- apply(
  assay(mouseBrainSubsetSCE, "counts"), 2, function(x) {
    x / (sum(x) / 1000000)
  })
mouseBrainSubsetSCE <- getTSNE(mouseBrainSubsetSCE, useAssay = "cpm",
                               reducedDimName = "TSNE_cpm")
reducedDims(mouseBrainSubsetSCE)

}
