% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTSNE.R
\name{getTSNE}
\alias{getTSNE}
\title{Run t-SNE dimensionality reduction method on a SingleCellExperiment Object}
\usage{
getTSNE(
  inSCE,
  useAssay = "logcounts",
  useAltExp = NULL,
  useReducedDim = NULL,
  reducedDimName = "TSNE",
  nIterations = 1000,
  perplexity = 30,
  run_pca = TRUE,
  ntop = NULL,
  seed = NULL
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{useAssay}{Assay to use for tSNE computation. If \code{useAltExp} is
specified, \code{useAssay} has to exist in
\code{assays(altExp(inSCE, useAltExp))}. Default \code{"logcounts"}}

\item{useAltExp}{The subset to use for tSNE computation, usually for the
selected.variable features. Default \code{NULL}.}

\item{useReducedDim}{The low dimension representation to use for UMAP
computation. Default \code{NULL}.}

\item{reducedDimName}{a name to store the results of the dimension
reductions. Default \code{"TSNE"}.}

\item{nIterations}{maximum iterations. Default \code{1000}.}

\item{perplexity}{perplexity parameter. Default \code{30}.}

\item{run_pca}{run tSNE on PCA components? Default \code{TRUE}.}

\item{ntop}{Number of top features to use as a further variable feature
selection. Default \code{NULL}.}

\item{seed}{Random seed for reproducibility of tSNE results. 
Default \code{NULL} will use global seed in use by the R environment.}
}
\value{
A \linkS4class{SingleCellExperiment} object with tSNE computation
updated in \code{reducedDim(inSCE, reducedDimName)}.
}
\description{
Run t-SNE dimensionality reduction method on a SingleCellExperiment Object
}
\examples{
data("mouseBrainSubsetSCE")
#add a CPM assay
assay(mouseBrainSubsetSCE, "cpm") <- apply(
  assay(mouseBrainSubsetSCE, "counts"), 2, function(x) {
    x / (sum(x) / 1000000)
  })
mouseBrainSubsetSCE <- getTSNE(mouseBrainSubsetSCE, useAssay = "cpm",
                               reducedDimName = "TSNE_cpm",
                               perplexity = NULL)
}
