% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSVA.R
\name{gsvaSCE}
\alias{gsvaSCE}
\alias{gsvaPlot}
\title{Run GSVA analysis on a \linkS4class{SingleCellExperiment} object}
\usage{
gsvaSCE(inSCE, useAssay = "logcounts", pathwaySource, pathwayNames, ...)

gsvaPlot(
  inSCE,
  gsvaData,
  plotType,
  condition = NULL,
  show_column_names = TRUE,
  show_row_names = TRUE,
  text_size = 12
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{useAssay}{Indicate which assay to use. The default is "logcounts"}

\item{pathwaySource}{The pathway source if "Manual Input", the pathwayNames
should be rowData annotations that are (0,1) vectors. If, "MSigDB c2 (Human,
Entrez ID only)", the pathwayNames should be pathways from MSigDB c2 or "ALL"
to run on all available pathways.}

\item{pathwayNames}{List of pathway names to run, depending on pathwaySource
parameter.}

\item{...}{Parameters to pass to gsva()}

\item{gsvaData}{GSVA data to plot. Required.}

\item{plotType}{The type of plot to use, "Violin" or "Heatmap". Required.}

\item{condition}{The condition(s) to use for the Violin plot, or the
condition(s) to add as color bars above the Heatmap. Required for Violin,
optional for Heatmap.}

\item{show_column_names}{Display the column labels on the heatmap. The
default is TRUE}

\item{show_row_names}{Display the row labels on the heatmap. The default
is TRUE.}

\item{text_size}{Text size for plots. The default is 12}
}
\value{
gsvaSCE(): A data.frame of pathway activity scores from GSVA.

gsvaPlot(): The requested plot of the GSVA results.
}
\description{
Run GSVA analysis on a \linkS4class{SingleCellExperiment} object
}
\section{Functions}{
\itemize{
\item \code{gsvaPlot}: Plot GSVA results.

Plot GSVA Results
}}

