% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCellRanger.R
\name{importCellRanger}
\alias{importCellRanger}
\alias{importCellRangerV2}
\alias{importCellRangerV3}
\title{Construct SCE object from Cell Ranger output}
\usage{
importCellRanger(
  cellRangerDirs = NULL,
  samples = NULL,
  cellRangerOuts = "outs/filtered_feature_bc_matrix/",
  matrixFileName = "matrix.mtx.gz",
  featuresFileName = "features.tsv.gz",
  barcodesFileName = "barcodes.tsv.gz",
  gzipped = TRUE,
  class = "DelayedArray"
)

importCellRangerV2(
  cellRangerDirs = NULL,
  samples = NULL,
  matrixFileName = "matrix.mtx",
  featuresFileName = "genes.tsv",
  barcodesFileName = "barcodes.tsv",
  gzipped = TRUE,
  class = "DelayedArray"
)

importCellRangerV3(
  cellRangerDirs = NULL,
  samples = NULL,
  matrixFileName = "matrix.mtx.gz",
  featuresFileName = "features.tsv.gz",
  barcodesFileName = "barcodes.tsv.gz",
  gzipped = TRUE,
  class = "DelayedArray"
)
}
\arguments{
\item{cellRangerDirs}{The root directories where Cell Ranger was run. These
folders should contain sample specific folders. Default \code{NULL},
meaning the paths for each sample will be specified in \emph{samples}
argument.}

\item{samples}{Default \code{NULL}. Can be one of
\itemize{
  \item \code{NULL}. All samples within \emph{cellRangerDirs} will be
   imported. The order of samples will be first determined by the order of
   \emph{cellRangerDirs} and then by \link[base]{list.files}. This is only
   for the case where \emph{cellRangerDirs} is specified.
  \item A list of vectors containing sample names to import. Each vector in
   the list corresponds to samples from one of \emph{cellRangerDirs}.
   These names are the same as the folder names under \emph{cellRangerDirs}.
   This is only for the case where \emph{cellRangerDirs} is specified.
  \item A vector of folder paths for the samples to import. This is only for
   the case where \emph{cellRangerDirs} is \code{NULL}.
}
The cells in the final SCE object will be ordered in the same order of
samples.}

\item{cellRangerOuts}{Character. It is the intermediate
path to filtered or raw feature count file saved in sparse matrix format
for each of \emph{samples}.
Reference genome names might need to be
appended for CellRanger version below 3.0.0 if reads were mapped to
multiple genomes when running Cell Ranger pipeline.}

\item{matrixFileName}{Filename for the Market Exchange Format (MEX) sparse
matrix file (.mtx file).}

\item{featuresFileName}{Filename for the feature annotation file. It can be
\emph{features.tsv.gz} or \emph{genes.tsv}.}

\item{barcodesFileName}{Filename for the cell barcode list file.}

\item{gzipped}{Boolean. \code{TRUE} if the Cell Ranger output files
(barcodes.tsv, features.tsv, and matrix.mtx) were
gzip compressed. \code{FALSE} otherwise. This is true after Cell Ranger
3.0.0 update. Default \code{TRUE}.}

\item{class}{Character. The class of the expression matrix stored in the SCE
object. Can be one of "DelayedArray" (as returned by
\link[DelayedArray]{DelayedArray} function), "Matrix" (as returned by
\link[Matrix]{readMM} function), or "matrix" (as returned by
\link[base]{matrix} function). Default "DelayedArray".}
}
\value{
A \code{SingleCellExperiment} object containing the combined count
 matrix, the feature annotations, and the cell annotation.
}
\description{
Read the filtered barcodes, features, and matrices for all
 samples from (preferably a single run of) Cell Ranger output. Import and
 combine them
 as one big \link[SingleCellExperiment]{SingleCellExperiment} object.
}
\details{
\code{importCellRangerV2} imports output from Cell Ranger V2.
 \code{importCellRangerV3} imports output from Cell Ranger V3. Some implicit
 assumptions are made in these two functions including \code{cellRangerOuts},
 \code{matrixFileName}, \code{featuresFileName}, and \code{barcodesFileName}.
 Alternatively, user can call \code{importCellRanger} to explicitly
 specify \code{cellRangerOuts}, \code{matrixFileName},
 \code{featuresFileName}, and \code{barcodesFileName}.
}
\examples{
# Example #1
# The following filtered feature, cell, and matrix files were downloaded from
# https://support.10xgenomics.com/single-cell-gene-expression/datasets/
# 3.0.0/hgmm_1k_v3
# The top 10 hg19 & mm10 genes are included in this example.
# Only the first 20 cells are included.
sce <- importCellRanger(
    cellRangerDirs = system.file("extdata", package = "singleCellTK"),
    samples = "hgmm_1k_v3_20x20")
sce <- importCellRangerV3(
    cellRangerDirs = system.file("extdata", package = "singleCellTK"),
    samples = "hgmm_1k_v3_20x20")
}
