% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCellRanger.R
\name{importCellRangerV2Sample}
\alias{importCellRangerV2Sample}
\title{Construct SCE object from Cell Ranger V2 output for a single sample}
\usage{
importCellRangerV2Sample(
  dataDir = NULL,
  sampleName = NULL,
  class = c("Matrix", "matrix"),
  delayedArray = TRUE
)
}
\arguments{
\item{dataDir}{A path to the directory containing the data files. Default "./".}

\item{sampleName}{A User-defined sample name. This will be prepended to all cell barcode IDs.
Default "sample".}

\item{class}{Character. The class of the expression matrix stored in the SCE
object. Can be one of "Matrix" (as returned by
\link[Matrix]{readMM} function), or "matrix" (as returned by
\link[base]{matrix} function). Default "Matrix".}

\item{delayedArray}{Boolean. Whether to read the expression matrix as
\link[DelayedArray]{DelayedArray} object or not. Default \code{TRUE}.}
}
\value{
A \code{SingleCellExperiment} object containing the count
 matrix, the feature annotations, and the cell annotation for the sample.
}
\description{
Read the filtered barcodes, features, and matrices for all
 samples from Cell Ranger V2 output. Files are assumed to be named
 "matrix.mtx", "genes.tsv", and "barcodes.tsv".
}
\examples{
sce <- importCellRangerV2Sample(
    dataDir = system.file("extdata/pbmc_4k_v2_20x20/outs/",
        "filtered_gene_bc_matrices/GRCh38", package = "singleCellTK"),
    sampleName = "pbmc4k_20")
}
