% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFromFiles.R
\name{importFromFiles}
\alias{importFromFiles}
\title{Create a SingleCellExperiment object from files}
\usage{
importFromFiles(
  assayFile,
  annotFile = NULL,
  featureFile = NULL,
  assayName = "counts",
  inputDataFrames = FALSE,
  class = c("Matrix", "matrix"),
  delayedArray = FALSE,
  annotFileHeader = FALSE,
  annotFileRowName = 1,
  annotFileSep = "\\t",
  featureHeader = FALSE,
  featureRowName = 1,
  featureSep = "\\t",
  gzipped = "auto"
)
}
\arguments{
\item{assayFile}{The path to a  file in .mtx, .txt, .csv, .tab, or .tsv format.}

\item{annotFile}{The path to a text file that contains columns of annotation
information for each sample in the assayFile. This file should have the same
number of rows as there are columns in the assayFile. If multiple samples are
represented in these files, this should be denoted by a column called \code{'sample'}
within the \code{annotFile}.}

\item{featureFile}{The path to a text file that contains columns of
annotation information for each gene in the count matrix. This file should
have the same genes in the same order as assayFile. This is optional.}

\item{assayName}{The name of the assay that you are uploading. The default
is "counts".}

\item{inputDataFrames}{If TRUE, assayFile and annotFile are read as data
frames instead of file paths. The default is FALSE.}

\item{class}{Character. The class of the expression matrix stored in the SCE
object. Can be one of "Matrix" (as returned by
\link[Matrix]{readMM} function), or "matrix" (as returned by
\link[base]{matrix} function). Default "Matrix".}

\item{delayedArray}{Boolean. Whether to read the expression matrix as
\link[DelayedArray]{DelayedArray} object or not. Default \code{TRUE}.}

\item{annotFileHeader}{Whether there's a header (colnames) in the cell annotation file. Default is FALSE}

\item{annotFileRowName}{Which column is used as the rownames for the cell annotation file. Default is 1 (first column).}

\item{annotFileSep}{Separater used for the cell annotation file. Default is "\\t".}

\item{featureHeader}{Whether there's a header (colnames) in the feature annotation file. Default is FALSE}

\item{featureRowName}{Which column is used as the rownames for the feature annotation file. Default is 1 (first column).}

\item{featureSep}{Separater used for the feature annotation file. Default is "\\t".}

\item{gzipped}{Whether the input file is gzipped. Default is "auto" and it will automatically detect whether the file is gzipped. Other options is TRUE or FALSE.}
}
\value{
a SingleCellExperiment object
}
\description{
Creates a SingleCellExperiment object from a counts file in various formats.
and a file of annotation information, .
}
