% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downsample.Matrix.R
\name{iterateSimulations}
\alias{iterateSimulations}
\title{Returns significance data from a snapshot.}
\usage{
iterateSimulations(originalData, realLabels, totalReads, cells, iterations)
}
\arguments{
\item{originalData}{TODO: document}

\item{realLabels}{TODO: document}

\item{totalReads}{TODO: document}

\item{cells}{TODO: document}

\item{iterations}{TODO: document}
}
\value{
A matrix of significance information from a snapshot
}
\description{
Returns significance data from a snapshot.
}
\examples{
data("GSE60361_subset_sce")
res <- iterateSimulations(GSE60361_subset_sce, realLabels = "level1class",
                          totalReads = 1000, cells = 10, iterations = 2)

}
