% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downsample.Matrix.R
\name{iterateSimulations}
\alias{iterateSimulations}
\title{Returns significance data from a snapshot.}
\usage{
iterateSimulations(originalData, realLabels, totalReads, cells, iterations)
}
\description{
Returns significance data from a snapshot.
}
