% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST.R
\name{MAST}
\alias{MAST}
\title{MAST}
\usage{
MAST(SCEdata, lbound = 0.1, FCTHRESHOLD = log2(1.5), freq_expressed = 0.2)
}
\arguments{
\item{FCTHRESHOLD}{the threshold for filetering the sample}

\item{freq_expressed}{}

\item{data}{SCE dataset for running the MAST}
}
\description{
MAST
}
