% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST.R
\name{MAST}
\alias{MAST}
\title{MAST}
\usage{
MAST(SCEdata, FCTHRESHOLD = log2(1.5), freq_expressed = 0.2,
  p.value = 0.05)
}
\arguments{
\item{SCEdata}{SCESet object}

\item{FCTHRESHOLD}{the threshold for filetering the sample}

\item{freq_expressed}{Filter genes that are expressed in at least this fraction of 
cells}

\item{p.value}{p-value cutoff}
}
\description{
Run MAST analysis on a SCESet object.
}
