% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST.R
\name{MAST}
\alias{MAST}
\title{MAST}
\usage{
MAST(SCEdata, freq_expressed = 0.1, condition = NULL,
  interest.level = NULL, FCTHRESHOLD = log2(1.5), p.value = 0.05)
}
\arguments{
\item{SCEdata}{SCESet object}

\item{freq_expressed}{Filter genes that are expressed in at least this fraction of 
cells}

\item{condition}{select varible (from the pdata) that is used for modle}

\item{interest.level}{interested level in the condition}

\item{FCTHRESHOLD}{the threshold for filetering the sample}

\item{p.value}{p values for selecting the hurdle reslut, default is 0.05}
}
\description{
#' Run MAST analysis on a SCESet object.
}
