% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST.R, R/MASTPlotFunctions.R
\name{MAST}
\alias{MAST}
\alias{thresholdGenes}
\alias{MASTviolin}
\alias{MASTregression}
\title{MAST}
\usage{
MAST(
  inSCE,
  condition = NULL,
  interest.level = NULL,
  freqExpressed = 0.1,
  fcThreshold = log2(1.5),
  p.value = 0.05,
  useThresh = FALSE,
  useAssay = "logcounts"
)

thresholdGenes(inSCE, useAssay = "logcounts")

MASTviolin(
  inSCE,
  useAssay = "logcounts",
  fcHurdleSig,
  samplesize = 49,
  threshP = FALSE,
  condition
)

MASTregression(
  inSCE,
  useAssay = "logcounts",
  fcHurdleSig,
  samplesize = 49,
  threshP = FALSE,
  condition
)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object. Required}

\item{condition}{select variable (from the colData) that is used for the
model.}

\item{interest.level}{If the condition of interest has more than two factors,
indicate which level should be used to compare to all other samples.}

\item{freqExpressed}{Filter genes that are expressed in at least this
fraction of cells. The default is expression in 0.1 of samples.}

\item{fcThreshold}{Minimum fold change for differentially expressed gene.}

\item{p.value}{p values for selecting the hurdle result, default is 0.05}

\item{useThresh}{Use adaptive thresholding to filter genes. The default is
FALSE.}

\item{useAssay}{The assay to use for the MAST calculations. The default is
"logcounts"}

\item{fcHurdleSig}{The filtered result from hurdle model}

\item{samplesize}{The number of most significant genes}

\item{threshP}{Plot threshold values from adaptive thresholding. Default is
FALSE}
}
\value{
MAST(): A data.frame of differentially expressed genes with p-values.

thresholdGenes(): list of thresholded counts (on natural scale),
thresholds, bins, densities estimated on each bin, and the original data from
MAST::thresholdSCRNACountMatrix

MASTviolin(): A ggplot object of MAST violin plots.

MASTregression(): A ggplot object of MAST linear regression plots.
}
\description{
Run and visualize MAST analysis on a SCtkExperiment object.
}
\section{Functions}{
\itemize{
\item \code{MAST}: Run MAST analysis.

\item \code{thresholdGenes}: Identify adaptive thresholds

\item \code{MASTviolin}: Visualize MAST results using violin plots

\item \code{MASTregression}: Visualize MAST results using linear model plots
}}

\examples{
data("mouseBrainSubsetSCE")
res <- thresholdGenes(mouseBrainSubsetSCE)
}
