% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST_plot_functions.R
\name{MASTregression}
\alias{MASTregression}
\title{MAST linear model}
\usage{
MASTregression(SCEdata, fcHurdleSig, samplesize = 49)
}
\arguments{
\item{SCEdata}{SingleCellExperiment object}

\item{fcHurdleSig}{The filtered result from hurdle model}

\item{samplesize}{The number of most significant genes}
}
\description{
Run MAST analysis on a SingleCellExperiment object.
}
