% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST_plot_functions.R
\name{MASTviolin}
\alias{MASTviolin}
\title{MAST Violin}
\usage{
MASTviolin(SCEdata, fcHurdleSig, samplesize = 49)
}
\arguments{
\item{SCEdata}{SCESet object}

\item{fcHurdleSig}{The filtered result from hurdle model}

\item{samplesize}{The number of most significant genes}
}
\description{
Run MAST analysis on a SCESet object.
}
