% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST_plot_functions.R
\name{MASTviolin}
\alias{MASTviolin}
\title{MAST Violin}
\usage{
MASTviolin(SCEdata, fcHurdleSig, samplesize = 49, threshP = FALSE, variable)
}
\arguments{
\item{SCEdata}{SingleCelltkExperiment object}

\item{fcHurdleSig}{The filtered result from hurdle model}

\item{samplesize}{The number of most significant genes}

\item{threshP}{Plot threshold values from adaptive thresholding. Default is
FALSE}

\item{variable}{Select the condition of interest}
}
\description{
Run MAST analysis on a SingleCelltkExperiment object.
}
