% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSCEColData.R
\name{mergeSCEColData}
\alias{mergeSCEColData}
\title{Merging colData from two singleCellExperiment objects}
\usage{
mergeSCEColData(sce1, sce2, id1 = "column_name", id2 = "column_name")
}
\arguments{
\item{sce1}{SingleCellExperiment object. The function will output this
singleCellExperiment object with a combined colData from sce1 and sce2.}

\item{sce2}{SingleCellExperiment object. colData from this object
will be merged with colData from sce1 and loaded into sce1.}

\item{id1}{Character vector. Column in colData of sce1 that will be 
used to combine sce1 and sce2. Default "column_name"}

\item{id2}{Character vector. Column in colData of sce2 that will be
used to combine sce1 and sce2. Default "column_name"}
}
\value{
SingleCellExperiment object containing combined colData from  
 both singleCellExperiment for samples in sce1.
}
\description{
Merges colData of the singleCellExperiment objects 
 obtained from the same dataset which contain differing colData. 
 (i.e. raw data and filtered data)
}
\examples{
sce1 <- importCellRanger(
    cellRangerDirs = system.file("extdata", package = "singleCellTK"),
    samples = "hgmm_1k_v3_20x20")
data(emptyDropsSceExample, package = "scruff")
sce2 <- emptyDropsSceExample
sce <- mergeSCEColData(sce1, sce2, id1 = "column_name", id2 = "column_name")
}
