% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{pca_variances<-}
\alias{pca_variances<-}
\title{Set PCA variances}
\usage{
pca_variances(x, ...) <- value
}
\arguments{
\item{x}{SCtkE object}

\item{...}{other parameters}

\item{value}{PCA variances DataFrame()}
}
\value{
A SCtkExperiment object with the pca_variances slot set.
}
\description{
Set PCA variances
}
\examples{
data("mouse_brain_subset_sce")
pca_variances(mouse_brain_subset_sce)
#getPCA() sets the pca_variances
newSCE <- getPCA(mouse_brain_subset_sce, use_assay = "counts")

#alternatively, set the pca_variances directly
pca <- prcomp(assay(mouse_brain_subset_sce, "logcounts"))
percentVar <- pca$sdev ^ 2 / sum(pca$sdev ^ 2)
pca_variances(mouse_brain_subset_sce) <- DataFrame(percentVar)
}
