% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getset.R
\docType{methods}
\name{pca_variances}
\alias{pca_variances}
\alias{pca_variances,SCtkExperiment-method}
\alias{pca_variances<-,SCtkExperiment-method}
\title{Get PCA variances}
\usage{
pca_variances(x, ...)

\S4method{pca_variances}{SCtkExperiment}(x)

\S4method{pca_variances}{SCtkExperiment}(x) <- value
}
\arguments{
\item{x}{SCtkE object}

\item{...}{other parameters}

\item{value}{The DataFrame of pca_variances}
}
\value{
A data frame of percent variation explained by each PC.

A SCtkExperiment object with the pca_variances object set.
}
\description{
Get PCA variances

Get PCA variances

Set PCA variances
}
\examples{
data("GSE60361_subset_sce")
pca_variances(GSE60361_subset_sce)
#getPCA() sets the pca_variances
newSCE <- getPCA(GSE60361_subset_sce, use_assay = "counts")

#alternatively, set the pca_variances directly
pca <- prcomp(assay(GSE60361_subset_sce, "logcounts"))
percentVar <- pca$sdev ^ 2 / sum(pca$sdev ^ 2)
pca_variances(GSE60361_subset_sce) <- DataFrame(percentVar)
}
