% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getset.R
\docType{methods}
\name{pca_variances}
\alias{pca_variances}
\alias{pca_variances,SCtkExperiment-method}
\alias{pca_variances<-,SCtkExperiment-method}
\title{Get PCA variances}
\usage{
pca_variances(x, ...)

\S4method{pca_variances}{SCtkExperiment}(x)

\S4method{pca_variances}{SCtkExperiment}(x) <- value
}
\arguments{
\item{x}{SCtkE object}

\item{...}{other parameters}

\item{value}{The DataFrame of pca_variances}
}
\description{
Get PCA variances

Get PCA variances

Set PCA variances
}
