% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{plot_d3DiffEx}
\alias{plot_d3DiffEx}
\title{Plot Interactive Differential Expression}
\usage{
plot_d3DiffEx(inSCESet, condition, geneList, clusterRow = TRUE,
  clusterCol = TRUE)
}
\arguments{
\item{inSCESet}{}

\item{condition}{}

\item{geneList}{}

\item{clusterRow}{}

\item{clusterCol}{}
}
\description{
Plot Interactive Differential Expression
}

