% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{plot_DiffEx}
\alias{plot_DiffEx}
\title{Plot Differential Expression}
\usage{
plot_DiffEx(inSCESet, condition, geneList, clusterRow = TRUE,
  clusterCol = TRUE)
}
\arguments{
\item{inSCESet}{Input data object that contains the data to be plotted.
Required}

\item{condition}{The condition used for plotting the heatmap. Required}

\item{geneList}{The list of genes to put in the heatmap. Required}

\item{clusterRow}{Cluster the rows. The default is TRUE}

\item{clusterCol}{Cluster the columns. The default is TRUE}
}
\description{
Plot Differential Expression
}

