% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{plot_DiffEx}
\alias{plot_DiffEx}
\title{Plot Differential Expression}
\usage{
plot_DiffEx(inSCESet, condition, geneList, clusterRow = TRUE,
  clusterCol = TRUE)
}
\arguments{
\item{inSCESet}{}

\item{condition}{}

\item{geneList}{}

\item{clusterRow}{}

\item{clusterCol}{}
}
\description{
Plot Differential Expression
}
