% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPerQCWrapper.R
\name{plotBarcodeRankDropsResults}
\alias{plotBarcodeRankDropsResults}
\title{Plots for runEmptyDrops outputs.}
\usage{
plotBarcodeRankDropsResults(
  inSCE,
  sample = NULL,
  defaultTheme = TRUE,
  dotSize = 1,
  titleSize = 18,
  axisLabelSize = 18,
  axisSize = 15,
  legendSize = 15
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object with saved
dimension reduction components or a variable with saved results from
\link{runBarcodeRankDrops}. Required.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.
Default NULL.}

\item{defaultTheme}{Removes grid in plot and sets axis title size to 10
when TRUE. Default TRUE.}

\item{dotSize}{Size of dots. Default 1.}

\item{titleSize}{Size of title of plot. Default 18.}

\item{axisLabelSize}{Size of x/y-axis labels. Default 18.}

\item{axisSize}{Size of x/y-axis ticks. Default 15.}

\item{legendSize}{size of legend. Default 15.}
}
\value{
list of .ggplot objects
}
\description{
A wrapper function which visualizes outputs from the
 runEmptyDrops function stored in the colData slot of the SingleCellExperiment
 object via plots.
}
\examples{
data(scExample, package="singleCellTK")
sce <- runBarcodeRankDrops(inSCE=sce)
plotBarcodeRankDropsResults(inSCE=sce)
}
