% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotting.R
\name{plotBarcodeRankScatter}
\alias{plotBarcodeRankScatter}
\title{Plots for runBarcodeRankDrops outputs.}
\usage{
plotBarcodeRankScatter(
  inSCE,
  sample = NULL,
  defaultTheme = TRUE,
  dotSize = 1,
  title = NULL,
  titleSize = 18,
  xlab = NULL,
  ylab = NULL,
  axisSize = 12,
  axisLabelSize = 15,
  legendSize = 10,
  combinePlot = "none",
  sampleRelHeights = 1,
  sampleRelWidths = 1
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object with saved
dimension reduction components or a variable with saved results from
\link{runBarcodeRankDrops}. Required.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.
Default NULL.}

\item{defaultTheme}{Removes grid in plot and sets axis title size to 10
when TRUE. Default TRUE.}

\item{dotSize}{Size of dots. Default 1.}

\item{title}{Title of plot. Default NULL.}

\item{titleSize}{Size of title of plot. Default 18.}

\item{xlab}{Character vector. Label for x-axis. Default NULL.}

\item{ylab}{Character vector. Label for y-axis. Default NULL.}

\item{axisSize}{Size of x/y-axis ticks. Default 12.}

\item{axisLabelSize}{Size of x/y-axis labels. Default 15.}

\item{legendSize}{size of legend. Default 10.}

\item{combinePlot}{Boolean. If multiple plots are generated (multiple
samples, etc.), will combined plots using `cowplot::plot_grid`.}

\item{sampleRelHeights}{If there are multiple samples and combining by "all",
the relative heights for each plot.}

\item{sampleRelWidths}{If there are multiple samples and combining by "all",
the relative widths for each plot.
Default TRUE.}
}
\value{
a ggplot object of the scatter plot.
}
\description{
A plotting function which visualizes outputs from the
 runBarcodeRankDrops function stored in the colData slot of the SingleCellExperiment
 object via scatterplot.
}
\examples{
data(scExample, package="singleCellTK")
sce <- runBarcodeRankDrops(inSCE=sce)
plotBarcodeRankScatter(inSCE=sce)
}
