% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBatchVariance.R
\name{plotBatchVariance}
\alias{plotBatchVariance}
\title{Plot the percent of the variation that is explained by batch and condition
in the data}
\usage{
plotBatchVariance(inSCESet, use_assay = "logcounts", batch,
  condition = NULL)
}
\arguments{
\item{inSCESet}{Input SCtkExperiment object. Required}

\item{use_assay}{Indicate which assay to use for PCA. Default is "logcounts"}

\item{batch}{The column in the annotation data that corresponds to batch.
Required}

\item{condition}{The column in the annotation data that corresponds to
condition. Optional}
}
\value{
A boxplot of variation explained by batch, condition, and
batch+condition (if applicable).
}
\description{
Visualize the percent variation in the data that is explained by batch and
condition if it is given.
}
\examples{
if(requireNamespace("bladderbatch", quietly = TRUE)) {
  library(bladderbatch)
  data(bladderdata)
  dat <- as(as(bladderEset, "SummarizedExperiment"), "SCtkExperiment")
  plotBatchVariance(dat, use_assay="exprs", batch="batch", condition = "cancer")
}

}
