% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBiomarker.R
\name{plotBiomarker}
\alias{plotBiomarker}
\title{plotBiomarker}
\usage{
plotBiomarker(count_data, gene, binary = "Binary", visual = "PCA",
  shape = "No Shape", x = "PC1", y = "PC2")
}
\arguments{
\item{count_data}{A SingleCelltkExperiment object}

\item{gene}{gene list}

\item{binary}{"Binary" for binary expression or "Continuous" for a gradient.
Default: "Binary"}

\item{visual}{Type of visualization (PCA or tSNE). Default: "PCA"}

\item{shape}{visualization shape}

\item{x}{x coordinate for PCA}

\item{y}{y coordinate for PCA}
}
\value{
A Biomarker plot
}
\description{
Given a set of genes, return a ggplot of expression values.
}
