% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBiomarker.R
\name{plotBiomarker}
\alias{plotBiomarker}
\title{plotBiomarker}
\usage{
plotBiomarker(count_data, gene, binary = "Binary", visual = "PCA",
  shape = "No Shape", x = "PC1", y = "PC2", use_assay = "counts",
  reducedDimName = "PCA")
}
\arguments{
\item{count_data}{A SCtkExperiment object}

\item{gene}{gene list}

\item{binary}{"Binary" for binary expression or "Continuous" for a gradient.
Default: "Binary"}

\item{visual}{Type of visualization (PCA or tSNE). Default: "PCA"}

\item{shape}{visualization shape}

\item{x}{x coordinate for PCA}

\item{y}{y coordinate for PCA}

\item{use_assay}{Indicate which assay to use for PCA. Default is "counts"}

\item{reducedDimName}{PCA dimension name. The default is PCA.
The toolkit will store data with the pattern <ASSAY>_<ALGORITHM>.}
}
\value{
A Biomarker plot
}
\description{
Given a set of genes, return a ggplot of expression values.
}
\examples{
data("mouse_brain_subset_sce")
plotBiomarker(mouse_brain_subset_sce, gene="C1qa", shape="level1class")

}
