% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBiomarker.R
\name{plotBiomarker}
\alias{plotBiomarker}
\title{Given a set of genes, return a ggplot of expression
values.}
\usage{
plotBiomarker(
  inSCE,
  gene,
  binary = "Binary",
  shape = "No Shape",
  useAssay = "counts",
  reducedDimName = "PCA",
  x = NULL,
  y = NULL,
  comp1 = NULL,
  comp2 = NULL
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{gene}{genelist to run the method on.}

\item{binary}{binary/continuous color for the expression.}

\item{shape}{shape parameter for the ggplot.}

\item{useAssay}{Indicate which assay to use. The default is "logcounts".}

\item{reducedDimName}{results name of dimension reduction
coordinates obtained from this method. This is stored in the SingleCellExperiment
object in the reducedDims slot. Required.}

\item{x}{PCA component to be used for plotting(if applicable).
Default is first PCA component for PCA data and NULL otherwise.}

\item{y}{PCA component to be used for plotting(if applicable).
Default is second PCA component for PCA data and NULL otherwise.}

\item{comp1}{label for x-axis}

\item{comp2}{label for y-axis}
}
\value{
A Biomarker plot
}
\description{
Given a set of genes, return a ggplot of expression
values.
}
\examples{
data("mouseBrainSubsetSCE")
plotBiomarker(mouseBrainSubsetSCE, gene="C1qa", shape="level1class", reducedDimName="TSNE_counts")

}
