% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBubble.R
\name{plotBubble}
\alias{plotBubble}
\title{Plot Bubble plot}
\usage{
plotBubble(
  inSCE,
  useAssay = "logcounts",
  gene,
  displayName = NULL,
  clusters = "cluster",
  title = "",
  colorLow = "white",
  colorHigh = "blue"
)
}
\arguments{
\item{inSCE}{The single cell experiment to use.}

\item{useAssay}{The assay to use.}

\item{gene}{A string or vector of strings with each gene to aggregate.}

\item{displayName}{A string that is the name of the column used for genes.}

\item{clusters}{The name of a colData entry that can be used as groups.}

\item{title}{The title of the bubble plot}

\item{colorLow}{The color to be used for lowest value of mean expression}

\item{colorHigh}{The color to be used for highest value of mean expression}
}
\value{
A ggplot of the bubble plot.
}
\description{
Plot a bubble plot with the color of the plot being the mean expression
and the size of the dot being the percent of cells in the cluster expressing the gene.
}
\examples{
data("scExample")
plotBubble(inSCE=sce, useAssay="counts", gene=c("B2M"), displayName="feature_name", 
clusters="type", title="cell type test", colorLow="white", colorHigh="blue")
}
