% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEAnalysis.R
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\title{Heatmap visualization of DEG result}
\usage{
plotDEGHeatmap(
  inSCE,
  useResult,
  onlyPos = FALSE,
  log2fcThreshold = 0.25,
  fdrThreshold = 0.05,
  useAssay = NULL,
  featureAnnotations = NULL,
  cellAnnotations = NULL,
  featureAnnotationColor = NULL,
  cellAnnotationColor = NULL,
  rowDataName = NULL,
  colDataName = NULL,
  colSplitBy = "condition",
  rowSplitBy = "regulation",
  title = paste0("MAST Result: ", useResult),
  ...
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object.
\code{runMAST()} has to be run in advance.}

\item{useResult}{character. A string specifying the \code{analysisName}
used when running a differential expression analysis function.}

\item{onlyPos}{logical. Whether to only plot DEG with positive log2_FC
value. Default \code{FALSE}.}

\item{log2fcThreshold}{numeric. Only plot DEGs with the absolute values of
log2FC larger than this value. Default \code{1}.}

\item{fdrThreshold}{numeric. Only plot DEGs with FDR value smaller than this
value. Default \code{0.05}.}

\item{useAssay}{character. A string specifying an assay of expression value
to plot. By default the assay used for \code{runMAST()} will be used.
Default \code{NULL}.}

\item{featureAnnotations}{\code{data.frame}, with \code{rownames} containing
all the features going to be plotted. Character columns should be factors.
Default \code{NULL}.}

\item{cellAnnotations}{\code{data.frame}, with \code{rownames} containing
all the cells going to be plotted. Character columns should be factors.
Default \code{NULL}.}

\item{featureAnnotationColor}{A named list. Customized color settings for
feature labeling. Should match the entries in the \code{featureAnnotations}
or \code{rowDataName}. For each entry, there should be a list/vector of
colors named with categories. Default \code{NULL}.}

\item{cellAnnotationColor}{A named list. Customized color settings for
cell labeling. Should match the entries in the \code{cellAnnotations} or
\code{colDataName}. For each entry, there should be a list/vector of colors
named with categories. Default \code{NULL}.}

\item{rowDataName}{character. The column name(s) in \code{rowData} that need
to be added to the annotation. Default \code{NULL}.}

\item{colDataName}{character. The column name(s) in \code{colData} that need
to be added to the annotation. Default \code{NULL}.}

\item{colSplitBy}{character. Do semi-heatmap based on the grouping of
this(these) annotation(s). Should exist in either \code{colDataName} or
\code{names(cellAnnotations)}. Default \code{"condition"}.}

\item{rowSplitBy}{character. Do semi-heatmap based on the grouping of
this(these) annotation(s). Should exist in either \code{rowDataName} or
\code{names(featureAnnotations)}. Default \code{"regulation"}.}

\item{title}{character. Main title of the heatmap. Default
\code{"MAST Result: <useResult>"}.}

\item{...}{Other arguments passed to \code{\link{plotSCEHeatmap}}}
}
\value{
A ComplexHeatmap::Heatmap object
}
\description{
A differential expression analysis function has to be run in advance so that
information is stored in the metadata of the input SCE object. This function
wraps plotSCEHeatmap.
A feature annotation basing on the log2FC level called \code{"regulation"}
will be automatically added. A cell annotation basing on the condition
selection while running the analysis called \code{"condition"}, and the
annotations used from \code{colData(inSCE)} while setting the condition and
covariates will also be added.
}
\examples{
data(scExample, package = "singleCellTK")
\dontrun{
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sce <- runDEAnalysis(inSCE = sce, groupName1 = "Sample1", method = "DESeq2",
 groupName2 = "Sample2", index1 = 1:100, index2 = 101:190, analysisName = "DESeq2")
plotDEGHeatmap(sce, useResult = "DESeq2", fdrThreshold = 1)
}

}
\author{
Yichen Wang
}
