% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDimRed.R
\name{plotDimRed}
\alias{plotDimRed}
\title{Plot dimensionality reduction}
\usage{
plotDimRed(inSCE, useReduction, showLegend = FALSE)
}
\arguments{
\item{inSCE}{Input SCE object}

\item{useReduction}{Reduction to plot}

\item{showLegend}{If legends should be plotted or not}
}
\value{
plot object
}
\description{
Plot dimensionality reduction
}
