% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDimRed.R
\name{plotDimRed}
\alias{plotDimRed}
\title{Plot PCA or tSNE
  
Use this function to plot PCA or tSNE results}
\usage{
plotDimRed(method, vals_method, count_data, colorClusters, pc1, pc2)
}
\arguments{
\item{method}{Either PCA or tSNE}

\item{vals_method}{PCA or tSNE data frame}

\item{count_data}{A SCE object}

\item{colorClusters}{The variable to color clusters by}

\item{pc1}{User choice for the first principal component}

\item{pc2}{User choice for the second prinicipal component}
}
\value{
A reduced dimension object
}
\description{
Plot PCA or tSNE
  
Use this function to plot PCA or tSNE results
}
