% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDimRed.R
\name{plotDimRed}
\alias{plotDimRed}
\title{Plot dimensionality reduction from computed metrics including PCA, ICA, tSNE
and UMAP}
\usage{
plotDimRed(
  inSCE,
  useReduction,
  showLegend = FALSE,
  xAxisLabel = NULL,
  yAxisLabel = NULL
)
}
\arguments{
\item{inSCE}{Input SCE object}

\item{useReduction}{Reduction to plot}

\item{showLegend}{If legends should be plotted or not}

\item{xAxisLabel}{Specify the label for x-axis. Default is \code{NULL} which
will specify the label as 'x'.}

\item{yAxisLabel}{Specify the label for y-axis. Default is \code{NULL} which
will specify the label as 'y'.}
}
\value{
plot object
}
\description{
Plot dimensionality reduction from computed metrics including PCA, ICA, tSNE
and UMAP
}
