% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPerQCWrapper.R
\name{plotEmptyDropsResults}
\alias{plotEmptyDropsResults}
\title{Plots for runEmptyDrops outputs.}
\usage{
plotEmptyDropsResults(
  inSCE,
  sample = NULL,
  combinePlot = "all",
  fdrCutoff = 0.01,
  defaultTheme = TRUE,
  dotSize = 1,
  titleSize = 18,
  axisLabelSize = 18,
  axisSize = 15,
  legendSize = 15,
  legendTitleSize = 16,
  relHeights = 1,
  relWidths = 1,
  samplePerColumn = TRUE,
  sampleRelHeights = 1,
  sampleRelWidths = 1
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object with saved
dimension reduction components or a variable with saved results from
\link{runScrublet}. Required.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.
Default NULL.}

\item{combinePlot}{Must be either "all", "sample", or "none". "all" will combine all plots into a single
.ggplot object, while "sample" will output a list of plots separated by sample. Default "all".}

\item{fdrCutoff}{Numeric. Thresholds barcodes based on the FDR values from
runEmptyDrops as "Empty Droplet" or "Putative Cell". Default 0.01.}

\item{defaultTheme}{Removes grid in plot and sets axis title size to 10
when TRUE. Default TRUE.}

\item{dotSize}{Size of dots. Default 1.}

\item{titleSize}{Size of title of plot. Default 18.}

\item{axisLabelSize}{Size of x/y-axis labels. Default 18.}

\item{axisSize}{Size of x/y-axis ticks. Default 15.}

\item{legendSize}{size of legend. Default 15.}

\item{legendTitleSize}{size of legend title. Default 16.}

\item{relHeights}{Relative heights of plots when combine is set.}

\item{relWidths}{Relative widths of plots when combine is set.}

\item{samplePerColumn}{If TRUE, when there are multiple samples and combining by "all",
the output .ggplot will have plots from each sample on a single column. Default TRUE.}

\item{sampleRelHeights}{If there are multiple samples and combining by "all",
the relative heights for each plot.}

\item{sampleRelWidths}{If there are multiple samples and combining by "all",
the relative widths for each plot.}
}
\value{
list of .ggplot objects
}
\description{
A wrapper function which visualizes outputs from the
 runEmptyDrops function stored in the colData slot of the SingleCellExperiment
 object via plots.
}
\examples{
data(scExample, package="singleCellTK")
sce <- runEmptyDrops(inSCE=sce)
plotEmptyDropsResults(inSCE=sce)
}
