% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot an interactive (D3) Heatmap
  
Use this function to run limma differential expression and load an interactive D3 heatmap}
\usage{
plotHeatmap(data)
}
\arguments{
\item{data}{a SCEset object}
}
\value{
D3 heatmap will load
}
\description{
Plot an interactive (D3) Heatmap
  
Use this function to run limma differential expression and load an interactive D3 heatmap
}
