% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASTPlotFunctions.R
\name{plotMASTHeatmap}
\alias{plotMASTHeatmap}
\title{Heatmap visualization of DEG result called by MAST}
\usage{
plotMASTHeatmap(
  inSCE,
  useResult,
  onlyPos = FALSE,
  log2fcThreshold = 1,
  fdrThreshold = 0.05,
  useAssay = NULL,
  featureAnnotations = NULL,
  cellAnnotations = NULL,
  featureAnnotationColor = NULL,
  cellAnnotationColor = NULL,
  rowDataName = NULL,
  colDataName = NULL,
  colSplitBy = "condition",
  rowSplitBy = "regulation",
  title = paste0("MAST Result: ", useResult),
  ...
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object.
\code{runMAST()} has to be run in advance.}

\item{useResult}{character. A string specifying the \code{comparisonName}
used when running \code{runMAST()}.}

\item{onlyPos}{logical. Whether to only plot DEG with positive log2_FC
value. Default \code{FALSE}.}

\item{log2fcThreshold}{numeric. Only plot DEGs with the absolute values of
log2FC larger than this value. Default \code{1}.}

\item{fdrThreshold}{numeric. Only plot DEGs with FDR value smaller than this
value. Default \code{0.05}.}

\item{useAssay}{character. A string specifying an assay of expression value
to plot. By default the assay used for \code{runMAST()} will be used.
Default \code{NULL}.}

\item{featureAnnotations}{\code{data.frame}, with \code{rownames} containing
all the features going to be plotted. Character columns should be factors.
Default \code{NULL}.}

\item{cellAnnotations}{\code{data.frame}, with \code{rownames} containing
all the cells going to be plotted. Character columns should be factors.
Default \code{NULL}.}

\item{featureAnnotationColor}{A named list. Customized color settings for
feature labeling. Should match the entries in the \code{featureAnnotations}
or \code{rowDataName}. For each entry, there should be a list/vector of
colors named with categories. Default \code{NULL}.}

\item{cellAnnotationColor}{A named list. Customized color settings for
cell labeling. Should match the entries in the \code{cellAnnotations} or
\code{colDataName}. For each entry, there should be a list/vector of colors
named with categories. Default \code{NULL}.}

\item{rowDataName}{character. The column name(s) in \code{rowData} that need
to be added to the annotation. Default \code{NULL}.}

\item{colDataName}{character. The column name(s) in \code{colData} that need
to be added to the annotation. Default \code{NULL}.}

\item{colSplitBy}{character. Do semi-heatmap based on the grouping of
this(these) annotation(s). Should exist in either \code{colDataName} or
\code{names(cellAnnotations)}. Default \code{"condition"}.}

\item{rowSplitBy}{character. Do semi-heatmap based on the grouping of
this(these) annotation(s). Should exist in either \code{rowDataName} or
\code{names(featureAnnotations)}. Default \code{"regulation"}.}

\item{title}{character. Main title of the heatmap. Default
\code{"MAST Result: <useResult>"}.}

\item{...}{Other arguments passed to `plotSCEHeatmap()`}
}
\value{
A ComplexHeatmap::Heatmap object
}
\description{
runMAST() has to be run in advance so that information is stored in the
metadata of the input SCE object. This function wraps plotSCEHeatmap.
}
\author{
Yichen Wang
}
