% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASTPlotFunctions.R
\name{plotMASTRegression}
\alias{plotMASTRegression}
\title{plot the linear regression to show visualize the expression the of DEGs
identified by MAST}
\usage{
plotMASTRegression(
  inSCE,
  useResult,
  threshP = FALSE,
  nrow = 6,
  ncol = 6,
  defaultTheme = TRUE
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object.
\code{runMAST()} has to be run in advance.}

\item{useResult}{character. A string specifying the \code{comparisonName}
used when running \code{runMAST()}.}

\item{threshP}{logical. Whether to plot threshold values from adaptive
thresholding, instead of using the assay used by \code{runMAST()}. Default
\code{FALSE}.}

\item{nrow}{Integer. Number of rows in the plot grid. Default \code{6}.}

\item{ncol}{Integer. Number of columns in the plot grid. Default \code{6}.}

\item{defaultTheme}{Logical scalar. Whether to use default SCTK theme in
ggplot. Default \code{TRUE}.}
}
\value{
A ggplot object of MAST linear regression
}
\description{
plot the linear regression to show visualize the expression the of DEGs
identified by MAST
}
