% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEAnalysis.R
\name{plotMASTThresholdGenes}
\alias{plotMASTThresholdGenes}
\title{MAST Identify adaptive thresholds}
\usage{
plotMASTThresholdGenes(inSCE, useAssay = "logcounts", check_sanity = TRUE)
}
\arguments{
\item{inSCE}{SingleCellExperiment object}

\item{useAssay}{character, default \code{"logcounts"}}

\item{check_sanity}{Logical scalar. Whether to perform MAST's sanity check
to see if the counts are logged. Default \code{TRUE}}
}
\value{
Plot the thresholding onto the plotting region.
}
\description{
Calculate and produce a list of thresholded counts (on natural scale),
thresholds, bins, densities estimated on each bin, and the original data from
\code{\link[MAST]{thresholdSCRNACountMatrix}}
}
\examples{
data("mouseBrainSubsetSCE")
plotMASTThresholdGenes(mouseBrainSubsetSCE)
}
