% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPathway.R
\name{plotPathway}
\alias{plotPathway}
\title{Generate violin plots for pathway analysis results}
\usage{
plotPathway(inSCE, resultName, geneset, groupby = NULL, boxplot = FALSE)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object. With 
\code{runGSVA()} or \code{runVAM()} applied in advance.}

\item{resultName}{A single character of the name of a score matrix, which 
should be found in \code{getPathwayResultNames(inSCE)}.}

\item{geneset}{A single character specifying the geneset of interest. Should
be found in the geneSetCollection used for performing the analysis.}

\item{groupby}{Either a single character specifying a column of 
\code{colData(inSCE)} or a vector of equal length as the number of cells. 
Default \code{NULL}.}

\item{boxplot}{Whether to add a boxplot. Default \code{FALSE}.}
}
\value{
A \code{ggplot} object for the violin plot
}
\description{
Generate violin plots for pathway analysis results
}
\details{
\code{runGSVA()} or \code{runVAM()} should be applied in advance of 
using this function. Users can group the data by specifying \code{groupby}.
}
\examples{
data("scExample", package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sce <- scaterlogNormCounts(sce, assayName = "logcounts")
gs1 <- rownames(sce)[seq(10)]
gs2 <- rownames(sce)[seq(11,20)]
gs <- list("geneset1" = gs1, "geneset2" = gs2)
sce <- importGeneSetsFromList(inSCE = sce, geneSetList = gs,
                              by = "rownames")
sce <- runVAM(inSCE = sce, geneSetCollectionName = "GeneSetCollection",
              useAssay = "logcounts")
plotPathway(sce, "VAM_GeneSetCollection_CDF", "geneset1")
}
