% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot PCA}
\usage{
plotPCA(countData, colorBy = "No Color", shape = "No Shape", pcX = "PC1",
  pcY = "PC2", reducedDimName = "PCA", runPCA = FALSE,
  useAssay = "logcounts")
}
\arguments{
\item{countData}{A SCE object}

\item{colorBy}{The variable to color clusters by}

\item{shape}{Shape of the points}

\item{pcX}{User choice for the first principal component}

\item{pcY}{User choice for the second principal component}

\item{reducedDimName}{PCA dimension name. The default is PCA.
The toolkit will store data with the pattern <ASSAY>_<ALGORITHM>.}

\item{runPCA}{Run PCA if the reducedDimName does not exist. the Default is
FALSE.}

\item{useAssay}{Indicate which assay to use for PCA. Default is "logcounts"}
}
\value{
A PCA plot
}
\description{
Use this function to plot PCA results
}
\examples{
data("mouseBrainSubsetSCE")
plotPCA(mouseBrainSubsetSCE, colorBy = "level1class",
        reducedDimName = "PCA_counts")

}
