% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot PCA
  
Use this function to plot PCA or tSNE results}
\usage{
plotPCA(count_data, pca_df = NULL, colorBy = NULL, shape = NULL,
  pcX = "PC1", pcY = "PC2")
}
\arguments{
\item{count_data}{A SCE object}

\item{pca_df}{PCA data frame}

\item{colorBy}{The variable to color clusters by}

\item{shape}{Shape of the points}

\item{pcX}{User choice for the first principal component}

\item{pcY}{User choice for the second prinicipal component}
}
\value{
A PCA plot
}
\description{
Plot PCA
  
Use this function to plot PCA or tSNE results
}
