% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot PCA}
\usage{
plotPCA(count_data, colorBy = "No Color", shape = "No Shape", pcX = "PC1",
  pcY = "PC2", reducedDimName = "PCA", runPCA = FALSE,
  use_assay = "logcounts")
}
\arguments{
\item{count_data}{A SCE object}

\item{colorBy}{The variable to color clusters by}

\item{shape}{Shape of the points}

\item{pcX}{User choice for the first principal component}

\item{pcY}{User choice for the second prinicipal component}

\item{reducedDimName}{PCA dimension name. The default is PCA.
The toolkit will store data with the pattern <ASSSAY>_<ALGORITHM>.}

\item{runPCA}{Run PCA if the reducedDimName doesn't exist. the Default is
FALSE.}

\item{use_assay}{Indicate which assay to use for PCA. Default is "counts"}
}
\value{
A PCA plot
}
\description{
Use this function to plot PCA or tSNE results
}
