% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpyFunctions.R
\name{plotScanpyHVG}
\alias{plotScanpyHVG}
\title{plotScanpyHVG}
\usage{
plotScanpyHVG(inSCE, log = FALSE)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{log}{Plot on logarithmic axes. Default \code{FALSE}.}
}
\value{
plot object
}
\description{
plotScanpyHVG
}
\examples{
data(scExample, package = "singleCellTK")
\dontrun{
sce <- runScanpyNormalizeData(sce, useAssay = "counts")
sce <- runScanpyFindHVG(sce, useAssay = "scanpyNormData", method = "seurat")
plotScanpyHVG(sce)
}
}
