% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpyFunctions.R
\name{plotScanpyViolin}
\alias{plotScanpyViolin}
\title{plotScanpyViolin}
\usage{
plotScanpyViolin(inSCE, features, groupBy, xlabel = "", ylabel = NULL)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{features}{Genes to plot. Sometimes is useful to pass a specific list of
var names (e.g. genes). The var_names could be a dictionary or a list.}

\item{groupBy}{The key of the observation grouping to consider.}

\item{xlabel}{Label of the x axis. Defaults to groupBy.}

\item{ylabel}{Label of the y axis. If NULL and groupBy is NULL, 
defaults to 'value'. If NULL and groupBy is not NULL, defaults to features.}
}
\value{
plot object
}
\description{
plotScanpyViolin
}
\examples{
data(scExample, package = "singleCellTK")
\dontrun{
sce <- runScanpyNormalizeData(sce, useAssay = "counts")
sce <- runScanpyPCA(sce, useAssay = "scanpyNormData")
sce <- runScanpyUMAP(sce, useReduction = "scanpyPCA")
sce <- runScanpyFindClusters(sce, useAssay = "counts")
marker <- c("ENSG00000102879" , "ENSG00000167283")
plotScanpyViolin(sce, features = markers, groupBy = "Scanpy_louvain_1")
}
}
