% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotting.R
\name{plotSCEBarAssayData}
\alias{plotSCEBarAssayData}
\title{Bar plot of assay data.}
\usage{
plotSCEBarAssayData(
  inSCE,
  useAssay = "counts",
  feature,
  groupby = NULL,
  dots = TRUE,
  xlab = NULL,
  ylab = NULL,
  axisSize = 10,
  dotSize = 1,
  transparency = 1,
  defaultTheme = TRUE,
  title = NULL,
  titleSize = NULL
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object with saved
dimension reduction components or a variable with saved results. Required.}

\item{useAssay}{Indicate which assay to use. Default "counts".}

\item{feature}{Name of feature stored in assay of SingleCellExperiment
object.}

\item{groupby}{Groupings for each numeric value. A user may input a vector
equal length to the number of the samples in the SingleCellExperiment
object, or can be retrieved from the colData slot. Default NULL.}

\item{dots}{Boolean. If TRUE, will plot dots for each violin plot.
Default TRUE.}

\item{xlab}{Character vector. Label for x-axis. Default NULL.}

\item{ylab}{Character vector. Label for y-axis. Default NULL.}

\item{axisSize}{Size of x/y-axis labels. Default 10.}

\item{dotSize}{Size of dots. Default 1.}

\item{transparency}{Transparency of the dots, values will be 0-1. Default 1.}

\item{defaultTheme}{Removes grid in plot and sets axis title size to 10
when TRUE. Default TRUE.}

\item{title}{Title of plot. Default NULL.}

\item{titleSize}{Size of title of plot. Default 15.}
}
\description{
Visualizes values stored in the assay slot of a
 SingleCellExperiment object via a violin plot.
}
\examples{
plotSCEBarAssayData(
  inSCE = mouseBrainSubsetSCE,
  feature = "Apoe", groupby = "sex"
)
}
