% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotting.R
\name{plotSCEBarColData}
\alias{plotSCEBarColData}
\title{Bar plot of colData.}
\usage{
plotSCEBarColData(
  inSCE,
  coldata,
  groupby = NULL,
  violin = TRUE,
  boxplot = TRUE,
  dots = TRUE,
  xlab = NULL,
  ylab = NULL,
  axisLabelSize = 10,
  dotSize = 1,
  transparency = 1,
  defaultTheme = TRUE,
  title = NULL,
  titleSize = NULL
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object with saved
dimension reduction components or a variable with saved results. Required.}

\item{coldata}{colData value that will be plotted.}

\item{groupby}{Groupings for each numeric value. A user may input a vector
equal length to the number of the samples in the SingleCellExperiment
object, or can be retrieved from the colData slot. Default NULL.}

\item{violin}{Boolean. Whether to plot the violon densities for each group.
Default \code{TRUE}.}

\item{boxplot}{Boolean. Whether to plot the boxes for each group.
Default \code{TRUE}.}

\item{dots}{Boolean. If TRUE, will plot dots for each violin plot.
Default TRUE.}

\item{xlab}{Character vector. Label for x-axis. Default NULL.}

\item{ylab}{Character vector. Label for y-axis. Default NULL.}

\item{axisLabelSize}{Size of x/y-axis labels. Default 10.}

\item{dotSize}{Size of dots. Default 1.}

\item{transparency}{Transparency of the dots, values will be 0-1. Default 1.}

\item{defaultTheme}{Removes grid in plot and sets axis title size to 10
when TRUE. Default TRUE.}

\item{title}{Title of plot. Default NULL.}

\item{titleSize}{Size of title of plot. Default 15.}

\item{axisSize}{Size of x/y-axis ticks. Default 10.}
}
\description{
Visualizes values stored in the colData slot of a
 SingleCellExperiment object via a violin plot.
}
\examples{
plotSCEBarColData(
  inSCE = mouseBrainSubsetSCE,
  coldata = "age", groupby = "sex"
)
}
