% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotting.R
\name{plotSCEDensityAssayData}
\alias{plotSCEDensityAssayData}
\title{Density plot of assay data.}
\usage{
plotSCEDensityAssayData(
  inSCE,
  sample = NULL,
  useAssay = "counts",
  feature,
  groupby = NULL,
  xlab = NULL,
  ylab = NULL,
  axisSize = 10,
  defaultTheme = TRUE,
  title = NULL,
  titleSize = NULL
)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object with saved dimension reduction
components or a variable with saved results. Required.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.}

\item{useAssay}{Indicate which assay to use. Default "counts".}

\item{feature}{Name of feature stored in assay of SingleCellExperiment
object.}

\item{groupby}{Groupings for each numeric value. A user may input a vector
equal length to the number of the samples in the SingleCellExperiment
object, or can be retrieved from the colData slot. Default NULL.}

\item{xlab}{Character vector. Label for x-axis. Default NULL.}

\item{ylab}{Character vector. Label for y-axis. Default NULL.}

\item{axisSize}{Size of x/y-axis labels. Default 10.}

\item{defaultTheme}{Removes grid in plot and sets axis title size to 10
when TRUE. Default TRUE.}

\item{title}{Title of plot. Default NULL.}

\item{titleSize}{Size of title of plot. Default 15.}
}
\description{
Visualizes values stored in the assay slot of a
 SingleCellExperiment object via a density plot.
}
\examples{
plotSCEDensityAssayData(
  inSCE = mouseBrainSubsetSCE,
  feature = "Sox2"
)
}
