% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotting.R
\name{plotSCEDensityColData}
\alias{plotSCEDensityColData}
\title{Density plot of colData.}
\usage{
plotSCEDensityColData(
  inSCE,
  sample = NULL,
  coldata,
  groupby = NULL,
  xlab = NULL,
  ylab = NULL,
  axisSize = 10,
  defaultTheme = TRUE,
  title = NULL,
  titleSize = NULL,
  cutoff = NULL
)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object with saved dimension reduction
components or a variable with saved results. Required.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.}

\item{coldata}{colData value that will be plotted.}

\item{groupby}{Groupings for each numeric value. A user may input a vector
equal length to the number of the samples in the SingleCellExperiment
object, or can be retrieved from the colData slot. Default NULL.}

\item{xlab}{Character vector. Label for x-axis. Default NULL.}

\item{ylab}{Character vector. Label for y-axis. Default NULL.}

\item{axisSize}{Size of x/y-axis labels. Default 10.}

\item{defaultTheme}{Removes grid in plot and sets axis title size to 10
when TRUE. Default TRUE.}

\item{title}{Title of plot. Default NULL.}

\item{titleSize}{Size of title of plot. Default 15.}

\item{cutoff}{Numeric value. The plot will be annotated with a vertical line
if set. Default NULL.}
}
\description{
Visualizes values stored in the colData slot of a
 SingleCellExperiment object via a density plot.
}
\examples{
plotSCEViolinColData(
  inSCE = mouseBrainSubsetSCE,
  coldata = "age", groupby = "sex"
)
}
