% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotting.R
\name{plotSCEDimReduceColData}
\alias{plotSCEDimReduceColData}
\title{Dimension reduction plot tool for colData}
\usage{
plotSCEDimReduceColData(
  inSCE,
  colorBy = "No Color",
  shape = "No Shape",
  reducedDimName = NULL,
  conditionClass = NULL,
  xlab = NULL,
  ylab = NULL,
  dim1 = NULL,
  dim2 = NULL,
  dotsize = 2,
  transparency = 1,
  defaultTheme = TRUE,
  title = NULL,
  titleSize = 15,
  labelClusters = TRUE,
  legendTitle = NULL
)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object with saved dimension reduction
components or a variable with saved results. Required}

\item{colorBy}{color by a condition(any column of the annotation data).}

\item{shape}{add shapes to each condition.}

\item{reducedDimName}{saved dimension reduction name in the SCtkExperiment
object. Required.}

\item{conditionClass}{class of the annotation data used in colorBy.
Options are NULL, "factor" or "numeric". If NULL, class will default to the
original class. Default NULL.}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{dim1}{1st dimension to be used for plotting. Default is NULL.}

\item{dim2}{2nd dimension to be used for plotting. Default is NULL.}

\item{dotsize}{size of dots. Default 2.}

\item{transparency}{transparency of the dots, values will be 0-1. Default 1.}

\item{defaultTheme}{adds grid to plot when TRUE. Default TRUE.}

\item{title}{title of plot. Default NULL.}

\item{titleSize}{size of title of plot. Default 15.}

\item{labelClusters}{Logical. Whether the cluster labels are plotted.}

\item{legendTitle}{title of legend. Default NULL.
Default FALSE.}
}
\value{
a ggplot of the reduced dimensions.
}
\description{
Plot results of reduced dimensions data and
 colors by annotation data stored in the colData slot.
}
\examples{
plotSCEDimReduceColData(
  inSCE = mouseBrainSubsetSCE, colorBy = "tissue",
  shape = "No Shape", conditionClass = "factor",
  reducedDimName = "TSNE_counts",
  xlab = "tSNE1", ylab = "tSNE2", labelClusters = TRUE
)
}
