% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotting.R
\name{plotSCEDimReduceColData}
\alias{plotSCEDimReduceColData}
\title{Dimension reduction plot tool for colData}
\usage{
plotSCEDimReduceColData(
  inSCE,
  sample = NULL,
  colorBy,
  conditionClass = NULL,
  shape = NULL,
  reducedDimName = NULL,
  xlab = NULL,
  ylab = NULL,
  dim1 = NULL,
  dim2 = NULL,
  bin = NULL,
  binLabel = NULL,
  dotSize = 2,
  transparency = 1,
  defaultTheme = TRUE,
  title = NULL,
  titleSize = 15,
  labelClusters = TRUE,
  legendTitle = NULL
)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object with saved dimension reduction
components or a variable with saved results. Required}

\item{sample}{Character vector. Indicates which sample each cell belongs to.}

\item{colorBy}{Color by a condition(any column of the annotation data).
Required.}

\item{conditionClass}{Class of the annotation data used in colorBy.
Options are NULL, "factor" or "numeric". If NULL, class will default to the
original class. Default NULL.}

\item{shape}{Add shapes to each condition.}

\item{reducedDimName}{Saved dimension reduction matrix name in the SCtkExperiment
object. Required.}

\item{xlab}{Character vector. Label for x-axis. Default NULL.}

\item{ylab}{Character vector. Label for y-axis. Default NULL.}

\item{dim1}{1st dimension to be used for plotting. Can either be a string which specifies
the name of the dimension to be plotted from reducedDims, or a numeric value which specifies
the index of the dimension to be plotted. Default is NULL.}

\item{dim2}{2nd dimension to be used for plotting. Can either be a string which specifies
the name of the dimension to be plotted from reducedDims, or a numeric value which specifies
the index of the dimension to be plotted. Default is NULL.}

\item{bin}{Numeric vector. If single value, will divide the numeric values into the `bin` groups.
If more than one value, will bin numeric values using values as a cut point.}

\item{binLabel}{Character vector. Labels for the bins created by the `bin` parameter.
Default NULL.}

\item{dotSize}{Size of dots. Default 2.}

\item{transparency}{Transparency of the dots, values will be 0-1. Default 1.}

\item{defaultTheme}{adds grid to plot when TRUE. Default TRUE.}

\item{title}{Title of plot. Default NULL.}

\item{titleSize}{Size of title of plot. Default 15.}

\item{labelClusters}{Logical. Whether the cluster labels are plotted.}

\item{legendTitle}{title of legend. Default NULL.
Default FALSE.}
}
\value{
a ggplot of the reduced dimensions.
}
\description{
Plot results of reduced dimensions data and
 colors by annotation data stored in the colData slot.
}
\examples{
plotSCEDimReduceColData(
  inSCE = mouseBrainSubsetSCE, colorBy = "tissue",
  shape = NULL, conditionClass = "factor",
  reducedDimName = "TSNE_counts",
  xlab = "tSNE1", ylab = "tSNE2", labelClusters = TRUE
)

plotSCEDimReduceColData(
  inSCE = mouseBrainSubsetSCE, colorBy = "age",
  shape = NULL, conditionClass = "numeric",
  reducedDimName = "TSNE_counts", bin = c(-Inf, 20, 25, +Inf),
  xlab = "tSNE1", ylab = "tSNE2", labelClusters = FALSE
)
}
