% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotting.R
\name{plotSCEDimReduceFeatures}
\alias{plotSCEDimReduceFeatures}
\title{Dimension reduction plot tool for assay data}
\usage{
plotSCEDimReduceFeatures(
  inSCE,
  feature,
  shape = "No Shape",
  reducedDimName,
  useAssay = "logcounts",
  xlab = NULL,
  ylab = NULL,
  dim1 = NULL,
  dim2 = NULL,
  dotsize = 2,
  transparency = 1,
  defaultTheme = TRUE,
  title = NULL,
  titleSize = 15,
  legendTitle = NULL
)
}
\arguments{
\item{inSCE}{Input SCtkExperiment object with saved dimension reduction
components or a variable with saved results. Required}

\item{feature}{name of feature stored in assay of singleCellExperiment
object. Plot will be colored based on feature value.}

\item{shape}{add shapes to each condition.}

\item{reducedDimName}{saved dimension reduction name in the SCtkExperiment
object. Required.}

\item{useAssay}{Indicate which assay to use. The default is "logcounts"}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{dim1}{1st dimension to be used for plotting. Default is NULL.}

\item{dim2}{2nd dimension to be used for plotting. Default is NULL.
Default is second PCA component for PCA data and NULL otherwise.}

\item{dotsize}{size of dots. Default 2.}

\item{transparency}{transparency of the dots, values will be 0-1. Default 1.}

\item{defaultTheme}{adds grid to plot when TRUE. Default TRUE.}

\item{title}{title of plot. Default NULL.}

\item{titleSize}{size of title of plot. Default 15.}

\item{legendTitle}{title of legend. Default NULL.}
}
\value{
a ggplot of the reduced dimensions.
}
\description{
Plot results of reduced dimensions data and
 colors by feature data stored in the assays slot.
}
\examples{
plotSCEDimReduceFeatures(
  inSCE = mouseBrainSubsetSCE, feature = "Sox2",
  shape = "No Shape", reducedDimName = "TSNE_counts",
  useAssay = "counts", xlab = "tSNE1", ylab = "tSNE2"
)
}
