% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotting.R
\name{plotSCEDimReduceFeatures}
\alias{plotSCEDimReduceFeatures}
\title{Dimension reduction plot tool for assay data}
\usage{
plotSCEDimReduceFeatures(
  inSCE,
  sample = NULL,
  feature,
  shape = NULL,
  reducedDimName,
  useAssay = "logcounts",
  xlab = NULL,
  ylab = NULL,
  axisSize = 10,
  axisLabelSize = 10,
  dim1 = NULL,
  dim2 = NULL,
  bin = NULL,
  binLabel = NULL,
  dotSize = 2,
  transparency = 1,
  colorLow = "white",
  colorMid = "gray",
  colorHigh = "blue",
  defaultTheme = TRUE,
  title = NULL,
  titleSize = 15,
  legendTitle = NULL,
  legendSize = 10,
  legendTitleSize = 12,
  groupBy = NULL,
  combinePlot = "none",
  plotLabels = NULL
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object with saved
dimension reduction components or a variable with saved results. Required}

\item{sample}{Character vector. Indicates which sample each cell belongs to.}

\item{feature}{name of feature stored in assay of singleCellExperiment
object. Plot will be colored based on feature value.}

\item{shape}{add shapes to each condition. Default NULL.}

\item{reducedDimName}{saved dimension reduction name in the
\linkS4class{SingleCellExperiment} object. Required.}

\item{useAssay}{Indicate which assay to use. The default is "logcounts"}

\item{xlab}{Character vector. Label for x-axis. Default NULL.}

\item{ylab}{Character vector. Label for y-axis. Default NULL.}

\item{axisSize}{Size of x/y-axis ticks. Default 10.}

\item{axisLabelSize}{Size of x/y-axis labels. Default 10.}

\item{dim1}{1st dimension to be used for plotting. Can either be a string which specifies
the name of the dimension to be plotted from reducedDims, or a numeric value which specifies
the index of the dimension to be plotted. Default is NULL.}

\item{dim2}{2nd dimension to be used for plotting. Can either be a string which specifies
the name of the dimension to be plotted from reducedDims, or a numeric value which specifies
the index of the dimension to be plotted. Default is NULL.
Default is second PCA component for PCA data and NULL otherwise.}

\item{bin}{Numeric vector. If single value, will divide the numeric values into the `bin` groups.
If more than one value, will bin numeric values using values as a cut point.}

\item{binLabel}{Character vector. Labels for the bins created by the `bin` parameter.
Default NULL.}

\item{dotSize}{Size of dots. Default 2.}

\item{transparency}{Transparency of the dots, values will be 0-1. Default 1.}

\item{colorLow}{Character. A color available from `colors()`.
The color will be used to signify the lowest values on the scale.
Default 'white'.}

\item{colorMid}{Character. A color available from `colors()`.
The color will be used to signify the midpoint on the scale.
Default 'gray'.}

\item{colorHigh}{Character. A color available from `colors()`.
The color will be used to signify the highest values on the scale.
Default 'blue'.}

\item{defaultTheme}{adds grid to plot when TRUE. Default TRUE.}

\item{title}{Title of plot. Default NULL.}

\item{titleSize}{Size of title of plot. Default 15.}

\item{legendTitle}{title of legend. Default NULL.}

\item{legendSize}{size of legend. Default 10.}

\item{legendTitleSize}{size of legend title. Default 12.}

\item{groupBy}{Facet wrap the scatterplot based on value.
Default \code{NULL}.}

\item{combinePlot}{Must be either "all", "sample", or "none". "all" will combine all plots into a single
.ggplot object, while "sample" will output a list of plots separated by sample. Default "none".}

\item{plotLabels}{labels to each plot. If set to "default", will use the name of the samples
as the labels. If set to "none", no label will be plotted.}
}
\value{
a ggplot of the reduced dimension plot of feature data.
}
\description{
Plot results of reduced dimensions data and
 colors by feature data stored in the assays slot.
}
\examples{
plotSCEDimReduceFeatures(
  inSCE = mouseBrainSubsetSCE, feature = "Apoe",
  shape = NULL, reducedDimName = "TSNE_counts",
  useAssay = "counts", xlab = "tSNE1", ylab = "tSNE2"
)
}
