% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSCEHeatmap.R
\name{plotSCEHeatmap}
\alias{plotSCEHeatmap}
\title{Plot heatmap of using data stored in SingleCellExperiment Object}
\usage{
plotSCEHeatmap(
  inSCE,
  useAssay = "logcounts",
  featureIndex = NULL,
  cellIndex = NULL,
  featureAnnotations = NULL,
  cellAnnotations = NULL,
  featureAnnotationColor = NULL,
  cellAnnotationColor = NULL,
  rowDataName = NULL,
  colDataName = NULL,
  rowSplitBy = NULL,
  colSplitBy = NULL,
  rowLabel = FALSE,
  colLabel = FALSE,
  rowDend = TRUE,
  colDend = TRUE,
  scale = TRUE,
  trim = c(-2, 2),
  title = "SCE Heatmap",
  colorScheme = NULL,
  ...
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object.}

\item{useAssay}{character. A string indicating the assay name that
provides the expression level to plot.}

\item{featureIndex}{A vector that can subset the input SCE object by rows
(features). Default \code{NULL}.}

\item{cellIndex}{A vector that can subset the input SCE object by columns
(cells). Default \code{NULL}.}

\item{featureAnnotations}{\code{data.frame}, with \code{rownames} containing
all the features going to be plotted. Character columns should be factors.
Default \code{NULL}.}

\item{cellAnnotations}{\code{data.frame}, with \code{rownames} containing
all the cells going to be plotted. Character columns should be factors.
Default \code{NULL}.}

\item{featureAnnotationColor}{A named list. Customized color settings for
feature labeling. Should match the entries in the \code{featureAnnotations}
or \code{rowDataName}. For each entry, there should be a list/vector of
colors named with categories. Default \code{NULL}.}

\item{cellAnnotationColor}{A named list. Customized color settings for
cell labeling. Should match the entries in the \code{cellAnnotations} or
\code{colDataName}. For each entry, there should be a list/vector of colors
named with categories. Default \code{NULL}.}

\item{rowDataName}{character. The column name(s) in \code{rowData} that need
to be added to the annotation. Default \code{NULL}.}

\item{colDataName}{character. The column name(s) in \code{colData} that need
to be added to the annotation. Default \code{NULL}.}

\item{rowSplitBy}{character. Do semi-heatmap based on the grouping of
this(these) annotation(s). Should exist in either \code{rowDataName} or
\code{names(featureAnnotations)}. Default \code{NULL}.}

\item{colSplitBy}{character. Do semi-heatmap based on the grouping of
this(these) annotation(s). Should exist in either \code{colDataName} or
\code{names(cellAnnotations)}. Default \code{NULL}.}

\item{rowLabel}{Whether to display all the feature names. Default
\code{FALSE}.}

\item{colLabel}{Whether to display all the cell names. Default \code{FALSE}.}

\item{rowDend}{Whether to display row dendrogram. Default \code{TRUE}.}

\item{colDend}{Whether to display column dendrogram. Default \code{TRUE}.}

\item{scale}{Whether to perform z-score scaling on each row. Default
\code{TRUE}.}

\item{trim}{A 2-element numeric vector. Values outside of this range will be
trimmed to their nearst bound. Default \code{c(-2, 2)}}

\item{title}{The main title of the whole plot. Default \code{"SCE Heatmap"}}

\item{colorScheme}{function. A function that generates color code by giving
a value. Can be generated by \code{\link{circlize::colorRamp2}}.
Default \code{NULL}.}

\item{...}{Other arguments passed to \code{\link{ComplexHeatmap::Heatmap}}.}
}
\value{
An \code{\link{ComplexHeatmap::Heatmap}} object
}
\description{
Plot heatmap of using data stored in SingleCellExperiment Object
}
\author{
Yichen Wang
}
