% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSCEHeatmap.R
\name{plotSCEHeatmap}
\alias{plotSCEHeatmap}
\title{Plot heatmap of using data stored in SingleCellExperiment Object}
\usage{
plotSCEHeatmap(
  inSCE,
  useAssay = "logcounts",
  featureIndex = NULL,
  cellIndex = NULL,
  featureAnnotations = NULL,
  cellAnnotations = NULL,
  featureAnnotationColor = NULL,
  cellAnnotationColor = NULL,
  rowDataName = NULL,
  colDataName = NULL,
  rowSplitBy = NULL,
  colSplitBy = NULL,
  rowLabel = FALSE,
  colLabel = FALSE,
  rowDend = TRUE,
  colDend = TRUE,
  scaleRow = scale,
  trim = c(-2, 2),
  title = "SCE Heatmap",
  colorScheme = NULL,
  ...
)
}
\arguments{
\item{inSCE}{SingleCellExperiment object.}

\item{useAssay}{character. A string indicating the assay name that 
provides the expression level to plot.}

\item{featureIndex}{vector, default `NULL`, that can subset the input SCE
object by rows (features)}

\item{cellIndex}{vector, default `NULL`, that can subset the input SCE
object by columns (cells)}

\item{featureAnnotations}{data.frame, with `rownames` containing all the
features going to be plotted, default NULL. Character columns should be
factors.}

\item{cellAnnotations}{data.frame, with `rownames` containing all the
cells going to be plotted, default NULL. Character columns should be
factors.}

\item{featureAnnotationColor}{A named list, default `NULL`. Customized color
setting for features. Should match the entries in the `featureAnnotations`
or `rowDataName`. For each entry, there should be a list/vector of colors
named with classes.}

\item{cellAnnotationColor}{A named list, default `NULL`. Customized color
setting for cells. Should match the entries in the `cellAnnotations` or
`colDataName`. For each entry, there should be a list/vector of colors named
with classes.}

\item{rowDataName}{character, default `NULL`. The column name(s) in
`rowData` that need to be added to the annotation.}

\item{colDataName}{character, default `NULL`. The column name(s) in
`colData` that need to be added to the annotation}

\item{rowSplitBy}{A single character, default `NULL`. Do semi-heatmap based
on the grouping of this annotation. Should exist in either `rowDataName` or
`names(featureAnnotations)`.}

\item{colSplitBy}{A single character, default `NULL`. Do semi-heatmap based
on the grouping of this annotation. Should exist in either `colDataName` or
`names(cellAnnotations)`.}

\item{rowLabel}{A single logical, default `FALSE`. Whether to display all
the feature names.}

\item{colLabel}{A single logical, default `FALSE`. Whether to display all
the cell names.}

\item{rowDend}{A single logical, default `True`. Whether to display row
dendrogram.}

\item{colDend}{A single logical, default `True`. Whether to display column
dendrogram.}

\item{scaleRow}{function, default `scale`. A function to be applied to each
row to scale the data.}

\item{trim}{A 2-element numeric vector, default `c(-2, 2)`. Values outside
of this range will be their nearst bound.}

\item{title}{character, default `"SCE Heatmap"`. The main title of the whole 
plot.}

\item{colorScheme}{function, default NULL. A function that generates color 
code by giving a value. Can be generated by `colorRamp2()`.}

\item{...}{Arguments passed to ComplexHeatmap::Heatmap()}
}
\value{
An ComplexHeatmap::Heatmap object
}
\description{
Plot heatmap of using data stored in SingleCellExperiment Object
}
\author{
Yichen Wang
}
