% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotting.R
\name{plotSCEScatter}
\alias{plotSCEScatter}
\title{Dimension reduction plot tool for all types of data}
\usage{
plotSCEScatter(
  inSCE,
  slot,
  annotation,
  feature = NULL,
  shape = "No Shape",
  reducedDimName = NULL,
  conditionClass = NULL,
  xlab = NULL,
  ylab = NULL,
  dim1 = NULL,
  dim2 = NULL,
  dotsize = 2,
  transparency = 1,
  defaultTheme = TRUE,
  title = NULL,
  titleSize = 15,
  labelClusters = TRUE,
  legendTitle = NULL
)
}
\arguments{
\item{inSCE}{Input SingleCellExperiment object with saved dimension reduction
components or a variable with saved results. Required}

\item{slot}{Desired slot of SingleCellExperiment used for plotting. Possible
options: "assays", "colData", "metadata"}

\item{annotation}{Desired vector within the slot used for plotting.}

\item{feature}{name of feature stored in assay of SingleCellExperiment
object. Will be used only if "assays" slot is chosen. Deafult NULL.}

\item{shape}{add shapes to each condition.}

\item{reducedDimName}{saved dimension reduction name in the SCtkExperiment
object. Required.}

\item{conditionClass}{class of the annotation data used in colorBy. Options
are NULL, "factor" or "numeric". If NULL, class will default to the original
class. Default NULL.}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{dim1}{1st dimension to be used for plotting. Default is NULL.}

\item{dim2}{2nd dimension to be used for plotting. Default is NULL.
Default is second PCA component for PCA data and NULL otherwise.}

\item{dotsize}{size of dots. Default 2.}

\item{transparency}{transparency of the dots, values will be 0-1. Default 1.}

\item{defaultTheme}{adds grid to plot when TRUE. Default TRUE.}

\item{title}{title of plot. Default NULL.}

\item{titleSize}{size of title of plot. Default 15.}

\item{labelClusters}{Logical. Whether the cluster labels are plotted.}

\item{legendTitle}{title of legend. Default NULL.}
}
\value{
a ggplot of the reduced dimensions.
}
\description{
Plot results of reduced dimensions data of counts stored in any
slot in the SingleCellExperiment object.
}
\examples{
plotSCEScatter(
  inSCE = mouseBrainSubsetSCE, legendTitle = NULL,
  slot = "assays", annotation = "counts", feature = "Tspan12",
  reducedDimName = "TSNE_counts", labelClusters = FALSE
)
}
