% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopHVG.R
\name{plotTopHVG}
\alias{plotTopHVG}
\title{Plot highly variable genes}
\usage{
plotTopHVG(
  inSCE,
  method = c("vst", "mean.var.plot", "dispersion", "modelGeneVar"),
  hvgNumber = 2000,
  labelsCount = 20,
  featureDisplay = metadata(inSCE)$featureDisplay
)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object containing the
computations.}

\item{method}{Select either \code{"vst"}, \code{"mean.var.plot"}, 
\code{"dispersion"} or \code{"modelGeneVar"}.}

\item{hvgNumber}{Specify the number of top genes to highlight in red. Default
\code{2000}.}

\item{labelsCount}{Specify the number of data points/genes to label. Should 
be less than \code{hvgNumber}. Default \code{20}.}

\item{featureDisplay}{A character string for the \code{rowData} variable name
to indicate what type of feature ID should be displayed. If set by 
\code{\link{setSCTKDisplayRow}}, will by default use it. If \code{NULL}, will
use \code{rownames(inSCE)}.}
}
\value{
ggplot of HVG metrics
}
\description{
Plot highly variable genes
}
\examples{
data("mouseBrainSubsetSCE", package = "singleCellTK")
mouseBrainSubsetSCE <- runModelGeneVar(mouseBrainSubsetSCE)
plotTopHVG(mouseBrainSubsetSCE, method = "modelGeneVar")
}
\seealso{
\code{\link{runFeatureSelection}}, \code{\link{runSeuratFindHVG}},
\code{\link{runModelGeneVar}}, \code{\link{getTopHVG}}
}
