% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopHVG.R
\name{plotTopHVG}
\alias{plotTopHVG}
\title{Plot highly variable genes}
\usage{
plotTopHVG(inSCE, method = "vst", hvgList)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object containing the computations.}

\item{method}{Select either "vst", "mean.var.plot", "dispersion" or "modelGeneVar".}

\item{hvgList}{Character vector indicating the labels of highly variable genes.}
}
\value{
plot object
}
\description{
Plot highly variable genes
}
