% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopHVG.R
\name{plotTopHVG}
\alias{plotTopHVG}
\title{Plot highly variable genes}
\usage{
plotTopHVG(
  inSCE,
  method = c("vst", "mean.var.plot", "dispersion", "modelGeneVar"),
  hvgList = NULL,
  n = NULL,
  labelsCount = NULL
)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object containing the
computations.}

\item{method}{Select either "vst", "mean.var.plot", "dispersion" or
"modelGeneVar".}

\item{hvgList}{Character vector indicating the labels of highly variable
genes.}

\item{n}{Specify the number of top genes to highlight in red. If
\code{hvgList}
 parameter is not provided, this parameter can be used simply to specify
 the number of top genes to highlight in red.}

\item{labelsCount}{Specify the number of data points/genes to label.
By default, all top genes will be labeled.}
}
\value{
plot object
}
\description{
Plot highly variable genes
}
\examples{
data("mouseBrainSubsetSCE", package = "singleCellTK")
mouseBrainSubsetSCE <- scranModelGeneVar(mouseBrainSubsetSCE, "logcounts")
plotTopHVG(mouseBrainSubsetSCE, method = "modelGeneVar",
           n = 1000, labelsCount = 0)
}
