% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSCAN.R
\name{plotTSCANDEgenes}
\alias{plotTSCANDEgenes}
\title{Run plotTSCANDEgenes function to plot cells colored by the expression of a gene of interest}
\usage{
plotTSCANDEgenes(inSCE, geneSymbol, useClusters = NULL, reducedDimName)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{geneSymbol}{Choose the gene of interest from the DE genes in order to know the level of expression of gene in clusters.}

\item{useClusters}{Choose the cluster containing the branch point in the data in order to recompute the pseudotimes so that the root lies at the cluster center, allowing us to detect genes that are associated with the divergence of the branches.}

\item{reducedDimName}{Saved dimension reduction name in the SingleCellExperiment object. Required.}
}
\value{
A plots with the cells colored by the expression of a gene of interest.
}
\description{
A wrapper function which plots all the cells in the cluster containing the branch point of the MST in the dataset. Each point is a cell colored by the expression of a gene of interest and the relevant edges of the MST are overlaid on top.
}
\examples{
sce <- runTSCAN (sce, reducedDimName = "PCA", seed = NULL)
sce <- runTSCANDEG(sce, pathIndex = 1, discardCluster = 8)
sce <- runTSCANClusterDEAnalysis(sce, useClusters)
plotTSCANDEgenes(sce, geneSymbol, useClusters=NULL, reducedDimName = "TSNE")
}
\author{
Nida Pervaiz
}
