% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSCAN.R
\name{plotTSCANPseudotimeGenes}
\alias{plotTSCANPseudotimeGenes}
\title{Run plotTSCANPseudotimeGenes function to plot genes with significant changes}
\usage{
plotTSCANPseudotimeGenes(
  inSCE,
  pathIndex,
  direction = c("increasing", "decreasing"),
  n = 10
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{pathIndex}{Path number for which the pseudotime values should be used. PathIndex corresponds to one path from the root node to one of the terminal nodes.}

\item{direction}{Which direction to use. Choices are increasing or decreasing.}

\item{n}{An integer. Only to plot this number of top genes that are increasing/decreasing in expression with increasing pseudotime along the path in the MST. Default 10}
}
\value{
A plot with the top genes that increase/decrease in expression with increasing pseudotime along the path in the MST
}
\description{
A wrapper function which visualizes outputs from the runTSCANDEG function. Plots the genes that increase or decrease in expression with increasing pseudotime along the path in the MST
}
\examples{
data("scExample", package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
rowData(sce)$Symbol = rowData(sce)$feature_name
rownames(sce) = rowData(sce)$Symbol
sce <- singleCellTK::scaterlogNormCounts(sce, assayName = "logcounts")
sce <- scater::runPCA(sce)
sce <- scater::runTSNE (sce, dimred = "PCA")
sce <- runTSCAN (inSCE = sce, useReducedDim = "PCA", seed = NULL)
sce <- runTSCANDEG(inSCE = sce, pathIndex = 3)
plotTSCANPseudotimeGenes(inSCE = sce, pathIndex = 3, direction = "increasing")
}
\author{
Nida Pervaiz
}
