% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSCAN.R
\name{plotTSCANPseudotimeHeatmap}
\alias{plotTSCANPseudotimeHeatmap}
\title{Run plotTSCANPseudotimeHeatmap function to plot heatmap for top genes}
\usage{
plotTSCANPseudotimeHeatmap(inSCE, pathIndex, topN = 50)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{pathIndex}{Path number for which the pseudotime values should be used. PathIndex corresponds to one path from the root node to one of the terminal nodes.}

\item{topN}{An integer. Only to plot this number of top genes along the path in the MST, in terms of log2FC value. Use NULL to cancel the top N subscription. Default 50}
}
\value{
A plot with the top genes that increase in expression with increasing pseudotime along the path in the MST
}
\description{
A wrapper function which visualizes outputs from the runTSCANDEG function. Plots the top genes that increase in expression with increasing pseudotime along the path in the MST
}
\examples{
sce <- runTSCAN (sce, reducedDimName = "PCA", seed = NULL)
sce <- runTSCANDEG(sce, pathIndex = 1, discardCluster = 8)
plotTSCANPseudotimeHeatmap(sce, pathIndex = 1, topN = 10)
}
\author{
Nida Pervaiz
}
