% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSCAN.R
\name{plotTSCANResults}
\alias{plotTSCANResults}
\title{Run plotTSCANResults function to plot MST pseudotime values for cells}
\usage{
plotTSCANResults(inSCE, useReducedDim)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{useReducedDim}{Saved dimension reduction name in the SingleCellExperiment object. Required.}
}
\value{
A plot with the pseudotime ordering of the cells by projecting them onto the MST.
}
\description{
A wrapper function which visualizes outputs from the runTSCAN function. Plots the pseudotime ordering of the cells by projecting them onto the MST
}
\examples{
sce <- runTSCAN (inSCE = sce, useReducedDim = "PCA", seed = NULL)
plotTSCANResults(inSCE = sce, useReducedDim = "TSNE")
}
\author{
Nida Pervaiz
}
