% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSCAN.R
\name{plotTSCANResults}
\alias{plotTSCANResults}
\title{Plot MST pseudotime values for cells}
\usage{
plotTSCANResults(inSCE, useReducedDim)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{useReducedDim}{Saved dimension reduction name in \code{inSCE} object. 
Required.}
}
\value{
A plot with the pseudotime ordering of the cells by projecting them 
onto the MST.
}
\description{
A wrapper function which visualizes outputs from the 
\code{\link{runTSCAN}} function. Plots the pseudotime ordering of the cells 
by projecting them onto the MST
}
\examples{
data("scExample", package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
rowData(sce)$Symbol <- rowData(sce)$feature_name
rownames(sce) <- rowData(sce)$Symbol
sce <- scaterlogNormCounts(sce, assayName = "logcounts")
sce <- runDimReduce(inSCE = sce, method = "scaterPCA", 
                    useAssay = "logcounts", reducedDimName = "PCA")
sce <- runDimReduce(inSCE = sce, method = "rTSNE", useReducedDim = "PCA", 
                    reducedDimName = "TSNE")
sce <- runTSCAN (inSCE = sce, useReducedDim = "PCA", seed = NULL)
plotTSCANResults(inSCE = sce, useReducedDim = "TSNE")
}
\author{
Nida Pervaiz
}
