% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTSNE.R
\name{plotTSNE}
\alias{plotTSNE}
\title{Plot TSNE}
\usage{
plotTSNE(count_data, colorBy = "No Color", shape = "No Shape")
}
\arguments{
\item{count_data}{A SCE object}

\item{colorBy}{The variable to color clusters by}

\item{shape}{Shape of the points}
}
\value{
A TSNE plot
}
\description{
Use this function to plot PCA or tSNE results
}
