% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTSNE.R
\name{plotTSNE}
\alias{plotTSNE}
\title{Plot t-SNE}
\usage{
plotTSNE(countData, colorBy = "No Color", shape = "No Shape",
  reducedDimName = "TSNE", runTSNE = FALSE, useAssay = "logcounts")
}
\arguments{
\item{countData}{A SCE object}

\item{colorBy}{The variable to color clusters by}

\item{shape}{Shape of the points}

\item{reducedDimName}{t-SNE dimension name. The default is TSNE
The toolkit will store data with the pattern <ASSAY>_<ALGORITHM>.}

\item{runTSNE}{Run t-SNE if the reducedDimName does not exist. the Default is
FALSE.}

\item{useAssay}{Indicate which assay to use for t-SNE if you are running it.
Default is "counts"}
}
\value{
A t-SNE plot
}
\description{
Use this function to plot t-SNE results
}
\examples{
data("mouseBrainSubsetSCE")
plotTSNE(mouseBrainSubsetSCE, colorBy = "level1class",
         reducedDimName = "TSNE_counts")

}
