% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTSNE.R
\name{plotTSNE}
\alias{plotTSNE}
\title{Plot TSNE}
\usage{
plotTSNE(count_data, colorBy = "No Color", shape = "No Shape",
  reducedDimName = "TSNE", runTSNE = FALSE, use_assay = "logcounts")
}
\arguments{
\item{count_data}{A SCE object}

\item{colorBy}{The variable to color clusters by}

\item{shape}{Shape of the points}

\item{reducedDimName}{TSNE dimension name. The default is PCA.
The toolkit will store data with the pattern <ASSSAY>_<ALGORITHM>.}

\item{runTSNE}{Run t-SNE if the reducedDimName doesn't exist. the Default is
FALSE.}

\item{use_assay}{Indicate which assay to use for t-SNE if you are running it.
Default is "counts"}
}
\value{
A TSNE plot
}
\description{
Use this function to plot PCA or tSNE results
}
