% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlReports.R
\name{reportCellQC}
\alias{reportCellQC}
\title{Get runCellQC .html report}
\usage{
reportCellQC(inSCE, output_file = NULL, output_dir = NULL, study = NULL)
}
\arguments{
\item{inSCE}{A \link[SingleCellExperiment]{SingleCellExperiment} object containing
the filtered count matrix with the output from runCellQC function}

\item{output_file}{name of the generated file. If NULL/default then the output file name will be based on the name of the Rmarkdown template.}

\item{output_dir}{name of the output directory to save the rendered file. If NULL/default the file is stored to the current working directory}

\item{study}{description of the data set and experiment design. Default is NULL.}
}
\value{
.html file
}
\description{
A  function to generate .html Rmarkdown report containing the visualizations of the runCellQC function output
}
\examples{
\dontrun{
data(scExample, package = "singleCellTK")
sce <- sce[, colData(sce)$type != 'EmptyDroplet']
sce <- runCellQC(sce)
reportCellQC(inSCE = sce)
}
}
