% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlReports.R
\name{reportFindMarker}
\alias{reportFindMarker}
\title{Get findMarkerDiffExp .html report}
\usage{
reportFindMarker(inSCE, output_file = NULL, output_dir = NULL)
}
\arguments{
\item{inSCE}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}
object containing the output from \code{\link{findMarkerDiffExp}} function}

\item{output_file}{name of the generated file. If \code{NULL} then the output
file name will be based on the name of the Rmarkdown template. Default
\code{NULL}.}

\item{output_dir}{name of the output directory to save the rendered file. If
\code{NULL} the file is stored to the current working directory.
Default \code{NULL}.}
}
\value{
An HTML file of the report will be generated at the path specified
in the arguments.
}
\description{
A  function to generate .html Rmarkdown report containing the
visualizations of the \code{\link{findMarkerDiffExp}} function output
}
