% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlReports.R
\name{reportSeuratFeatureSelection}
\alias{reportSeuratFeatureSelection}
\title{Computes an HTML report from the Seurat workflow and returns the output SCE
 object with the computations stored in it.}
\usage{
reportSeuratFeatureSelection(
  inSCE,
  outputFile = NULL,
  outputDir = NULL,
  subtitle = NULL,
  authors = NULL,
  sce = NULL,
  variable.features = 2000,
  runHVG = TRUE,
  plotHVG = TRUE,
  forceRun = FALSE,
  showSession = FALSE,
  pdf = FALSE
)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{outputFile}{Specify the name of the generated output HTML file. If \code{NULL} then the output
file name will be based on the name of the Rmarkdown template. Default
\code{NULL}.}

\item{outputDir}{Specify the name of the output directory to save the
rendered HTML file. If \code{NULL} the file is stored to the current
working directory.}

\item{subtitle}{A \code{character} value specifying the subtitle to use in the
Seurat report. Default \code{NULL}.}

\item{authors}{A \code{character} value specifying the names of the authors
to use in the Seurat report. Default \code{NULL}.}

\item{sce}{A \code{character} value specifying the path of the input
\code{SingleCellExperiment} object.}

\item{variable.features}{A \code{numeric} value indicating the number of
top variable genes to identify in the seurat report. Default is \code{2000}.}

\item{showSession}{A \code{logical} value indicating if session information
should be displayed or not. Default is \code{TRUE}.}

\item{pdf}{A \code{logical} value indicating if a pdf should also be
generated for each figure in the report. Default is \code{TRUE}.}
}
\value{
A \code{SingleCellExperiment} object that has the seurat computations
 stored and can be used to interactively visualize the plots by importing
 in the \code{singleCellTK} user interface.
}
\description{
Computes an HTML report from the Seurat workflow and returns the output SCE
 object with the computations stored in it.
}
