% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBBKNN.R
\name{runBBKNN}
\alias{runBBKNN}
\title{Apply BBKNN batch effect correction method to SingleCellExperiment object}
\usage{
runBBKNN(
  inSCE,
  exprs = "logcounts",
  batchKey = "batch",
  reducedDimName = "BBKNN",
  nComponents = 50L
)
}
\arguments{
\item{inSCE}{SingleCellExperiment object. An object that stores your dataset
and analysis procedures.}

\item{exprs}{character, default `"logcounts"`. A string indicating the name 
of the assay requiring batch correction in "inSCE", should exist in 
`assayNames(inSCE)`.}

\item{batchKey}{character, default `"batch"`. A string indicating the 
field of `colData(inSCE)` that defines different batches.}

\item{reducedDimName}{character, default `"BBKNN"`. The name for the 
corrected low-dimensional representation.}

\item{nComponents}{integer, default `50L`. Number of principle components or 
dimensionality, adopted in the pre-PCA-computation step, the BBKNN step (for 
how many PCs the algorithm takes into account), and the final UMAP 
combination step where the value represent the dimensionality of the updated 
reducedDim.}
}
\value{
SingleCellExperiment object with `reducedDim(inSCE, reducedDimName)` 
updated with corrected low-dimentional representation
}
\description{
BBKNN, an extremely fast graph-based data integration algorithm. It modifies 
the neighbourhood construction step to produce a graph that is balanced 
across all batches of the data.
}
\examples{
 
\dontrun{
data('sceBatches', package = 'singleCellTK')
sceCorr <- runBBKNN(sceBatches)
}
}
\references{
Krzysztof Polański et al., 2020
}
