% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scds_doubletdetection.R
\name{runBcds}
\alias{runBcds}
\title{Find doublets/multiplets using \link[scds]{bcds}.}
\usage{
runBcds(
  inSCE,
  sample = NULL,
  seed = 12345,
  ntop = 500,
  srat = 1,
  verb = FALSE,
  retRes = FALSE,
  nmax = "tune",
  varImp = FALSE,
  estNdbl = FALSE
)
}
\arguments{
\item{inSCE}{A \link[SingleCellExperiment]{SingleCellExperiment} object.
Needs \code{counts} in assays slot.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.
\link[scds]{bcds} will be run on cells from each sample
separately. If NULL, then all cells will be processed together.
Default NULL.}

\item{seed}{Seed for the random number generator. Default 12345.}

\item{...}{Additional arguments passed to \link[scds]{bcds}.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object with
 \link[scds]{bcds} output appended to the
 \link[SummarizedExperiment]{colData} slot. The columns include
 \emph{bcds_score} and optionally \emph{bcds_call}.
 Please refer to the documentation of \link[scds]{bcds} for details.
}
\description{
A wrapper function for \link[scds]{bcds}. Annotate
 doublets/multiplets using a binary classification approach to discriminate
 artificial doublets from original data. Generate a doublet
 score for each cell. Infer doublets if \code{estNdbl} is \code{TRUE}.
}
\examples{
data(sce_chcl, package = "scds")
sce <- runBcds(sce_chcl)
}
