% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQC.R
\name{runCellQC}
\alias{runCellQC}
\title{Perform comprehensive single cell QC}
\usage{
runCellQC(
  sce,
  algorithms = c("doubletCells", "cxds", "bcds", "cxds_bcds_hybrid"),
  sample = NULL,
  assayName = "counts",
  seed = 12345,
  ...
)
}
\arguments{
\item{sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object.}

\item{algorithms}{Character vector. Specify which QC algorithms to run.
Available options are "decontX" and "doubletCells".}

\item{sample}{Character vector. Indicates which sample each cell belongs to.
Algorithms will be run on cells from each sample separately.}

\item{assayName}{A string specifying which assay contains the count
matrix for cells. This argument is not used for algorithms
\link[scds]{cxds}, \link[scds]{bcds}, and \link[scds]{cxds_bcds_hybrid}.}

\item{seed}{Seed for the random number generator. Default 12345.}
}
\value{
SingleCellExperiment object containing the outputs of the
 specified algorithms in the \link[SummarizedExperiment]{colData}
of \code{sce}.
}
\description{
A wrapper function to run several QC algorithms on a SingleCellExperiment
object containing cells after empty droplets have been removed.
}
\examples{
data(sce_chcl, package = "scds")
sce <- runCellQC(sce_chcl)
}
