% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQC.R
\name{runCellQC}
\alias{runCellQC}
\title{Perform comprehensive single cell QC}
\usage{
runCellQC(
  inSCE,
  algorithms = c("QCMetrics", "doubletCells", "cxds", "bcds", "cxds_bcds_hybrid",
    "scrublet", "doubletFinder", "decontX"),
  sample = NULL,
  collectionName = NULL,
  geneSetList = NULL,
  geneSetListLocation = "rownames",
  geneSetCollection = NULL,
  useAssay = "counts",
  seed = 12345,
  paramsList = NULL
)
}
\arguments{
\item{inSCE}{A \link[SingleCellExperiment]{SingleCellExperiment} object.}

\item{algorithms}{Character vector. Specify which QC algorithms to run.
Available options are "QCMetrics", "scrublet", "doubletCells", "cxds", "bcds", "cxds_bcds_hybrid", and "decontX".}

\item{sample}{Character vector. Indicates which sample each cell belongs to.
Algorithms will be run on cells from each sample separately.}

\item{collectionName}{Character. Name of a \code{GeneSetCollection} obtained by using one of the importGeneSet* functions. Default \code{NULL}.}

\item{geneSetList}{See \code{runPerCellQC}. Default NULL.}

\item{geneSetListLocation}{See \code{runPerCellQC}. Default NULL.}

\item{geneSetCollection}{See \code{runPerCellQC}. Default NULL.}

\item{useAssay}{A string specifying which assay contains the count
matrix for cells.}

\item{seed}{Seed for the random number generator. Default 12345.}

\item{paramsList}{A list containing parameters for QC functions. Default NULL.}
}
\value{
SingleCellExperiment object containing the outputs of the
 specified algorithms in the \link[SummarizedExperiment]{colData}
of \code{inSCE}.
}
\description{
A wrapper function to run several QC algorithms on a
 SingleCellExperiment
 object containing cells after empty droplets have been removed.
}
\examples{
data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
\dontrun{
sce <- runCellQC(sce)
}
}
