% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBatchCorrection.R, R/runComBat.R
\name{runComBat}
\alias{runComBat}
\title{Apply ComBat batch effect correction method to SingleCellExperiment object}
\usage{
runComBat(
  inSCE,
  useAssay = "logcounts",
  batch = "batch",
  par.prior = TRUE,
  covariates = NULL,
  mean.only = FALSE,
  ref.batch = NULL,
  assayName = "ComBat"
)

runComBat(
  inSCE,
  useAssay = "logcounts",
  batch = "batch",
  par.prior = TRUE,
  covariates = NULL,
  mean.only = FALSE,
  ref.batch = NULL,
  assayName = "ComBat"
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object. Required.}

\item{useAssay}{A single character indicating the name of the assay requiring
batch correction. Default \code{"logcounts"}.}

\item{batch}{A single character indicating a field in
\code{\link[SummarizedExperiment]{colData}} that annotates the batches.
Default \code{"batch"}.}

\item{par.prior}{A logical scalar. TRUE indicates parametric adjustments
will be used, FALSE indicates non-parametric adjustments will be used.
Default \code{TRUE}.}

\item{covariates}{List of other column names in colData to be added to the
ComBat model as covariates. Default \code{NULL}.}

\item{mean.only}{If TRUE ComBat only corrects the mean of the batch effect.
Default \code{FALSE}.}

\item{ref.batch}{If given, will use the selected batch as a reference for
batch adjustment. Default \code{NULL}.}

\item{assayName}{A single characeter. The name for the corrected assay. Will
be saved to \code{\link[SummarizedExperiment]{assay}}. Default
\code{"ComBat"}.}
}
\value{
The input \linkS4class{SingleCellExperiment} object with
\code{assay(inSCE, assayName)} updated.

The input \linkS4class{SingleCellExperiment} object with
\code{assay(inSCE, assayName)} updated.
}
\description{
The ComBat batch adjustment approach assumes that batch effects represent
non-biological but systematic shifts in the mean or variability of genomic
features for all samples within a processing batch. It uses either parametric
or non-parametric empirical Bayes frameworks for adjusting data for batch
effects.

The ComBat batch adjustment approach assumes that batch effects represent
non-biological but systematic shifts in the mean or variability of genomic
features for all samples within a processing batch. It uses either parametric
or non-parametric empirical Bayes frameworks for adjusting data for batch
effects.
}
\examples{
\dontrun{
data('sceBatches', package = 'singleCellTK')
# parametric adjustment
sceCorr <- runComBat(sceBatches)
# non-parametric adjustment, mean-only version
sceCorr <- runComBat(sceBatches, par.prior=FALSE, mean.only=TRUE)
# reference-batch version, with covariates
sceCorr <- runComBat(sceBatches, covariates = "cell_type", ref.batch = 'w')
}
data('sceBatches', package = 'singleCellTK')
\dontrun{
# parametric adjustment
sceCorr <- runComBat(sceBatches)
# non-parametric adjustment, mean-only version
sceCorr <- runComBat(sceBatches, par.prior=FALSE, mean.only=TRUE)
# reference-batch version, with covariates
sceCorr <- runComBat(sceBatches, covariates = "cell_type", ref.batch = 'w')
}
}
