% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scds_doubletdetection.R
\name{runCxds}
\alias{runCxds}
\title{Find doublets/multiplets using \link[scds]{cxds}.}
\usage{
runCxds(
  inSCE,
  sample = NULL,
  seed = 12345,
  ntop = 500,
  binThresh = 0,
  verb = FALSE,
  retRes = FALSE,
  estNdbl = FALSE,
  useAssay = "counts"
)
}
\arguments{
\item{inSCE}{A \linkS4class{SingleCellExperiment} object.}

\item{sample}{Character vector or colData variable name. Indicates which 
sample each cell belongs to. Default \code{NULL}.}

\item{seed}{Seed for the random number generator, can be \code{NULL}. Default
\code{12345}.}

\item{ntop}{See \link[scds]{cxds} for more information. Default \code{500}.}

\item{binThresh}{See \link[scds]{cxds} for more information. Default 
\code{0}.}

\item{verb}{See \link[scds]{cxds} for more information. Default \code{FALSE}.}

\item{retRes}{See \link[scds]{cxds} for more information. Default 
\code{FALSE}.}

\item{estNdbl}{See \link[scds]{cxds} for more information. Default 
\code{FALSE}.}

\item{useAssay}{A string specifying which assay in the SCE to use. Default 
\code{"counts"}}
}
\value{
A \linkS4class{SingleCellExperiment} object with \link[scds]{cxds} 
output appended to the \link{colData} slot. The columns include 
\emph{cxds_score} and optionally \emph{cxds_call}.
}
\description{
A wrapper function for \link[scds]{cxds}. Annotate
doublets/multiplets using co-expression based approach. Generate a doublet
score for each cell. Infer doublets if \code{estNdbl} is \code{TRUE}.
}
\details{
When the argument \code{sample} is specified, \link[scds]{cxds} will
be run on cells from each sample separately. If \code{sample = NULL}, then 
all cells will be processed together.
}
\examples{
data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sce <- runCxds(sce)
}
\seealso{
\code{\link[scds]{cxds}}, \code{\link{plotCxdsResults}}, 
\code{\link{runCellQC}}

Other run functions: 
\code{\link{computeZScore}()},
\code{\link{detectCellOutlier}()},
\code{\link{diffAbundanceFET}()},
\code{\link{runBBKNN}()},
\code{\link{runBarcodeRankDrops}()},
\code{\link{runBcds}()},
\code{\link{runCellQC}()},
\code{\link{runClusterSummaryMetrics}()},
\code{\link{runComBatSeq}()},
\code{\link{runCxdsBcdsHybrid}()},
\code{\link{runDEAnalysis}()},
\code{\link{runDecontX}()},
\code{\link{runDimReduce}()},
\code{\link{runDoubletFinder}()},
\code{\link{runDropletQC}()},
\code{\link{runEmptyDrops}()},
\code{\link{runEnrichR}()},
\code{\link{runFastMNN}()},
\code{\link{runFeatureSelection}()},
\code{\link{runFindMarker}()},
\code{\link{runGSVA}()},
\code{\link{runHarmony}()},
\code{\link{runKMeans}()},
\code{\link{runLimmaBC}()},
\code{\link{runMNNCorrect}()},
\code{\link{runModelGeneVar}()},
\code{\link{runNormalization}()},
\code{\link{runPerCellQC}()},
\code{\link{runSCANORAMA}()},
\code{\link{runSCMerge}()},
\code{\link{runScDblFinder}()},
\code{\link{runScanpyFindClusters}()},
\code{\link{runScanpyFindHVG}()},
\code{\link{runScanpyFindMarkers}()},
\code{\link{runScanpyNormalizeData}()},
\code{\link{runScanpyPCA}()},
\code{\link{runScanpyScaleData}()},
\code{\link{runScanpyTSNE}()},
\code{\link{runScanpyUMAP}()},
\code{\link{runScranSNN}()},
\code{\link{runScrublet}()},
\code{\link{runSeuratFindClusters}()},
\code{\link{runSeuratFindHVG}()},
\code{\link{runSeuratFindMarkers}()},
\code{\link{runSeuratHeatmap}()},
\code{\link{runSeuratICA}()},
\code{\link{runSeuratIntegration}()},
\code{\link{runSeuratJackStraw}()},
\code{\link{runSeuratNormalizeData}()},
\code{\link{runSeuratPCA}()},
\code{\link{runSeuratSCTransform}()},
\code{\link{runSeuratScaleData}()},
\code{\link{runSeuratTSNE}()},
\code{\link{runSeuratUMAP}()},
\code{\link{runSingleR}()},
\code{\link{runSoupX}()},
\code{\link{runTSCAN}()},
\code{\link{runTSCANClusterDEAnalysis}()},
\code{\link{runTSCANDEG}()},
\code{\link{runTSNE}()},
\code{\link{runUMAP}()},
\code{\link{runVAM}()},
\code{\link{runZINBWaVE}()},
\code{\link{scaterCPM}()},
\code{\link{scaterPCA}()},
\code{\link{scaterlogNormCounts}()}
}
\concept{run functions}
