% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scds_doubletdetection.R
\name{runCxds}
\alias{runCxds}
\title{Find doublets/multiplets using \link[scds]{cxds}.}
\usage{
runCxds(sce, sample = NULL, seed = 12345, ...)
}
\arguments{
\item{sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object.
Needs \code{counts} in assays slot.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.
\link[DropletUtils]{emptyDrops} will be run on cells from each sample
separately. If NULL, then all cells will be processed together.
Default NULL.}

\item{seed}{Seed for the random number generator. Default 12345.}

\item{...}{Additional arguments passed to \link[scds]{cxds}.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object with
 \link[scds]{cxds} output appended to the
 \link[SummarizedExperiment]{colData} slot. The columns include
 \emph{cxds_score} and optionally \emph{cxds_call}.
 Please refer to the documentation of \link[scds]{cxds} for details.
}
\description{
A wrapper function for \link[scds]{cxds}. Annotate
 doublets/multiplets using co-expression based approach. Generate a doublet
 score for each cell. Infer doublets if \code{estNdbl} is \code{TRUE}.
}
\examples{
data(sce_chcl, package = "scds")
sce <- runCxds(sce_chcl)
}
