% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scds_doubletdetection.R
\name{runCxds}
\alias{runCxds}
\title{Find doublets/multiplets using \link[scds]{cxds}.}
\usage{
runCxds(
  inSCE,
  sample = NULL,
  seed = 12345,
  ntop = 500,
  binThresh = 0,
  verb = FALSE,
  retRes = FALSE,
  estNdbl = FALSE,
  useAssay = "counts"
)
}
\arguments{
\item{inSCE}{A \link[SingleCellExperiment]{SingleCellExperiment} object.
Needs \code{counts} in assays slot.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.
\link[scds]{cxds} will be run on cells from each sample
separately. If NULL, then all cells will be processed together.
Default NULL.}

\item{seed}{Seed for the random number generator. Default 12345.}

\item{ntop}{See \link[scds]{cxds} for more information. Default \code{500}.}

\item{binThresh}{See \link[scds]{cxds} for more information. Default \code{0}.}

\item{verb}{See \link[scds]{cxds} for more information. Default \code{FALSE}.}

\item{retRes}{See \link[scds]{cxds} for more information. Default \code{FALSE}.}

\item{estNdbl}{See \link[scds]{cxds} for more information. Default \code{FALSE}.}

\item{useAssay}{A string specifying which assay in the SCE to use.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object with
 \link[scds]{cxds} output appended to the
 \link[SummarizedExperiment]{colData} slot. The columns include
 \emph{cxds_score} and optionally \emph{cxds_call}.
 Please refer to the documentation of \link[scds]{cxds} for details.
}
\description{
A wrapper function for \link[scds]{cxds}. Annotate
 doublets/multiplets using co-expression based approach. Generate a doublet
 score for each cell. Infer doublets if \code{estNdbl} is \code{TRUE}.
}
\examples{
data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sce <- runCxds(sce)
}
