% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scds_doubletdetection.R
\name{runCxdsBcdsHybrid}
\alias{runCxdsBcdsHybrid}
\title{Find doublets/multiplets using \link[scds]{cxds_bcds_hybrid}.}
\usage{
runCxdsBcdsHybrid(sce, sample = NULL, seed = 12345, ...)
}
\arguments{
\item{sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object.
Needs \code{counts} in assays slot.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.
\link[scds]{cxds_bcds_hybrid} will be run on cells from each sample
separately. If NULL, then all cells will be processed together.
Default NULL.}

\item{seed}{Seed for the random number generator. Default 12345.}

\item{...}{Additional arguments passed to \link[scds]{cxds_bcds_hybrid}.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object with
 \link[scds]{cxds_bcds_hybrid} output appended to the
 \link[SummarizedExperiment]{colData} slot. The columns include
 \emph{hybrid_score} and optionally \emph{hybrid_call}.
 Please refer to the documentation of \link[scds]{cxds_bcds_hybrid} for
 details.
}
\description{
A wrapper function for \link[scds]{cxds_bcds_hybrid}. Annotate
 doublets/multiplets using a binary classification approach to discriminate
 artificial doublets from original data. Generate a doublet
 score for each cell. Infer doublets if \code{estNdbl} is \code{TRUE}.
}
\examples{
data(sce_chcl, package = "scds")
sce <- runCxdsBcdsHybrid(sce_chcl)
}
