% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEAnalysis.R
\name{runDEAnalysis}
\alias{runDEAnalysis}
\title{Perform differential expression analysis on SCE with specified method
Method supported: 'MAST', 'DESeq2', 'Limma', 'ANOVA'}
\usage{
runDEAnalysis(method = "MAST", ...)
}
\arguments{
\item{method}{A single character for specific method. Choose from 'MAST',
'DESeq2', 'Limma', 'ANOVA'. Required}

\item{...}{Other arguments passed to specific functions. Refer to
\code{\link{runMAST}}, \code{\link{runDESeq2}}, \code{\link{runLimmaDE}},
\code{\link{runANOVA}}}
}
\value{
Input SCE object with \code{metadata(inSCE)} updated with name
\code{"diffExp"} as a \code{list} object. Detail refers to the four child
functions.
}
\description{
Perform differential expression analysis on SCE with specified method
Method supported: 'MAST', 'DESeq2', 'Limma', 'ANOVA'
}
\examples{
data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sce <- runDEAnalysis(inSCE = sce, groupName1 = "Sample1", method = "DESeq2",
 groupName2 = "Sample2", index1 = 1:20, index2 = 21:40,
 analysisName = "DESeq2")
}
