% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDimReduce.R
\name{runDimReduce}
\alias{runDimReduce}
\title{Generic Wrapper function for running dimensionality reduction}
\usage{
runDimReduce(
  inSCE,
  method = c("scaterPCA", "seuratPCA", "seuratICA", "scanpyPCA", "rTSNE", "seuratTSNE",
    "scaterUMAP", "seuratUMAP", "scanpyUMAP", "scanpyTSNE"),
  useAssay = NULL,
  useReducedDim = NULL,
  useAltExp = NULL,
  reducedDimName = method,
  nComponents = 20,
  useFeatureSubset = NULL,
  scale = FALSE,
  seed = 12345,
  ...
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{method}{One from \code{"scaterPCA"}, \code{"seuratPCA"},
\code{"seuratICA"}, \code{"rTSNE"}, \code{"seuratTSNE"}, \code{"scaterUMAP"},
\code{"seuratUMAP"}, \code{"scanpyPCA"}, \code{"scanpyUMAP"} and \code{"scanpyTSNE"}.}

\item{useAssay}{Assay to use for computation. If \code{useAltExp} is
specified, \code{useAssay} has to exist in
\code{assays(altExp(inSCE, useAltExp))}. Default \code{"counts"}.}

\item{useReducedDim}{The low dimension representation to use for embedding
computation. Default \code{NULL}.}

\item{useAltExp}{The subset to use for computation, usually for the
selected variable features. Default \code{NULL}.}

\item{reducedDimName}{The name of the result matrix. Required.}

\item{nComponents}{Specify the number of dimensions to compute with the
selected method in case of PCA/ICA and the number of components to
use in the case of TSNE/UMAP methods.}

\item{useFeatureSubset}{Subset of feature to use for dimension reduction. A
character string indicating a \code{rowData} variable that stores the logical
vector of HVG selection, or a vector that can subset the rows of
\code{inSCE}. Default \code{NULL}.}

\item{scale}{Logical scalar, whether to standardize the expression values.
Default \code{TRUE}.}

\item{seed}{Random seed for reproducibility of results.
Default \code{NULL} will use global seed in use by the R environment.}

\item{...}{The other arguments for running a specific algorithm. Please refer
to the one you use.}
}
\value{
The input \linkS4class{SingleCellExperiment} object with
\code{reducedDim} updated with the result.
}
\description{
Generic Wrapper function for running dimensionality reduction
}
\details{
Wrapper function to run one of the available dimensionality
reduction algorithms integrated within SCTK from \code{\link{scaterPCA}},
\code{\link{runSeuratPCA}}, \code{\link{runSeuratICA}}, \code{\link{runTSNE}},
\code{\link{runSeuratTSNE}}, \code{\link{runUMAP}} and
\code{\link{runSeuratUMAP}}. Users can use an assay by specifying
\code{useAssay}, use the assay in an altExp by specifying both
\code{useAltExp} and \code{useAssay}, or use a low-dimensionality
representation by specifying \code{useReducedDim}.
}
\examples{
data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sce <- runNormalization(sce, useAssay = "counts",
                        outAssayName = "logcounts",
                        normalizationMethod = "logNormCounts")
sce <- runDimReduce(inSCE = sce, method = "scaterPCA",
                    useAssay = "logcounts", scale = TRUE,
                    reducedDimName = "PCA")
}
\seealso{
Other run functions: 
\code{\link{computeZScore}()},
\code{\link{detectCellOutlier}()},
\code{\link{diffAbundanceFET}()},
\code{\link{runBBKNN}()},
\code{\link{runBarcodeRankDrops}()},
\code{\link{runBcds}()},
\code{\link{runCellQC}()},
\code{\link{runClusterSummaryMetrics}()},
\code{\link{runComBatSeq}()},
\code{\link{runCxds}()},
\code{\link{runCxdsBcdsHybrid}()},
\code{\link{runDEAnalysis}()},
\code{\link{runDecontX}()},
\code{\link{runDoubletFinder}()},
\code{\link{runDropletQC}()},
\code{\link{runEmptyDrops}()},
\code{\link{runEnrichR}()},
\code{\link{runFastMNN}()},
\code{\link{runFeatureSelection}()},
\code{\link{runFindMarker}()},
\code{\link{runGSVA}()},
\code{\link{runHarmony}()},
\code{\link{runKMeans}()},
\code{\link{runLimmaBC}()},
\code{\link{runMNNCorrect}()},
\code{\link{runModelGeneVar}()},
\code{\link{runNormalization}()},
\code{\link{runPerCellQC}()},
\code{\link{runSCANORAMA}()},
\code{\link{runSCMerge}()},
\code{\link{runScDblFinder}()},
\code{\link{runScanpyFindClusters}()},
\code{\link{runScanpyFindHVG}()},
\code{\link{runScanpyFindMarkers}()},
\code{\link{runScanpyNormalizeData}()},
\code{\link{runScanpyPCA}()},
\code{\link{runScanpyScaleData}()},
\code{\link{runScanpyTSNE}()},
\code{\link{runScanpyUMAP}()},
\code{\link{runScranSNN}()},
\code{\link{runScrublet}()},
\code{\link{runSeuratFindClusters}()},
\code{\link{runSeuratFindHVG}()},
\code{\link{runSeuratFindMarkers}()},
\code{\link{runSeuratHeatmap}()},
\code{\link{runSeuratICA}()},
\code{\link{runSeuratIntegration}()},
\code{\link{runSeuratJackStraw}()},
\code{\link{runSeuratNormalizeData}()},
\code{\link{runSeuratPCA}()},
\code{\link{runSeuratSCTransform}()},
\code{\link{runSeuratScaleData}()},
\code{\link{runSeuratTSNE}()},
\code{\link{runSeuratUMAP}()},
\code{\link{runSingleR}()},
\code{\link{runSoupX}()},
\code{\link{runTSCAN}()},
\code{\link{runTSCANClusterDEAnalysis}()},
\code{\link{runTSCANDEG}()},
\code{\link{runTSNE}()},
\code{\link{runUMAP}()},
\code{\link{runVAM}()},
\code{\link{runZINBWaVE}()},
\code{\link{scaterCPM}()},
\code{\link{scaterPCA}()},
\code{\link{scaterlogNormCounts}()}
}
\concept{run functions}
