% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scran_doubletCells.R
\name{runDoubletCells}
\alias{runDoubletCells}
\title{Detect doublet cells using \link[scran]{doubletCells}.}
\usage{
runDoubletCells(inSCE, sample = NULL, useAssay = "counts", seed = 12345, ...)
}
\arguments{
\item{inSCE}{A \link[SingleCellExperiment]{SingleCellExperiment} object.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.
\link[scran]{doubletCells} will be run on cells from each sample separately.}

\item{useAssay}{A string specifying which assay in the SCE to use.}

\item{seed}{Seed for the random number generator. Default 12345.}

\item{...}{Additional arguments to pass to \link[scran]{doubletCells}.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object with the
 'scran_doubletCell_Score' column added to the
 \link[SummarizedExperiment]{colData} slot.
}
\description{
A wrapper function for \link[scran]{doubletCells}. Identify
 potential doublet cells based on simulations of putative doublet expression
 profiles. Generate a doublet score for each cell.
}
\details{
This function is a wrapper function for \link[scran]{doubletCells}.
 \code{runDoubletCells} runs \link[scran]{doubletCells} for each
 \code{sample} within \code{inSCE} iteratively. The
 resulting doublet scores for all cells will be appended to the
 \link[SummarizedExperiment]{colData} of \code{inSCE}.
}
\examples{
data(emptyDropsSceExample, package = "singleCellTK")
sce <- runDoubletCells(emptyDropsSceExample)
}
\references{
Lun ATL (2018). Detecting doublet cells with scran.
 \url{https://ltla.github.io/SingleCellThoughts/software/
doublet_detection/bycell.html}
}
\seealso{
\link[scran]{doubletCells}
}
