% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletCells.R
\name{runDoubletCells}
\alias{runDoubletCells}
\title{Detect doublet cells using \link[scran]{doubletCells}.}
\usage{
runDoubletCells(sce, sampleColname = "sample", ..., assayType = "counts")
}
\arguments{
\item{sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object. Must
contain a count matrix.}

\item{sampleColname}{Character. The column name which specifies the sample
origin in the \link[SummarizedExperiment]{colData} of the provided
\code{sce} object. Default "sample".}

\item{...}{Additional arguments to pass to \link[scran]{doubletCells}.}

\item{assayType}{A string specifying which assay values contain the count
matrix.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object with the
 'scran_doublet_score' column added to the
 \link[SummarizedExperiment]{colData} slot.
}
\description{
A wrapper function for \link[scran]{doubletCells}. Identify
 potential doublet cells based on simulations of putative doublet expression
 profiles. Generate a doublet score for each cell.
}
\details{
This function is a wrapper function for \link[scran]{doubletCells}.
 It assumes the provided \code{sce} was generated using the import
 functions in this package so there
 will be a "sample" column in the \link[SummarizedExperiment]{colData}
 specifying the sample origin of each cell. The column name can be specified
 using the \code{sampleColname} argument.
 \link{runDoubletCells} runs \link[scran]{doubletCells} for each
 sample within \code{sce} iteratively. The
 resulting doublet scores for all cells will be appended to the
 \link[SummarizedExperiment]{colData} of \code{sce}.
}
\examples{
data(emptyDropsSceExample, package = "scruff")
sce <- runDoubletCells(emptyDropsSceExample)
}
\references{
Lun ATL (2018). Detecting doublet cells with scran.
 \url{https://ltla.github.io/SingleCellThoughts/software/
doublet_detection/bycell.html}
}
\seealso{
\link[scran]{doubletCells}
}
