% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scran_doubletCells.R
\name{runDoubletCells}
\alias{runDoubletCells}
\title{Detect doublet cells using \link[scDblFinder]{scDblFinder}.}
\usage{
runDoubletCells(
  inSCE,
  sample = NULL,
  useAssay = "counts",
  nNeighbors = 50,
  simDoublets = max(10000, ncol(inSCE)),
  seed = 12345,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{inSCE}{A \link[SingleCellExperiment]{SingleCellExperiment} object.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.
\link[scDblFinder]{scDblFinder} will be run on cells from each sample separately.}

\item{useAssay}{A string specifying which assay in the SCE to use.}

\item{nNeighbors}{Number of nearest neighbors used to calculate density for
doublet detection. Default 50.}

\item{simDoublets}{Number of simulated doublets created for doublet
detection. Default 10000.}

\item{seed}{Seed for the random number generator. Default 12345.}

\item{BPPARAM}{A \code{\link{BiocParallelParam}} object specifying whether the
neighbour searches should be parallelized.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object with the
 'scran_doubletCells_score' column added to the
 \link{colData} slot.
}
\description{
A wrapper function for \link[scDblFinder]{scDblFinder}. Identify
 potential doublet cells based on simulations of putative doublet expression
 profiles. Generate a doublet score for each cell.
}
\details{
This function is a wrapper function for \link[scDblFinder]{scDblFinder}.
 \code{runDoubletCells} runs \link[scDblFinder]{scDblFinder} for each
 \code{sample} within \code{inSCE} iteratively. The
 resulting doublet scores for all cells will be appended to the
 \link{colData} of \code{inSCE}.
}
\examples{
data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sce <- runDoubletCells(sce)
}
\references{
Lun ATL (2018). Detecting doublet cells with scran.
 \url{https://ltla.github.io/SingleCellThoughts/software/
doublet_detection/bycell.html}
}
\seealso{
\link[scran]{doubletCells}
}
