% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletFinder_doubletDetection.R
\name{runDoubletFinder}
\alias{runDoubletFinder}
\title{Generates a doublet score for each cell via doubletFinder}
\usage{
runDoubletFinder(sce, sample = NULL, seed = 12345, seurat.pcs = 1:15,
  seurat.res = 2, seurat.nfeatures = 2000, verbose = FALSE)
}
\arguments{
\item{sce}{SingleCellExperiment object. Must contain a counts matrix}

\item{sample}{Numeric vector. Each cell will be assigned a sample number.}

\item{seed}{Seed for the random number generator. Default 12345.}

\item{seurat.pcs}{Numeric vector. The PCs used in seurat function to 
determine number of clusters. Default 1:15.}

\item{seurat.res}{Numeric vector. The resolution parameter used in seurat 
which adjusts the number of clusters determined via the algorithm. Default 2.}

\item{seurat.nfeatures}{Integer. Number of highly variable genes to use. Default 2000.}

\item{verbose}{Boolean. Wheter to print messages from Seurat and DoubletFinder. Default FALSE.}
}
\value{
SingleCellExperiment object containing the
 'doublet_finder_doublet_score'.
}
\description{
Uses doubletFinder to determine cells within the dataset
 suspected to be doublets.
}
\examples{
data(sce_chcl, package = "scds")
sce <- runDoubletFinder(sce_chcl)
}
