% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletFinder_doubletDetection.R
\name{runDoubletFinder}
\alias{runDoubletFinder}
\title{Generates a doublet score for each cell via doubletFinder}
\usage{
runDoubletFinder(
  inSCE,
  useAssay = "counts",
  sample = NULL,
  seed = 12345,
  seuratNfeatures = 2000,
  seuratPcs = 1:15,
  seuratRes = 1.5,
  formationRate = 0.075,
  nCores = NULL,
  verbose = FALSE
)
}
\arguments{
\item{inSCE}{Input SingleCellExperiment object. Must contain a counts matrix}

\item{useAssay}{Indicate which assay to use. Default "counts".}

\item{sample}{Numeric vector. Each cell will be assigned a sample number.}

\item{seed}{Seed for the random number generator. Default 12345.}

\item{seuratNfeatures}{Integer. Number of highly variable genes to use.
Default 2000.}

\item{seuratPcs}{Numeric vector. The PCs used in seurat function to
determine number of clusters. Default 1:15.}

\item{seuratRes}{Numeric vector. The resolution parameter used in seurat,
which adjusts the number of clusters determined via the algorithm.
Default 1.5.}

\item{formationRate}{Doublet formation rate used within algorithm.
Default 0.075.}

\item{nCores}{Number of cores used for running the function.}

\item{verbose}{Boolean. Wheter to print messages from Seurat and
DoubletFinder. Default FALSE.}
}
\value{
SingleCellExperiment object containing the
 'doublet_finder_doublet_score'.
}
\description{
Uses doubletFinder to determine cells within the dataset
 suspected to be doublets.
}
\examples{
data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sce <- runDoubletFinder(sce)
}
