% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQC.R
\name{runDropletQC}
\alias{runDropletQC}
\title{Perform comprehensive droplet QC}
\usage{
runDropletQC(
  sce,
  algorithms = c("QCMetrics", "emptyDrops", "barcodeRanks"),
  sample = NULL,
  assayName = "counts"
)
}
\arguments{
\item{sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object containing
the full droplet count matrix}

\item{algorithms}{Character vector. Specify which QC algorithms to run.
Available options are "emptyDrops" and "barcodeRanks".}

\item{sample}{Character vector. Indicates which sample each cell belongs to.
Algorithms will be run on cells from each sample separately.}

\item{assayName}{A string specifying which assay contains the count
matrix for droplets.}
}
\value{
SingleCellExperiment object containing the outputs of the
 specified algorithms in the \link[SummarizedExperiment]{colData}
of \code{sce}.
}
\description{
A wrapper function to run several QC algorithms for determining
empty droplets in single cell RNA-seq data
}
\examples{
data(emptyDropsSceExample, package = "singleCellTK")
sce <- runDropletQC(emptyDropsSceExample,
  sample = colData(emptyDropsSceExample)$sample)
}
