% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyDrops.R
\name{runEmptyDrops}
\alias{runEmptyDrops}
\title{Identify empty droplets using \link[DropletUtils]{emptyDrops}.}
\usage{
runEmptyDrops(sce, sampleColname = "sample", ..., assayType = "counts")
}
\arguments{
\item{sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object.
Must contain an unfiltered raw counts matrix.}

\item{sampleColname}{Character. The column name which specifies the sample
origin in the \link[SummarizedExperiment]{colData} of the provided
\code{sce} object. Default "sample".}

\item{...}{Additional arguments to pass to \link[DropletUtils]{emptyDrops}.}

\item{assayType}{A string specifying which assay values contain the count
matrix.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object with the
 \link[DropletUtils]{emptyDrops} output table appended to the
 \link[SingleCellExperiment]{colData} slot. The columns include
 \emph{emptyDrops_total}, \emph{emptyDrops_logprob},
 \emph{emptyDrops_pvalue}, \emph{emptyDrops_limited}, \emph{emptyDrops_fdr}.
 Please refer to the documentation of \link[DropletUtils]{emptyDrops} for
 details.
}
\description{
Run \link[DropletUtils]{emptyDrops} on the count matrix in the
 provided \link[SingleCellExperiment]{SingleCellExperiment} object.
 Distinguish between droplets containing cells and ambient RNA in a
 droplet-based single-cell RNA sequencing experiment. This function assumes
 the provided
 \link[SingleCellExperiment]{SingleCellExperiment} object contains cells
 from only one sample and any neccesary subsetting of cells happened
 beforehand.
}
\examples{
# The following unfiltered PBMC_1k_v3 data were downloaded from
# https://support.10xgenomics.com/single-cell-gene-expression/datasets/3.0.0
# /pbmc_1k_v3
# Only the top 10 cells with most counts and the last 10 cells with non-zero
# counts are included in this example.
# This example only serves as an proof of concept and a tutoriol on how to
# run the function. The results should not be
# used for drawing scientific conclusions.
data(emptyDropsSceExample, package = "singleCellTK")
sce <- runEmptyDrops(sce = emptyDropsSceExample)
}
