% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropletUtils_emptyDrops.R
\name{runEmptyDrops}
\alias{runEmptyDrops}
\title{Identify empty droplets using \link[DropletUtils]{emptyDrops}.}
\usage{
runEmptyDrops(
  inSCE,
  sample = NULL,
  useAssay = "counts",
  lower = 100,
  niters = 10000,
  testAmbient = FALSE,
  ignore = NULL,
  alpha = NULL,
  retain = NULL,
  barcodeArgs = list(),
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{inSCE}{Input \link[SingleCellExperiment]{SingleCellExperiment} object.
Must contain a raw counts matrix before empty droplets have been removed.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.
\link[DropletUtils]{emptyDrops} will be run on cells from each sample separately.
If NULL, then all cells will be processed together. Default NULL.}

\item{useAssay}{A string specifying which assay in the SCE to use.}

\item{lower}{See \link[DropletUtils]{emptyDrops} for more information.}

\item{niters}{See \link[DropletUtils]{emptyDrops} for more information.}

\item{testAmbient}{See \link[DropletUtils]{emptyDrops} for more information.}

\item{ignore}{See \link[DropletUtils]{emptyDrops} for more information.}

\item{alpha}{See \link[DropletUtils]{emptyDrops} for more information.}

\item{retain}{See \link[DropletUtils]{emptyDrops} for more information.}

\item{barcodeArgs}{See \link[DropletUtils]{emptyDrops} for more information.}

\item{BPPARAM}{See \link[DropletUtils]{emptyDrops} for more information.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object with the
 \link[DropletUtils]{emptyDrops} output table appended to the
 \link[SummarizedExperiment]{colData} slot. The columns include
 \emph{emptyDrops_total}, \emph{emptyDrops_logprob},
 \emph{emptyDrops_pvalue}, \emph{emptyDrops_limited}, \emph{emptyDrops_fdr}.
 Please refer to the documentation of \link[DropletUtils]{emptyDrops} for
 details.
}
\description{
Run \link[DropletUtils]{emptyDrops} on the count matrix in the
 provided \link[SingleCellExperiment]{SingleCellExperiment} object.
 Distinguish between droplets containing cells and ambient RNA in a
 droplet-based single-cell RNA sequencing experiment.
}
\examples{
# The following unfiltered PBMC_1k_v3 data were downloaded from
# https://support.10xgenomics.com/single-cell-gene-expression/datasets/3.0.0
# /pbmc_1k_v3
# Only the top 10 cells with most counts and the last 10 cells with non-zero
# counts are included in this example.
# This example only serves as an proof of concept and a tutorial on how to
# run the function. The results should not be
# used for drawing scientific conclusions.
data(scExample, package = "singleCellTK")
sce <- runEmptyDrops(inSCE = sce)
}
