% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichRSCE.R
\name{runEnrichR}
\alias{runEnrichR}
\title{Run EnrichR on SCE object}
\usage{
runEnrichR(
  inSCE,
  features,
  analysisName,
  db = NULL,
  by = "rownames",
  featureName = NULL
)
}
\arguments{
\item{inSCE}{A \linkS4class{SingleCellExperiment} object.}

\item{features}{Character vector, selected genes for enrichment analysis.}

\item{analysisName}{A string that identifies each specific analysis.}

\item{db}{Character vector. Selected database name(s) from the enrichR 
database list. If \code{NULL} then EnrichR will be run on all the available 
databases on the enrichR database. See details. Default \code{NULL}}

\item{by}{Character. From where should we find the \code{features}? 
\code{"rownames"} for from \code{rownames(inSCE)}, otherwise, from a column
of feature metadata (\code{rowData(inSCE)[[by]]}). See details. Default 
\code{"rownames"}.}

\item{featureName}{Character. Indicates the actual feature identifiers to be
passed to EnrichR. Can be \code{"rownames"}, a column in feature metadata 
(\code{rowData(inSCE)[[featureName]]}), or a character vector with its length
equals to \code{nrow(inSCE)}. See details. Default \code{"rownames"}.}
}
\value{
Updates \code{inSCE} metadata with a data.frame of enrichment terms 
overlapping in the respective databases along with p-values, z-scores etc.
}
\description{
Run EnrichR on SCE object
}
\details{
EnrichR works by querying the specified \code{features} to its online 
databases, thus it requires the Internet connection. 

Available \code{db} options could be shown by running 
\code{enrichR::listEnrichrDbs()$libraryName}

This function checks for the existence of features in the SCE object. When 
\code{features} do not have a match in \code{rownames(inSCE)}, users may 
try to specify \code{by} to pass the check. 

EnrichR expects gene symbols/names as the input (i.e. Ensembl ID might not 
work). When specified \code{features} are not qualified for this, users may 
try to specify \code{featureName} to change the identifier type to pass to 
EnrichR.
}
\examples{
data("mouseBrainSubsetSCE")
if (Biobase::testBioCConnection()) {
  mouseBrainSubsetSCE <- runEnrichR(mouseBrainSubsetSCE, features = "Cmtm5", 
                                    db = "GO_Cellular_Component_2017",
                                    analysisName = "analysis1")
}

}
\seealso{
\code{\link{getEnrichRResult}}

Other run functions: 
\code{\link{computeZScore}()},
\code{\link{detectCellOutlier}()},
\code{\link{diffAbundanceFET}()},
\code{\link{runBBKNN}()},
\code{\link{runBarcodeRankDrops}()},
\code{\link{runBcds}()},
\code{\link{runCellQC}()},
\code{\link{runClusterSummaryMetrics}()},
\code{\link{runComBatSeq}()},
\code{\link{runCxds}()},
\code{\link{runCxdsBcdsHybrid}()},
\code{\link{runDEAnalysis}()},
\code{\link{runDecontX}()},
\code{\link{runDimReduce}()},
\code{\link{runDoubletFinder}()},
\code{\link{runDropletQC}()},
\code{\link{runEmptyDrops}()},
\code{\link{runFastMNN}()},
\code{\link{runFeatureSelection}()},
\code{\link{runFindMarker}()},
\code{\link{runGSVA}()},
\code{\link{runHarmony}()},
\code{\link{runKMeans}()},
\code{\link{runLimmaBC}()},
\code{\link{runMNNCorrect}()},
\code{\link{runModelGeneVar}()},
\code{\link{runNormalization}()},
\code{\link{runPerCellQC}()},
\code{\link{runSCANORAMA}()},
\code{\link{runSCMerge}()},
\code{\link{runScDblFinder}()},
\code{\link{runScanpyFindClusters}()},
\code{\link{runScanpyFindHVG}()},
\code{\link{runScanpyFindMarkers}()},
\code{\link{runScanpyNormalizeData}()},
\code{\link{runScanpyPCA}()},
\code{\link{runScanpyScaleData}()},
\code{\link{runScanpyTSNE}()},
\code{\link{runScanpyUMAP}()},
\code{\link{runScranSNN}()},
\code{\link{runScrublet}()},
\code{\link{runSeuratFindClusters}()},
\code{\link{runSeuratFindHVG}()},
\code{\link{runSeuratFindMarkers}()},
\code{\link{runSeuratHeatmap}()},
\code{\link{runSeuratICA}()},
\code{\link{runSeuratIntegration}()},
\code{\link{runSeuratJackStraw}()},
\code{\link{runSeuratNormalizeData}()},
\code{\link{runSeuratPCA}()},
\code{\link{runSeuratSCTransform}()},
\code{\link{runSeuratScaleData}()},
\code{\link{runSeuratTSNE}()},
\code{\link{runSeuratUMAP}()},
\code{\link{runSingleR}()},
\code{\link{runSoupX}()},
\code{\link{runTSCAN}()},
\code{\link{runTSCANClusterDEAnalysis}()},
\code{\link{runTSCANDEG}()},
\code{\link{runTSNE}()},
\code{\link{runUMAP}()},
\code{\link{runVAM}()},
\code{\link{runZINBWaVE}()},
\code{\link{scaterCPM}()},
\code{\link{scaterPCA}()},
\code{\link{scaterlogNormCounts}()}
}
\concept{run functions}
