% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichRSCE.R
\name{runEnrichR}
\alias{runEnrichR}
\title{Run EnrichR on SCE object}
\usage{
runEnrichR(
  inSCE,
  glist = NULL,
  geneSetCollectionName = NULL,
  geneSetName = NULL,
  db = NULL
)
}
\arguments{
\item{inSCE}{A \linkS4class{SingleCellExperiment} object.}

\item{glist}{Character vector, selected genes for enrichment analysis using. 
Default \code{NULL}.}

\item{geneSetCollectionName}{Character. The name of an imported geneset 
collection. Default \code{NULL}.}

\item{geneSetName}{Character. The name of a geneset that should be found in 
the geneset collection specified by \code{geneSetCollectionName}. Default 
\code{NULL}.}

\item{db}{Character vector. Selected database name(s) from the enrichR 
database list. If \code{NULL} then enrichR will be run on all the available 
databases on the enrichR database. Default \code{NULL}}
}
\value{
Updates \code{inSCE} metadata with a data.frame of enrichment terms 
overlapping in the respective databases along with p-values, z-scores etc.
}
\description{
Run EnrichR on SCE object
}
\details{
The list of gene to be analyzed could be specified in two ways. 
Either by directly passing a character vector to \code{glist} and leave 
\code{geneSetCollectionName} and \code{geneSetName} as \code{NULL}, or by 
specifying the latter two arguments and leave \code{glist} as \code{NULL}.

Available \code{db} options could be shown by running 
\code{enrichR::listEnrichrDbs()$libraryName}
}
\examples{
data("mouseBrainSubsetSCE")
mouseBrainSubsetSCE <- runEnrichR(mouseBrainSubsetSCE, glist = "Cmtm5", 
                                  db = "GO_Cellular_Component_2017")
}
\seealso{
\code{\link{getEnrichRResult}}
}
