% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFastMNN.R
\name{runFastMNN}
\alias{runFastMNN}
\title{Apply a fast version of the mutual nearest neighbors (MNN) batch effect 
correction method to SingleCellExperiment object}
\usage{
runFastMNN(
  inSCE,
  useAssay = "logcounts",
  reducedDimName = "MNN",
  batch = "batch",
  pcInput = FALSE
)
}
\arguments{
\item{inSCE}{SingleCellExperiment object. An object that stores your dataset
and analysis procedures.}

\item{useAssay}{character, default `"logcounts"`. A string indicating the name 
of the assay requiring batch correction in "inSCE", should exist in 
`assayNames(inSCE)`. Alternatively, see `pcInput` parameter.}

\item{reducedDimName}{character, default `"fastMNN"`. The name for the 
corrected low-dimensional representation.}

\item{batch}{character, default `"batch"`. A string indicating the 
field of `colData(inSCE)` that defines different batches.}

\item{pcInput}{bool, default `FALSE`. Whether to use a reduceDim matrix for 
batch effect correction. If TRUE, `useAssay` should exist in 
`reduceDimNames(inSCE)`. Note that more dimensions in precomputed reducedDim
have shown better correction results.}
}
\value{
SingleCellExperiment object with `reducedDim(inSCE, reducedDimName)` 
updated with corrected low-dimentional representation.
}
\description{
`fastMNN` is a variant of the `mnnCorrect` function, modified for speed and 
more robust performance. For introduction of MNN, see `runMNNCorrect`
}
\examples{
 
data('sceBatches', package = 'singleCellTK')
sceCorr <- runFastMNN(sceBatches, useAssay = 'PCA', pcInput = TRUE)
}
\references{
Lun ATL, et al., 2016
}
