% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFastMNN.R
\name{runFastMNN}
\alias{runFastMNN}
\title{Apply a fast version of the mutual nearest neighbors (MNN) batch effect 
correction method to SingleCellExperiment object}
\usage{
runFastMNN(inSCE, exprs = "logcounts", reducedDimName = "MNN",
  batchKey = "batch", pcInput = FALSE)
}
\arguments{
\item{inSCE}{SingleCellExperiment object. An object that stores your dataset
and analysis procedures.}

\item{exprs}{character, default `"logcounts"`. A string indicating the name 
of the assay requiring batch correction in "inSCE", should exist in 
`assayNames(inSCE)`. Alternatively, see `pcInput` parameter.}

\item{reducedDimName}{character, default `"fastMNN"`. The name for the 
corrected low-dimensional representation.}

\item{batchKey}{character, default `"batch"`. A string indicating the 
field of `colData(inSCE)` that defines different batches.}

\item{pcInput}{bool, default `FALSE`. Whether to use a reduceDim matrix for 
batch effect correction. If TRUE, `exprs` should exist in 
`reduceDimNames(inSCE)`. Note that more dimensions in precomputed reducedDim
have shown better correction results.}
}
\value{
SingleCellExperiment object with `reducedDim(inSCE, reducedDimName)` 
updated with corrected low-dimentional representation.
}
\description{
`fastMNN` is a variant of the `mnnCorrect` function, modified for speed and 
more robust performance. For introduction of MNN, see `runMNNCorrect`
}
\examples{
 
data('sceBatches', package = 'singleCellTK')
sceBatches
## class: SingleCellExperiment 
## dim: 27610 1820 
## metadata(0):
## assays(3): normcounts logcounts
## rownames(27610): GCG MALAT1 ... LOC102724004 LOC102724238
## rowData names(0):
## colnames(1820): reads.12732 reads.12733 ... Sample_1598 Sample_1600
## colData names(2): cell_type1 batch
## reducedDimNames(5): PCA
## spikeNames(0):
sceCorr <- runFastMNN(sceBatches, exprs = 'PCA', pcInput = TRUE)
}
\references{
Lun ATL, et al., 2016
}
