% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBatchCorrection.R, R/runFastMNN.R
\name{runFastMNN}
\alias{runFastMNN}
\title{Apply a fast version of the mutual nearest neighbors (MNN) batch effect
correction method to SingleCellExperiment object}
\usage{
runFastMNN(
  inSCE,
  useAssay = "logcounts",
  reducedDimName = "fastMNN",
  batch = "batch",
  pcInput = FALSE
)

runFastMNN(
  inSCE,
  useAssay = "logcounts",
  reducedDimName = "fastMNN",
  batch = "batch",
  pcInput = FALSE
)
}
\arguments{
\item{inSCE}{inherited object. Required.}

\item{useAssay}{A single character indicating the name of the assay requiring
batch correction. Default \code{"logcounts"}. Alternatively, see
\code{pcInput} parameter.}

\item{reducedDimName}{A single character. The name for the corrected
low-dimensional representation. Will be saved to \code{reducedDim(inSCE)}.
Default \code{"fastMNN"}.}

\item{batch}{A single character indicating a field in
\code{\link[SummarizedExperiment]{colData}} that annotates the batches.
Default \code{"batch"}.}

\item{pcInput}{A logical scalar. Whether to use a low-dimension matrix for
batch effect correction. If \code{TRUE}, \code{useAssay} will be searched
from \code{reducedDimNames(inSCE)}. Default \code{FALSE}.}
}
\value{
The input \linkS4class{SingleCellExperiment} object with
\code{reducedDim(inSCE, reducedDimName)} updated.

The input \linkS4class{SingleCellExperiment} object with
\code{reducedDim(inSCE, reducedDimName)} updated.
}
\description{
fastMNN is a variant of the classic MNN method, modified for speed and more
robust performance. For introduction of MNN, see \code{\link{runMNNCorrect}}.

fastMNN is a variant of the classic MNN method, modified for speed and more
robust performance. For introduction of MNN, see \code{\link{runMNNCorrect}}.
}
\examples{
\dontrun{
data('sceBatches', package = 'singleCellTK')
sceCorr <- runFastMNN(sceBatches, useAssay = 'logcounts', pcInput = FALSE)
}
data('sceBatches', package = 'singleCellTK')
\dontrun{
sceCorr <- runFastMNN(sceBatches, useAssay = 'logcounts', pcInput = FALSE)
}
}
\references{
Lun ATL, et al., 2016

Lun ATL, et al., 2016
}
