% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFeatureSelection.R
\name{runFeatureSelection}
\alias{runFeatureSelection}
\title{Run Variable Feature Detection Methods}
\usage{
runFeatureSelection(
  inSCE,
  useAssay,
  method = c("vst", "mean.var.plot", "dispersion", "modelGeneVar"),
  hvgNumber = 2000,
  rowSubsetName = paste0("HVG_", method, hvgNumber)
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{useAssay}{Specify the name of the assay that should be used. Should use
raw counts for \code{"vst"} method, or a normalized assay for other methods.}

\item{method}{Specify the method to use for variable gene selection.
Options include \code{"vst"}, \code{"mean.var.plot"} or \code{"dispersion"}
from Seurat and \code{"modelGeneVar"} from Scran.}

\item{hvgNumber}{An integer for the number of top HVG to select. Default 
\code{2000}.}

\item{rowSubsetName}{A character string for the \code{rowData} variable name
to store a logical index of selected HVGs. Default 
\code{paste0("HVG_", method, hvgNumber)}}
}
\value{
The input \linkS4class{SingleCellExperiment} object that contains 
\item{}{The computed statistics in the \code{rowData} slot} 
\item{}{A logical vector indicating the selected HVG in the 
\code{rowData} slot, named by \code{rowSubsetName}}
}
\description{
Wrapper function to run all of the feature selection methods 
integrated within the singleCellTK package including three methods from
Seurat (\code{"vst"}, \code{"mean.var.plot"} or \code{dispersion}) and the 
Scran \code{modelGeneVar} method.

This function does not return the names of the variable features but only 
computes the statistics and the selection which will be stored in the 
\code{rowData} slot. To get the names of the variable features, users should 
call \code{\link{getTopHVG}} function after computing the statistics.
}
\examples{
data("mouseBrainSubsetSCE", package = "singleCellTK")
mouseBrainSubsetSCE <- runFeatureSelection(mouseBrainSubsetSCE,
                                           "logcounts",
                                           "modelGeneVar")
}
\seealso{
\code{\link{runModelGeneVar}}, \code{\link{runSeuratFindHVG}},
\code{\link{getTopHVG}}, \code{\link{plotTopHVG}}
}
