% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFeatureSelection.R
\name{runFeatureSelection}
\alias{runFeatureSelection}
\title{Wrapper function to run all of the feature selection methods integrated
within the singleCellTK package including three methods from
Seurat (`vst`, `mean.var.plot` or `dispersion`) and the Scran `modelGeneVar`
method.}
\usage{
runFeatureSelection(
  inSCE,
  useAssay,
  hvgMethod = c("vst", "mean.var.plot", "dispersion", "modelGeneVar")
)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{useAssay}{Specify the name of the assay that should be used.
A normalized assay is recommended for use with this function.}

\item{hvgMethod}{Specify the method to use for variable gene selection.
Options include \code{"vst"}, \code{"mean.var.plot"} or \code{"dispersion"}
from Seurat and \code{"modelGeneVar"} from Scran.}
}
\value{
A \code{SingleCellExperiment} object that contains the computed
 statistics in the \code{rowData} slot of the output object. This function
 does not return the names of the variable features but only computes the
 statistics that are stored in the \code{rowData} slot of the. To get the
 names of the variable features \code{getTopHVG} function should be used
 after computing these statistics.
}
\description{
Wrapper function to run all of the feature selection methods integrated
within the singleCellTK package including three methods from
Seurat (`vst`, `mean.var.plot` or `dispersion`) and the Scran `modelGeneVar`
method.
}
\examples{
data("mouseBrainSubsetSCE", package = "singleCellTK")
mouseBrainSubsetSCE <- runFeatureSelection(mouseBrainSubsetSCE,
                                           "logcounts",
                                           "modelGeneVar")
}
