% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHarmony.R
\name{runHarmony}
\alias{runHarmony}
\title{Apply Harmony batch effect correction method to SingleCellExperiment object}
\usage{
runHarmony(inSCE, exprs = "logcounts", pcInput = FALSE,
  batchKey = "batch", reducedDimName = "HARMONY", nComponents = 50,
  theta = 5, nIter = 10)
}
\arguments{
\item{inSCE}{SingleCellExperiment object. An object that stores your dataset
and analysis procedures.}

\item{exprs}{character, default `"logcounts"`. A string indicating the name 
of the assay requiring batch correction in "inSCE", should exist in 
`assayNames(inSCE)`.}

\item{pcInput}{bool, default `FALSE`. Whether to do correction on a low-dim
matrix. If `TRUE`, will look for `exprs` in `reducedDim(inSCE)`.}

\item{batchKey}{character, default `"batch"`. A string indicating the 
field of `colData(inSCE)` that defines different batches.}

\item{reducedDimName}{character, default `"HARMONY"`. The name for the 
corrected low-dimensional representation.}

\item{nComponents}{integer, default `20`. Number of principle components or 
dimensionality to generate in the resulting reducedDim. If `pcInput`, will
directly follow the dimensionality of the specified reducedDim.}

\item{theta}{Numeric, default 5. Diversity clustering penalty parameter, 
Larger value results in more diverse clusters.}

\item{nIter}{integer, default `10`. The max number of iterations to perform.}
}
\value{
SingleCellExperiment object with `reducedDim(inSCE, reducedDimName)` 
updated with corrected low-dimentional representation.
}
\description{
Harmony is an algorithm that projects cells into a shared embedding in which 
cells group by cell type rather than dataset-specific conditions.
}
\examples{
 
data('sceBatches', package = 'singleCellTK')
sceBatches
## class: SingleCellExperiment 
## dim: 27610 1820 
## metadata(0):
## assays(3): normcounts logcounts
## rownames(27610): GCG MALAT1 ... LOC102724004 LOC102724238
## rowData names(0):
## colnames(1820): reads.12732 reads.12733 ... Sample_1598 Sample_1600
## colData names(2): cell_type1 batch
## reducedDimNames(5): PCA
## spikeNames(0):
sceCorr <- runHarmony(sceBatches, nComponents = 10)
}
\references{
Ilya Korsunsky, et al., 2019
}
