% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBatchCorrection.R
\name{runHarmony}
\alias{runHarmony}
\title{Apply Harmony batch effect correction method to SingleCellExperiment object}
\usage{
runHarmony(
  inSCE,
  useAssay = "logcounts",
  pcInput = FALSE,
  batch = "batch",
  reducedDimName = "HARMONY",
  nComponents = 50L,
  theta = 5,
  nIter = 10L
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object. Required.}

\item{useAssay}{A single character indicating the name of the assay requiring
batch correction. Default \code{"logcounts"}.}

\item{pcInput}{A logical scalar. Whether to use a low-dimension matrix for
batch effect correction. If \code{TRUE}, \code{useAssay} will be searched
from \code{reducedDimNames(inSCE)}. Default \code{FALSE}.}

\item{batch}{A single character indicating a field in
\code{\link[SummarizedExperiment]{colData}} that annotates the batches.
Default \code{"batch"}.}

\item{reducedDimName}{A single character. The name for the corrected
low-dimensional representation. Will be saved to \code{reducedDim(inSCE)}.
Default \code{"HARMONY"}.}

\item{nComponents}{An integer. The number of principle components or
dimensionality to generate in the resulting matrix. If \code{pcInput} is set
to \code{TRUE}, the output dimension will follow the low-dimension matrix,
so this argument will be ignored. Default \code{50L}.}

\item{theta}{A Numeric scalar. Diversity clustering penalty parameter,
Larger value results in more diverse clusters. Default \code{5}}

\item{nIter}{An integer. The max number of iterations to perform. Default
\code{10L}.}
}
\value{
The input \linkS4class{SingleCellExperiment} object with
\code{reducedDim(inSCE, reducedDimName)} updated.
}
\description{
Harmony is an algorithm that projects cells into a shared embedding in which
cells group by cell type rather than dataset-specific conditions.
}
\examples{
data('sceBatches', package = 'singleCellTK')
sceCorr <- runHarmony(sceBatches, nComponents = 10L)
}
\references{
Ilya Korsunsky, et al., 2019
}
