% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBatchCorrection.R
\name{runLIGER}
\alias{runLIGER}
\title{Apply LIGER batch effect correction method to SingleCellExperiment object}
\usage{
runLIGER(
  inSCE,
  useAssay = "logcounts",
  batch = "batch",
  reducedDimName = "LIGER",
  nComponents = 20L,
  lambda = 5,
  resolution = 1
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object. Required.}

\item{useAssay}{A single character indicating the name of the assay requiring
batch correction. Default \code{"logcounts"}.}

\item{batch}{A single character indicating a field in
\code{\link[SummarizedExperiment]{colData}} that annotates the batches.
Default \code{"batch"}.}

\item{reducedDimName}{A single character. The name for the corrected
low-dimensional representation. Will be saved to \code{reducedDim(inSCE)}.
Default \code{"LIGER"}.}

\item{nComponents}{An integer. The number of principle components or
dimensionality to generate in the resulting matrix. Default \code{20L}.}

\item{lambda}{A numeric scalar. Algorithmic parameter, the penalty
parameter which limits the dataset-specific component of the factorization.
Default \code{5.0}.}

\item{resolution}{A numeric scalar. Algorithmic paramter, the clustering
resolution, increasing this increases the number of communities detected.
Default \code{1.0}}
}
\value{
The input \linkS4class{SingleCellExperiment} object with
\code{reducedDim(inSCE, reducedDimName)} updated.
}
\description{
LIGER relies on integrative non-negative matrix factorization to identify
shared and dataset-specific factors.
}
\examples{
\dontrun{
data('sceBatches', package = 'singleCellTK')
sceCorr <- runLIGER(sceBatches)
}
}
\references{
Joshua Welch, et al., 2018
}
