% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLimmaBC.R
\name{runLimmaBC}
\alias{runLimmaBC}
\title{Apply Limma's batch effect correction method to SingleCellExperiment object}
\usage{
runLimmaBC(inSCE, exprs = "logcounts", assayName = "LIMMA",
  batchKey = "batch")
}
\arguments{
\item{inSCE}{SingleCellExperiment object. An object that stores your dataset
and analysis procedures.}

\item{exprs}{character, default `"logcounts"`. A string indicating the name 
of the assay requiring batch correction in "inSCE", should exist in 
`assayNames(inSCE)`.}

\item{assayName}{character, default `"LIMMA"`. The name for the corrected 
full-sized expression matrix.}

\item{batchKey}{character, default `"batch"`. A string indicating the 
field of `colData(inSCE)` that defines different batches.}
}
\value{
SingleCellExperiment object with `assay(inSCE, assayName)` updated 
with corrected full-sized expression matrix.
}
\description{
Limma's batch effect removal function fits a linear model to the data, then 
removes the component due to the batch effects.
}
\examples{
 
data('sceBatches', package = 'singleCellTK')
sceBatches
## class: SingleCellExperiment 
## dim: 27610 1820 
## metadata(0):
## assays(3): normcounts logcounts
## rownames(27610): GCG MALAT1 ... LOC102724004 LOC102724238
## rowData names(0):
## colnames(1820): reads.12732 reads.12733 ... Sample_1598 Sample_1600
## colData names(2): cell_type1 batch
## reducedDimNames(5): PCA
## spikeNames(0):
sceCorr <- runLimmaBC(sceBatches)
}
\references{
Gordon K Smyth, et al., 2003
}
