% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLimmaBC.R
\name{runLimmaBC}
\alias{runLimmaBC}
\title{Apply Limma's batch effect correction method to SingleCellExperiment object}
\usage{
runLimmaBC(inSCE, useAssay = "logcounts", assayName = "LIMMA",
  batch = "batch")
}
\arguments{
\item{inSCE}{SingleCellExperiment object. An object that stores your dataset
and analysis procedures.}

\item{useAssay}{character, default `"logcounts"`. A string indicating the name 
of the assay requiring batch correction in "inSCE", should exist in 
`assayNames(inSCE)`.}

\item{assayName}{character, default `"LIMMA"`. The name for the corrected 
full-sized expression matrix.}

\item{batch}{character, default `"batch"`. A string indicating the 
field of `colData(inSCE)` that defines different batches.}
}
\value{
SingleCellExperiment object with `assay(inSCE, assayName)` updated 
with corrected full-sized expression matrix.
}
\description{
Limma's batch effect removal function fits a linear model to the data, then 
removes the component due to the batch effects.
}
\examples{
 
data('sceBatches', package = 'singleCellTK')
sceCorr <- runLimmaBC(sceBatches)
}
\references{
Gordon K Smyth, et al., 2003
}
