% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMAST.R
\name{runMAST}
\alias{runMAST}
\title{runMAST}
\usage{
runMAST(
  inSCE,
  useAssay = "logcounts",
  index1 = NULL,
  index2 = NULL,
  class = NULL,
  classGroup1 = NULL,
  classGroup2 = NULL,
  comparisonName,
  groupName1,
  groupName2,
  useThresh = FALSE,
  freqExpressed = 0.1,
  onlyPos = FALSE,
  log2fcThreshold = NULL,
  fdrThreshold = 1
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object.}

\item{useAssay}{character. A string specifying which assay to use for the
MAST calculations. Default \code{"logcounts"}.}

\item{index1}{Any type of indices that can subset a
\linkS4class{SingleCellExperiment} inherited object by cells. Specifies
which cells are of interests. Default \code{NULL}.}

\item{index2}{Any type of indices that can subset a
\linkS4class{SingleCellExperiment} inherited object by cells. specifies
which cells are being compared with those specified by \code{index1}. If
\code{NULL}, \code{index1} cells will be comapred with all other cells.
Default \code{NULL}.}

\item{class}{A vector/factor with \code{ncol(inSCE)} elements, or a single
string that specifies a column name of \code{colData(inSCE)}. Default
\code{NULL}.}

\item{classGroup1}{a vector specifying which "levels" given in \code{class}
are of interests. Default \code{NULL}.}

\item{classGroup2}{a vector specifying which "levels" given in \code{class}
will be compared with those specified by \code{classGroup1}. If \code{NULL},
\code{classGroup1} cells will be compared with all other cells.}

\item{comparisonName}{A string naming the DEG experiment. Will be required
in downstream plotting functions.}

\item{groupName1}{A string naming the first comparison group - the condition
you are interested in.}

\item{groupName2}{A string naming the second comparison group - the
condition that the first group is being compared with.}

\item{useThresh}{Whether to use adaptive thresholding to filter genes.
Default \code{FALSE}.}

\item{freqExpressed}{A numeric threshold that the genes expressed in less
than this fraction of cells will be removed. Default \code{0.1}.}

\item{onlyPos}{Whether to only output DEG with positive log2_FC value.
Default \code{FALSE}.}

\item{log2fcThreshold}{Only out put DEGs with the absolute values of log2FC
larger than this value. Default \code{NULL}}

\item{fdrThreshold}{Only out put DEGs with FDR value smaller than this
value. Default \code{1}}
}
\value{
The input \linkS4class{SingleCellExperiment} object with
\code{metadata(inSCE)$MAST} updated with the results: a list named by
\code{comparisonName}, with \code{$groupNames} containing the naming of the
two conditions, \code{$useAssay} storing the assay name that was used for
calculation, \code{$select} storing the cell selection indices (logical) for
each condition, and \code{$result} storing a \code{\link{data.frame}} of
the DEGs summary.
}
\description{
Run MAST for differential expression analysis on a
\linkS4class{SingleCellExperiment} inherited object.
}
\details{
Condition group can be set by one and only one of the two ways: setting
\code{index1} and \code{index2}, or setting \code{class}, \code{classGroup1}
and \code{classGroup2}.
}
\examples{
data(sceBatch)
sce.deg <- runMAST(sceBatch, class = "cell_type", comparisonName = 'aVSb',
                   classGroup1 = 'alpha', classGroup2 = 'beta',
                   groupName1 = 'a', groupName2 = 'b')
}
