% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNormalization.R
\name{runNormalization}
\alias{runNormalization}
\title{Wrapper function to run any of the integrated normalization/transformation
methods in the singleCellTK.}
\usage{
runNormalization(
  inSCE,
  normalizationMethod = NULL,
  useAssay = "counts",
  normAssayName = "customNormalizedAssay",
  scale = FALSE,
  seuratScaleFactor = 10,
  transformation = NULL,
  pseudocountsNorm = NULL,
  pseudocountsTransform = NULL,
  trim = NULL
)
}
\arguments{
\item{inSCE}{input}

\item{normalizationMethod}{normal}

\item{useAssay}{use}

\item{normAssayName}{npr}

\item{scale}{scale}

\item{seuratScaleFactor}{factor}
}
\value{
object
}
\description{
Wrapper function to run any of the integrated normalization/transformation
methods in the singleCellTK.
}
\examples{
data(sce_chcl, package = "scds")
sce_chcl <- runNormalization(
 inSCE = sce_chcl, 
 "SCTransform", 
 "counts", 
 "sctCounts")
}
