% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNormalization.R
\name{runNormalization}
\alias{runNormalization}
\title{Title}
\usage{
runNormalization(
  inSCE,
  normalizationMethod,
  useAssay,
  normAssayName,
  scale = FALSE,
  seuratScaleFactor
)
}
\arguments{
\item{inSCE}{input}

\item{normalizationMethod}{normal}

\item{useAssay}{use}

\item{normAssayName}{npr}

\item{scale}{scale}

\item{seuratScaleFactor}{factor}
}
\value{
object
}
\description{
Title
}
\examples{
data(sce_chcl, package = "scds")
sce_chcl <- runNormalization(
 inSCE = sce_chcl, 
 "SCTransform", 
 "counts", 
 "sctCounts")
}
