% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scater_addPerCellQC.R
\name{runPerCellQC}
\alias{runPerCellQC}
\title{Wrapper for calculating QC metrics with scater.}
\usage{
runPerCellQC(sce, assayName = "counts", geneSetList = NULL,
  geneSetListLocation = "rownames", geneSetCollection = NULL, ...)
}
\arguments{
\item{sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object.}

\item{assayName}{A string specifying which assay in the SCE to use. Default 
'counts'.}

\item{geneSetListLocation}{Character or numeric vector. If set to 'rownames', then the genes in 'geneSetList' will be looked up in \code{rownames(sce)}.
If another character is supplied, then genes will be looked up in the column names of \code{rowData(sce)}. A character vector with the same length as \code{geneSetList} can be supplied if the IDs for different 
gene sets are found in different places, including a mixture of 'rownames' and \code{rowData(sce)}. An integer or integer vector can be supplied to denote the column index in \code{rowData(sce)}. Default 'rownames'.}

\item{...}{Additional arguments to pass to \link[scran]{addPerCellQC}.}

\item{geneSetList.}{List of gene sets to be quantified. The genes in the assays will be matched to the genes in the list based on \code{geneSetListLocation}.}

\item{geneSetCollection.}{Class of \code{GeneSetCollection} from package \link[GSEABase]. The location of the gene IDs in \code{sce} should be in the \code{description} slot of each gene set and should follow the 
same notation as \code{geneSetListLocation}. The function \link[GSEABase]{getGmt} can be used to read in gene sets from a GMT file. If reading a GMT file, the second column for each gene set should be the description denoting the location
of the gene IDs in \code{sce}. These gene sets will be included with those from \code{geneSetList} if both parameters are provided.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object with
 cell QC metrics added to the \link[SummarizedExperiment]{colData} slot. If \code{geneSetList} or \code{geneSetCollection} are provided, then the rownames for each gene set will be saved in \code{metadata(sce)$scran$addPerCellQC$geneSets}.
}
\description{
A wrapper function for \link[celda]{decontX}. Identify
 potential contamination from experimental factors such as ambient RNA.
}
\examples{
data(emptyDropsSceExample, package = "singleCellTK")
mito.ix = grep("^MT-", rowData(emptyDropsSceExample)$feature_name)
geneSet <- list("Mito"=rownames(emptyDropsSceExample)[mito.ix])
sce <- runPerCellQC(emptyDropsSceExample, geneSet = geneSet)
}
