% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQC.R
\name{runQC}
\alias{runQC}
\title{Generates a SingleCellExperiment object containing the output of
 specified QC functions}
\usage{
runQC(
  sce,
  algorithms = c("emptyDrops", "doubletCells"),
  sampleColname = "sample",
  assayType = "counts",
  ...
)
}
\arguments{
\item{sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object. Must
contain a count matrix.}

\item{algorithms}{Character vector. Specify which QC algorithms to run.
Possible options are: "emptyDrops", "doubletCells".}

\item{sampleColname}{Character. The column name which specifies the sample
origin in the \link[SummarizedExperiment]{colData} of the provided
\code{sce} object. Default "sample".}

\item{assayType}{A string specifying which assay values contain the count
matrix.}

\item{...}{Additional arguments to pass to \code{algorithms}.}
}
\value{
SingleCellExperiment object containing the outputs of the
 specified algorithms.
}
\description{
A wrapper function for the individual QC algorithms.
}
\examples{
data(emptyDropsSceExample, package = "scruff")
sce <- runQC(emptyDropsSceExample,
    algorithms = c("emptyDrops", "doubletCells"))
}
