% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBatchCorrection.R, R/runSCANORAMA.R
\name{runSCANORAMA}
\alias{runSCANORAMA}
\title{Apply the mutual nearest neighbors (MNN) batch effect correction method to
SingleCellExperiment object}
\usage{
runSCANORAMA(
  inSCE,
  useAssay = "logcounts",
  batch = "batch",
  SIGMA = 15,
  ALPHA = 0.1,
  KNN = 20L,
  assayName = "SCANORAMA"
)

runSCANORAMA(
  inSCE,
  useAssay = "logcounts",
  batch = "batch",
  SIGMA = 15,
  ALPHA = 0.1,
  KNN = 20L,
  assayName = "SCANORAMA"
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object. Required.}

\item{useAssay}{A single character indicating the name of the assay requiring
batch correction. Default \code{"logcounts"}.}

\item{batch}{A single character indicating a field in
\code{\link[SummarizedExperiment]{colData}} that annotates the batches.
Default \code{"batch"}.}

\item{SIGMA}{A numeric scalar. Algorithmic parameter, correction smoothing
parameter on Gaussian kernel. Default \code{15}.}

\item{ALPHA}{A numeric scalar. Algorithmic parameter, alignment score
minimum cutoff. Default \code{0.1}.}

\item{KNN}{An integer. Algorithmic parameter, number of nearest neighbors to
use for matching. Default \code{20L}.}

\item{assayName}{A single characeter. The name for the corrected assay. Will
be saved to \code{\link[SummarizedExperiment]{assay}}. Default
\code{"SCANORAMA"}.}
}
\value{
The input \linkS4class{SingleCellExperiment} object with
\code{assay(inSCE, assayName)} updated.

The input \linkS4class{SingleCellExperiment} object with
\code{assay(inSCE, assayName)} updated.
}
\description{
SCANORAMA is analogous to computer vision algorithms for panorama stitching
that identify images with overlapping content and merge these into a larger
panorama.

SCANORAMA is analogous to computer vision algorithms for panorama stitching
that identify images with overlapping content and merge these into a larger
panorama.
}
\examples{
\dontrun{
data('sceBatches', package = 'singleCellTK')
sceCorr <- runSCANORAMA(sceBatches)
}
data('sceBatches', package = 'singleCellTK')
\dontrun{
sceCorr <- runSCANORAMA(sceBatches)
}
}
\references{
Brian Hie et al, 2019

Brian Hie et al, 2019
}
