% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpyFunctions.R
\name{runScanpyFindHVG}
\alias{runScanpyFindHVG}
\title{runScanpyFindHVG
Find highly variable genes and store in the input sce object}
\usage{
runScanpyFindHVG(
  inSCE,
  useAssay = "counts",
  method = c("seurat", "cell_ranger", "seurat_v3"),
  hvgNumber = 2000
)
}
\arguments{
\item{inSCE}{(sce) object to compute highly variable genes from and to store
back to it}

\item{useAssay}{Specify the name of the assay to use for computation
of variable genes. It is recommended to use a raw counts assay with the}

\item{method}{selected method to use for computation of highly variable
genes. One of \code{'seurat'}, \code{'cell_ranger'}, or \code{'seurat_v3'}.
Default \code{"seurat"}.}

\item{hvgNumber}{numeric value of how many genes to select as highly
variable. Default \code{2000}}
}
\value{
Updated \code{SingleCellExperiment} object with highly variable genes
computation stored
\code{\link{getTopHVG}}, \code{\link{plotTopHVG}}
}
\description{
runScanpyFindHVG
Find highly variable genes and store in the input sce object
}
\examples{
data(scExample, package = "singleCellTK")
sce <- runScanpyFindHVG(sce)
}
