% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpyFunctions.R
\name{runScanpyNormalizeData}
\alias{runScanpyNormalizeData}
\title{runScanpyNormalizeData
Wrapper for NormalizeData() function from scanpy library
Normalizes the sce object according to the input parameters}
\usage{
runScanpyNormalizeData(
  inSCE,
  useAssay,
  countsPerCellAfter = 10000,
  countsPerCell = NULL,
  minCount = 0,
  normAssayName = "scanpyNormData"
)
}
\arguments{
\item{inSCE}{(sce) object to normalize}

\item{useAssay}{Assay containing raw counts to use for normalization.}

\item{countsPerCellAfter}{If None, after normalization, each cell has a total
count equal to the median of the counts_per_cell before normalization.}

\item{countsPerCell}{Precomputed counts per cell.}

\item{minCount}{Cells with counts less than min_counts are filtered out 
during normalization.}

\item{normAssayName}{Name of new assay containing normalized data. Default
\code{scanpyNormData}.}
}
\value{
Normalized \code{SingleCellExperiment} object
}
\description{
runScanpyNormalizeData
Wrapper for NormalizeData() function from scanpy library
Normalizes the sce object according to the input parameters
}
\examples{
data(scExample, package = "singleCellTK")
\dontrun{
sce <- runScanpyNormalizeData(sce, useAssay = "counts")
}
}
