% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpyFunctions.R
\name{runScanpyNormalizeData}
\alias{runScanpyNormalizeData}
\title{runScanpyNormalizeData
Wrapper for NormalizeData() function from scanpy library
Normalizes the sce object according to the input parameters}
\usage{
runScanpyNormalizeData(
  inSCE,
  useAssay,
  normAssayName = "scanpyNormData",
  normalizationMethod = "LogNormalize"
)
}
\arguments{
\item{inSCE}{(sce) object to normalize}

\item{useAssay}{Assay containing raw counts to use for normalization.}

\item{normAssayName}{Name of new assay containing normalized data. Default
\code{scanpyNormData}.}

\item{normalizationMethod}{selected normalization method. Default
\code{"LogNormalize"}.}
}
\value{
Normalized \code{SingleCellExperiment} object
}
\description{
runScanpyNormalizeData
Wrapper for NormalizeData() function from scanpy library
Normalizes the sce object according to the input parameters
}
\examples{
data(scExample, package = "singleCellTK")
\dontrun{
sce <- runScanpyNormalizeData(sce, useAssay = "counts")
}
}
