% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpyFunctions.R
\name{runScanpyTSNE}
\alias{runScanpyTSNE}
\title{runScanpyTSNE
Computes tSNE from the given sce object and stores the tSNE computations back
into the sce object}
\usage{
runScanpyTSNE(
  inSCE,
  useReduction = "scanpyPCA",
  reducedDimName = "scanpyTSNE",
  dims = 10L,
  perplexity = 15L,
  externalReduction = NULL
)
}
\arguments{
\item{inSCE}{(sce) object on which to compute the tSNE}

\item{useReduction}{selected reduction algorithm to use for computing tSNE.
Default \code{"pca"}.}

\item{reducedDimName}{Name of new reducedDims object containing Scanpy tSNE
Default \code{scanpyTSNE}.}

\item{dims}{Number of reduction components to use for tSNE computation.
Default \code{10}.}

\item{perplexity}{Adjust the perplexity tuneable parameter for the underlying
tSNE call. Default \code{15}.}

\item{externalReduction}{Pass DimReduc object if PCA computed through
other libraries. Default \code{NULL}.}
}
\value{
Updated sce object with tSNE computations stored
}
\description{
runScanpyTSNE
Computes tSNE from the given sce object and stores the tSNE computations back
into the sce object
}
