% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCGEN.R
\name{runSCGEN}
\alias{runSCGEN}
\title{Apply scGen batch effect correction method to SingleCellExperiment object}
\usage{
runSCGEN(
  inSCE,
  useAssay = "logcounts",
  batch = "batch",
  cellType = "cell_type",
  nEpochs = 50L,
  assayName = "SCGEN"
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object. Required.}

\item{useAssay}{A single character indicating the name of the assay requiring
batch correction. Default \code{"logcounts"}.}

\item{batch}{A single character indicating a field in
\code{\link[SummarizedExperiment]{colData}} that annotates the batches.
Default \code{"batch"}.}

\item{cellType}{A single character. A string indicating a field in
\code{colData(inSCE)} that defines different cell types. Default
\code{'cell_type'}.}

\item{nEpochs}{An integer. Algorithmic parameter, the number of epochs to
iterate and optimize network weights. Default \code{50L}.}

\item{assayName}{A single characeter. The name for the corrected assay. Will
be saved to \code{\link[SummarizedExperiment]{assay}}. Default
\code{"SCGEN"}.}
}
\value{
The input \linkS4class{SingleCellExperiment} object with
\code{assay(inSCE, assayName)} updated.
}
\description{
scGen is a generative model to predict single-cell perturbation response
across cell types, studies and species. It works by combining variational
autoencoders and latent space vector arithmetics for high-dimensional single-
cell gene expression data.
}
\examples{
\dontrun{
data('sceBatches', package = 'singleCellTK')
sceCorr <- runSCGEN(sceBatches)
}
}
\references{
Lotfollahi, Mohammad et al., 2019
}
