% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCMerge.R
\name{runSCMerge}
\alias{runSCMerge}
\title{Apply scMerge batch effect correction method to SingleCellExperiment object}
\usage{
runSCMerge(
  inSCE,
  useAssay = "logcounts",
  batch = "batch",
  assayName = "scMerge",
  seg = NULL,
  kmeansK = NULL,
  cellType = "cell_type",
  nCores = parallel::detectCores()
)
}
\arguments{
\item{inSCE}{SingleCellExperiment object. An object that stores your dataset
and analysis procedures.}

\item{useAssay}{character, default `"logcounts"`. A string indicating the name 
of the assay requiring batch correction in "inSCE", should exist in 
`assayNames(inSCE)`.}

\item{batch}{character, default `"batch"`. A string indicating the field 
of `colData(inSCE)` that defines different batches.}

\item{assayName}{character, default `"scMerge"`. The name for the corrected 
full-sized expression matrix.}

\item{seg}{array, default `NULL`. An array of gene names or indices that 
specifies SEG (Stably Expressed Genes) set as negative control. Pre-defined 
dataset with human and mouse SEG lists is available to user by running 
`data('SEG')`.}

\item{kmeansK}{vector of int, default `NULL`. A vector indicating the 
kmeans' K-value for each batch, in order to construct pseudo-replicates. The 
length of `kmeansK` needs to be the same as the number of batches.}

\item{cellType}{character, default `"cell_type"`. A string indicating the 
field of `colData(inSCE)` that defines different cell types.}

\item{nCores}{integer, default `parallel::detectCores()`. The number of 
cores of processors to allocate for the task. By default it takes all the 
cores available to the user.}
}
\value{
SingleCellExperiment object with `assay(inSCE, assayName)` updated 
with corrected full-sized expression matrix.
}
\description{
The scMerge method leverages factor analysis, stably expressed genes (SEGs) 
and (pseudo-) replicates to remove unwanted variations and merge multiple 
scRNA-Seq data.
}
\examples{
data('sceBatches', package = 'singleCellTK')
sceCorr <- runSCMerge(sceBatches, species = 'human')
}
\references{
Hoa, et al., 2020
}
