% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCMerge.R
\name{runSCMerge}
\alias{runSCMerge}
\title{Apply scMerge batch effect correction method to SingleCellExperiment object}
\usage{
runSCMerge(inSCE, exprs = "logcounts", batchKey = "batch",
  assayName = "scMerge", kmeansK = NULL, cellTypeKey = "cell_type",
  species = NULL, seg = NULL, nHVG = 1000)
}
\arguments{
\item{inSCE}{SingleCellExperiment object. An object that stores your dataset
and analysis procedures.}

\item{exprs}{character, default `"logcounts"`. A string indicating the name 
of the assay requiring batch correction in "inSCE", should exist in 
`assayNames(inSCE)`.}

\item{batchKey}{character, default `"batch"`. A string indicating the field 
of `colData(inSCE)` that defines different batches.}

\item{assayName}{character, default `"scMerge"`. The name for the corrected 
full-sized expression matrix.}

\item{kmeansK}{vector of int, default `NULL`. Indicates the kmeans's K for 
each batch. The length of `kmeansK` needs to be the same as the number of 
batches. If not given, this vector will be identified by counting cell types 
in each batch.}

\item{cellTypeKey}{character, default `"cell_type"`. A string indicating the 
field of `colData(inSCE)` that defines different cell types. Only needed 
when `kmeansK` is left to `NULL`.}

\item{species}{character, default `NULL`. Choose from `{"human", "mouse"}`. 
If given, the algorithm will take default species specific SEG (Stably 
Expressed Genes) set as negative control.}

\item{seg}{character vecter, default `NULL`. A gene list that specifies SEG 
(Stably Expressed Genes) set as negative control. If `species` set, will be 
ignored; if both `species` and `seg` are `NULL`, will take some time to 
automatically detect top 1000 SEGs from `assay(inSCE, exprs)`.}

\item{nHVG}{integer, default `1000`. The number of top highly variable genes 
to select per batch.}
}
\value{
SingleCellExperiment object with `assay(inSCE, assayName)` updated 
with corrected full-sized expression matrix.
}
\description{
The scMerge method leverages factor analysis, stably expressed genes (SEGs) 
and (pseudo-) replicates to remove unwanted variations and merge multiple 
scRNA-Seq data.
}
\examples{
data('sceBatches', package = 'singleCellTK')
sceBatches
## class: SingleCellExperiment 
## dim: 27610 1820 
## metadata(0):
## assays(3): normcounts logcounts
## rownames(27610): GCG MALAT1 ... LOC102724004 LOC102724238
## rowData names(0):
## colnames(1820): reads.12732 reads.12733 ... Sample_1598 Sample_1600
## colData names(2): cell_type1 batch
## reducedDimNames(5): PCA
## spikeNames(0):
sceCorr <- runSCMerge(sceBatches, species = 'human')
}
\references{
Hoa, et al., 2020
}
