% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBatchCorrection.R
\name{runSCMerge}
\alias{runSCMerge}
\title{Apply scMerge batch effect correction method to SingleCellExperiment object}
\usage{
runSCMerge(
  inSCE,
  useAssay = "logcounts",
  batch = "batch",
  assayName = "scMerge",
  seg = NULL,
  kmeansK = NULL,
  cellType = "cell_type",
  nCores = 1L
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object. Required.}

\item{useAssay}{A single character indicating the name of the assay requiring
batch correction. Default \code{"logcounts"}.}

\item{batch}{A single character indicating a field in
\code{\link[SummarizedExperiment]{colData}} that annotates the batches.
Default \code{"batch"}.}

\item{assayName}{A single characeter. The name for the corrected assay. Will
be saved to \code{\link[SummarizedExperiment]{assay}}. Default
\code{"scMerge"}.}

\item{seg}{A vector of gene names or indices that specifies SEG (Stably
Expressed Genes) set as negative control. Pre-defined dataset with human and
mouse SEG lists is available to user by running \code{data('SEG')}. Default
\code{NULL}, and this value will be auto-detected by default with
\code{\link[scMerge]{scSEGIndex}}.}

\item{kmeansK}{An integer vector. Indicating the kmeans' K-value for each
batch (i.e. how many subclusters in each batch should exist), in order to
construct pseudo-replicates. The length of code{kmeansK} needs to be the same
as the number of batches. Default \code{NULL}, and this value will be
auto-detected by default, depending on \code{cellType}.}

\item{cellType}{A single character. A string indicating a field in
\code{colData(inSCE)} that defines different cell types. Default
\code{'cell_type'}.}

\item{nCores}{An integer. The number of cores of processors to allocate for
the task. Default \code{1L}.}
}
\value{
The input \linkS4class{SingleCellExperiment} object with
\code{assay(inSCE, assayName)} updated.
}
\description{
The scMerge method leverages factor analysis, stably expressed genes (SEGs)
and (pseudo-) replicates to remove unwanted variations and merge multiple
scRNA-Seq data.
}
\examples{
data('sceBatches', package = 'singleCellTK')
sceCorr <- runSCMerge(sceBatches)
}
\references{
Hoa, et al., 2020
}
