% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubletDetection_scrublet.R,
%   R/scrublet_doubletDetection.R
\name{runScrublet}
\alias{runScrublet}
\title{Estimates doublet rates using Scrublet}
\usage{
runScrublet(sce, sample = NULL, assayName = "counts", seed = 12345)

runScrublet(sce, sample = NULL, assayName = "counts", seed = 12345)
}
\arguments{
\item{sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object.
Needs \code{counts} in assays slot.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.
Scrublet will be run on cells from each sample separately. If NULL, then
all cells will be processed together. Default \code{NULL}.}

\item{seed}{Seed for the random number generator. Default 12345.}

\item{counts_matrix}{full path to a counts matrix file - raw or filtered}

\item{genes}{full path to a gene list usually included in the raw sequecing folder}

\item{output}{full path to an output directory of preference. All plots, output files and log files will be written here}

\item{...}{Any of expected_doublet_rate,min_counts,min_cells, min_gene_variability_pctl, n_prin_comps,umap_n_neighbors, min_dist, angle,}
}
\value{
NA

A \link[SingleCellExperiment]{SingleCellExperiment} object with
 \code{scrub_doublets} output appended to the
 \link[SingleCellExperiment]{colData} slot. The columns include
 \emph{scrublet_score} and \emph{scrublet_call}.
}
\description{
Builds k-nearest neighbors using simulated doublets to estimate likelihood of doublets

A wrapper function that calls \code{scrub_doublets} from python
 module \code{scrublet}. Simulates doublets from the observed data and uses
 a k-nearest-neighbor classifier to calculate a continuous
 \code{scrublet_score} (between 0 and 1) for each transcriptome. The score
 is automatically thresholded to generate \code{scrublet_call}, a boolean
 array that is \code{TRUE} for predicted doublets and \code{FALSE} otherwise.
}
\examples{
runScrublet(counts_matrix=mtx,genes=genes,output=outpath,expected_doublet_rate=0.2)
data(sce_chcl, package = "scds")
sce <- runScrublet(sce_chcl)
}
