% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrublet_doubletDetection.R
\name{runScrublet}
\alias{runScrublet}
\title{Find doublets using \code{scrublet}.}
\usage{
runScrublet(inSCE, sample = NULL, useAssay = "counts", seed = 12345)
}
\arguments{
\item{inSCE}{A \link[SingleCellExperiment]{SingleCellExperiment} object.
Needs \code{counts} in assays slot.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.
Scrublet will be run on cells from each sample separately. If NULL, then
all cells will be processed together. Default \code{NULL}.}

\item{useAssay}{A string specifying which assay in the SCE to use. Default 'counts'.}

\item{seed}{Seed for the random number generator. Default 12345.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object with
 \code{scrub_doublets} output appended to the
 \link[SummarizedExperiment]{colData} slot. The columns include
 \emph{scrublet_score} and \emph{scrublet_call}.
}
\description{
A wrapper function that calls \code{scrub_doublets} from python
 module \code{scrublet}. Simulates doublets from the observed data and uses
 a k-nearest-neighbor classifier to calculate a continuous
 \code{scrublet_score} (between 0 and 1) for each transcriptome. The score
 is automatically thresholded to generate \code{scrublet_call}, a boolean
 array that is \code{TRUE} for predicted doublets and \code{FALSE} otherwise.
}
\examples{
\dontrun{
data(sce_chcl, package = "scds")
inSCE <- runScrublet(sce_chcl)
}
}
