% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSeurat3Integration.R
\name{runSeurat3Integration}
\alias{runSeurat3Integration}
\title{Apply Integration batch effect correction method from Seurat v3 to
SingleCellExperiment object}
\usage{
runSeurat3Integration(
  inSCE,
  useAssay = "logcounts",
  batch = "batch",
  altExpName = "Seurat3Int",
  nAnchors = nrow(inSCE),
  verbose = TRUE
)
}
\arguments{
\item{inSCE}{\linkS4class{SingleCellExperiment} inherited object. Required.}

\item{useAssay}{A single character indicating the name of the assay requiring
batch correction. Default \code{"logcounts"}.}

\item{batch}{A single character indicating a field in
\code{\link[SummarizedExperiment]{colData}} that annotates the batches.
Default \code{"batch"}.}

\item{altExpName}{A single character. The name for the
\code{\link[SingleCellExperiment]{altExp}} that stores the corrected assay.
The name of this assay has the same name. Default \code{"Seurat3Int"}.}

\item{nAnchors}{An integer. The number of features to anchor. The final
number of the corrected features depends on this value. Default
\code{nrow(inSCE)}.}

\item{verbose}{A logical scalar. Whether to show detail information of
the process. Default \code{TRUE}.}
}
\value{
The input \linkS4class{SingleCellExperiment} object with
\code{altExp(inSCE, altExpName)} updated.
}
\description{
Can get either a full-sized corrected assay or a dimension reduced corrected
matrix.
}
\details{
This method aims to first identify anchors between pairs of datasets, that
represent pairwise correspondences between individual cells in each
dataset, that is hypothesized to originate from the same cell state. These
anchors are then used to harmonize the datasets, or transfer information
from one dataset to another.
}
\examples{
data('sceBatches', package = 'singleCellTK')
\dontrun{
sceCorr <- runSeurat3Integration(sceBatches, nAnchors = 100)
}
}
\references{
Stuart et al. 2019
}
