% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{runSeuratICA}
\alias{runSeuratICA}
\title{runSeuratICA
Computes ICA on the input sce object and stores the calculated independent
components within the sce object}
\usage{
runSeuratICA(
  inSCE,
  useAssay,
  reducedDimName = "seuratICA",
  features = NULL,
  nics = 20,
  seed = NULL
)
}
\arguments{
\item{inSCE}{(sce) object on which to compute ICA}

\item{useAssay}{Assay containing scaled counts to use in ICA.}

\item{reducedDimName}{Name of new reducedDims object containing Seurat ICA
Default \code{seuratICA}.}

\item{features}{Specify the feature names or rownames which should be used
for computation of ICA. Default is \code{NULL} which will use the previously
stored variable features.}

\item{nics}{Number of independent components to compute. Default \code{20}.}

\item{seed}{Random seed for reproducibility of results.
Default \code{NULL} will use global seed in use by the R environment.}
}
\value{
Updated \code{SingleCellExperiment} object which now contains the
computed independent components
}
\description{
runSeuratICA
Computes ICA on the input sce object and stores the calculated independent
components within the sce object
}
\examples{
data(scExample, package = "singleCellTK")
\dontrun{
sce <- runSeuratNormalizeData(sce, useAssay = "counts")
sce <- runSeuratFindHVG(sce, useAssay = "counts")
sce <- runSeuratScaleData(sce, useAssay = "counts")
sce <- runSeuratICA(sce, useAssay = "counts")
}
}
