% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{runSeuratPCA}
\alias{runSeuratPCA}
\title{runSeuratPCA
Computes PCA on the input sce object and stores the calculated principal
components within the sce object}
\usage{
runSeuratPCA(
  inSCE,
  useAssay = "seuratScaledData",
  reducedDimName = "seuratPCA",
  nPCs = 20,
  features = NULL,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{inSCE}{(sce) object on which to compute PCA}

\item{useAssay}{Assay containing scaled counts to use in PCA.}

\item{reducedDimName}{Name of new reducedDims object containing Seurat PCA.
Default \code{seuratPCA}.}

\item{nPCs}{numeric value of how many components to compute. Default
\code{20}.}

\item{features}{Specify the feature names or rownames which should be used
for computation of PCA. Default is \code{NULL} which will use the previously
stored variable features.}

\item{seed}{Random seed for reproducibility of results.
Default \code{NULL} will use global seed in use by the R environment.}

\item{verbose}{Logical value indicating if informative messages should
be displayed. Default is \code{TRUE}.}
}
\value{
Updated \code{SingleCellExperiment} object which now contains the
computed principal components
}
\description{
runSeuratPCA
Computes PCA on the input sce object and stores the calculated principal
components within the sce object
}
\examples{
data(scExample, package = "singleCellTK")
\dontrun{
sce <- runSeuratNormalizeData(sce, useAssay = "counts")
sce <- runSeuratFindHVG(sce, useAssay = "counts")
sce <- runSeuratScaleData(sce, useAssay = "counts")
sce <- runSeuratPCA(sce, useAssay = "counts")
}
}
