% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelCellType.R
\name{runSingleR}
\alias{runSingleR}
\title{label cell types with SingleR}
\usage{
runSingleR(
  inSCE,
  useAssay = "logcounts",
  useSCERef = NULL,
  labelColName = NULL,
  useBltinRef = c("hpca", "bpe", "mp", "dice", "immgen", "mouse", "zeisel"),
  level = c("main", "fine", "ont"),
  featureType = c("symbol", "ensembl"),
  labelByCluster = NULL
)
}
\arguments{
\item{inSCE}{In put SCE object}

\item{useAssay}{The assay to calculate}

\item{useSCERef}{A customized reference dataset in SCE format}

\item{labelColName}{The colData column for cell type labeling, stored in
useSCERef}

\item{useBltinRef}{Specify an reference provided by SingleR}

\item{level}{Cell type labeling level, only supported by a few SingleR
references.}

\item{featureType}{Use gene symbols or Ensembl IDs, only supported by SingleR
references.}

\item{labelByCluster}{Label inSCE clusters already identified, instead of
labeling each cells.}
}
\value{
An SCE object
}
\description{
to be filled in
}
