% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelCellType.R
\name{runSingleR}
\alias{runSingleR}
\title{label cell types with SingleR}
\usage{
runSingleR(
  inSCE,
  useAssay = "logcounts",
  useRef = c("hpca", "be", "dice", "nh", "mi", "ig", "mrs"),
  level = c("main", "fine", "ont"),
  featureType = c("symbol", "ensembl")
)
}
\arguments{
\item{inSCE}{In put SCE object}

\item{useAssay}{The assay to calculate}

\item{useRef}{Specify an reference provided by SingleR}

\item{level}{Cell type labeling level}

\item{featureType}{Gene symbols or Ensembl IDs}
}
\value{
An SCE object
}
\description{
to be filled in
}
