% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSCAN.R
\name{runTSCAN}
\alias{runTSCAN}
\title{Run runTSCAN function to obtain pseudotime values for cells}
\usage{
runTSCAN(inSCE, useReducedDim, cluster = NULL, seed = 12345)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{useReducedDim}{Character. Saved dimension reduction name in the SingleCellExperiment object. Required. Used for specifying which low-dimension representation to perform the clustering algorithm and building nearest neighbor graph on. Default "PCA"}

\item{cluster}{Grouping for each cell in inSCE. A user may input a vector equal length to the number of the samples in the SingleCellExperiment object, or can be retrieved from the colData slot. Default NULL.}

\item{seed}{An integer. Set the seed for random process that happens only in "random" generation. Default 12345.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object with pseudotime ordering of the cells along the paths
}
\description{
Wrapper for obtaining a pseudotime ordering of the cells by projecting them onto the MST
}
\examples{
sce <- singleCellTK::importExampleData("pbmc3k")
sce <- singleCellTK::scaterlogNormCounts(sce, assayName = "logcounts")
sce <- scater::runPCA(sce)
sce <- scater::runTSNE (sce, dimred = "PCA")
sce <- runTSCAN (sce, useReducedDim = "PCA", seed = NULL)
}
\author{
Nida Pervaiz
}
