% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSCAN.R
\name{runTSCAN}
\alias{runTSCAN}
\title{Run runTSCAN function to obtain pseudotime values for cells}
\usage{
runTSCAN(inSCE, useReducedDim, cluster = NULL, seed = 12345)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{useReducedDim}{Character. Saved dimension reduction name in 
\code{inSCE} object. Required. Used for specifying which low-dimension 
representation to perform the clustering algorithm and building nearest 
neighbor graph on. Default \code{"PCA"}}

\item{cluster}{Grouping for each cell in \code{inSCE}. A user may input a 
vector equal length to the number of the samples in \code{inSCE}, or can be 
retrieved from the \code{colData} slot. Default \code{NULL}.}

\item{seed}{An integer. Set the seed for random process that happens only in 
"random" generation. Default \code{12345}.}
}
\value{
A \linkS4class{SingleCellExperiment} object with pseudotime ordering 
of the cells along the paths
}
\description{
Wrapper for obtaining a pseudotime ordering of the cells by 
projecting them onto the MST
}
\examples{
data("scExample", package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
rowData(sce)$Symbol <- rowData(sce)$feature_name
rownames(sce) <- rowData(sce)$Symbol
sce <- scaterlogNormCounts(sce, assayName = "logcounts")
sce <- runDimReduce(inSCE = sce, method = "scaterPCA", 
                    useAssay = "logcounts", reducedDimName = "PCA")
sce <- runDimReduce(inSCE = sce, method = "rTSNE", useReducedDim = "PCA", 
                    reducedDimName = "TSNE")
sce <- runTSCAN (inSCE = sce, useReducedDim = "PCA", seed = NULL)
}
\author{
Nida Pervaiz
}
