% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSCAN.R
\name{runTSCANClusterDEAnalysis}
\alias{runTSCANClusterDEAnalysis}
\title{Run runTSCANClusterDEAnalysis function to observe changes between 
paths and to obtain DE genes}
\usage{
runTSCANClusterDEAnalysis(
  inSCE,
  useClusters,
  useAssay = "logcounts",
  fdrThreshold = 0.05
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{useClusters}{Choose the cluster containing the branch point in the data
in order to recompute the pseudotimes so that the root lies at the cluster 
center, allowing us to detect genes that are associated with the divergence 
of the branches.}

\item{useAssay}{Character. The name of the assay to use. This assay should 
contain log normalized counts.}

\item{fdrThreshold}{Only out put DEGs with FDR value smaller than this value.
Default \code{0.05}.}
}
\value{
A \linkS4class{SingleCellExperiment} object with DE genes that are 
significant in our path of interest and are not significant and/or changing 
in the opposite direction in the other paths.
}
\description{
Wrapper for looking for differences in expression between paths 
of a branched trajectory. The differential expression analysis may highlight 
genes which are responsible for the branching event
}
\examples{
data("scExample", package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
rowData(sce)$Symbol <- rowData(sce)$feature_name
rownames(sce) <- rowData(sce)$Symbol
sce <- scaterlogNormCounts(sce, assayName = "logcounts")
sce <- runDimReduce(inSCE = sce, method = "scaterPCA", 
                    useAssay = "logcounts", reducedDimName = "PCA")
sce <- runDimReduce(inSCE = sce, method = "rTSNE", useReducedDim = "PCA", 
                    reducedDimName = "TSNE")
sce <- runTSCAN (inSCE = sce, useReducedDim = "PCA", seed = NULL)
sce <- runTSCANDEG(inSCE = sce, pathIndex = 4)
sce <- runTSCANClusterDEAnalysis(inSCE = sce, useClusters = 5)
}
\author{
Nida Pervaiz
}
