% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVAM.R
\name{runVAM}
\alias{runVAM}
\title{Run VAM to score gene sets in single cell data}
\usage{
runVAM(
  inSCE,
  geneSetCollectionName = "H",
  useAssay = "logcounts",
  resultNamePrefix = NULL,
  center = FALSE,
  gamma = TRUE
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{geneSetCollectionName}{Character. The name of the gene set collection
to use. Default \code{"H"}.}

\item{useAssay}{Character. The name of the assay to use. This assay should
contain log normalized counts. Default \code{"logcounts"}.}

\item{resultNamePrefix}{Character. Prefix to the name the VAM results which
will be stored in the reducedDim slot of \code{inSCE}. The names of the
output matrices will be \code{resultNamePrefix_Distance} and
\code{resultNamePrefix_CDF}. If this parameter is set to \code{NULL}, then
\code{"VAM_geneSetCollectionName_"} will be used. Default \code{NULL}.}

\item{center}{Boolean. If \code{TRUE}, values will be mean centered when
computing the Mahalanobis statistic. Default \code{FALSE}.}

\item{gamma}{Boolean. If \code{TRUE}, a gamma distribution will be fit to
the non-zero squared Mahalanobis distances computed from a row-permuted
version of the gene expression matrix. The estimated gamma distribution will
be used to compute a one-sided p-value for each cell. If \code{FALSE}, the
p-value will be computed using the standard chi-square approximation for the
squared Mahalanobis distance (or non-central if \code{center = FALSE}).
Default \code{TRUE}.}
}
\value{
A \linkS4class{SingleCellExperiment} object with VAM metrics stored
in \code{reducedDim} as \code{VAM_NameOfTheGeneset_Distance} and
\code{VAM_NameOfTheGeneset_CDF}.
}
\description{
Wrapper for the Variance-adjusted Mahalanobis (VAM), which is a
fast and accurate method for cell-specific gene set scoring of single cell
data. This algorithm computes distance statistics and one-sided p-values for
all cells in the specified single cell gene expression matrix. Gene sets
should already be imported and stored in the meta data using functions such
as \link{importGeneSetsFromList} or \link{importGeneSetsFromMSigDB}
}
\examples{
data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sce <- scaterlogNormCounts(sce, assayName = "logcounts")
gs1 <- rownames(sce)[seq(10)]
gs2 <- rownames(sce)[seq(11,20)]
gs <- list("geneset1" = gs1, "geneset2" = gs2)
sce <- importGeneSetsFromList(inSCE = sce,geneSetList = gs,
                              by = "rownames")
sce <- runVAM(inSCE = sce,
              geneSetCollectionName = "GeneSetCollection",
              useAssay = "logcounts")
}
\seealso{
\link{importGeneSetsFromList}, \link{importGeneSetsFromMSigDB},
\link{importGeneSetsFromGMT}, \link{importGeneSetsFromCollection} for
importing gene sets. \link{sctkListGeneSetCollections},
\link{getPathwayResultNames} and \link{getGenesetNamesFromCollection} for
available related information in \code{inSCE}.

Other run functions: 
\code{\link{computeZScore}()},
\code{\link{detectCellOutlier}()},
\code{\link{diffAbundanceFET}()},
\code{\link{runBBKNN}()},
\code{\link{runBarcodeRankDrops}()},
\code{\link{runBcds}()},
\code{\link{runCellQC}()},
\code{\link{runClusterSummaryMetrics}()},
\code{\link{runComBatSeq}()},
\code{\link{runCxds}()},
\code{\link{runCxdsBcdsHybrid}()},
\code{\link{runDEAnalysis}()},
\code{\link{runDecontX}()},
\code{\link{runDimReduce}()},
\code{\link{runDoubletFinder}()},
\code{\link{runDropletQC}()},
\code{\link{runEmptyDrops}()},
\code{\link{runEnrichR}()},
\code{\link{runFastMNN}()},
\code{\link{runFeatureSelection}()},
\code{\link{runFindMarker}()},
\code{\link{runGSVA}()},
\code{\link{runHarmony}()},
\code{\link{runKMeans}()},
\code{\link{runLimmaBC}()},
\code{\link{runMNNCorrect}()},
\code{\link{runModelGeneVar}()},
\code{\link{runNormalization}()},
\code{\link{runPerCellQC}()},
\code{\link{runSCANORAMA}()},
\code{\link{runSCMerge}()},
\code{\link{runScDblFinder}()},
\code{\link{runScanpyFindClusters}()},
\code{\link{runScanpyFindHVG}()},
\code{\link{runScanpyFindMarkers}()},
\code{\link{runScanpyNormalizeData}()},
\code{\link{runScanpyPCA}()},
\code{\link{runScanpyScaleData}()},
\code{\link{runScanpyTSNE}()},
\code{\link{runScanpyUMAP}()},
\code{\link{runScranSNN}()},
\code{\link{runScrublet}()},
\code{\link{runSeuratFindClusters}()},
\code{\link{runSeuratFindHVG}()},
\code{\link{runSeuratFindMarkers}()},
\code{\link{runSeuratHeatmap}()},
\code{\link{runSeuratICA}()},
\code{\link{runSeuratIntegration}()},
\code{\link{runSeuratJackStraw}()},
\code{\link{runSeuratNormalizeData}()},
\code{\link{runSeuratPCA}()},
\code{\link{runSeuratSCTransform}()},
\code{\link{runSeuratScaleData}()},
\code{\link{runSeuratTSNE}()},
\code{\link{runSeuratUMAP}()},
\code{\link{runSingleR}()},
\code{\link{runSoupX}()},
\code{\link{runTSCAN}()},
\code{\link{runTSCANClusterDEAnalysis}()},
\code{\link{runTSCANDEG}()},
\code{\link{runTSNE}()},
\code{\link{runUMAP}()},
\code{\link{runZINBWaVE}()},
\code{\link{scaterCPM}()},
\code{\link{scaterPCA}()},
\code{\link{scaterlogNormCounts}()}
}
\author{
Nida Pervaiz
}
\concept{run functions}
