% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runZINBWaVE.R
\name{runZINBWaVE}
\alias{runZINBWaVE}
\title{Apply ZINBWaVE Batch effect correction mathod to SingleCellExperiment object}
\usage{
runZINBWaVE(
  inSCE,
  exprs = "logcounts",
  batchKey = "batch",
  reducedDimName = "zinbwave",
  nHVG = 3000,
  filterParams = NULL,
  nComponents = 50,
  epsilon = 1000,
  nIter = 10
)
}
\arguments{
\item{inSCE}{SingleCellExperiment object. An object that stores your dataset
and analysis procedures.}

\item{exprs}{character, default `"logcounts"`. A string indicating the name 
of the assay requiring batch correction in "inSCE", should exist in 
`assayNames(inSCE)`.}

\item{batchKey}{character, default `"batch"`. A string indicating the 
field of `colData(inSCE)` that defines different batches.}

\item{reducedDimName}{character, default `"zinbwave"`. The name for the 
corrected low-dimensional representation.}

\item{nHVG}{integer, default `1000`. Number of highly variable genes to use 
when fitting the model}

\item{filterParams}{vector of two number, default `c(5, 5)`. In the form of 
c(a, b), and a filter like `rowSums(assay > a) > b` will be applied, meaning
that we only want the genes that has at least `a` counts in at least `b` 
cells.}

\item{nComponents}{integer, default `50L`. Number of principle components or 
dimensionality to generate in the resulting reducedDim.}

\item{epsilon}{integer, default `1000`. Algorithmic parameter, by default, the 
epsilon parameter is set to the number of genes. We empirically found that a 
high epsilon is often required to obtained a good low-level representation.}

\item{nIter}{integer, default `10`. The max number of iterations to perform.}
}
\description{
A general and flexible zero-inflated negative binomial model that can be 
used to provide a low-dimensional representations of scRNAseq data. The 
model accounts for zero inflation (dropouts), over-dispersion, and the count 
nature of the data. The model also accounts for the difference in library 
sizes and optionally for batch effects and/or other covariates.
}
\examples{
data('sceBatches', package = 'singleCellTK')
sceBatches
## class: SingleCellExperiment 
## dim: 27610 1820 
## metadata(0):
## assays(3): normcounts logcounts
## rownames(27610): GCG MALAT1 ... LOC102724004 LOC102724238
## rowData names(0):
## colnames(1820): reads.12732 reads.12733 ... Sample_1598 Sample_1600
## colData names(2): cell_type1 batch
## reducedDimNames(5): PCA
## spikeNames(0):
sceCorr <- runZINBWaVE(sceBatches, nIter=5)
}
\references{
Pollen, Alex A et al., 2014
}
