% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSummaryStats.R
\name{sampleSummaryStats}
\alias{sampleSummaryStats}
\title{Plot table of SCTK QC outputs.}
\usage{
sampleSummaryStats(
  inSCE,
  sample = NULL,
  useAssay = "counts",
  simple = TRUE,
  output = "kable",
  fileName = "sctkSummaryStats.csv"
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object with saved
dimension reduction components or a variable with saved results. Required}

\item{sample}{Character vector. Indicates which sample each cell belongs to.}

\item{useAssay}{A string specifying which assay in the SCE to use. Default
'counts'.}

\item{simple}{Boolean. Indicates whether to generate a table of only
basic QC stats (ex. library size), or to generate a summary table of all
QC stats stored in the inSCE.}

\item{output}{Character vector. Options are "kable",
which a table will be generated via kable to an html file,
"csv", which a csv file will be generated, or "dataframe",
which outputs a dataframe.}

\item{fileName}{String. File name if output set to "csv".}
}
\description{
Plot QC metrics generated from QC algorithms via either kable or csv file.
}
\examples{
data(scExample, package = "singleCellTK")
sce <- sce[, colData(sce)$type != 'EmptyDroplet']
sampleSummaryStats(sce, simple = TRUE, output = "dataframe")
}
