% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSummaryStats.R
\name{sampleSummaryStats}
\alias{sampleSummaryStats}
\title{Generate table of SCTK QC outputs.}
\usage{
sampleSummaryStats(inSCE, sample = NULL, useAssay = "counts", simple = TRUE)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object with saved
\link[SummarizedExperiment]{assay} data and/or
\link[SummarizedExperiment]{colData} data. Required.}

\item{sample}{Character vector. Indicates which sample each cell belongs to.}

\item{useAssay}{A string specifying which assay in the SCE to use. Default
'counts'.}

\item{simple}{Boolean. Indicates whether to generate a table of only
basic QC stats (ex. library size), or to generate a summary table of all
QC stats stored in the inSCE.}
}
\value{
A matrix/array object.
}
\description{
Creates a table of QC metrics generated from
 QC algorithms via either kable or csv file.
}
\examples{
data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sampleSummaryStats(sce, simple = TRUE)
}
