% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scater_CPM.R
\name{scaterCPM}
\alias{scaterCPM}
\title{scaterCPM
Uses CPM from scater library to compute counts-per-million.}
\usage{
scaterCPM(inSCE, assayName = "ScaterCPMCounts", useAssay = "counts")
}
\arguments{
\item{inSCE}{Input SingleCellExperiment object}

\item{assayName}{New assay name for cpm data.}

\item{useAssay}{Input assay}
}
\value{
inSCE Updated SingleCellExperiment object
}
\description{
scaterCPM
Uses CPM from scater library to compute counts-per-million.
}
\examples{
data(sce_chcl, package = "scds")
sce_chcl <- scaterCPM(sce_chcl,"countsCPM", "counts")
}
\author{
Irzam Sarfraz
}
