% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scater_logNormCounts.R
\name{scaterlogNormCounts}
\alias{scaterlogNormCounts}
\title{scaterlogNormCounts
Uses \link{logNormCounts} to log normalize input data}
\usage{
scaterlogNormCounts(
  inSCE,
  assayName = "ScaterLogNormCounts",
  useAssay = "counts"
)
}
\arguments{
\item{inSCE}{Input SingleCellExperiment object}

\item{assayName}{New assay name for log normalized data}

\item{useAssay}{Input assay}
}
\value{
inSCE Updated SingleCellExperiment object that contains the new log normalized data
}
\description{
scaterlogNormCounts
Uses \link{logNormCounts} to log normalize input data
}
\examples{
data(sce_chcl, package = "scds")
sce_chcl <- scaterlogNormCounts(sce_chcl,"logcounts", "counts")
}
\seealso{
Other run functions: 
\code{\link{computeZScore}()},
\code{\link{detectCellOutlier}()},
\code{\link{diffAbundanceFET}()},
\code{\link{runBBKNN}()},
\code{\link{runBarcodeRankDrops}()},
\code{\link{runBcds}()},
\code{\link{runCellQC}()},
\code{\link{runClusterSummaryMetrics}()},
\code{\link{runComBatSeq}()},
\code{\link{runCxds}()},
\code{\link{runCxdsBcdsHybrid}()},
\code{\link{runDEAnalysis}()},
\code{\link{runDecontX}()},
\code{\link{runDimReduce}()},
\code{\link{runDoubletFinder}()},
\code{\link{runDropletQC}()},
\code{\link{runEmptyDrops}()},
\code{\link{runEnrichR}()},
\code{\link{runFastMNN}()},
\code{\link{runFeatureSelection}()},
\code{\link{runFindMarker}()},
\code{\link{runGSVA}()},
\code{\link{runHarmony}()},
\code{\link{runKMeans}()},
\code{\link{runLimmaBC}()},
\code{\link{runMNNCorrect}()},
\code{\link{runModelGeneVar}()},
\code{\link{runNormalization}()},
\code{\link{runPerCellQC}()},
\code{\link{runSCANORAMA}()},
\code{\link{runSCMerge}()},
\code{\link{runScDblFinder}()},
\code{\link{runScanpyFindClusters}()},
\code{\link{runScanpyFindHVG}()},
\code{\link{runScanpyFindMarkers}()},
\code{\link{runScanpyNormalizeData}()},
\code{\link{runScanpyPCA}()},
\code{\link{runScanpyScaleData}()},
\code{\link{runScanpyTSNE}()},
\code{\link{runScanpyUMAP}()},
\code{\link{runScranSNN}()},
\code{\link{runScrublet}()},
\code{\link{runSeuratFindClusters}()},
\code{\link{runSeuratFindHVG}()},
\code{\link{runSeuratFindMarkers}()},
\code{\link{runSeuratHeatmap}()},
\code{\link{runSeuratICA}()},
\code{\link{runSeuratIntegration}()},
\code{\link{runSeuratJackStraw}()},
\code{\link{runSeuratNormalizeData}()},
\code{\link{runSeuratPCA}()},
\code{\link{runSeuratSCTransform}()},
\code{\link{runSeuratScaleData}()},
\code{\link{runSeuratTSNE}()},
\code{\link{runSeuratUMAP}()},
\code{\link{runSingleR}()},
\code{\link{runSoupX}()},
\code{\link{runTSCAN}()},
\code{\link{runTSCANClusterDEAnalysis}()},
\code{\link{runTSCANDEG}()},
\code{\link{runTSNE}()},
\code{\link{runUMAP}()},
\code{\link{runVAM}()},
\code{\link{runZINBWaVE}()},
\code{\link{scaterCPM}()},
\code{\link{scaterPCA}()}
}
\author{
Irzam Sarfraz
}
\concept{run functions}
