% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scater_logNormCounts.R
\name{scaterlogNormCounts}
\alias{scaterlogNormCounts}
\title{scaterlogNormCounts
Uses \link{logNormCounts} to log normalize input data}
\usage{
scaterlogNormCounts(
  inSCE,
  assayName = "ScaterLogNormCounts",
  useAssay = "counts"
)
}
\arguments{
\item{inSCE}{Input SingleCellExperiment object}

\item{assayName}{New assay name for log normalized data}

\item{useAssay}{Input assay}
}
\value{
inSCE Updated SingleCellExperiment object that contains the new log normalized data
}
\description{
scaterlogNormCounts
Uses \link{logNormCounts} to log normalize input data
}
\examples{
data(sce_chcl, package = "scds")
sce_chcl <- scaterlogNormCounts(sce_chcl,"logcounts", "counts")
}
\author{
Irzam Sarfraz
}
