% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scater_PCA.R
\name{scaterPCA}
\alias{scaterPCA}
\title{Perform scater PCA on a SingleCellExperiment Object}
\usage{
scaterPCA(
  inSCE,
  useAssay = "logcounts",
  useHVG = NULL,
  scale = TRUE,
  reducedDimName = "PCA",
  nComponents = 50,
  useAltExp = NULL,
  seed = NULL,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{useAssay}{Assay to use for PCA computation. If \code{useAltExp} is
specified, \code{useAssay} has to exist in
\code{assays(altExp(inSCE, useAltExp))}. Default \code{"logcounts"}}

\item{useHVG}{A character string indicating a \code{rowData} variable that
stores the logical vector of HVG selection. Default \code{NULL}.}

\item{scale}{Logical scalar, whether to standardize the expression values.
Default \code{TRUE}.}

\item{reducedDimName}{Name to use for the reduced output assay. Default
\code{"PCA"}.}

\item{nComponents}{Number of principal components to obtain from the PCA
computation. Default \code{50}.}

\item{useAltExp}{The subset to use for PCA computation, usually for the
selected.variable features. Default \code{NULL}.}

\item{seed}{Integer, random seed for reproducibility of PCA results. 
Default \code{NULL}.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object specifying whether 
the PCA should be parallelized.}
}
\value{
A \linkS4class{SingleCellExperiment} object with PCA computation
updated in \code{reducedDim(inSCE, reducedDimName)}.
}
\description{
A wrapper to \link[scater]{runPCA} function to compute principal
component analysis (PCA) from a given \linkS4class{SingleCellExperiment} 
object.
}
\examples{
data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sce <- scaterlogNormCounts(sce, "logcounts")
sce <- scaterPCA(sce, "logcounts", scale = TRUE)
}
