% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scater_PCA.R
\name{scaterPCA}
\alias{scaterPCA}
\title{Perform PCA on a SingleCellExperiment Object
A wrapper to \link[scater]{runPCA} function to compute principal component
analysis (PCA) from a given \linkS4class{SingleCellExperiment} object.}
\usage{
scaterPCA(
  inSCE,
  useAssay = "logcounts",
  useAltExp = NULL,
  reducedDimName = "PCA",
  nComponents = 50,
  scale = FALSE,
  ntop = NULL,
  seed = NULL
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{useAssay}{Assay to use for PCA computation. If \code{useAltExp} is
specified, \code{useAssay} has to exist in
\code{assays(altExp(inSCE, useAltExp))}. Default \code{"logcounts"}}

\item{useAltExp}{The subset to use for PCA computation, usually for the
selected.variable features. Default \code{NULL}.}

\item{reducedDimName}{Name to use for the reduced output assay. Default
\code{"PCA"}.}

\item{nComponents}{Number of principal components to obtain from the PCA
computation. Default \code{50}.}

\item{scale}{Logical scalar, whether to standardize the expression values.
Default \code{FALSE}.}

\item{ntop}{Number of top features to use as a further variable feature
selection. Default \code{NULL}.}

\item{seed}{Random seed for reproducibility of PCA results.}
}
\value{
A \linkS4class{SingleCellExperiment} object with PCA computation
updated in \code{reducedDim(inSCE, reducedDimName)}.
}
\description{
Perform PCA on a SingleCellExperiment Object
A wrapper to \link[scater]{runPCA} function to compute principal component
analysis (PCA) from a given \linkS4class{SingleCellExperiment} object.
}
\examples{
data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sce <- scaterlogNormCounts(sce, "logcounts")
sce <- scaterPCA(sce, "logcounts")
}
