% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{scDiffEx}
\alias{scDiffEx}
\title{Create a heatmap for differential expression analysis}
\usage{
scDiffEx(inSCESet, use_assay = "logcounts", condition, covariates = NULL,
  significance = 0.05, ntop = 500, usesig = TRUE, diffexmethod,
  clusterRow = TRUE, clusterCol = TRUE, levelofinterest = NULL,
  analysis_type = NULL, controlLevel = NULL)
}
\arguments{
\item{inSCESet}{Input SCtkExperiment object. Required}

\item{use_assay}{Indicate which assay to use. Default is "logcounts"}

\item{condition}{The name of the condition to use for differential
expression. Required}

\item{covariates}{Additional covariates to add to the model. Currently only
supported for ANOVA. Default is NULL}

\item{significance}{FDR corrected significance cutoff for differentially
expressed genes. Required}

\item{ntop}{Number of top differentially expressed genes to display in the
heatmap. Required}

\item{usesig}{If TRUE, only display genes that meet the significance cutoff,
up to ntop genes. Required}

\item{diffexmethod}{The method for performing differential expression
analysis. Available options are DESeq2, limma, and ANOVA. Required}

\item{clusterRow}{Cluster the rows. The default is TRUE}

\item{clusterCol}{Cluster the columns. The default is TRUE}

\item{levelofinterest}{If the condition has more than two labels,
levelofinterest should contain one factor for condition. The differential
expression results will compare the factor in levelofinterest to all other
data.}

\item{analysis_type}{Choose "biomarker" to compare the levelofinterest to all
other samples. Choose "contrast" to compare the levelofinterest to a
controlLevel. For DESeq2, choose "anova" to perform DESeq2 in LRT mode
comparing the model with condition to a model without condition.}

\item{controlLevel}{If the condition has more than two labels, controlLevel
should contain one factor from condition to use as the control.}
}
\value{
A list of differentially expressed genes.
}
\description{
Create a heatmap for differential expression analysis
}
\examples{
data("GSE60361_subset_sce")
res <- scDiffEx(GSE60361_subset_sce,
                use_assay = "logcounts",
                "level1class",
                diffexmethod = "limma")

}
