% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{scDiffEx}
\alias{scDiffEx}
\title{Create a heatmap for differential expression analysis}
\usage{
scDiffEx(inSCESet, condition, significance = 0.05, ntop = 500,
  usesig = TRUE, diffexmethod, clusterRow = TRUE, clusterCol = TRUE,
  displayRowLabels = TRUE, levelofinterest)
}
\arguments{
\item{inSCESet}{Input SCtkExperiment object. Required}

\item{condition}{The name of the condition to use for differential
expression. Required}

\item{significance}{FDR corrected significance cutoff for differentially
expressed genes. Required}

\item{ntop}{Number of top differentially expressed genes to display in the
heatmap. Required}

\item{usesig}{If TRUE, only display genes that meet the significance cutoff,
up to ntop genes. Required}

\item{diffexmethod}{The method for performing differential expression
analysis. Available options are DESeq, DESeq2, and limma. Required}

\item{clusterRow}{Cluster the rows. The default is TRUE}

\item{clusterCol}{Cluster the columns. The default is TRUE}

\item{displayRowLabels}{Display the row labels on the heatmap. The default
is TRUE.}

\item{levelofinterest}{If the condition has more than two labels,
levelofinterest should contain one factor for condition. The differential
expression results will compare the factor in levelofinterest to all other
data.}
}
\value{
A list of differentially expressed genes.
}
\description{
Create a heatmap for differential expression analysis
}
\examples{
\dontrun{
scDiffEx(newsceset_david, "age", "0.1")
}

}
