% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{scDiffEx_deseq2}
\alias{scDiffEx_deseq2}
\title{Perform differential expression analysis with DESeq2}
\usage{
scDiffEx_deseq2(inSCESet, condition)
}
\arguments{
\item{inSCESet}{Input SingleCellExperiment object. Required}

\item{condition}{The name of the condition to use for differential
expression. Must be a name of a column from pData that contains two labels.
Required}
}
\value{
A data frame of gene names and adjusted p-values
}
\description{
Returns a data frame of gene names and adjusted p-values
}
