% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{scDiffEx_limma}
\alias{scDiffEx_limma}
\title{Perform differential expression analysis with limma}
\usage{
scDiffEx_limma(inSCESet, condition)
}
\arguments{
\item{inSCESet}{Input SingleCellExperiment object. Required}

\item{condition}{The name of the condition to use for differential
expression. Must be a name of a column from colData that contains two labels.
Required}
}
\value{
A data frame of gene names and adjusted p-values
}
\description{
Returns a data frame of gene names and adjusted p-values
}
