% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDiffEx.R
\name{scDiffEx_limma}
\alias{scDiffEx_limma}
\title{Perform differential expression analysis with limma}
\usage{
scDiffEx_limma(inSCESet, use_assay = "logcounts", condition,
  analysis_type = "biomarker", levelofinterest = NULL, covariates = NULL,
  adjust = "fdr")
}
\arguments{
\item{inSCESet}{Input SCtkExperiment object. Required}

\item{use_assay}{Indicate which assay to use. Default is "logcounts"}

\item{condition}{The name of the condition to use for differential
expression. Must be a name of a column from colData that contains at least
two labels. Required}

\item{analysis_type}{If there are more than two levels in your condition
variable, select the analysis_type. Choose "biomarker" to compare the
levelofinterest to all other samples. Choose "coef" to select a coefficient
of interset with levelofinterest (see below). Choose "allcoef" to test if
any coefficient is different from zero.}

\item{levelofinterest}{If the condition has more than two labels,
levelofinterest should contain one factor of interest from condition.}

\item{covariates}{Additional covariates to add to the model.}

\item{adjust}{Method for p-value correction. See options in p.adjust().
The default is fdr.}
}
\value{
A data frame of gene names and adjusted p-values
}
\description{
Returns a data frame of gene names and adjusted p-values
}
\examples{
data("mouse_brain_subset_sce")
res <- scDiffEx_limma(mouse_brain_subset_sce, condition = "level1class")

}
