% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SCtkExperiment-class}
\alias{SCtkExperiment-class}
\title{A lightweight S4 extension to the SingleCellExperiment class to store
additional information.}
\arguments{
\item{value}{The DataFrame of pca_variances}
}
\value{
A SingleCellExperiment like object with an addition pca_variances
slot.
}
\description{
A lightweight S4 extension to the SingleCellExperiment class to store
additional information.
}
\section{Slots}{

\describe{
\item{\code{pca_variances}}{The percent variation contained in each PCA dimension}
}}

