% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SCtkExperiment-class}
\alias{SCtkExperiment-class}
\title{A lightweight S4 extension to the SingleCellExperiment class to store
additional information.}
\arguments{
\item{value}{The DataFrame of pca_variances}
}
\value{
A SingleCellExperiment like object with an addition pca_variances
slot.
}
\description{
A lightweight S4 extension to the SingleCellExperiment class to store
additional information.
}
\section{Slots}{

\describe{
\item{\code{pca_variances}}{The percent variation contained in each PCA dimension}
}}

\examples{
data("mouse_brain_subset_sce")
counts_mat <- assay(mouse_brain_subset_sce, "counts")
sample_annot <- colData(mouse_brain_subset_sce)
row_annot <- rowData(mouse_brain_subset_sce)
newSCE <- SCtkExperiment(assays=list(counts=counts_mat),
                         colData=sample_annot,
                         rowData=row_annot)
newSCE <- getPCA(newSCE, use_assay = "counts")
#View the percent variation of the PCA
pca_variances(newSCE)

}
