% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{SCtkExperiment}
\alias{SCtkExperiment}
\title{Create a SCtkExperiment}
\usage{
SCtkExperiment(..., pcaVariances = S4Vectors::DataFrame())
}
\arguments{
\item{...}{SingleCellExperiment and SummarizedExperiment components}

\item{pcaVariances}{The percent variation contained in each PCA dimension}
}
\value{
A SingleCellExperiment like object with an addition pcaVariances
slot.
}
\description{
Create a SCtkExperiment
}
\examples{
data("mouseBrainSubsetSCE")
counts_mat <- assay(mouseBrainSubsetSCE, "counts")
sample_annot <- colData(mouseBrainSubsetSCE)
row_annot <- rowData(mouseBrainSubsetSCE)
newSCE <- SCtkExperiment(assays=list(counts=counts_mat),
                         colData=sample_annot,
                         rowData=row_annot)
newSCE <- getPCA(newSCE, useAssay = "counts")
#View the percent variation of the PCA
pcaVariances(newSCE)

}
