% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{SCtkExperiment}
\alias{SCtkExperiment}
\title{Create a SCtkExperiment}
\usage{
SCtkExperiment(..., pca_variances = DataFrame())
}
\arguments{
\item{...}{SingleCellExperiment and SummarizedExperiment components}

\item{pca_variances}{The percent variation contained in each PCA dimension}
}
\description{
Create a SCtkExperiment
}
