% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate_setup.R
\name{selectSCTKConda}
\alias{selectSCTKConda}
\title{Selects a Conda environment}
\usage{
selectSCTKConda(envname = "sctk-reticulate")
}
\arguments{
\item{envname}{Character. Name of the conda environment to activate.}
}
\value{
None. Selects Conda environment.
}
\description{
Selects a Conda environment with Python packages used in \code{\link{singleCellTK}}.
}
\examples{
\dontrun{
sctkPythonInstallConda(envname = "sctk-reticulate", selectConda = FALSE)
selectSCTKConda(envname = "sctk-reticulate")
}
}
\seealso{
\code{\link[reticulate]{conda-tools}} for more information on using Conda environments with package \code{\link{reticulate}}.
See \url{https://rstudio.github.io/reticulate/} for more information on package \code{\link{reticulate}}.

See \code{\link[singleCellTK]{sctkPythonInstallConda}} for installation of Python modules into a Conda environment.
See\code{\link[reticulate]{conda-tools}} for more information on using Conda environments with package \code{\link{reticulate}}.
See \url{https://rstudio.github.io/reticulate/} for more information on package \code{\link{reticulate}}.
See \url{https://docs.conda.io/en/latest/} for more information on Conda environments.

Other Python environment: 
\code{\link{sctkPythonInstallConda}()},
\code{\link{sctkPythonInstallVirtualEnv}()},
\code{\link{selectSCTKVirtualEnvironment}()}
}
\concept{Python environment}
