% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate_setup.R
\name{selectSCTKVirtualEnvironment}
\alias{selectSCTKVirtualEnvironment}
\title{Selects a virtual environment}
\usage{
selectSCTKVirtualEnvironment(envname = "sctk-reticulate")
}
\arguments{
\item{envname}{Character. Name of the virtual environment to activate.}
}
\value{
None. Selects virtual environment.
}
\description{
Selects a virtual environment with Python packages used in \code{\link{singleCellTK}}
}
\examples{
\dontrun{
sctkPythonInstallVirtualEnv(envname = "sctk-reticulate", selectEnvironment = FALSE)
selectSCTKVirtualEnvironment(envname = "sctk-reticulate")
}
}
\seealso{
See \code{\link[singleCellTK]{sctkPythonInstallVirtualEnv}} for installation of Python modules into a virtual environment.
See\code{\link[reticulate]{virtualenv-tools}} for more information on using virtual environments with package \code{\link{reticulate}}.
See \url{https://rstudio.github.io/reticulate/} for more information on package \code{\link{reticulate}}.
}
