% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratComputeHeatmap}
\alias{seuratComputeHeatmap}
\title{seuratComputeHeatmap
Computes the heatmap plot object from the pca slot in the input sce object}
\usage{
seuratComputeHeatmap(
  inSCE,
  useAssay,
  useReduction = c("pca", "ica"),
  dims = NULL,
  nfeatures = 30,
  fast = TRUE,
  combine = TRUE,
  raster = TRUE
)
}
\arguments{
\item{inSCE}{(sce) object from which to compute heatmap (pca should be computed)}

\item{useAssay}{Assay containing scaled counts to use in heatmap.}

\item{useReduction}{Reduction method to use for computing clusters. One of "pca" or "ica". Default \code{"pca"}.}

\item{dims}{Number of components to generate heatmap plot objects. If \code{NULL}, a heatmap will be generated for all components. Default \code{NULL}.}

\item{nfeatures}{Numer of features to include in the heatmap. Default \code{30}.}

\item{fast}{See \link[Seurat]{DimHeatmap} for more information. Default \code{TRUE}.}

\item{combine}{See \link[Seurat]{DimHeatmap} for more information. Default \code{TRUE}.}

\item{raster}{See \link[Seurat]{DimHeatmap} for more information. Default \code{TRUE}.}
}
\value{
plot object
}
\description{
seuratComputeHeatmap
Computes the heatmap plot object from the pca slot in the input sce object
}
\examples{
data(scExample, package = "singleCellTK")
\dontrun{
sce <- seuratNormalizeData(sce, useAssay = "counts")
sce <- seuratFindHVG(sce, useAssay = "counts")
sce <- seuratScaleData(sce, useAssay = "counts")
sce <- seuratPCA(sce, useAssay = "counts")
heatmap <- seuratComputeHeatmap(sce, useAssay = "counts")
seuratHeatmapPlot(heatmap)
}
}
