% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratComputeJackStraw}
\alias{seuratComputeJackStraw}
\title{seuratComputeJackStraw
Compute jackstraw plot and store the computations in the input sce object}
\usage{
seuratComputeJackStraw(inSCE, useAssay, dims = NULL)
}
\arguments{
\item{inSCE}{(sce) object on which to compute and store jackstraw plot}

\item{useAssay}{Assay containing scaled counts to use in JackStraw calculation.}

\item{dims}{Number of components to test in Jackstraw. If \code{NULL}, then all components are used. Default \code{NULL}.}
}
\value{
Updated \code{SingleCellExperiment} object with jackstraw computations stored in it
}
\description{
seuratComputeJackStraw
Compute jackstraw plot and store the computations in the input sce object
}
\examples{
data(scExample, package = "singleCellTK")
\dontrun{
sce <- seuratNormalizeData(sce, useAssay = "counts")
sce <- seuratFindHVG(sce, useAssay = "counts")
sce <- seuratScaleData(sce, useAssay = "counts")
sce <- seuratPCA(sce, useAssay = "counts")
sce <- seuratComputeJackStraw(sce, useAssay = "counts")
}
}
