% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratElbowPlot}
\alias{seuratElbowPlot}
\title{seuratElbowPlot
Computes the plot object for elbow plot from the pca slot in the input sce object}
\usage{
seuratElbowPlot(inSCE, significantPC = NULL, reduction = "pca")
}
\arguments{
\item{inSCE}{(sce) object from which to compute the elbow plot (pca should be computed)}

\item{significantPC}{Number of significant principal components to plot. This is used to alter the color of the points for the corresponding PCs. If \code{NULL}, all points will be the same color. Default \code{NULL}.}

\item{reduction}{Reduction to use for elbow plot generation. Either \code{"pca"} or \code{"ica"}. Default \code{"pca"}.}
}
\value{
plot object
}
\description{
seuratElbowPlot
Computes the plot object for elbow plot from the pca slot in the input sce object
}
\examples{
data(scExample, package = "singleCellTK")
\dontrun{
sce <- seuratNormalizeData(sce, useAssay = "counts")
sce <- seuratFindHVG(sce, useAssay = "counts")
sce <- seuratScaleData(sce, useAssay = "counts")
sce <- seuratPCA(sce, useAssay = "counts")
seuratElbowPlot(sce)
}
}
