% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratFindClusters}
\alias{seuratFindClusters}
\title{seuratFindClusters
Computes the clusters from the input sce object and stores them back in sce object}
\usage{
seuratFindClusters(inSCE, useAssay, geneNamesSeurat, reduction, dims,
  algorithm, groupSingletons)
}
\arguments{
\item{inSCE}{(sce) object from which clusters should be computed and stored in}

\item{useAssay}{which assay to use from sce object}

\item{geneNamesSeurat}{a list of rowames/genenames/featurenames for seurat object for consistent formatting#'}

\item{reduction}{selected reduction method to use for computing clusters ("pca" or "ica", default is "pca")}

\item{dims}{numeric value of how many components to use for computing clusters (default is 10)}

\item{algorithm}{selected algorithm to compute clusters (default is "original Louvain algorithm")}

\item{groupSingletons}{boolean if singletons should be grouped together or not (TRUE or FALSE, default is TRUE)}
}
\value{
sceObject; updated sce object which now contains the computed clusters
}
\description{
seuratFindClusters
Computes the clusters from the input sce object and stores them back in sce object
}
