% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratFindMarkers}
\alias{seuratFindMarkers}
\title{seuratFindMarkers}
\usage{
seuratFindMarkers(inSCE, cells1, cells2, group1, group2)
}
\arguments{
\item{inSCE}{Input \code{SingleCellExperiment} object.}

\item{cells1}{A \code{list} of sample names included in group1.}

\item{cells2}{A \code{list} of sample names included in group2.}

\item{group1}{Name of group1.}

\item{group2}{Name of group2.}
}
\value{
Nothing yet.
}
\description{
seuratFindMarkers
}
