% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratPCA}
\alias{seuratPCA}
\title{seuratPCA
Computes PCA on the input sce object and stores the calculated principal components within the sce object}
\usage{
seuratPCA(
  inSCE,
  useAssay = "seuratScaledData",
  reducedDimName = "seuratPCA",
  nPCs = 20,
  verbose = TRUE
)
}
\arguments{
\item{inSCE}{(sce) object on which to compute PCA}

\item{useAssay}{Assay containing scaled counts to use in PCA.}

\item{reducedDimName}{Name of new reducedDims object containing Seurat PCA. Default \code{seuratPCA}.}

\item{nPCs}{numeric value of how many components to compute. Default \code{20}.}

\item{verbose}{Logical value indicating if informative messages should 
be displayed. Default is \code{TRUE}.}
}
\value{
Updated \code{SingleCellExperiment} object which now contains the computed principal components
}
\description{
seuratPCA
Computes PCA on the input sce object and stores the calculated principal components within the sce object
}
\examples{
data(scExample, package = "singleCellTK")
\dontrun{
sce <- seuratNormalizeData(sce, useAssay = "counts")
sce <- seuratFindHVG(sce, useAssay = "counts")
sce <- seuratScaleData(sce, useAssay = "counts")
sce <- seuratPCA(sce, useAssay = "counts")
}
}
