% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratPCA}
\alias{seuratPCA}
\title{seuratPCA
Computes PCA on the input sce object and stores the calculated principal components within the sce object}
\usage{
seuratPCA(sceObject, geneNamesSeurat, npcs)
}
\arguments{
\item{sceObject;}{sce object on which to compute PCA}

\item{geneNamesSeurat;}{a list of rowames/genenames/featurenames for seurat object for consistent formatting}

\item{npcs;}{numeric value of how many components to compute (default is 20)}
}
\value{
sceObject; updated sce object which now contains the computed principal components
}
\description{
seuratPCA
Computes PCA on the input sce object and stores the calculated principal components within the sce object
}
