% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratReductionPlot}
\alias{seuratReductionPlot}
\title{seuratReductionPlot
Plots the selected dimensionality reduction method}
\usage{
seuratReductionPlot(inSCE, useAssay, geneNamesSeurat, reduction)
}
\arguments{
\item{inSCE}{(sce) object which has the selected dimensionality reduction algorithm already computed and stored}

\item{useAssay}{which assay to use from sce object}

\item{geneNamesSeurat}{a list of rowames/genenames/featurenames for seurat object for consistent formatting}

\item{reduction}{one of selected algorithm from pca, ica, tsne and umap}
}
\value{
plot object
}
\description{
seuratReductionPlot
Plots the selected dimensionality reduction method
}
