% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratReductionPlot}
\alias{seuratReductionPlot}
\title{seuratReductionPlot
Plots the selected dimensionality reduction method}
\usage{
seuratReductionPlot(
  inSCE,
  useReduction = c("pca", "ica", "tsne", "umap"),
  showLegend = FALSE
)
}
\arguments{
\item{inSCE}{(sce) object which has the selected dimensionality reduction algorithm already computed and stored}

\item{useReduction}{Dimentionality reduction to plot. One of "pca", "ica", "tsne", or "umap". Default \code{"umap"}.}

\item{showLegend}{Select if legends should be shown on the output plot or not. Either "TRUE" or "FALSE". Default \code{FALSE}.}
}
\value{
plot object
}
\description{
seuratReductionPlot
Plots the selected dimensionality reduction method
}
\examples{
data(scExample, package = "singleCellTK")
\dontrun{
sce <- seuratNormalizeData(sce, useAssay = "counts")
sce <- seuratFindHVG(sce, useAssay = "counts")
sce <- seuratScaleData(sce, useAssay = "counts")
sce <- seuratPCA(sce, useAssay = "counts")
seuratReductionPlot(sce, useReductionPlot = "pca")}
}
