% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratRunTSNE}
\alias{seuratRunTSNE}
\title{seuratRunTSNE
Computes tSNE from the given sce object and stores the tSNE computations back into the sce object}
\usage{
seuratRunTSNE(
  inSCE,
  useReduction = c("pca", "ica"),
  reducedDimName = "seuratTSNE",
  dims = 10,
  perplexity = 30
)
}
\arguments{
\item{inSCE}{(sce) object on which to compute the tSNE}

\item{useReduction}{selected reduction algorithm to use for computing tSNE. One of "pca" or "ica". Default \code{"pca"}.}

\item{reducedDimName}{Name of new reducedDims object containing Seurat tSNE Default \code{seuratTSNE}.}

\item{dims}{Number of reduction components to use for tSNE computation. Default \code{10}.}

\item{perplexity}{Adjust the preplexity tuneable parameter for the underlying tSNE call. Default \code{30}.}
}
\value{
Updated sce object with tSNE computations stored
}
\description{
seuratRunTSNE
Computes tSNE from the given sce object and stores the tSNE computations back into the sce object
}
