% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratRunUMAP}
\alias{seuratRunUMAP}
\title{seuratRunUMAP
Computes UMAP from the given sce object and stores the UMAP computations back into the sce object}
\usage{
seuratRunUMAP(sceObject, geneNamesSeurat, reduction, dims)
}
\arguments{
\item{sceObject}{sce object on which to compute the UMAP}

\item{geneNamesSeurat}{a list of rowames/genenames/featurenames for seurat object for consistent formatting}

\item{reduction}{selected reduction algorithm to use for computing UMAP ("pca" or "ica", default is "pca")}

\item{dims}{numerical value of how many reduction components to use for UMAP computation (default is 10)}
}
\value{
Updated sce object with UMAP computations stored
}
\description{
seuratRunUMAP
Computes UMAP from the given sce object and stores the UMAP computations back into the sce object
}
