% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratScaleData}
\alias{seuratScaleData}
\title{seuratScaleData
Scales the input sce object according to the input parameters}
\usage{
seuratScaleData(
  inSCE,
  useAssay,
  scaledAssayName = "seuratScaledData",
  model = "linear",
  scale = TRUE,
  center = TRUE,
  scaleMax = 10
)
}
\arguments{
\item{inSCE}{(sce) object to scale}

\item{useAssay}{Assay containing normalized counts to scale.}

\item{scaledAssayName}{Name of new assay containing scaled data. Default \code{seuratScaledData}.}

\item{model}{selected model to use for scaling data. Default \code{"linear"}.}

\item{scale}{boolean if data should be scaled or not. Default \code{TRUE}.}

\item{center}{boolean if data should be centered or not. Default \code{TRUE}}

\item{scaleMax}{maximum numeric value to return for scaled data. Default \code{10}.}
}
\value{
sceObject scaled sce object
}
\description{
seuratScaleData
Scales the input sce object according to the input parameters
}
