% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratSCTransform}
\alias{seuratSCTransform}
\title{seuratSCTransform
Runs the SCTransform function from seurat library to transform the input data}
\usage{
seuratSCTransform(inSCE, newAssayName = "SCTCounts", useAssay, geneNamesSeurat)
}
\arguments{
\item{inSCE}{the input sce object containing the counts}

\item{newAssayName}{name for the transformed counts assay (default is "SCTCounts")}

\item{useAssay}{which counts assay to use from input sce object for transformation}

\item{geneNamesSeurat}{genenames for consistent formatting}
}
\description{
seuratSCTransform
Runs the SCTransform function from seurat library to transform the input data
}
