% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{seuratSCTransform}
\alias{seuratSCTransform}
\title{seuratSCTransform
Runs the \link[Seurat]{SCTransform} function to transform/normalize the input data}
\usage{
seuratSCTransform(inSCE, normAssayName = "SCTCounts", useAssay = "counts")
}
\arguments{
\item{inSCE}{Input SingleCellExperiment object}

\item{normAssayName}{Name for the output data assay. Default \code{"SCTCounts"}.}

\item{useAssay}{Name for the input data assay. Default \code{"counts"}.}
}
\value{
Updated SingleCellExperiment object containing the transformed data
}
\description{
seuratSCTransform
Runs the \link[Seurat]{SCTransform} function to transform/normalize the input data
}
\examples{
data(sce_chcl, package = "scds")
sce_chcl <- seuratSCTransform(sce_chcl, "SCTCounts", "counts")
}
