% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellTK.R
\name{singleCellTK}
\alias{singleCellTK}
\title{Run the single cell analysis app}
\usage{
singleCellTK(inputData = NULL)
}
\arguments{
\item{inputData}{The input SCtkExperiment class object}
}
\value{
The shiny app will open
}
\description{
Use this function to run the single cell analysis app.
}
\examples{
#Upload data through the app
if(interactive()){
  singleCellTK()
}

#Load the app with a SCtkExperiment object
if(interactive()){
  data("mouseBrainSubsetSCE")
  singleCellTK(mouseBrainSubsetSCE)
}

}
