% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellTK.R
\name{singleCellTK}
\alias{singleCellTK}
\title{Run the single cell analysis app}
\usage{
singleCellTK(inSCE = NULL, includeVersion = TRUE, theme = "yeti")
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{includeVersion}{Include the version number in the SCTK header. The
default is TRUE.}

\item{theme}{The bootswatch theme to use for the singleCellTK UI. The default
is 'flatly'.}
}
\value{
The shiny app will open
}
\description{
Use this function to run the single cell analysis app.
}
\examples{
#Upload data through the app
if(interactive()){
  singleCellTK()
}

# Load the app with a SingleCellExperiment object
if(interactive()){
  data("mouseBrainSubsetSCE")
  singleCellTK(mouseBrainSubsetSCE)
}

}
