% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downsample.Matrix.R
\name{subDiffEx_anova}
\alias{subDiffEx_anova}
\title{Runs ANOVA on all genes in a simulated dataset with more than 2 conditions, and adjusts for FDR.}
\usage{
subDiffEx_anova(countMatrix, condition)
}
\arguments{
\item{countMatrix}{Matrix. A simulated counts matrix, sans labels.}

\item{condition}{Factor. The condition labels for the simulated cells.}
}
\value{
A vector of fdr-adjusted p-values for all genes. Nonviable results (such as for genes with 0 counts in a simulated dataset) are coerced to 1.
}
\description{
Runs ANOVA on all genes in a simulated dataset with more than 2 conditions, and adjusts for FDR.
}
