% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{summarizeSCE}
\alias{summarizeSCE}
\title{Summarize SCtkExperiment}
\usage{
summarizeSCE(inSCE, useAssay = "counts", sampleVariableName = "sample")
}
\arguments{
\item{inSCE}{Input SCtkExperiment object.}

\item{useAssay}{Indicate which assay to summarize. Default \code{"counts"}.}

\item{sampleVariableName}{Variable name in \code{colData} denoting which
sample each cell belongs to. If \code{NULL}, all cells will be assumed
to come from the same sample. Default \code{"sample"}.}
}
\value{
A data.frame of summary metrics per sample.
}
\description{
Creates a table of summary metrics from an input SCtkExperiment.
}
\examples{
data("mouseBrainSubsetSCE")
summarizeSCE(mouseBrainSubsetSCE, sample = NULL)

}
