% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{summarizeTable}
\alias{summarizeTable}
\title{Summarize SCtkExperiment}
\usage{
summarizeTable(indata, useAssay = "counts", expressionCutoff = 1700)
}
\arguments{
\item{indata}{Input SCtkExperiment}

\item{useAssay}{Indicate which assay to summarize. Default is "counts"}

\item{expressionCutoff}{Count number of samples with fewer than
expressionCutoff genes. The default is 1700.}
}
\value{
A data.frame object of summary metrics.
}
\description{
Creates a table of summary metrics from an input SCtkExperiment.
}
\examples{
data("mouseBrainSubsetSCE")
summarizeTable(mouseBrainSubsetSCE)

}
