% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{summarizeTable}
\alias{summarizeTable}
\title{Summarize SCtkExperiment}
\usage{
summarizeTable(indata, use_assay = "counts", expression_cutoff = 1700)
}
\arguments{
\item{indata}{Input SCtkExperiment}

\item{use_assay}{Indicate which assay to summarize. Default is "counts"}

\item{expression_cutoff}{Count number of samples with fewer than
expression_cutoff genes. The default is 1700.}
}
\value{
A data.frame object of summary metrics.
}
\description{
Creates a table of summary metrics from an input SCtkExperiment.
}
