% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST_plot_functions.R
\name{thresholdGenes}
\alias{thresholdGenes}
\title{thresholdGenes}
\usage{
thresholdGenes(SCEdata, use_assay = "logcounts")
}
\arguments{
\item{SCEdata}{SingleCelltkExperiment object}

\item{use_assay}{The assay to use for the MAST calculations. The default is
"logcounts"}
}
\description{
thresholdGenes
}
