% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST_plot_functions.R
\name{thresholdGenes}
\alias{thresholdGenes}
\title{thresholdGenes}
\usage{
thresholdGenes(SCEdata, use_assay = "logcounts")
}
\arguments{
\item{SCEdata}{SCtkExperiment object}

\item{use_assay}{The assay to use for the MAST calculations. The default is
"logcounts"}
}
\value{
list of thresholded counts (on natural scale), thresholds, bins,
densities estimated on each bin, and the original data from
MAST::thresholdSCRNACountMatrix
}
\description{
thresholdGenes
}
\examples{
data("GSE60361_subset_sce")
res <- thresholdGenes(GSE60361_subset_sce)

}
