% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEAnalysis.R
\name{thresholdGenes}
\alias{thresholdGenes}
\title{MAST Identify adaptive thresholds}
\usage{
thresholdGenes(inSCE, useAssay = "logcounts")
}
\arguments{
\item{inSCE}{SingleCellExperiment object}

\item{useAssay}{character, default \code{"logcounts"}}
}
\value{
list of thresholded counts (on natural scale), thresholds, bins,
densities estimated on each bin, and the original data from
\code{\link[MAST]{thresholdSCRNACountMatrix}}
}
\description{
MAST Identify adaptive thresholds
}
\examples{
data("mouseBrainSubsetSCE")
res <- thresholdGenes(mouseBrainSubsetSCE)
}
