% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visPlot.R
\name{visPlot}
\alias{visPlot}
\title{visPlot}
\usage{
visPlot(
  inSCE,
  useAssay,
  method,
  condition = NULL,
  glist,
  facetWrap = TRUE,
  scaleHMap = TRUE,
  convertFactor = FALSE
)
}
\arguments{
\item{inSCE}{Input \linkS4class{SingleCellExperiment} object.}

\item{useAssay}{The assay to use in the visualization plot. Required}

\item{method}{Visualization method. Available options are boxplot,
scatterplot, or heatmap. Required}

\item{condition}{colData annotation of the experiment. Required}

\item{glist}{selected genes for visualization. Maximum 25 genes. Required}

\item{facetWrap}{facet wrap according to genes for boxplot, scatterplot and
barplot. Default is FALSE. Optional}

\item{scaleHMap}{scale heatmap expression values. Default is TRUE. Optional}

\item{convertFactor}{If the condition is not a factor, convert it to a factor
before plotting. The default is FALSE}
}
\value{
A visualization plot
}
\description{
Given a plotting method with condition and gene list, return
the respective visualization plot(s).
}
\examples{
visPlot(mouseBrainSubsetSCE, "logcounts", "boxplot", "level1class", "C1qa")
visPlot(mouseBrainSubsetSCE, "counts", "scatterplot", "age", "Cmtm5")
visPlot(mouseBrainSubsetSCE, "counts", "heatmap", "level1class",
        c("Cmtm5", "C1qa"))
}
